cidre::am::device::base

Struct Error

#[repr(transparent)]
pub struct Error(pub i32);

Tuple Fields
0: i32
Implementations

impl Error

pub const SUCCESS: Self

pub const UNDEFINED: Self

pub const BAD_HEADER: Self

pub const NO_RESOURCES: Self

pub const READ: Self

pub const WRITE: Self

pub const UNKNOWN_PACKET: Self

pub const INVALID_ARGUMENT: Self

pub const NOT_FOUND: Self

pub const IS_DIRECTORY: Self

pub const PERMISSION: Self

pub const NOT_CONNECTED: Self

pub const TIME_OUT: Self

pub const OVERRUN: Self

pub const EOF: Self

pub const UNSUPPORTED: Self

pub const FILE_EXISTS: Self

pub const BUSY: Self

pub const CRYPTO: Self

pub const TOO_BIG: Self

pub const INVALID_RESPONSE: Self

Error codes received from Lockdownd over the connection. MobileDevice itself should only use these to relay information about Lockdownd errors to clients.

There should be an analogous AMDError for each kLDErrorXXX declared at the top of MobileDevice.c.

pub const MISSING_KEY: Self

pub const MISSING_VALUE: Self

pub const GET_PROHIBITED: Self

pub const SET_PROHIBITED: Self

pub const REMOVE_PROHIBITED: Self

pub const IMMUTABLE_VALUE: Self

pub const PASSWORD_PROTECTED: Self

pub const USER_DENIED_PAIRING: Self

pub const PAIRING_DIALOG_RESPONSE_PENDING: Self

pub const MISSING_HOST_ID: Self

pub const INVALID_HOST_ID: Self

pub const SESSION_ACTIVE: Self

pub const SESSION_INACTIVE: Self

pub const MISSING_SESSION_ID: Self

pub const INVALID_SESSION_ID: Self

pub const MISSING_SERVICE: Self

pub const INVALID_SERVICE: Self

pub const SERVICE_LIMIT: Self

pub const CHECKIN_SETUP_FAILED: Self

pub const INVALID_CHECKIN: Self

pub const CHECKIN_TIMEOUT: Self

pub const CHECKIN_CONNECTION_FAILED: Self

pub const CHECKIN_RECEIVE_FAILED: Self

pub const CHECKIN_RESPONSE_FAILED: Self

pub const CHECKIN_OUT_OF_MEMORY: Self

pub const CHECKIN_SEND_FAILED: Self

pub const MISSING_PAIR_RECORD: Self

pub const INVALID_PAIR_RECORD: Self

pub const SAVE_PAIR_RECORD_FAILED: Self

pub const INVALID_ACTIVATION_RECORD: Self

pub const MISSING_ACTIVATION_RECORD: Self

pub const SERVICE_PROHIBITED: Self

pub const ESCROW_LOCKED: Self

pub const PAIRING_PROHIBITED: Self

pub const PROHIBITED_BY_SUPERVISION: Self

pub const FMI_PPROTECTED: Self

pub const MC_PROTECTED: Self

pub const MC_CHALLENGE_REQUIRED: Self

pub const WRONG_DROID: Self

pub const SU_VERIFICATION: Self

pub const SU_PATCH: Self

pub const SU_FIRMWARE: Self

pub const PROVISIONING_PROFILE_NOT_VALID: Self

pub const SEND_MESSAGE: Self

pub const RECEIVE_MESSAGE: Self

pub const MISSING_OPTIONS: Self

pub const MISSING_IMAGE_TYPE: Self

pub const DIGEST_FAILED: Self

pub const START_SERVICE: Self

pub const INVALID_DISK_IMAGE: Self

pub const MISSING_DIGEST: Self

pub const MUX: Self

pub const APPLICATION_ALREADY_INSTALLED: Self

Errors defined by either the MobileInstallationProxy or MobileInstallation.framework.

pub const APPLICATION_MOVE_FAILED: Self

pub const APPLICATION_SINFCAPTURE_FAILED: Self

pub const APPLICATION_SANDBOX_FAILED: Self

OBSOLETE

pub const APPLICATION_VERIFICATION_FAILED: Self

pub const ARCHIVE_DESTRUCTION_FAILED: Self

OBSOLETE

pub const BUNDLE_VERIFICATION_FAILED: Self

OBSOLETE

pub const CARRIER_BUNDLE_COPY_FAILED: Self

pub const CARRIER_BUNDLE_DIRECTORY_CREATION_FAILED: Self

pub const CARRIER_BUNDLE_MISSING_SUPPORTED_SIMS: Self

pub const COMM_CENTER_NOTIFICATION_FAILED: Self

pub const CONTAINER_CREATION_FAILED: Self

pub const CONTAINER_P0WN_FAILED: Self

pub const CONTAINER_REMOVAL_FAILED: Self

OBSOLETE

pub const EMBEDDED_PROFILE_INSTALL_FAILED: Self

OBSOLETE

pub const ERROR: Self

pub const EXECUTABLE_TWIDDLE_FAILED: Self

pub const EXISTENCE_CHECK_FAILED: Self

OBSOLETE

pub const INSTALL_MAP_UPDATE_FAILED: Self

pub const MANIFEST_CAPTURE_FAILED: Self

pub const MAP_GENERATION_FAILED: Self

OBSOLETE

pub const MISSING_BUNDLE_EXECUTABLE: Self

pub const MISSING_BUNDLE_IDENTIFIER: Self

pub const MISSING_BUNDLE_PATH: Self

pub const MISSING_CONTAINER: Self

pub const NOTIFICATION_FAILED: Self

OBSOLETE

pub const PACKAGE_EXTRACTION_FAILED: Self

pub const PACKAGE_INSPECTION_FAILED: Self

pub const PACKAGE_MOVE_FAILED: Self

OBSOLETE

pub const PATH_CONVERSION_FAILED: Self

pub const RESTORE_CONTAINER_FAILED: Self

OBSOLETE

pub const SEATBELT_PROFILE_REMOVAL_FAILED: Self

OBSOLETE

pub const STAGE_CREATION_FAILED: Self

pub const SYMLINK_FAILED: Self

pub const ITUNES_ARTWORK_CAPTURE_FAILED: Self

pub const ITUNES_METADATA_CAPTURE_FAILED: Self

pub const ALREADY_ARCHIVED: Self

pub const UNKNOWN_COMMAND: Self

pub const API_INTERNAL: Self

pub const DEVICE_OS_VERSION_TOO_LOW: Self

pub const DEVICE_FAMILY_NOT_SUPPORTED: Self

pub const PACKAGE_PATCH_FAILED: Self

pub const INCORRECT_ARCHITECTURE: Self

pub const PLUGIN_COPY_FAILED: Self

pub const BREADCRUMB_FAILED: Self

pub const BREADCRUMB_UNLOCK: Self

pub const GEO_JSON_CAPTURE_FAILED: Self

pub const NEWSSTAND_ARTWORK_CAPTURE_FAILED: Self

pub const MISSING_COMMAND: Self

pub const NOT_ENTITLED: Self

pub const MISSING_PACKAGE_PATH: Self

pub const MISSING_CONTAINER_PATH: Self

pub const MISSING_APPLICATION_IDENTIFIER: Self

pub const MISSING_ATTRIBUTE_VALUE: Self

pub const LOOKUP_FAILED: Self

pub const DICT_CREATION_FAILED: Self

pub const INSTALL_PROHIBITED: Self

pub const UNINSTALL_PROHIBITED: Self

pub const MISSING_BUNDLE_VERSION: Self

pub const APP_BLACKLISTED: Self

pub const APPEX_BUNDLE_ID_NOT_PREFIXED: Self

pub const APPEX_BUNDLE_ILLEGAL_XPC_SERVICE_DICT: Self

pub const APPEX_BUNDLE_MISSING_NS_EXTENSION_DICT: Self

pub const APPEX_BUNDLE_ILLEGAL_PACKAGE_TYPE_VALUE: Self

pub const APPEX_BUNDLE_MISSING_CLASS_OR_STORYBOARD: Self

pub const APPEX_BUNDLE_ILLEGAL_EXTENSION_CONTEXT_CLASS: Self

pub const APPEX_BUNDLE_ILLEGAL_EXTENSION_CONTEXT_HOST_CLASS: Self

pub const APPEX_BUNDLE_ILLEGAL_EXTENSION_VIEW_CONTROLLER_HOST_CLASS: Self

pub const APPEX_BUNDLE_MISSING_EXTENSION_POINT_IDENTIFIER_STRING: Self

pub const APPEX_BUNDLE_ILLEGAL_EXTENSION_POINT_IDENTIFIER_STRING: Self

pub const APPEX_BUNDLE_ILLEGAL_EXTENSION_ATTRIBUTES_DICT: Self

pub const APPEX_BUNDLE_ILLEGAL_EXTENSION_POINT_NAME_STRING: Self

pub const APPEX_BUNDLE_ILLEGAL_EXTENSION_POINT_VERSION_STRING: Self

pub const MISSING_BUNDLE_NAME_STRING: Self

pub const MISSING_BUNDLE_DISPLAY_NAME_STRING: Self

pub const ILLEGAL_BUNDLE_SHORT_VERSION_STRING: Self

pub const ILLEGAL_XPCSERVICE_RUN_LOOP_TYPE: Self

pub const ILLEGAL_XPCSERVICE_SERVICE_TYPE: Self

pub const DUPLICATE_IDENTIFIER: Self

pub const APPEX_BUNDLE_UNKNOWN_EXTENSION_POINT_IDENTIFIER_STRING: Self

pub const MULTIPLE_FILE_PROVIDER_APPEX_BUNDLES: Self

pub const MUX_CREATE_LISTENER_ERROR: Self

pub const MUX_GET_LISTENER: Self

pub const MUX_CONNECT: Self

pub const DEVICE_TOO_NEW: Self

pub const DEVICE_REF_NO_GOOD: Self

pub const DEVICE_DISCONNECTED: Self

pub const CANNOT_TRANSLATE: Self

pub const MOBILE_IMAGE_MOUNTER_MISSING_IMAGE_SIGNATURE: Self

pub const MOBILE_IMAGE_MOUNTER_RESPONSE_CREATION_FAILED: Self

pub const MOBILE_IMAGE_MOUNTER_MISSING_IMAGE_TYPE: Self

pub const MOBILE_IMAGE_MOUNTER_MISSING_IMAGE_PATH: Self

pub const MOBILE_IMAGE_MOUNTER_IMAGE_MAP_LOAD_FAILED: Self

pub const MOBILE_IMAGE_MOUNTER_ALREADY_MOUNTED: Self

pub const MOBILE_IMAGE_MOUNTER_IMAGE_MOVE_FAILED: Self

pub const MOBILE_IMAGE_MOUNTER_MOUNT_PATH_MISSING: Self

pub const MOBILE_IMAGE_MOUNTER_MOUNT_PATH_NOT_EMPTY: Self

pub const MOBILE_IMAGE_MOUNTER_IMAGE_MOUNT_FAILED: Self

pub const MOBILE_IMAGE_MOUNTER_TRUST_CACHE_LOAD_FAILED: Self

pub const MOBILE_IMAGE_MOUNTER_DIGEST_FAILED: Self

pub const MOBILE_IMAGE_MOUNTER_DIGEST_CREATION_FAILED: Self

pub const MOBILE_IMAGE_MOUNTER_IMAGE_VERIFICATION_FAILED: Self

pub const MOBILE_IMAGE_MOUNTER_IMAGE_INFO_CREATION_FAILED: Self

pub const MOBILE_IMAGE_MOUNTER_IMAGE_MAP_STORE_FAILED: Self

pub const BONJOUR_SETUP: Self

pub const NO_WIFI_SYNC_SUPPORT: Self

pub const MAXIMUM: Self

This should always be the last error. When adding new error code, steal whatever value is currently being used by kAMDMaximumError, and then bump the value of kAMDMaximumError by one.

pub const fn make(num: i32) -> Error

pub fn text(&self) -> Option<R<String>>

pub fn is_ok(&self) -> bool

pub fn is_err(&self) -> bool

pub unsafe fn to_result<T>(self, option: Option<T>) -> Result<T, Self>

pub fn result(self) -> Result<(), Self>