cidre::am::device::discovery

Type Alias NotificationCallback

pub type NotificationCallback<T> = extern "C" fn(info: &NotificationInfo, context: *mut T);

Notification callback. Ownership of the notification info struct info DOES NOT pass to the callback function.

The Action field of the ‘info’ parameter describes the notification being sent.

kAMDeviceAttached: A device has attached. The device reference belongs to the client. It must be explicitly released, or else it will leak.

kAMDeviceDetached: A device has detached. The device object delivered will be the same as the one delivered in the kAMDeviceAttached notification. This device reference does not need to be released.

kAMDeviceNotificationStopped: This notification is delivered in response to

A call to AMDeviceNotificationUnsubscribe().

An error occurred on one of the underlying notification systems (i.e. usbmuxd or mDNSResponder crashed or stopped responding). Unsubcribing and resubscribing may recover the notification system.