cidre::am::device

Struct Session

pub struct Session<'a>(/* private fields */);

Implementations

impl<'a> Session<'a>

pub unsafe fn mound_disk_with_callback<T>( &self, image: &String, options: &Dictionary, callback: *const MountCallback<T>, ctx: *mut T, ) -> Error

pub fn mount_developer_image(&self) -> Result<(), Error>

pub fn mound_disk( &self, image_path: &String, options: &Dictionary, ) -> Result<(), Error>

pub fn relay_file(&self, source: &String, options: &Dictionary) -> Error

Copy a file from the relay service. Options must contain:

kAMDRelayTypeKey: one of kAMDRelayTypeFileDescriptor kAMDRelayTypeData

impl<'a> Session<'a>

pub fn lookup_apps(&self, options: &Dictionary) -> Result<R<Dictionary>, Error>

Lookup the set of installed applications on the device.

The info argument will point to a cf::Dictionary. Each key in the dictionary is an application identifier and the value is the Info.plist of the installed application.

The client can specify options to filter the results. If the value of kLookupAttributeKey is: IMPORTANT NOTE: You should specify kLookupReturnAttributesKey in the options dictionary with an array value containing those keys you would like returned. If you do not specify this attribute, MobileDevice will log every time you make an offending call and performance will be poor. When specifying this key, you MUST include at least kCFBundleIdentifierKey in the array of keys to be returned or the call will not work properly.

pub fn apps_lookup(&'a self) -> AppsLookupBuilder<'a>

impl<'a> Session<'a>

pub fn secure_start_service( &self, name: &String, ) -> Result<R<ServiceConnection>, Error>

Securely start a service on the device specifying options.

Starts a service on the device. Requires that a session with the device be active. Attempting to start a service without an active session will result in kAMDInvalidArgumentError. Fails with kAMDNoWifiSyncSupportError if the device side service does not support SSL. Fails with kAMDServiceProhibitedError if the service is not allowed to run. Some services are only allowed to run when a device has been activated.

The device may request that a connection be encrypted. If so, this call will also perform the initial SSL handshake. See <MobileDevice/AMDServiceConnection.h> for more information on using the service connection.

Options

kAMDServiceOptionTimeoutConnection - The service connection will have SO_SNDTIMEO and SO_RCVTIMEO set. It will be closed if either input or output operations fail to complete within 25 seconds (use carefully if you expect your service to perform long-lived operations on your behalf). Valid values are boolean true or false. This is off by default.

kAMDServiceOptionUnlockEscrowBag - Attempt the passcode-unlock the device when starting a service.

kAMDOptionCloseOnInvalidate - The returned service connection will take ownership of the underlying file descriptor and SSL context. Manually closing the fd or SSL_free()’ing the ssl context will result in an double free()s and close()s when the service connection object is invalidate or ultimately released. Valid values are boolean true or false. This is on by default.

On success, the AMDServiceConnection returned in service_out must be CFRelease()’d by the caller.

pub fn start_debug_server(&self) -> Result<R<ServiceConnection>, Error>

pub fn battery_level(&self) -> Option<R<Number>>