cidre::arc

Type Alias A

pub type A<T> = Allocated<T>;

Aliased Type

pub struct A<T>(/* private fields */);

Implementations

impl A<AudioUnit>

pub unsafe fn init_with_comp_desc_opts_err<'ear>( self, desc: ComponentDesc, opts: ComponentInstantiationOpts, err: *mut Option<&'ear Error>, ) -> Option<R<AudioUnit>>

pub unsafe fn init_with_comp_desc_err<'ear>( self, desc: ComponentDesc, err: *mut Option<&'ear Error>, ) -> Option<R<AudioUnit>>

impl A<SysZoomSlider>

pub unsafe fn init_with_device_throws( self, device: &CaptureDevice, ) -> R<SysZoomSlider>

pub fn init_with_device<'ear>( self, device: &CaptureDevice, ) -> ExResult<'ear, R<SysZoomSlider>>

pub unsafe fn init_with_device_action_throws( self, device: &CaptureDevice, action: &mut EscBlock<fn(video_zoom_factor: Float)>, ) -> R<SysZoomSlider>

pub fn init_with_device_action<'ear>( self, device: &CaptureDevice, action: &mut EscBlock<fn(video_zoom_factor: Float)>, ) -> ExResult<'ear, R<SysZoomSlider>>

impl A<SysExposureBiasSlider>

pub unsafe fn init_with_device_throws( self, device: &CaptureDevice, ) -> R<SysExposureBiasSlider>

pub fn init_with_device<'ear>( self, device: &CaptureDevice, ) -> ExResult<'ear, R<SysExposureBiasSlider>>

pub unsafe fn init_with_device_action_throws( self, device: &CaptureDevice, action: &mut EscBlock<fn(video_zoom_factor: Float)>, ) -> R<SysExposureBiasSlider>

pub fn init_with_device_action<'ear>( self, device: &CaptureDevice, action: &mut EscBlock<fn(video_zoom_factor: Float)>, ) -> ExResult<'ear, R<SysExposureBiasSlider>>

impl A<Slider>

pub unsafe fn init_with_range_throws( self, localized_title: &String, symbol_name: &String, min_val: f32, max_val: f32, ) -> R<Slider>

pub fn init_with_range<'ear>( self, localized_title: &String, symbol_name: &String, min_val: f32, max_val: f32, ) -> ExResult<'ear, R<Slider>>

pub unsafe fn init_with_range_step_throws( self, localized_title: &String, symbol_name: &String, min_val: f32, max_val: f32, step: f32, ) -> R<Slider>

pub fn init_with_range_step<'ear>( self, localized_title: &String, symbol_name: &String, min_val: f32, max_val: f32, step: f32, ) -> ExResult<'ear, R<Slider>>

pub unsafe fn init_with_vals_throws( self, localized_title: &String, symbol_name: &String, vals: &Array<Number>, ) -> R<Slider>

pub fn init_with_vals<'ear>( self, localized_title: &String, symbol_name: &String, vals: &Array<Number>, ) -> ExResult<'ear, R<Slider>>

impl A<IndexPicker>

pub unsafe fn init_with_indexes_n_throws( self, localized_title: &String, symbol_name: &String, n: isize, ) -> R<IndexPicker>

pub fn init_with_indexes_n<'ear>( self, localized_title: &String, symbol_name: &String, n: isize, ) -> ExResult<'ear, R<IndexPicker>>

pub unsafe fn init_with_indexes_titles_throws( self, localized_title: &String, symbol_name: &String, titles: &Array<String>, ) -> R<IndexPicker>

pub fn init_with_titles<'ear>( self, localized_title: &String, symbol_name: &String, titles: &Array<String>, ) -> ExResult<'ear, R<IndexPicker>>

impl A<RotationCoordinator>

pub fn init_with_device_preview_layer( &self, device: &Device, preview_layer: Option<&Layer>, ) -> R<RotationCoordinator>

impl A<DeviceInput>

pub fn init_with_device_err<'ear>( self, device: &CaptureDevice, error: *mut Option<&'ear Error>, ) -> Option<R<DeviceInput>>

impl A<Session>

pub fn init(self) -> Retained<Session>

impl A<MultiCamSession>

pub fn init(self) -> Retained<MultiCamSession>

impl A<Connection>

pub fn init_with_ports( self, input: &Array<CaptureInputPort>, output: &CaptureOutput, ) -> R<Connection>

pub fn init_with_port_preview_layer( self, input: &CaptureInputPort, layer: &CaptureVideoPreviewLayer, ) -> R<Connection>

impl A<MetadataOutput>

pub fn init(self) -> Retained<MetadataOutput>

impl A<VideoDataOutput>

pub fn init(self) -> Retained<VideoDataOutput>

impl A<VideoPreviewLayer>

pub fn init_with_session(self, session: &CaptureSession) -> R<VideoPreviewLayer>

pub fn init_with_session_no_connection( self, session: &CaptureSession, ) -> R<VideoPreviewLayer>

impl A<PhotoOutput>

pub fn init(self) -> Retained<PhotoOutput>

impl A<PhotoSettings>

pub fn init(self) -> Retained<PhotoSettings>

impl A<AudioDataOutput>

pub fn init(self) -> Retained<AudioDataOutput>

impl A<ItemVideoOutput>

pub fn init_with_pixel_buf_attrs( self, pixel_buf_attrs: Option<Dictionary<String, Id>>, ) -> R<ItemVideoOutput>

pub fn init_with_output_settings( self, output_settings: Option<Dictionary<String, Id>>, ) -> R<ItemVideoOutput>

impl A<Player>

pub fn init_with_url(self, url: &Url) -> R<Player>

pub unsafe fn init_with_player_item_throws( self, item: Option<&PlayerItem>, ) -> R<Player>

impl A<QueuePlayer>

pub fn init(self) -> Retained<QueuePlayer>

impl A<QueuePlayer>

pub unsafe fn init_with_items_throws( self, items: &Array<PlayerItem>, ) -> R<QueuePlayer>

impl A<PlayerLayer>

pub fn init(self) -> Retained<PlayerLayer>

impl A<Reader>

pub unsafe fn init_with_assert_err<'ear>( self, asset: &Asset, error: *mut Option<&'ear Error>, ) -> Option<R<Reader>>

impl A<ReaderTrackOutput>

pub fn init_with_track_throws( self, track: &Track, output_settings: Option<&Dictionary<String, Id>>, ) -> R<ReaderTrackOutput>

impl A<Writer>

pub fn init_with_url_file_type_err<'ear>( self, url: &Url, file_type: &FileType, error: *mut Option<&'ear Error>, ) -> Option<R<Writer>>

pub unsafe fn init_with_content_type_throws( self, output_content_type: &Type, ) -> R<Writer>

impl A<WriterInput>

pub unsafe fn init_media_type_output_settings_throws( self, media_type: &MediaType, output_settings: Option<&Dictionary<String, Id>>, ) -> R<WriterInput>

pub unsafe fn with_media_type_output_settings_source_format_hint_throws( self, media_type: &MediaType, output_settings: Option<&Dictionary<String, Id>>, source_format_hint: Option<&FormatDesc>, ) -> R<WriterInput>

impl A<WriterInputPixelBufAdaptor>

pub unsafe fn init_with_asset_writer_input_throws( self, input: &WriterInput, src_pixel_buf_attrs: Option<Dictionary<String, Id>>, ) -> R<WriterInputPixelBufAdaptor>

impl A<ImageGenerator>

pub fn init_with_asset(self, asset: &Asset) -> R<ImageGenerator>

impl A<UrlAsset>

pub fn init_with_url_opts( self, url: &Url, options: Option<&Dictionary<String, Id>>, ) -> Option<R<UrlAsset>>

impl A<Effect>

pub fn init_with_audio_component_desc( self, description: ComponentDesc, ) -> R<Effect>

impl A<UnitEq>

pub fn init_with_number_of_bands(self, number_of_bands: usize) -> R<UnitEq>

impl A<TimeEffect>

pub fn init_with_audio_component_desc( self, description: ComponentDesc, ) -> R<TimeEffect>

impl A<Mix>

pub fn init(self) -> Retained<Mix>

impl A<MixMut>

pub fn init(self) -> Retained<MixMut>

impl A<InputParams>

pub fn init(self) -> Retained<InputParams>

impl A<InputParamsMut>

pub fn init(self) -> Retained<InputParamsMut>

impl A<MixerNode>

pub fn init(self) -> Retained<MixerNode>

impl A<Time>

pub fn init_with_audio_ts_sample_rate( self, ts: &AudioTimeStamp, sample_rate: f64, ) -> R<Time>

pub fn init_with_host_time(self, host_time: u64) -> R<Time>

pub fn init_with_host_time_sample_rate_at_rate( self, host_time: u64, sample_time: FramePos, at_rate: f64, ) -> R<Time>

pub fn init_with_sample_time_at_rate( self, time: FramePos, at_rate: f64, ) -> R<Time>

impl A<Engine>

pub fn init(self) -> Retained<Engine>

impl A<Player>

pub unsafe fn init_with_url_err<'ear>( self, url: &Url, err: *mut Option<&'ear Error>, ) -> Option<R<Player>>

pub unsafe fn init_with_data_err<'ear>( self, data: &Data, err: *mut Option<&'ear Error>, ) -> Option<R<Player>>

pub unsafe fn init_with_url_hint_err<'ear>( self, url: &Url, hint: Option<&String>, err: *mut Option<&'ear Error>, ) -> Option<R<Player>>

pub unsafe fn init_with_data_hint_err<'ear>( self, data: &Data, hint: Option<&String>, err: *mut Option<&'ear Error>, ) -> Option<R<Player>>

impl A<PlayerNode>

pub fn init(self) -> Retained<PlayerNode>

impl A<PcmBuf>

pub fn init_with_pcm_format_frame_capacity( self, format: &Format, frame_capacity: FrameCount, ) -> Option<R<PcmBuf>>

pub fn init_with_pcm_format_buf_list_no_copy<const N: usize>( self, format: &Format, buf_list: &AudioBufList<N>, deallocator: Option<&mut EscBlock<fn(buf_list: *const AudioBufList<N>)>>, ) -> Option<R<PcmBuf>>

impl A<CompressedBuf>

pub fn init_with_format_and_packet_capacity( self, format: &Format, packet_capacity: PacketCount, ) -> R<CompressedBuf>

pub fn init_with_format_packet_capacity_and_maximum_packet_size( self, format: &Format, packet_capacity: PacketCount, maximum_packet_size: isize, ) -> R<CompressedBuf>

impl A<Format>

pub fn init_with_stream_desc(self, asbd: &StreamBasicDesc) -> Option<R<Format>>

pub fn init_with_stream_desc_channel_layout( self, asbd: &StreamBasicDesc, layout: Option<&ChannelLayout>, ) -> Option<R<Format>>

pub fn init_standard_with_sample_rate_channels_ar( self, sample_rate: f64, channels: ChannelCount, ) -> Option<Rar<Format>>

pub fn init_standard_with_sample_rate_channels( self, sample_rate: f64, channels: ChannelCount, ) -> Option<R<Format>>

pub fn init_standard_with_sample_rate_channel_layout_ar( self, sample_rate: f64, layout: &ChannelLayout, ) -> Rar<Format>

pub fn init_standard_with_sample_rate_channel_layout( self, sample_rate: f64, layout: &ChannelLayout, ) -> R<Format>

pub fn init_with_common_format_sample_rate_channels_interleaved( self, format: CommonFormat, sample_rate: f64, channels: ChannelCount, interleaved: bool, ) -> Option<R<Format>>

pub fn init_with_common_format_sample_rate_interleaved_channel_layout( self, format: CommonFormat, sample_rate: f64, interleaved: bool, channel_layout: &ChannelLayout, ) -> R<Format>

pub fn init_with_settings( self, settings: &Dictionary<String, Id>, ) -> Option<R<Format>>

impl A<File>

pub unsafe fn init_for_reading_err_ar<'ear>( self, file_url: &Url, err: *mut Option<&'ear Error>, ) -> Option<Rar<File>>

pub unsafe fn init_for_reading_err<'ear>( self, file_url: &Url, err: *mut Option<&'ear Error>, ) -> Option<R<File>>

pub unsafe fn init_for_reading_common_format_err_ar<'ear>( self, file_url: &Url, common_format: CommonFormat, interleaved: bool, err: *mut Option<&'ear Error>, ) -> Option<Rar<File>>

pub unsafe fn init_for_reading_common_format_err<'ear>( self, file_url: &Url, common_format: CommonFormat, interleaved: bool, err: *mut Option<&'ear Error>, ) -> Option<R<File>>

pub unsafe fn init_for_writing_err_ar<'ear>( self, file_url: &Url, settings: &Dictionary<String, Id>, err: *mut Option<&'ear Error>, ) -> Option<Rar<File>>

pub unsafe fn init_for_writing_err<'ear>( self, file_url: &Url, settings: &Dictionary<String, Id>, err: *mut Option<&'ear Error>, ) -> Option<R<File>>

pub unsafe fn init_for_writing_common_format_err_ar<'ear>( self, file_url: &Url, settings: &Dictionary<String, Id>, common_format: CommonFormat, interleaved: bool, err: *mut Option<&'ear Error>, ) -> Option<Rar<File>>

pub unsafe fn init_for_writing_common_format_err<'ear>( self, file_url: &Url, settings: &Dictionary<String, Id>, common_format: CommonFormat, interleaved: bool, err: *mut Option<&'ear Error>, ) -> Option<R<File>>

impl A<ChannelLayout>

pub fn init_with_layout<const N: usize>( self, layout: &AudioChannelLayout<N>, ) -> R<ChannelLayout>

pub fn init_with_layout_tag( self, tag: AudioChannelLayoutTag, ) -> R<ChannelLayout>

impl A<ConnectionPoint>

pub fn init_with_node_bus(self, node: &Node, bus: NodeBus) -> R<ConnectionPoint>

impl A<Converter>

pub fn init_from_format_to_format_ar( self, from: &AudioFormat, to: &AudioFormat, ) -> Option<Rar<Converter>>

pub fn init_from_format_to_format( self, from: &AudioFormat, to: &AudioFormat, ) -> Option<R<Converter>>

impl A<Voice>

pub fn init(self) -> Retained<Voice>

impl A<Utterance>

pub fn init(self) -> Retained<Utterance>

impl A<Utterance>

pub fn init_with_string(self, val: &String) -> R<Utterance>

pub fn init_with_attr_string(self, val: &AttrString) -> R<Utterance>

pub fn init_with_ssml(self, val: &String) -> R<Utterance>

impl A<Synthesizer>

pub fn init(self) -> Retained<Synthesizer>

impl A<DisplayLayer>

pub fn init(self) -> Retained<DisplayLayer>

impl A<VideoRenderer>

pub fn init(self) -> Retained<VideoRenderer>

impl A<PipController>

pub fn init_with_content_src( self, content_src: &PipControllerContentSrc, ) -> R<PipController>

pub fn sel_init_with_content_src() -> &'static Sel

@selector(initWithContentSource:) but dynamic use this function to check if object responds to selector

pub fn init_with_player_layer( self, player_layer: &PlayerLayer, ) -> Option<R<PipController>>

impl A<PipControllerContentSrc>

pub fn init_with_player_layer( self, player_layer: &PlayerLayer, ) -> R<PipControllerContentSrc>

pub fn sel_init_with_player_layer() -> &'static Sel

@selector(initWithPlayerLayer:) but dynamic use this function to check if object responds to selector

impl A<PipControllerContentSrc>

pub fn init_with_sample_buf_display_layer<D: PipSampleBufPlayerDelegate>( self, sbuf_display_layer: &SampleBufDisplayLayer, playback_delegate: &D, ) -> R<PipControllerContentSrc>

impl A<PlaybackSpeed>

pub fn init_with_rate( self, rate: f32, localized_name: &String, ) -> R<PlaybackSpeed>

impl A<DisplayLink>

pub fn init(self) -> Retained<DisplayLink>

impl A<Animation>

pub fn init(self) -> Retained<Animation>

impl A<MediaTimingFn>

pub fn init_with_ctrl_points( self, c1x: f32, c1y: f32, c2x: f32, c2y: f32, ) -> R<MediaTimingFn>

impl A<Layer>

pub fn init(self) -> Retained<Layer>

impl A<MetalLayer>

pub fn init(self) -> Retained<MetalLayer>

impl A<Transaction>

pub fn init(self) -> Retained<Transaction>

impl A<Context>

pub fn init(self) -> Retained<Context>

impl A<Context>

pub fn init_with_opts( self, options: Option<&Dictionary<String, Id>>, ) -> Option<R<Context>>

impl A<Image>

pub fn init_with_mlt_texture_options( self, texture: &Texture, options: Option<&DictionaryOf<ImageOpt, Type>>, ) -> Option<Retained<Image>>

pub fn init_with_cv_image_buf_options( self, image_buf: &ImageBuf, options: Option<&DictionaryOf<ImageOpt, Type>>, ) -> Option<Retained<Image>>

pub fn init_with_cv_pixel_buf_options( self, pixel_buf: &PixelBuf, options: Option<&DictionaryOf<ImageOpt, Type>>, ) -> Option<Retained<Image>>

impl A<Vec>

pub fn init_with_vals(self, values: *const Float, count: usize) -> R<Vec>

pub fn init_ns_string(self, val: &String) -> R<Vec>

impl A<LocationManager>

pub fn init(self) -> Retained<LocationManager>

impl A<Location>

pub fn init(self) -> Retained<Location>

impl A<Location>

pub fn init_with_lat_lon( self, lat: LocationDegrees, lon: LocationDegrees, ) -> R<Location>

impl A<TapDesc>

pub fn init(self) -> Retained<TapDesc>

impl A<TapDesc>

pub fn init_stereo_mixdown_of_processes_ar( self, processes_obj_ids_to_include: &Array<Number>, ) -> Rar<TapDesc>

pub fn init_stereo_mixdown_of_processes( self, processes_obj_ids_to_include: &Array<Number>, ) -> R<TapDesc>

pub fn init_stereo_global_tap_but_exclude_processes_ar( self, processes_obj_ids_to_exclude: &Array<Number>, ) -> Rar<TapDesc>

pub fn init_stereo_global_tap_but_exclude_processes( self, processes_obj_ids_to_exclude: &Array<Number>, ) -> R<TapDesc>

pub fn init_mono_mixdown_of_processes_ar( self, processes_obj_ids_to_include: &Array<Number>, ) -> Rar<TapDesc>

pub fn init_mono_mixdown_of_processes( self, processes_obj_ids_to_include: &Array<Number>, ) -> R<TapDesc>

pub fn init_mono_global_tap_but_exclude_processes_ar( self, processes_obj_ids_to_exclude: &Array<Number>, ) -> Rar<TapDesc>

pub fn init_mono_global_tap_but_exclude_processes( self, processes_obj_ids_to_exclude: &Array<Number>, ) -> R<TapDesc>

pub fn init_with_processes_and_device( self, processes_obj_ids_to_include: &Array<Number>, device_uid: &String, stream: isize, ) -> R<TapDesc>

pub fn init_excluding_processes_and_device_ar( self, processes_obj_ids_to_exclude: &Array<Number>, device_uid: &String, stream: isize, ) -> Rar<TapDesc>

pub fn init_excluding_processes_and_device( self, processes_obj_ids_to_exclude: &Array<Number>, device_uid: &String, stream: isize, ) -> R<TapDesc>

impl A<HeadphoneMotionManager>

pub fn init(self) -> Retained<HeadphoneMotionManager>

impl A<MotionManager>

pub fn init(self) -> Retained<MotionManager>

impl A<Pedometer>

pub fn init(self) -> Retained<Pedometer>

impl A<Color>

pub fn init_with_rgb(self, r: f32, g: f32, b: f32) -> R<Color>

impl A<BrowserViewController>

pub unsafe fn init_with_service_type_throws( self, service_type: &String, session: &Session, ) -> R<BrowserViewController>

pub unsafe fn init_with_browser_throws( self, browser: &NearbyServiceBrowser, session: &Session, ) -> R<BrowserViewController>

impl A<PeerId>

pub unsafe fn init_with_display_name_throws( self, display_name: &String, ) -> R<PeerId>

impl A<Session>

pub unsafe fn init_with_peer_throws(self, peer: &PeerId) -> R<Session>

pub unsafe fn init_with_encryption_throws( self, peer: &PeerId, sec_identity: Option<&Array<Id>>, encryption_preference: EncryptionPreference, ) -> R<Session>

impl A<AdvertiserAssistant>

pub unsafe fn init_service_type_throws_ar( self, peer: &PeerId, service_type: &String, ) -> Rar<AdvertiserAssistant>

pub unsafe fn init_service_type_throws( self, peer: &PeerId, service_type: &String, ) -> R<AdvertiserAssistant>

impl A<NearbyServiceAdvertiser>

pub unsafe fn init_with_peer_throws( self, my_peer: &PeerId, discorvery_info: Option<&Dictionary<String, String>>, service_type: &String, ) -> R<NearbyServiceAdvertiser>

impl A<NearbyServiceBrowser>

pub unsafe fn init_with_peer_and_service_type_throws( self, peer: &PeerId, service_type: &String, ) -> R<NearbyServiceBrowser>

impl A<GeometryDesc>

pub fn init(self) -> Retained<GeometryDesc>

impl A<TriangleGeometryDesc>

pub fn init(self) -> Retained<TriangleGeometryDesc>

impl A<MotionBoundingBoxGeometryDesc>

pub fn init(self) -> Retained<MotionBoundingBoxGeometryDesc>

impl A<TensorExtents>

pub fn init(self) -> Retained<TensorExtents>

impl A<TensorExtents>

pub fn sel_init_with_rank() -> &'static Sel

@selector(initWithRank:values:) but dynamic use this function to check if object responds to selector

pub unsafe fn init_with_rank( self, rank: usize, values: *const usize, ) -> Option<R<TensorExtents>>

impl A<TensorDesc>

pub fn init(self) -> Retained<TensorDesc>

impl A<Tensor>

pub fn init(self) -> Retained<Tensor>

impl A<ResidencySetDesc>

pub fn init(self) -> Retained<ResidencySetDesc>

impl A<CompileOpts>

pub fn init(self) -> Retained<CompileOpts>

impl A<Desc>

pub fn init(self) -> Retained<Desc>

impl A<FnConstValues>

pub fn init(self) -> Retained<FnConstValues>

impl A<FnStitchingAttrAlwaysInline>

pub fn init(self) -> Retained<FnStitchingAttrAlwaysInline>

impl A<FnStitchingInputNode>

pub fn init(self) -> Retained<FnStitchingInputNode>

impl A<FnStitchingInputNode>

pub fn init_with_arg_index(self, index: usize) -> R<FnStitchingInputNode>

impl A<FnStitchingFnNode>

pub fn init_with_name( self, name: &String, args: &Array<FnStitchingNode>, control_deps: &Array<FnStitchingFnNode>, ) -> R<FnStitchingFnNode>

impl A<FnStitchingGraph>

pub fn init_with_fn_name( self, fn_name: &String, nodes: &Array<FnStitchingFnNode>, output_node: Option<&FnStitchingFnNode>, attribtues: &Array<FnStitchingAttr>, ) -> R<FnStitchingGraph>

impl A<FnStitchedLibDesc>

pub fn init(self) -> Retained<FnStitchedLibDesc>

impl A<FnDesc>

pub fn init(self) -> Retained<FnDesc>

impl A<Desc>

pub fn init(self) -> Retained<Desc>

impl A<Desc>

pub fn init(self) -> Retained<Desc>

impl A<TileRenderPipelineDesc>

pub fn init(self) -> Retained<TileRenderPipelineDesc>

impl A<SharedEventListener>

pub fn init_with_dispatch_queue(self, queue: &Queue) -> R<SharedEventListener>

impl A<Desc>

pub fn init(self) -> Retained<Desc>

impl A<StencilDesc>

pub fn init(self) -> Retained<StencilDesc>

impl A<DepthStencilDesc>

pub fn init(self) -> Retained<DepthStencilDesc>

impl A<VertexBufLayoutDesc>

pub fn init(self) -> Retained<VertexBufLayoutDesc>

impl A<VertexAttrDesc>

pub fn init(self) -> Retained<VertexAttrDesc>

impl A<Desc>

pub fn init(self) -> Retained<Desc>

impl A<Desc>

pub fn init(self) -> Retained<Desc>

impl A<Desc>

pub fn init(self) -> Retained<Desc>

impl A<Desc>

pub fn init(self) -> Retained<Desc>

impl A<Desc>

pub fn init(self) -> Retained<Desc>

impl A<Desc>

pub fn init(self) -> Retained<Desc>

impl A<Desc>

pub fn init(self) -> Retained<Desc>

impl A<PipelineBufDesc>

pub fn init(self) -> Retained<PipelineBufDesc>

impl A<CaptureDesc>

pub fn init(self) -> Retained<CaptureDesc>

impl A<ArgTableDesc>

pub fn init(self) -> Retained<ArgTableDesc>

impl A<CounterHeapDesc>

pub fn init(self) -> Retained<CounterHeapDesc>

impl A<CmdAllocatorDesc>

pub fn init(self) -> Retained<CmdAllocatorDesc>

impl A<CmdBufOpts>

pub fn init(self) -> Retained<CmdBufOpts>

impl A<CommitOpts>

pub fn init(self) -> Retained<CommitOpts>

impl A<CmdQueueDesc>

pub fn init(self) -> Retained<CmdQueueDesc>

impl A<RenderPassDesc>

pub fn init(self) -> Retained<RenderPassDesc>

impl A<MlPipelineDesc>

pub fn init(self) -> Retained<MlPipelineDesc>

impl A<PipelineOpts>

pub fn init(self) -> Retained<PipelineOpts>

impl A<TextureLoader>

pub fn init_with_device(self, device: &Device) -> R<TextureLoader>

impl A<Device>

pub fn init(self) -> Retained<Device>

impl A<Graph>

pub fn init(self) -> Retained<Graph>

impl A<TrainingGraph>

pub fn init(self) -> Retained<TrainingGraph>

impl A<ExecutionDesc>

pub fn init(self) -> Retained<ExecutionDesc>

impl A<Executable>

pub fn init(self) -> Retained<Executable>

impl A<Executable>

pub fn init_with_mps_graph_package_at_url( self, mps_graph_package_url: &Url, compilation_descriptor: Option<&CompilationDesc>, ) -> R<Executable>

pub fn sel_init_with_mps_graph_package_at_url() -> &'static Sel

@selector(initWithMPSGraphPackageAtURL:compilationDescriptor:) but dynamic use this function to check if object responds to selector

pub fn init_with_core_ml_package_at_url( self, core_ml_package_url: &Url, compilation_descriptor: Option<&CompilationDesc>, ) -> R<Executable>

pub fn sel_init_with_core_ml_package_at_url() -> &'static Sel

@selector(initWithCoreMLPackageAtURL:compilationDescriptor:) but dynamic use this function to check if object responds to selector

impl A<TensorData>

pub fn init(self) -> Retained<TensorData>

impl A<TensorData>

pub fn init_with_device_data_shape_data_type( self, device: &Device, data: &Data, shape: &Shape, data_type: DType, ) -> R<TensorData>

pub fn init_with_mtl_buffer_shape_data_type( self, buffer: &Buf, shape: &Shape, data_type: DType, ) -> R<TensorData>

pub fn init_with_mtl_buffer_shape_data_type_row_bytes( self, buffer: &Buf, shape: &Shape, data_type: DType, row_bytes: usize, ) -> R<TensorData>

pub fn sel_init_with_mtl_buffer_shape_data_type_row_bytes() -> &'static Sel

@selector(initWithMTLBuffer:shape:dataType:rowBytes:) but dynamic use this function to check if object responds to selector

pub fn init_with_mps_nd_array(self, ndarray: &NdArray) -> R<TensorData>

impl A<SingleGateRnnDesc>

pub fn init(self) -> Retained<SingleGateRnnDesc>

impl A<LstmDesc>

pub fn init(self) -> Retained<LstmDesc>

impl A<GruDescr>

pub fn init(self) -> Retained<GruDescr>

impl A<CompilationDesc>

pub fn init(self) -> Retained<CompilationDesc>

impl A<ExecutionDesc>

pub fn init(self) -> Retained<ExecutionDesc>

impl A<Graph>

pub fn init(self) -> Retained<Graph>

impl A<ArrayBatchProvider>

pub fn init_with_feature_provider_array<P: FeatureProvider>( self, arr: &Array<P>, ) -> R<ArrayBatchProvider>

pub unsafe fn init_with_dictionary<'ear>( self, dictionary: &Dictionary<String, Id>, err: *mut Option<&'ear Error>, ) -> Option<R<ArrayBatchProvider>>

impl A<DictionaryFeatureProvider>

pub unsafe fn init_with_dictionary_err<'ear>( self, dictionary: &Dictionary<String, Id>, err: *mut Option<&'ear Error>, ) -> Option<R<DictionaryFeatureProvider>>

impl A<FeatureValue>

pub fn init(self) -> Retained<FeatureValue>

impl A<MultiArray>

pub fn init(self) -> Retained<MultiArray>

impl A<MultiArray>

pub unsafe fn init_with_shape_err<'ear>( self, shape: &Array<Number>, d_type: MultiArrayDType, err: *mut Option<&'ear Error>, ) -> Option<R<MultiArray>>

pub fn init_with_shape_strides( self, shape: &Array<Number>, d_type: MultiArrayDType, strides: &Array<Number>, ) -> R<MultiArray>

pub unsafe fn init_with_ptr<'ear>( self, data_ptr: *mut c_void, shape: &Array<Number>, d_type: MultiArrayDType, strides: &Array<Number>, deallocator: Option<&mut EscBlock<fn(ptr: *mut c_void)>>, error: *mut Option<&'ear Error>, ) -> Option<R<MultiArray>>

pub unsafe fn init_with_pixel_buf_throws( self, pixel_buf: &PixelBuf, shape: &Array<Number>, ) -> R<MultiArray>

impl A<PredictionOpts>

pub fn init(self) -> Retained<PredictionOpts>

impl A<ModelCfg>

pub fn init(self) -> Retained<ModelCfg>

impl A<Bundle>

pub fn init_with_path(self, path: &String) -> Option<R<Bundle>>

impl A<Predicate>

pub fn init(self) -> Retained<Predicate>

impl A<DateFormatter>

pub fn init(self) -> Retained<DateFormatter>

impl A<Iso8601DateFormatter>

pub fn init(self) -> Retained<Iso8601DateFormatter>

impl A<Port>

pub fn init(self) -> Retained<Port>

impl A<MachPort>

pub fn init(self) -> Retained<MachPort>

impl A<UrlRequest>

pub fn init_with_url(self, url: &Url) -> R<UrlRequest>

pub fn init_with_url_cache_policy_timeout( self, url: &Url, cache_policy: CachePolicy, timeout: TimeInterval, ) -> R<UrlRequest>

impl A<UrlRequestMut>

pub fn init_with_url(self, url: &Url) -> R<UrlRequestMut>

pub fn init_with_url_cache_policy_timeout( self, url: &Url, cache_policy: CachePolicy, timeout: TimeInterval, ) -> R<UrlRequestMut>

impl A<WebSocketMessage>

pub fn init_with_data(self, data: &Data) -> R<WebSocketMessage>

pub fn init_with_string(self, string: &String) -> R<WebSocketMessage>

impl A<UrlCache>

pub fn init_with_capacity( self, mem_capacity: usize, disk_capacity: usize, directory_url: Option<&Url>, ) -> R<UrlCache>

impl A<UserActivity>

pub fn init_with_activity_type(self, activity_type: &String) -> R<UserActivity>

impl A<UserDefaults>

pub fn init(self) -> Retained<UserDefaults>

impl A<UserDefaults>

pub fn init_with_suite_name( self, name: Option<&String>, ) -> Option<R<UserDefaults>>

impl A<IndexPath>

pub fn init(self) -> Retained<IndexPath>

impl A<IndexPath>

pub fn init_with_index(self, index: usize) -> R<IndexPath>

pub fn init_with_indexes( self, indexes: *const usize, lenght: usize, ) -> R<IndexPath>

impl A<IndexSet>

pub fn init_with_index(self, val: usize) -> R<IndexSet>

pub fn init_with_index_set(self, val: &IndexSet) -> R<IndexSet>

pub fn init_with_range(self, val: Range) -> R<IndexSet>

impl A<IndexSetMut>

pub fn init_with_index(self, val: usize) -> R<IndexSetMut>

pub fn init_with_index_set(self, val: &IndexSet) -> R<IndexSetMut>

pub fn init_with_range(self, val: Range) -> R<IndexSetMut>

impl A<IndexSet>

pub fn init(self) -> Retained<IndexSet>

impl A<IndexSetMut>

pub fn init(self) -> Retained<IndexSetMut>

impl<T: Obj> A<Array<T>>

pub fn init_ar(self) -> Rar<Array<T>>

Initializes an empty array.

pub fn init(self) -> R<Array<T>>

Initializes an empty array.

pub unsafe fn init_with_objs(self, ptr: *const &T, count: usize) -> R<Array<T>>

Initializes from a raw pointer to Objective-C object references.

Safety

ptr must point to count valid references for the duration of the call.

impl<T: Obj> A<ArrayMut<T>>

pub fn init_with_capacity(self, capacity: usize) -> R<ArrayMut<T>>

pub unsafe fn init_with_objs( &self, ptr: *const &T, count: usize, ) -> R<ArrayMut<T>>

impl<K: Obj, V: Obj> A<Dictionary<K, V>>

pub fn init_ar(self) -> Rar<Dictionary<K, V>>

pub fn init(self) -> R<Dictionary<K, V>>

pub unsafe fn init_with_objs_for_keys_count_throws( self, objects: *const V, keys: *const K, count: usize, ) -> R<Dictionary<K, V>>

impl<K: Obj, V: Obj> A<DictionaryMut<K, V>>

pub fn init_ar(self) -> Rar<DictionaryMut<K, V>>

pub fn init(self) -> R<DictionaryMut<K, V>>

pub fn init_with_capacity(self, capacity: usize) -> R<DictionaryMut<K, V>>

impl<T: Obj> A<Set<T>>

pub fn init_ar(self) -> Rar<Set<T>>

pub fn init(self) -> R<Set<T>>

pub fn init_with_objs_count(self, ptr: *const c_void, count: usize) -> R<Set<T>>

impl<T: Obj> A<SetMut<T>>

pub fn init_ar(self) -> Rar<SetMut<T>>

pub fn init(self) -> R<SetMut<T>>

pub fn init_with_capacity(self, capacity: usize) -> R<SetMut<T>>

impl A<Data>

pub fn init(self) -> Retained<Data>

impl A<DataMut>

pub fn init(self) -> Retained<DataMut>

impl A<Data>

pub fn init_with_contents_of_file_opts_err<'ear>( self, path: &String, options: ReadOpts, error: *mut Option<&'ear Error>, ) -> Option<R<Data>>

pub fn init_with_contents_of_url_opts_err<'ear>( self, url: &Url, options: ReadOpts, error: *mut Option<&'ear Error>, ) -> Option<R<Data>>

pub fn init_with_bytes(self, bytes: *const u8, length: usize) -> R<Data>

impl A<Regex>

pub fn init_with_pattern_opts_err<'ear>( self, pattern: &String, opts: Opts, error: *mut Option<&'ear Error>, ) -> Option<R<Regex>>

impl A<String>

pub fn init(self) -> Retained<String>

impl A<StringMut>

pub fn init(self) -> Retained<StringMut>

impl A<String>

pub fn init_with_bytes_length_encoding( self, bytes: *const u8, length: usize, encoding: StringEncoding, ) -> Option<R<String>>

pub fn init_with_bytes_no_copy_length_encoding_free_when_done( self, bytes: *const u8, length: usize, encoding: StringEncoding, free_when_done: bool, ) -> Option<R<String>>

impl A<StringMut>

pub fn with_capacity(self, capacity: usize) -> R<StringMut>

impl A<AttrString>

pub fn init(self) -> Retained<AttrString>

impl A<AttrString>

pub fn init_with_string(self, str: &String) -> R<AttrString>

pub fn init_with_attr_string(self, str: &AttrString) -> R<AttrString>

impl A<AttrStringMut>

pub fn init(self) -> Retained<AttrStringMut>

impl A<AttrStringMut>

pub fn init_with_string(self, str: &String) -> R<AttrStringMut>

pub fn init_with_attr_string(self, str: &AttrString) -> R<AttrStringMut>

impl A<Url>

pub fn init_with_path_is_dir_relative_to_url_ar( self, path: &String, is_dir: bool, relative_to: Option<&Url>, ) -> Rar<Url>

pub fn init_with_path_is_dir_relative_to_url( self, path: &String, is_dir: bool, relative_to: Option<&Url>, ) -> R<Url>

pub fn init_with_string_relative_to( self, string: &String, relative_to: Option<&Url>, ) -> Option<R<Url>>

impl A<Uuid>

pub fn init(self) -> Retained<Uuid>

impl A<RunLoop>

pub fn init(self) -> Retained<RunLoop>

impl A<Date>

pub fn init(self) -> Retained<Date>

impl A<Date>

pub fn init_with_time_interval_since_now(self, secs: TimeInterval) -> R<Date>

pub fn init_with_time_interval_since_1970(self, secs: TimeInterval) -> R<Date>

impl A<Error>

pub fn init_with_domain( self, domain: &ErrorDomain, code: Integer, user_info: Option<&Dictionary<String, Id>>, ) -> R<Error>

impl A<Thread>

pub fn init(self) -> Retained<Thread>

impl A<Timer>

pub fn init_with_fire_date_block( self, date: &Date, interval: TimeInterval, repeats: bool, block: &mut SyncBlock<fn(timer: &mut Timer)>, ) -> R<Timer>

pub fn init_with_fire_date_sel( self, date: &Date, interval: TimeInterval, target: &Id, selector: &Sel, user_info: Option<&Id>, repeats: bool, ) -> R<Timer>

impl A<Notification>

pub fn init_with_name( self, name: &NotificationName, object: Option<&Id>, user_info: Option<&Dictionary<Id, Id>>, ) -> R<Notification>

pub fn init_with_coder(self, coder: &Coder) -> R<Notification>

impl A<NotificationCenter>

pub fn init(self) -> Retained<NotificationCenter>

impl A<OpQueue>

pub fn init(self) -> Retained<OpQueue>

impl A<Coder>

pub fn init(self) -> Retained<Coder>

impl A<Locale>

pub fn init_with_locale_id(self, id: &String) -> R<Locale>

impl<S: Obj, I: Obj> A<DiffableDataSrcSnapshot<S, I>>

pub fn init_ar(self) -> Rar<DiffableDataSrcSnapshot<S, I>>

pub fn init(self) -> R<DiffableDataSrcSnapshot<S, I>>

impl<T: Obj> A<OrderedCollectionChange<T>>

pub fn init_with_object_type_index( self, object: Option<&T>, type_: CollectionChangeType, index: UInteger, ) -> R<OrderedCollectionChange<T>>

pub fn init_with_object_type_index_associated( self, object: Option<&T>, type_: CollectionChangeType, index: UInteger, associated_index: UInteger, ) -> R<OrderedCollectionChange<T>>

impl<T: Obj> A<OrderedCollectionDiff<T>>

pub fn init_with_changes( self, changes: &Array<OrderedCollectionChange<T>>, ) -> R<OrderedCollectionDiff<T>>

impl A<AnimationCtx>

pub fn init(self) -> Retained<AnimationCtx>

impl A<Button>

pub fn init(self) -> Retained<Button>

impl A<Cell>

pub fn init_text_cell_ar(self, string: &String) -> Rar<Cell>

pub fn init_text_cell(self, string: &String) -> R<Cell>

pub fn init_image_cell_ar(self, image: Option<&Image>) -> Rar<Cell>

pub fn init_image_cell(self, image: Option<&Image>) -> R<Cell>

impl A<Cell>

pub fn init(self) -> Retained<Cell>

impl A<CollectionView>

pub fn init(self) -> Retained<CollectionView>

impl A<CollectionViewItem>

pub fn init(self) -> Retained<CollectionViewItem>

impl A<ColorSpace>

pub fn init(self) -> Retained<ColorSpace>

impl A<ColorSpace>

pub fn init_with_cg_color_space(self, val: &ColorSpace) -> Option<R<ColorSpace>>

impl A<Color>

pub fn init(self) -> Retained<Color>

impl<S: Obj, I: Obj> A<CollectionViewDiffableDataSrc<S, I>>

pub fn init_with_collection_view_item_provider( self, collection_view: &CollectionView, item_provider: &mut CollectionViewDiffableDataSrcItemProvider<I>, ) -> R<CollectionViewDiffableDataSrc<S, I>>

impl A<Event>

pub fn init(self) -> Retained<Event>

impl A<Font>

pub fn init(self) -> Retained<Font>

impl A<FontManager>

pub fn init(self) -> Retained<FontManager>

impl A<GestureRecognizer>

pub fn init_with_target( self, target: Option<&Id>, action: Option<&Sel>, ) -> R<GestureRecognizer>

impl A<GestureRecognizer>

pub fn init(self) -> Retained<GestureRecognizer>

impl A<Image>

pub fn init(self) -> Retained<Image>

impl A<Menu>

pub fn init(self) -> Retained<Menu>

impl A<Menu>

pub fn init_with_title(self, title: &String) -> R<Menu>

impl A<Pasteboard>

pub fn init(self) -> Retained<Pasteboard>

impl A<Responder>

pub fn init(self) -> Retained<Responder>

impl A<RunningApp>

pub fn init(self) -> Retained<RunningApp>

impl A<Screen>

pub fn init(self) -> Retained<Screen>

impl A<SplitViewItem>

pub fn init(self) -> Retained<SplitViewItem>

impl A<TextField>

pub fn init(self) -> Retained<TextField>

impl A<TitlebarAccessoryViewController>

pub fn init(self) -> Retained<TitlebarAccessoryViewController>

impl A<ViewController>

pub fn init(self) -> Retained<ViewController>

impl A<View>

pub fn init(self) -> Retained<View>

impl A<View>

pub fn init_with_frame(self, frame: Rect) -> R<View>

impl A<Toolbar>

pub fn init(self) -> Retained<Toolbar>

impl A<Toolbar>

pub fn init_with_id(self, id: &ToolbarId) -> R<Toolbar>

impl A<ToolbarItem>

pub fn init(self) -> Retained<ToolbarItem>

impl A<ToolbarItem>

pub fn init_with_item_id(self, id: &ToolbarItemId) -> R<ToolbarItem>

impl A<StatusBar>

pub fn init(self) -> Retained<StatusBar>

impl A<WindowController>

pub fn init_with_window(self, window: Option<&Window>) -> R<WindowController>

impl A<WindowController>

pub fn init(self) -> Retained<WindowController>

impl A<Window>

pub fn init(self) -> Retained<Window>

impl A<WorkspaceOpenCfg>

pub fn init(self) -> Retained<WorkspaceOpenCfg>

impl A<Connection>

pub fn init_with_service_name(self, name: &String) -> R<Connection>

pub fn init_with_mach_service( self, name: &String, options: ConnectionOptions, ) -> R<Connection>

pub fn init_with_listener_endpoint( self, endpoint: &ListenerEndpoint, ) -> R<Connection>

impl A<Listener>

pub fn init_with_mach_service_name(self, name: &String) -> R<Listener>

impl A<TextAttachment>

pub fn init(self) -> Retained<TextAttachment>

impl A<TextAttachment>

pub fn init_with_data( self, data: Option<&Data>, type_: Option<&String>, ) -> R<TextAttachment>

impl A<LangRecognizer>

pub fn init(self) -> Retained<LangRecognizer>

impl A<Embedding>

pub fn init(self) -> Retained<Embedding>

impl A<Cfg>

pub fn init(self) -> Retained<Cfg>

impl A<Cfg>

pub fn init(self) -> Retained<Cfg>

impl A<ContentFilter>

pub fn init_with_desktop_independent_window( self, window: &Window, ) -> Retained<ContentFilter>

pub fn init_with_display_excluding_windows( self, display: &Display, windows: &Array<Window>, ) -> Retained<ContentFilter>

pub fn init_with_display_including_windows( self, display: &Display, windows: &Array<Window>, ) -> Retained<ContentFilter>

Creates a content filter that includes only the specified windows on the given display.

Use this initializer to capture specific windows from a display. Unlike init_with_desktop_independent_window, this method works reliably for all window types including regular application windows.

pub fn sel_init_with_display_including_windows() -> &'static Sel

@selector(initWithDisplay:includingWindows:) but dynamic use this function to check if object responds to selector

pub fn init_with_display_including_apps_excepting_windows( self, display: &Display, apps: &Array<RunningApp>, excepting_windows: &Array<Window>, ) -> Retained<ContentFilter>

Creates a content filter that includes content from the specified applications, optionally excluding specific windows.

pub fn sel_init_with_display_including_apps_excepting_windows() -> &'static Sel

@selector(initWithDisplay:includingApplications:exceptingWindows:) but dynamic use this function to check if object responds to selector

pub fn init_with_display_excluding_apps_excepting_windows( self, display: &Display, apps: &Array<RunningApp>, excepting_windows: &Array<Window>, ) -> Retained<ContentFilter>

Creates a content filter that excludes content from the specified applications, optionally excepting specific windows that should still be included.

pub fn sel_init_with_display_excluding_apps_excepting_windows() -> &'static Sel

@selector(initWithDisplay:excludingApplications:exceptingWindows:) but dynamic use this function to check if object responds to selector

impl A<Stream>

pub fn init_with_filter_configuration_delegate<D: Delegate>( self, filter: &ContentFilter, configuration: &Cfg, delegate: Option<&D>, ) -> Retained<Stream>

impl A<RecordingOutputCfg>

pub fn init(self) -> Retained<RecordingOutputCfg>

impl A<RecordingOutput>

pub fn init_with_cfg_delegate<D: Delegate>( self, cfg: &RecordingOutputCfg, delegate: &D, ) -> Retained<RecordingOutput>

impl A<AudioStreamAnalyzer>

pub fn init(self) -> Retained<AudioStreamAnalyzer>

impl A<AudioStreamAnalyzer>

pub fn init_with_format(self, format: &AudioFormat) -> R<AudioStreamAnalyzer>

impl A<AudioFileAnalyzer>

pub fn init(self) -> Retained<AudioFileAnalyzer>

impl A<AudioFileAnalyzer>

pub unsafe fn init_with_url_err<'ear>( self, url: &Url, error: *mut Option<&'ear Error>, ) -> Option<R<AudioFileAnalyzer>>

impl A<Constraint>

pub fn init_with_duration_range( self, duration_range: TimeRange, ) -> R<Constraint>

pub fn init_with_enumerated_durations( self, durations: &Array<Value>, ) -> R<Constraint>

impl A<ClassifySoundRequest>

pub unsafe fn init_with_classifier_id_err<'ear>( self, id: &Id, err: *mut Option<&'ear Error>, ) -> Option<R<ClassifySoundRequest>>

impl A<ClassificationResult>

pub fn init(self) -> Retained<ClassificationResult>

impl A<NotificationContent>

pub fn init(self) -> Retained<NotificationContent>

impl A<NotificationContentMut>

pub fn init(self) -> Retained<NotificationContentMut>

impl A<CalcImageAestheticsScoresRequest>

pub fn init_with_ch( self, ch: Option<&mut RequestCh<CalcImageAestheticsScoresRequest>>, ) -> R<CalcImageAestheticsScoresRequest>

impl A<CalcImageAestheticsScoresRequest>

pub fn init(self) -> Retained<CalcImageAestheticsScoresRequest>

impl A<DetectHorizonRequest>

pub fn init(self) -> Retained<DetectHorizonRequest>

impl A<ImageRequestHandler>

pub fn init_with_url_options( self, url: &Url, options: Option<&Dictionary<Id, Id>>, ) -> R<ImageRequestHandler>

pub fn init_with_url_orientation_options( self, url: &Url, orientation: ImagePropOrientation, options: Option<&Dictionary<Id, Id>>, ) -> R<ImageRequestHandler>

pub fn init_with_cv_pixel_buf_options( self, pb: &PixelBuf, options: Option<&Dictionary<Id, Id>>, ) -> Option<R<ImageRequestHandler>>

pub fn init_with_cv_pixel_buf_orientaion_options( self, pb: &PixelBuf, orientation: ImagePropOrientation, options: Option<&Dictionary<Id, Id>>, ) -> Option<R<ImageRequestHandler>>

impl A<SequenceRequestHandler>

pub fn init(self) -> Retained<SequenceRequestHandler>

impl A<StatefulRequest>

pub fn init_with_frame_analysis_spacing_ch( self, frame_analysis_spacing: Time, ch: Option<&mut RequestCh>, ) -> R<StatefulRequest>

impl A<RecognizeTextRequest>

pub fn init(self) -> Retained<RecognizeTextRequest>

impl A<RecognizeAnimalsRequest>

pub fn init(self) -> Retained<RecognizeAnimalsRequest>

impl A<GenForegroundInstanceMaskRequest>

pub fn init(self) -> Retained<GenForegroundInstanceMaskRequest>

impl A<GenPersonSegmentationRequest>

pub fn init(self) -> Retained<GenPersonSegmentationRequest>

impl A<DetectDocumentSegmentationRequest>

pub fn init(self) -> Retained<DetectDocumentSegmentationRequest>

impl A<GenAttentionBasedSaliencyImageRequest>

pub fn init(self) -> Retained<GenAttentionBasedSaliencyImageRequest>

impl A<GenObjectnessBasedSaliencyImageRequest>

pub fn init(self) -> Retained<GenObjectnessBasedSaliencyImageRequest>

impl A<GenImageFeaturePrintRequest>

pub fn init_with_ch( self, ch: Option<&mut RequestCh<GenImageFeaturePrintRequest>>, ) -> R<GenImageFeaturePrintRequest>

impl A<GenImageFeaturePrintRequest>

pub fn init(self) -> Retained<GenImageFeaturePrintRequest>

impl A<ClassifyImageRequest>

pub fn init(self) -> Retained<ClassifyImageRequest>

impl A<DetectFaceRectanglesRequest>

pub fn init(self) -> Retained<DetectFaceRectanglesRequest>

impl A<DetectFaceCaptureQualityRequest>

pub fn init(self) -> Retained<DetectFaceCaptureQualityRequest>

impl A<DetectBarcodesRequest>

pub fn init(self) -> Retained<DetectBarcodesRequest>

impl A<FrameRateCadence>

pub fn init_with_frame_rate(self, frame_rate: isize) -> R<FrameRateCadence>

impl A<TimeIntervalCadence>

pub fn init_with_time_interval( self, time_interval: TimeInterval, ) -> R<TimeIntervalCadence>

impl A<RequestProcessingOpts>

pub fn init(self) -> Retained<RequestProcessingOpts>

impl A<VideoProcessor>

pub fn init_with_url(self, url: &Url) -> R<VideoProcessor>

pub fn sel_init_with_url() -> &'static Sel

@selector(initWithURL:) but dynamic use this function to check if object responds to selector

impl A<WebView>

pub fn init(self) -> Retained<WebView>

impl A<WebView>

pub fn init_with_frame_cfg(self, frame: Rect, cfg: &WebViewCfg) -> R<WebView>

impl A<WebViewCfg>

pub fn init(self) -> Retained<WebViewCfg>

impl A<ProcessPool>

pub fn init(self) -> Retained<ProcessPool>

impl A<Preferences>

pub fn init(self) -> Retained<Preferences>

impl A<UserContentController>

pub fn init(self) -> Retained<UserContentController>

impl A<UserScript>

pub fn init_with_src_injection_time_for_main_frame_only( self, src: &String, injection_time: UserScriptInjectionTime, for_main_frame_only: bool, ) -> R<UserScript>