cidre::at

Module audio

Re-exports

pub use component::Component;

pub use component::Desc as ComponentDesc;

pub use component::Flags as ComponentFlags;

pub use component::InitializedState as ComponentInitializedState;

pub use component::Instance as ComponentInstance;

pub use component::InstanceRef as ComponentInstanceRef;

pub use component::InstantiationOpts as ComponentInstantiationOpts;

pub use component::UninitializedState as ComponentUnitializedState;

pub use unit::Element as UnitElement;

pub use unit::Manufacturer as UnitManufacturer;

pub use unit::Param as UnitParam;

pub use unit::ParamId as UnitParamId;

pub use unit::ParamValue as UnitParamValue;

pub use unit::SubType as UnitSubType;

pub use unit::Type as UnitType;

pub use unit::Unit;

pub use unit::UnitRef;

pub use crate::cat::audio::*;

Modules

asbd_prop

channel_layout_prop

codec_quality

Constants to be used with kAudioCodecPropertyQualitySetting

component

converter_err

file_err

id3_prop

queue_err

unit

Structs

BalanceFade

this struct is used with kAudioFormatProperty_BalanceFade

Codec

AudioCodec components translate audio data from one format to another. There are three kinds of AudioCodec components. Decoder components (‘adec’) translate data that isn’t in linear PCM into linear PCM formatted data. Encoder components (‘aenc’) translate linear PCM data into some other format. Unity codecs (‘acdc’) translate between different flavors of the same type (e.g. 16 bit signed integer linear PCM into 32 bit floating point linear PCM).

CodecGlobalPropId

These properties reflect the capabilities of the underlying codec. The values of these properties are independent of the codec’s internal state.

CodecInstancePropId

Properties which can be set or read on an instance of the underlying audio codec. These properties are dependent on the codec’s current state. A property may be read/write or read only, depending on the data format of the codec.

CodecMagicCookieInfo

Structure holding the magic cookie information.

CodecRef

Converter

ConverterOpts

ConverterPrimeInfo

When using AudioConverterFillComplexBuffer() (either a single call or a series of calls), some conversions, particularly involving sample-rate conversion, ideally require a certain number of input frames previous to the normal start input frame and beyond the end of the last expected input frame in order to yield high-quality results.

ConverterQuality

ConverterRef

Useful link on formats https://tech.ebu.ch/docs/techreview/trev_305-moser.pdf

ConverterSampleRateConverterComplexity

DitherAlgorithm

kAudioConverterPropertyDithering

ExtendedFormatInfo

This struct is used as a specifier for the kAudioFormatProperty_FormatList property

FileFlags

FileId

FilePropId

FileTypeId

Identifier for an audio file type.

FormatInfo

FormatPropId

PanningInfo

Queue

QueueBuf

QueueLevelMeterState

Specifies the current level metering information for one channel of an audio queue.

QueueParam

QueueParamEvent

QueueProcessingTapFlags

QueueProp

QueueTimePitchAlgorithm

QueueTimeline

Enums

BalanceFadeType

CodecBitRateControlMode

Constants defining various bit rate control modes to be used with kAudioCodecPropertyBitRateControlMode. These modes are only applicable to encoders that can produce variable packet sizes, such as AAC.

CodecProduceOutputPacketStatus

ConverterPrimeMethod

DynamicRangeCompressionProfile

Constants to be used with kAudioCodecPropertyAdjustCompressionProfile to control the DRC Effect Type based on ISO/IEC 23003-4

FilePermissions

PanningMode

AudioPanningMode

ProgramTargetLevel

Constants to be used with kAudioCodecPropertyProgramTargetLevelConstant to set the target loudness

Constants

DECODER_COMPONENT_TYPE

ENCODER_COMPONENT_TYPE

UNITY_CODEC_COMPONENT_TYPE

Type Aliases

QueueInputCb

QueueInputCbBlock

Defines a pointer to a block that is called when a recording audio queue has finished filling a buffer.

QueueOutputCb

QueueOutputCbBlock

Defines a pointer to a block that is called when a playback audio queue has finished taking data from a buffer.

QueueParamValue

QueuePropListenerProc