cidre::at::audio

Module asbd_prop

Constants

ASBD_FROM_ESDS

Returns an audio stream description for a given ESDS. The specifier is an ESDS. The value is a AudioStreamBasicDescription. If multiple formats can be described by the ESDS this property will return the base level format.

ASBD_FROM_MPEG_PACKET

Returns an audio stream description for a given MPEG Packet. The specifier is an MPEG Packet. The value is a AudioStreamBasicDescription.

AVAILABLE_DECODE_NUMBER_CHANNELS

The specifier is an AudioStreamBasicDescription with at least the format_id field set. The property value is an array of UInt32 indicating the number of channels that can be decoded.

AVAILABLE_ENCODE_BIT_RATES

The specifier is a u32 ‘audio::Format’. The property value is an array of AudioValueRange describing all available bit rates.

AVAILABLE_ENCODE_CHANNEL_LAYOUT_TAGS

The specifier is an AudioStreamBasicDescription with at least the format_id and channels_per_frame fields set. The property value is an array of AudioChannelLayoutTags for the format and number of channels specified. If mChannelsPerFrame is zero, then all layouts supported by the format are returned.

AVAILABLE_ENCODE_NUMBER_CHANNELS

The specifier is an AudioStreamBasicDescription with at least the format_id field set. The property value is an array of UInt32 indicating the number of channels that can be encoded. A value of 0xFFFFFFFF indicates that any number of channels may be encoded.

AVAILABLE_ENCODE_SAMPLE_RATES

The specifier is a u32 format ID. The property value is an array of AudioValueRange describing all available sample rates.

CHANNEL_LAYOUT_FROM_ESDS

Returns an audio channel layout for a given ESDS. The specifier is an ESDS. The value is a AudioChannelLayout.

DECODERS

The specifier is the format that you are interested in, e.g. ’aac ’ Returns an array of AudioClassDescriptions for all installed decoders for the given format

DECODE_FORMAT_IDS

No specifier needed. Must be set to NULL. Returns an array of UInt32 format IDs for formats that are valid input formats

ENCODERS

The specifier is the format that you are interested in, e.g. ’aac ’ Returns an array of AudioClassDescriptions for all installed encoders for the given format

ENCODE_FORMAT_IDS

No specifier needed. Must be set to NULL. Returns an array of u32 format IDs for formats that are valid output formats for a converter.

FIRST_PLAYABLE_FORMAT_FROM_LIST

The specifier is a list of 1 or more AudioFormatListItem. Generally it is the list of these items returned from kAudioFormatProperty_FormatList. The property value retrieved is an UInt32 that specifies an index into that list. The list that the caller provides is generally sorted with the first item as the best format (most number of channels, highest sample rate), and the returned index represents the first item in that list that can be played by the system. Thus, the property is typically used to determine the best playable format for a given (layered) audio stream

FORMAT_EMPLOYS_DEPENDENT_PACKETS

Returns whether or not a format is capable of combining independently decodable packets with dependent packets. The specifier is an AudioStreamBasicDescription describing the format to ask about. The value is a UInt32 where zero means that all packets in streams of the specified format are independently decodable and non-zero means that streams of the specified format may include dependent packets.

FORMAT_INFO

Retrieves general information about a format. The specifier is a magic cookie, or NULL. On input, the property value is an AudioStreamBasicDescription which should have at least the ‘format’ filled out. On output it will be filled out as much as possible given the information known about the format and the contents of the magic cookie (if any is given). If multiple formats can be described by the AudioStreamBasicDescription and the associated magic cookie, this property will return the base level format.

FORMAT_IS_ENCRYPTED

Returns whether or not a format is encrypted. The specifier is a u32 format ID. The value is a u32 where non-zero means the format is encrypted.

FORMAT_IS_EXTERNALLY_FRAMED

Returns whether or not a format requires external framing information, i.e. AudioStreamPacketDescriptions. The specifier is an AudioStreamBasicDescription describing the format to ask about. The value is a UInt32 where non-zero means the format is externally framed. Any format which has variable byte sized packets requires AudioStreamPacketDescriptions.

FORMAT_IS_VBR

Returns whether or not a format has a variable number of bytes per packet. The specifier is an AudioStreamBasicDescription describing the format to ask about. The value is a u32 where non-zero means the format is variable bytes per packet.

FORMAT_LIST

Returns a list of AudioFormatListItem structs describing the audio formats contained within the compressed bit stream as described by the magic cookie. The specifier is an AudioFormatInfo struct. The ‘format’ member of the

FORMAT_NAME

Returns a name for a given format. The specifier is an AudioStreamBasicDescription describing the format to ask about. The value is a cf::String. The caller is responsible for releasing the returned string. For some formats (eg, Linear PCM) you will get back a descriptive string (e.g. 16-bit, interleaved, etc…)

OUTPUT_FORMAT_LIST

Returns a list of AudioFormatListItem structs describing the audio formats which may be obtained by decoding the format described by the specifier.