cidre::at::audio

Module file_err

Constants

BAD_PROPERTY_SIZE

0x2173697A, 561211770 The size of the property data was not correct.

DOES_NOT_ALLOW64_BIT_DATA_SIZE

0x6F66663F, 1868981823 The file offset was too large for the file type. The AIFF and WAVE file format types have 32-bit file size limits.

END_OF_FILE

End of file.

FILE_NOT_FOUND

File not found.

INVALID_CHUNK

0x63686B3F, 1667787583 Either the chunk does not exist in the file or it is not supported by the file.

INVALID_FILE

0x6474613F, 1685348671 The file is malformed, or otherwise not a valid instance of an audio file of its type.

INVALID_PACKET_DEPENDENCY

0x6465703F, 1684369471 The file offset was too large for the file type. The AIFF and WAVE file format types have 32-bit file size limits.

INVALID_PACKET_OFFSET

0x70636B3F, 1885563711 The file offset was too large for the file type. The AIFF and WAVE file format types have 32-bit file size limits.

NOT_OPEN

The file is closed.

NOT_OPTIMIZED

0x6F70746D, 1869640813 The chunks following the audio data chunk are preventing the extension of the audio data chunk. To write more data, you must optimize the file.

OPERATION_NOT_SUPPORTED

0x6F703F3F The file is malformed, or otherwise not a valid instance of an audio file of its type.

PERMISSIONS

0x70726D3F, 1886547263 The operation violated the file permissions. For example, an attempt was made to write to a file opened with the kAudioFileReadPermission constant.

POS

Invalid file position.

UNSPECIFIED

0x7768743F, 2003334207 An unspecified error has occurred.

UNSUPPORTED_DATA_FORMAT

0x666D743F, 1718449215 The data format is not supported by this file type.

UNSUPPORTED_FILE_TYPE

0x7479703F, 1954115647 The file type is not supported.

UNSUPPORTED_PROPERTY

0x7074793F, 1886681407 The property is not supported.