cidre::at::audio

Struct Codec

pub struct Codec(/* private fields */);

AudioCodec components translate audio data from one format to another. There are three kinds of AudioCodec components. Decoder components (‘adec’) translate data that isn’t in linear PCM into linear PCM formatted data. Encoder components (‘aenc’) translate linear PCM data into some other format. Unity codecs (‘acdc’) translate between different flavors of the same type (e.g. 16 bit signed integer linear PCM into 32 bit floating point linear PCM).

AudioCodec components are standard components and are managed by the Component Manager.

Once an AudioCodec is found that implements the translation in question, it has to be set up to do the translation. This can be done by setting the appropriate properties or by calling AudioCodecInitialize. If the translation is specified by properties, AudioCodecInitialize still needs to be called prior to appending input data or producing output data. AudioCodecInitialize puts the codec into the “initialized” state. In this state, the format information for the translation cannot be changed. The codec has to be in the initialized state for AudioCodecAppendInputData and AudioCodecProduceOutputData to work. They will return kAudioCodecStateError if the codec isn’t initialized.

AudioCodecUninitialize will return the codec to the uninitialized state and release any allocated resources. The codec may then be configured freely. It is not necessary to call AudioCodecUninitialize prior to closing the codec.

Once in the initialized state, the codec is ready to receive input and produce output using the append_data and produce_data routines. Input data can be fed into an encoder and some decoders in any size (even byte by byte). Input data fed to a decoder should be in terms of whole packets in the encoded format if the format is variable bit rate and is not self framing (e.g. MPEG-4 AAC). Output data can only be produced in whole packet sizes. Both routines will return the amount of data they consume/produce.

AudioCodecProduceOutputData also returns a status code to the caller that indicates the result of the operation (success or failure) as well as the state of the input buffer.

The combination of append_data and produce_packets can be thought of a “push-pull” model of data handling. First, the input data is pushed into the component and the resulting output data gets pulled out of that same component.

Basic Workflow

Find the appropriate codec component

Open the codec component

Configure it (AudioCodecGetPropertyInfo, AudioCodecGetProperty, AudioCodecSetProperty)

AudioCodecInitialize

Loop a. append_data (EOF is signaled by passing a 0-sized buffer) b. produce_packets

Close the codec component

Implementations

impl Codec

pub unsafe fn init_codec( &mut self, input_format: *const StreamBasicDesc, output_format: *const StreamBasicDesc, magic_cookie: Option<&[u8]>, ) -> Result

pub fn prop_info(&self, prop_id: u32) -> Result<(u32, bool)>

pub unsafe fn prop_vec<T: Sized + Default + Clone>( &self, prop_id: u32, ) -> Result<Vec<T>>

pub fn cf_prop<T: Release>(&self, prop_id: u32) -> Result<R<T>>

pub fn prop<T: Sized>(&self, prop_id: u32) -> Result<T>

pub unsafe fn set_prop<T: Sized>(&mut self, property_id: u32, val: &T) -> Result

pub fn quality(&self) -> Result<u32>

pub fn settings(&self) -> Result<R<DictionaryOf<String, Type>>>

pub fn set_quality(&mut self, val: u32) -> Result

pub fn bit_rate_control_mode(&self) -> Result<BitRateControlMode>

pub fn set_bit_rate_control_mode(&mut self, val: BitRateControlMode) -> Result

pub fn current_target_bit_rate(&self) -> Result<u32>

pub fn set_current_target_bit_rate(&mut self, val: u32) -> Result

pub fn current_input_sample_rate(&self) -> Result<f64>

pub fn set_current_input_sample_rate(&mut self, val: f64) -> Result

A f64 containing the current input sample rate in Hz. No Default. May be writable. If only one sample rate is supported it does not have to be.

pub fn current_output_sample_rate(&self) -> Result<f64>

A f64 containing the current output sample rate in Hz. No Default. May be writable. If only one sample rate is supported it does not have to be.

pub fn set_current_output_sample_rate(&mut self, val: f64) -> Result

pub fn set_current_input_channel_layout<const N: usize>( &mut self, val: &ChannelLayout<N>, ) -> Result

pub fn current_input_channel_layout<const N: usize>( &self, ) -> Result<ChannelLayout<N>>

pub fn set_current_output_channel_layout<const N: usize>( &mut self, val: &ChannelLayout<N>, ) -> Result

pub fn current_output_channel_layout<const N: usize>( &self, ) -> Result<ChannelLayout<N>>

pub fn applicable_input_sample_rates(&self) -> Result<Vec<ValueRange>>

pub fn applicable_output_sample_rates(&self) -> Result<Vec<ValueRange>>

pub fn recommended_bit_rate_range(&self) -> Result<Vec<ValueRange>>

pub fn applicable_bit_rate_range(&self) -> Result<Vec<ValueRange>>

pub fn supported_input_formats(&self) -> Result<Vec<StreamBasicDesc>>

pub fn input_formats_for_output_format( &self, output_format: &StreamBasicDesc, ) -> Result<Vec<StreamBasicDesc>>

pub fn supported_output_formats(&self) -> Result<Vec<StreamBasicDesc>>

pub fn output_formats_for_input_format( &self, input_format: &StreamBasicDesc, ) -> Result<Vec<StreamBasicDesc>>

pub fn reset(&mut self) -> Result

Flushes all the data in the codec and clears the input buffer. Note that the formats, and magic cookie will be retained so they won’t need to be set up again to decode the same data.