cidre::at::audio

Struct CodecGlobalPropId

#[repr(transparent)]
pub struct CodecGlobalPropId(pub u32);

These properties reflect the capabilities of the underlying codec. The values of these properties are independent of the codec’s internal state.

These properties can be read at any time the codec is open.

Tuple Fields
0: u32
Implementations

impl GlobalPropId

pub const SUPPORTED_INPUT_FORMATS: Self

An array of audio::StreamBasicDescription structs describing what formats the codec supports for input data

pub const SUPPORTED_OUTPUT_FORMATS: Self

An array of audio::StreamBasicDescription structs describing what formats the codec supports for output data

pub const AVAILABLE_INPUT_SAMPLE_RATES: Self

An array of audio::ValueRange indicating the valid ranges for the output sample rate of the codec for the current bit rate. This property is only relevant to encoders. See also kAudioCodecPropertyAvailableOutputSampleRates. Not writable.

pub const AVAILABLE_OUTPUT_SAMPLE_RATES: Self

An array of audio::ValueRange indicating the valid ranges for the output sample rate of the codec. Required for encoders. (see also kAudioCodecPropertyApplicableOutputSampleRates)

pub const AVAILABLE_BIT_RATE_RANGE: Self

An array of audio::ValueRange that indicate the target bit rates supported by the encoder. This can be total bit rate or bit rate per channel as appropriate. This property is only relevant to encoders. (see also kAudioCodecPropertyApplicableBitRateRange)

pub const MINIMUM_NUMBER_INPUT_PACKETS: Self

A u32 indicating the minimum number of input packets that need to be supplied to the codec. The actual input the codec accepts could be less than this. For most codecs this value will be 1.

pub const MINIMUM_NUMBER_OUTPUT_PACKETS: Self

A u32 indicating the minimum number of output packets that need to be handled from the codec. The actual output might be less than this. For most codecs this value will be 1.

pub const AVAILABLE_NUMBER_CHANNELS: Self

An array of u32 that specifies the number of channels the codec is capable of encoding or decoding to. 0xFFFFFFFF means any number of channels.

pub const DOES_SAMPLE_RATE_CONVERSION: Self

A u32 indicating if the codec wants to do a sample rate conversion (if necessary) because it can do it in a way that is meaningful for quality. Value is 1 if true, 0 otherwise.

pub const AVAILABLE_INPUT_CHANNEL_LAYOUT_TAGS: Self

An array of audio::ChannelLayoutTag that specifies what channel layouts the codec is capable of using on input.

pub const AVAILABLE_OUTPUT_CHANNEL_LAYOUT_TAGS: Self

An array of audio::ChannelLayoutTag that specifies what channel layouts the codec is capable of using on output.

pub const INPUT_FORMATS_FOR_OUTPUT_FORMAT: Self

An array of AudioStreamBasicDescription indicating what the codec supports for input data given an output format that’s passed in as the first member of the array (and is overwritten on the reply). Always a subset of kAudioCodecPropertySupportedInputFormats

pub const OUTPUT_FORMATS_FOR_INPUT_FORMAT: Self

An array of AudioStreamBasicDescription indicating what the codec supports for output data given an input format that’s passed in as the first member of the array (and is overwritten on the reply). Always a subset of kAudioCodecPropertySupportedOutputFormats

pub const FORMAT_INFO: Self

Takes an audio::FormatInfo on input. This audio::FormatInfo is validated either through the provided magic cookie or the AudioStreamBasicDescription and where applicable, wildcards are overwritten with default values.