cidre::at::audio

Struct CodecRef

pub struct CodecRef<S>(/* private fields */)
where
    S: State<Codec>;

Implementations

impl CodecRef<UninitializedState>

pub fn new_apple_aac_encoder() -> Result<Self>

pub fn new_apple_opus_encoder() -> Result<Self>

pub fn initialize( self, input_format: *const StreamBasicDesc, output_format: *const StreamBasicDesc, magic_cookie: Option<&[u8]>, ) -> Result<CodecRef<InitializedState>>

pub fn set_quality(&mut self, val: u32) -> Result

pub fn set_bit_rate_control_mode(&mut self, val: BitRateControlMode) -> Result

pub fn set_current_target_bit_rate(&mut self, val: u32) -> Result

pub fn input_formats_for_output_format( &self, output_format: &StreamBasicDesc, ) -> Result<Vec<StreamBasicDesc>>

pub fn output_formats_for_input_format( &self, input_format: &StreamBasicDesc, ) -> Result<Vec<StreamBasicDesc>>

impl CodecRef<InitializedState>

pub fn uninitialize(self) -> Result<CodecRef<UninitializedState>>

This call will move the codec from the initialized state back to the uninitialized state. The codec will release any resources it allocated or claimed in AudioCodecInitialize.

pub fn append(&mut self, data: &[u8]) -> Result<u32>

pub unsafe fn append_input( &mut self, data_ptr: *const u8, data_len: &mut u32, ) -> Result

pub fn append_data( &mut self, data: &[u8], packets: &[StreamPacketDesc], ) -> Result<Consumed>

Append as much of the given data to the codec’s input buffer as possible and return in (data_len, packets_len) the amount of data and packets used.

pub fn produce_packet( &mut self, data: &mut [u8], ) -> Result<(u32, ProduceOutputPacketStatus)>

pub fn produce_packets( &mut self, data: &mut [u8], out_packet_descriptions: &mut [StreamPacketDesc], ) -> Result<Produced>

pub fn append_buf_list( &mut self, in_buf_list: &BufList, packet_descriptions: &mut [StreamPacketDesc], ) -> Result<Consumed>

pub fn produce_buf_list( &mut self, buf_list: &mut BufList, number_of_packets: &mut u32, ) -> Result<Status>

pub fn produce_buf_list_with_descs( &mut self, buf_list: &mut BufList, packet_descriptions: &mut [StreamPacketDesc], ) -> Result<(u32, Status)>

impl<S> CodecRef<S>
where S: State<Codec>,

pub fn supported_input_formats(&self) -> Result<Vec<StreamBasicDesc>>

pub fn supported_output_formats(&self) -> Result<Vec<StreamBasicDesc>>

pub fn bit_rate_control_mode(&self) -> Result<BitRateControlMode>

pub fn recommended_bit_rate_range(&self) -> Result<Vec<ValueRange>>

pub fn applicable_output_sample_rates(&self) -> Result<Vec<ValueRange>>

pub fn applicable_input_sample_rates(&self) -> Result<Vec<ValueRange>>

pub fn applicable_bit_rate_range(&self) -> Result<Vec<ValueRange>>

pub fn current_target_bit_rate(&self) -> Result<u32>

pub fn magic_cookie(&self) -> Result<Vec<u8>>

pub fn current_output_format(&self) -> Result<StreamBasicDesc>

pub fn current_input_format(&self) -> Result<StreamBasicDesc>

pub fn set_current_input_format(&mut self, val: &StreamBasicDesc) -> Result

pub fn max_packet_byte_size(&self) -> Result<u32>

pub fn input_buf_size(&self) -> Result<u32>

pub fn quality(&self) -> Result<u32>

pub fn settings(&self) -> Result<R<DictionaryOf<String, Type>>>