cidre::at::audio

Struct FilePropId

#[repr(transparent)]
pub struct FilePropId(pub u32);

Tuple Fields
0: u32
Implementations

impl PropId

pub const FILE_FORMAT: Self

The format of the audio data file. An FileTypeID that identifies the format of the file

pub const DATA_FORMAT: Self

An audio::StreamBasicDescription describing the format of the audio data

pub const IS_OPTIMIZED: Self

A u32 indicating whether an Audio File has been optimized. Optimized means it is ready to start having sound data written to it. A value of 0 indicates the file needs to be optimized. A value of 1 indicates the file is currently optimized.

pub const MAGIC_COOKIE_DATA: Self

A *const c_void pointing to memory set up by the caller. Some file types require that a magic cookie be provided before packets can be written to the file, so this property should be set before calling write_bytes()/write_packets() if a magic cookie exists.

pub const DATA_BYTE_COUNT: Self

A u64 that indicates the number of bytes of audio data contained in the file

pub const DATA_PACKET_COUNT: Self

A u64 that indicates the number of packets of audio data contained in the file

pub const MAXIMUM_PACKET_SIZE: Self

A u32 that indicates the maximum size of a packet for the data contained in the file

pub const DATA_OFFSET: Self

a s64 that indicates the byte offset in the file of the audio data.

pub const CHANNEL_LAYOUT: Self

An audio::ChannelLayout struct.

pub const DEFER_SIZE_UPDATES: Self

A u32. If 1, then updating the files sizes in the header is not done for every write, but deferred until the file is read, optimized or closed. This is more efficient, but less safe since, if the application crashes before the size is updated, the file may not be readable. The default value is one, it doesn’t update the header.

pub const DATA_FORMAT_NAME: Self

pub const MARKER_LIST: Self

pub const REGION_LIST: Self

pub const PACKET_TO_FRAME: Self

pub const FRAME_TO_PACKET: Self

pub const RESTRICTS_RANDOM_ACCESS: Self

pub const PACKET_TO_ROLL_DISTANCE: Self

pub const PREVIOUS_INDEPENDENT_PACKET: Self

pub const NEXT_INDEPENDENT_PACKET: Self

Pass an AudioIndependentPacketTranslation with mPacket filled out and get mIndependentlyDecodablePacket back. A value of -1 means that no independent packet is present in the stream in the direction of interest. Otherwise, for kAudioFilePropertyPreviousIndependentPacket, mIndependentlyDecodablePacket will be less than mPacket, and for kAudioFilePropertyNextIndependentPacket, mIndependentlyDecodablePacket will be greater than mPacket.

pub const PACKET_TO_DEPENDENCY_INFO: Self

Pass an AudioPacketDependencyInfoTranslation with mPacket filled out and get mIsIndependentlyDecodable and mPrerollPacketCount back. A value of 0 for mIsIndependentlyDecodable indicates that the specified packet is not independently decodable. A value of 1 for mIsIndependentlyDecodable indicates that the specified packet is independently decodable. For independently decodable packets, mPrerollPacketCount indicates the count of packets that must be decoded after the packet with the specified number in order to refresh the decoder. If the value of kAudioFilePropertyRestrictsRandomAccess is 1, either kAudioFilePropertyPacketToRollDistance or kAudioFilePropertyPacketToDependencyInfo must be used in order to identify an appropriate random access point.

pub const PACKET_TO_BYTE: Self

pass an AudioBytePacketTranslation struct with mPacket filled out and get mByte back. mByteOffsetInPacket is ignored. If the mByte value is an estimate then kBytePacketTranslationFlag_IsEstimate will be set in the mFlags field.

pub const BYTE_TO_PACKET: Self

pass an AudioBytePacketTranslation struct with mByte filled out and get mPacket and mByteOffsetInPacket back. If the mPacket value is an estimate then kBytePacketTranslationFlag_IsEstimate will be set in the mFlags field

pub const CHUNK_IDS: Self

returns an array of os::Type four char codes for each kind of chunk in the file.

pub const INFO_DICTIONARY: Self

returns a CFDictionary filled with information about the data contained in the file. See dictionary key constants already defined for info string types. AudioFileComponents are free to add keys to the dictionaries that they return for this property… caller is responsible for releasing the CFObject

pub const PACKET_TABLE_INFO: Self

Gets or sets an AudioFilePacketTableInfo struct for the file types that support it. When setting, the sum of mNumberValidFrames, mPrimingFrames and mRemainderFrames must be the same as the total number of frames in all packets. If not you will get a kAudio_ParamError. The best way to ensure this is to get the value of the property and make sure the sum of the three values you set has the same sum as the three values you got.

pub const FORMAT_LIST: Self

In order to support formats such as AAC SBR where an encoded data stream can be decoded to multiple destination formats, this property returns an array of AudioFormatListItems (see AudioFormat.h) of those formats. The default behavior is to return the an AudioFormatListItem that has the same AudioStreamBasicDescription that kAudioFilePropertyDataFormat returns.

pub const PACKET_SIZE_UPPER_BOUND: Self

pub const PACKET_RANGE_BYTE_COUNT_UPPER_BOUND: Self

Pass an AudioPacketRangeByteCountTranslation with mPacket and mPacketCount filled out and get mByteCountUpperBound back. The value of mByteCountUpperBound can be used to allocate a buffer for use with AudioFileReadPacketData in order to accommodate the entire packet range. May require scanning in order to obtain the requested information, but even if so, no scanning will occur beyond the last packet in the specified range. For file formats in which packets are directly accessible and stored both contiguously and byte-aligned, the returned upper bound will be equal to the total size of the packets in the range. Otherwise the upper bound may reflect per-packet storage overhead.

pub const RESERVE_DURATION: Self

The value is a f64 of the duration in seconds of data that is expected to be written. Setting this property before any data has been written reserves space in the file header for a packet table and/or other information so that it can appear before the audio data. Otherwise the packet table may get written at the end of the file, preventing the file from being streamable.

pub const ESTIMATED_DURATION: Self

The value is a f64 representing an estimated duration in seconds. If duration can be calculated without scanning the entire file, or all the audio data packets have been scanned, the value will accurately reflect the duration of the audio data.

pub const BIT_RATE: Self

Returns the bit rate for the audio data as a u32. For some formats this will be approximate.

pub const ID3_TAG: Self

pub const ID3_TAG_OFFSET: Self

pub const SOURCE_BIT_DEPTH: Self

pub const ALBUM_ARTWORK: Self

pub const AUDIO_TRACK_COUNT: Self

A u32 that indicates the number of audio tracks contained in the file. (get property only)

pub const USE_AUDIO_TRACK: Self

A u32 that indicates the number of audio tracks contained in the file. (set property only)