cidre::at::audio

Struct QueueProp

#[repr(transparent)]
pub struct QueueProp(pub u32);

Tuple Fields
0: u32
Implementations

impl QueueProp

pub const IS_RUNNING: Self

A read-only property whose value is a u32 that indicates whether or not the queue is running. A notification is sent when the audio device starts or stops, which is not necessarily when the start or stop function is called.

pub const SAMPLE_RATE: Self

A read-only property whose value is a f64 that indicates the sampling rate of the associated audio device.

pub const NUMBER_CHANNELS: Self

A read-only property whose value is a u32 that indicates the number of channels in the associated audio device.

pub const CURRENT_DEVICE: Self

A read/write property whose value is a cf::String that contains the unique identifier (UID) of the associated audio device. If the audio queue is tracking the default system device and the device changes, it will generate a property changed notification for this property. You can then query the HAL for info on the new default system device.

pub const MAGIC_COOKIE: Self

A read/write property whose value is an audio format magic cookie. If the audio format requires a magic cookie, you must set this property before enqueuing any buffers.

pub const MAX_OUTPUT_PACKET_SIZE: Self

A read-only u32 that indicates the size in bytes of the largest single packet of data in the output format. This is mostly useful for recording/encoding variable bit rate compressed data.

pub const STREAM_DESCRIPTION: Self

A read-only AudioStreamBasicDescription that indicates the queue’s recording format. This is useful when recording, where you may specify a sample rate of 0 during construction.

pub const CHANNEL_LAYOUT: Self

A read/write property whose value is an audio channel layout structure that describes the audio queue’s channel layout. The number of channels must match the format of the queue. If more than two channels (for instance, a five-channel surround sound) exist in the queue, there may be a need to specify a channel layout. This layout indicates the specific order in which the channels do appear, such as left, then center, then right.

pub const ENABLE_LEVEL_METERING: Self

A read-write property whose value is a UInt32 that indicates whether metering of audio levels is enabled for the audio queue. (0=off, 1=on).

pub const CURRENT_LEVEL_METER: Self

A read-only property whose value is an array of AudioQueueLevelMeterState structures, one array element per audio channel. The values in the AudioQueueLevelMetersState are in the range 0-1.

pub const CURRENT_LEVEL_METER_DB: Self

A read-only property whose value is an array of AudioQueueLevelMeterState structures, one array element per audio channel. The values in the AudioQueueLevelMeterState are in decibels.

pub const DECODE_BUFFER_SIZE_FRAMES: Self

A read/write property whose value is a UInt32 that is the size of the buffer into which an output audio queue decodes buffers. A large buffer provides more reliability and better long-term performance at the expense of memory and decreased responsiveness in some situations.

pub const CONVERTER_ERROR: Self

A read-only property whose value is a UInt32 indicating the most recent error (if any) encountered by the queue’s internal encoding/decoding process.

pub const ENABLE_TIME_PITCH: Self

A read/write property whose value is a UInt32 describing whether there is a time/pitch unit inserted into the queue’s audio signal chain. This property may only be set while the queue is stopped.

pub const TIME_PITCH_ALGORITHM: Self

A read/write property whose value is a UInt32 describing the time/pitch algorithm in use. This property is only valid while a time/pitch has been inserted, and may only be changed when the queue is not running.

pub const TIME_PITCH_BYPASS: Self

A read/write property whose value is a UInt32 describing whether the time/pitch unit has been bypassed (1=bypassed, 0=not bypassed).

impl QueueProp

pub const HARDWARE_CODEC_POLICY: Self

pub const CHANNEL_ASSIGNMENTS: Self

A write-only property whose value is an array of AudioQueueChannelAssignment. There must be one array element for each channel of the queue’s format as specified in the AudioStreamBasicDescription passed to AudioQueueNewOutput or AudioQueueNewInput. (New in iOS 6.0).