cidre::at::audio

Module unit

Modules

component_err

err

Audio unit errors

preset_key

Keys contains in an audio unit preset (ClassInfo) dictionary These strings are used as keys in the AUPreset-“classInfo” dictionary

sample_rate_converter_complexity

voice_io_other_audio_ducking_level

Ducking level applied to other (i.e. non-voice) audio by AUVoiceIO.

Structs

AudioUnit

AudioUnitBus

AudioUnitBusArray

ChannelInfo

Define an audio unit’s channel handling capabilities

Connection

This structure contains the information needed to make a connection between a source and destination audio unit.

Element

Type used for audio unit elements.

ExternalBuf

Allow a host to tell an audio unit to use the provided memory for its input callback

FormatConverter

It is better to use AudioConverterRef directly

FrequencyResponseBin

Structure used to get the magnitude of the frequency response at a particular frequency via kAudioUnitProperty_FrequencyResponse.

Manufacturer

MeterClipping

MultiChannelMixer

Output

OutputStartAtTimeParams

Param

ParamEvent

ParamFlags

ParamId

Type used for audio unit parameters.

ParamInfo

Preset

Prop

PropId

Type used for audio unit properties.

RenderActionFlags

RenderCbStruct

Used by a host when registering a callback with the audio unit to provide input

ScheduledFileRegion

ScheduledSlice

ScheduledSliceFlags

Scope

Type used for audio unit scopes.

SubType

Apple input/output audio unit sub types

Type

Different types of audio units

Unit

Useful links https://github.com/apple/AudioUnitSDK/tree/main/ https://developer.apple.com/library/archive/documentation/MusicAudio/Conceptual/AudioUnitHostingGuide_iOS/Introduction/Introduction.html#//apple_ref/doc/uid/TP40009492-CH1-SW1

UnitRef

VoiceIoOtherAudioDuckingCfg

The configuration of ducking other (i.e. non-voice) audio

Enums

AudioUnitBusType

NBandEQFilterType

NetStatus

OfflinePreflight

ParamEventType

The type of a parameter event (see AudioUnitScheduleParameter)

ParamEventValue

ParamUnit

SoundIsolationSoundType

Constants available as values for parameter kAUSoundIsolation_SoundToIsolate.

SpatializationAlgorithm

VoiceIoSpeechActivityEvent

Type Aliases

AudioUnitStatus

InputSamplesInOutputCb

ParamValue

Type used for audio unit parameter values. The value of a given parameter is specified using this type

PropListenerProc

RenderCb

ScheduledFileRegionCompProc

ScheduledSliceCompProc

VoiceIoMutedSpeechActivityEventListener