cidre::at::audio::unit

Enum ParamUnit

#[repr(u32)]
pub enum ParamUnit {

    Generic = 0,
    Indexed = 1,
    Boolean = 2,
    Percent = 3,
    Seconds = 4,
    SampleFrames = 5,
    Phase = 6,
    Rate = 7,
    Hertz = 8,
    Cents = 9,
    RelativeSemiTones = 10,
    MidiNoteNumber = 11,
    MidiController = 12,
    Decibels = 13,
    LinearGain = 14,
    Degrees = 15,
    EqualPowerCrossfade = 16,
    MixerFaderCurve1 = 17,
    Pan = 18,
    Meters = 19,
    AbsoluteCents = 20,
    Octaves = 21,
    Bpm = 22,
    Beats = 23,
    Milliseconds = 24,
    Ratio = 25,
    CustomUnit = 26,
    Midi2Controller = 27,

}

Variants

Generic = 0

untyped value generally between 0.0 and 1.0

Indexed = 1

takes an integer value (good for menu selections)

Boolean = 2

0.0 means false, non-zero means true

Percent = 3

usually from 0 -> 100, sometimes -50 -> +50

Seconds = 4

absolute or relative time

SampleFrames = 5

one sample frame equals (1.0/sampleRate) seconds

Phase = 6

-180 to 180 degrees

Rate = 7

rate multiplier, for playback speed, etc. (e.g. 2.0 == twice as fast)

Hertz = 8

absolute frequency/pitch in cycles/second

Cents = 9

unit of relative pitch

RelativeSemiTones = 10

useful for coarse detuning

MidiNoteNumber = 11

absolute pitch as defined in the MIDI spec (exact freq may depend on tuning table)

MidiController = 12

a generic MIDI controller value from 0 -> 127

Decibels = 13

logarithmic relative gain

LinearGain = 14

linear relative gain

Degrees = 15

-180 to 180 degrees, similar to phase but more general (good for 3D coord system)

EqualPowerCrossfade = 16

0 -> 100, crossfade mix two sources according to sqrt(x) and sqrt(1.0 - x)

MixerFaderCurve1 = 17

0.0 -> 1.0, pow(x, 3.0) -> linear gain to simulate a reasonable mixer channel fader response

Pan = 18

standard left to right mixer pan

Meters = 19

distance measured in meters

AbsoluteCents = 20

absolute frequency measurement : if f is freq in hertz then absoluteCents = 1200 * log2(f / 440) + 6900

Octaves = 21

octaves in relative pitch where a value of 1 is equal to 1200 cents

Bpm = 22

beats per minute, ie tempo

Beats = 23

time relative to tempo, i.e., 1.0 at 120 BPM would equal 1/2 a second

Milliseconds = 24

parameter is expressed in milliseconds

Ratio = 25

for compression, expansion ratio, etc.

CustomUnit = 26

this is the parameter unit type for parameters that present a custom unit name

Midi2Controller = 27

a generic MIDI 2.0 controller value with 32-bit range