cidre::at::audio::unit

Module err

Audio unit errors

These are the various errors that can be returned by AudioUnit API calls

Constants

CANNOT_DO_IN_CURRENT_CONTEXT

Returned when an audio unit is in a state where it can’t perform the requested action now - but it could later. Its usually used to guard a render operation when a reconfiguration of its internal state is being performed.

COMPONENT_MANAGER_NOT_SUPPORTED

The operation can not be performed for a component instance instantiated using the deprecated Component Manager. A host application should use the API functions AudioComponentInstantiate or AudioComponentInstanceNew when rebuilding against the macOS 11 or later SDK.

EXTENSION_NOT_FOUND

The specified identifier did not match any Audio Unit Extensions.

FAILED_INITIALIZATION

The audio unit is unable to be initialized

FILE_NOT_SPECIFIED

If an audio unit uses external files as a data source, this error is returned if a file hasn’t been set on it (Apple’s DLS synth returns this error)

FORMAT_NOT_SUPPORTED

Returned if an input or output format is not supported

INITIALIZED

Indicates the operation cannot be performed because the audio unit is initialized.

INVALID_ELEMENT

The specified element is not valid

INVALID_FILE

If an audio unit uses external files as a data source, this error is returned if a file is invalid (Apple’s DLS synth returns this error)

INVALID_FILE_PATH

The file path that was passed is not supported. It is either too long or contains invalid characters.

INVALID_OFFLINE_RENDER

Used to indicate that the offline render operation is invalid. For instance, when the audio unit needs to be pre-flighted, but it hasn’t been.

INVALID_PARAM

The parameter is not supported

INVALID_PARAMETER_VALUE

The parameter value is not supported, e.g. the value specified is NaN or infinite.

INVALID_PROPERTY

The property is not supported

INVALID_PROPERTY_VALUE

The property is valid, but the value of the property being provided is not

INVALID_SCOPE

The specified scope is invalid

MIDI_OUTPUT_BUFFER_FULL

Returned during the render call, if the audio unit produces more MIDI output, than the default allocated buffer. The audio unit can provide a size hint, in case it needs a larger buffer. See the documentation for AUAudioUnit’s MIDIOutputBufferSizeHint property.

MISSING_KEY

A required key is missing from a dictionary object.

MULTIPLE_VOICE_PROCESSORS

On some platforms, this error is returned when a client attempts to initialize a voice processor instance while another is initialized

NO_CONNECTION

There is no connection (generally an audio unit is asked to render but it has not input from which to gather data)

PROPERTY_NOT_IN_USE

Returned when a property is valid, but it hasn’t been set to a valid value at this time.

PROPERTY_NOT_WRITABLE

The property cannot be written

RENDER_TIMEOUT

The audio unit did not satisfy the render request in time.

TOO_MANY_FRAMES_TO_PROCESS

When an audio unit is initialized it has a value which specifies the max number of frames it will be asked to render at any given time. If an audio unit is asked to render more than this, this error is returned.

UNAUTHORIZED

Returned by either Open or Initialize, this error is used to indicate that the audio unit is not authorised, that it cannot be used. A host can then present a UI to notify the user the audio unit is not able to be used in its current state.

UNINITIALIZED

Returned if an operation requires an audio unit to be initialized and it is not.

UNKNOWN_FILE_TYPE

If an audio unit uses external files as a data source, this error is returned if a file is invalid (Apple’s DLS synth returns this error)