cidre::at::audio::unit

Module preset_key

Keys contains in an audio unit preset (ClassInfo) dictionary These strings are used as keys in the AUPreset-“classInfo” dictionary

Functions

cpu_load

data

element_name

file_refs

manufacturer

mas_data

name

part

This key if present, distinguishes a global preset that is set on the global scope with a part-based preset that is set on the part scope. The value of this key is audio unit defined

preset_number

render_quality

subtype

type_

version

vst_data

vst_preset