cidre::at::audio::unit

Struct Output

pub struct Output<S>(/* private fields */)
where
    S: State<Unit>;

Implementations

impl<S> Output<S>
where S: State<Unit>,

pub fn unit(&self) -> &Unit

pub fn unit_mut(&mut self) -> &mut Unit

pub fn last_render_err(&self) -> Result<Status>

pub fn set_input_cb<const N: usize, T>( &mut self, cb: RenderCb<N, T>, ref_con: *const T, ) -> Result

pub fn set_output_cb<const N: usize, T>( &mut self, cb: RenderCb<N, T>, ref_con: *const T, ) -> Result

pub fn set_render_cb<const N: usize, T>( &mut self, scope: Scope, bus: u32, cb: RenderCb<N, T>, ref_con: *const T, ) -> Result

pub fn output_stream_format(&self, bus: u32) -> Result<StreamBasicDesc>

pub fn input_stream_format(&self, bus: u32) -> Result<StreamBasicDesc>

pub fn is_running(&self) -> Result<bool>

pub fn is_io_enabled(&self, scope: Scope, bus: u32) -> Result<bool>

pub fn set_io_enabled(&mut self, scope: Scope, bus: u32, val: bool) -> Result

pub fn has_output_io(&self) -> Result<bool>

pub fn has_input_io(&self) -> Result<bool>

pub fn set_should_allocate_input_buf(&mut self, arg: bool) -> Result

pub fn set_should_allocate_output_buf(&mut self, arg: bool) -> Result

pub fn vp_mute_output(&self) -> Result<bool>

pub fn vp_set_mute_output(&mut self, arg: bool) -> Result

pub fn vp_set_muted_speech_activity_event_listener_block( &mut self, cb: Option<&mut VoiceIoMutedSpeechActivityEventListener>, ) -> Result

pub fn vp_set_muted_speech_activity_event_listener( &mut self, cb: impl FnMut(VoiceIoSpeechActivityEvent) + Send + 'static, ) -> Result

pub fn vp_enable_agc(&self) -> Result<bool>

pub fn vp_set_enable_agc(&mut self, arg: bool) -> Result

pub fn vp_bypass_voice_processing(&self) -> Result<bool>

pub fn vp_set_bypass_voice_processing(&mut self, arg: bool) -> Result

pub fn vp_other_audio_ducking_cfg( &mut self, ) -> Result<VoiceIoOtherAudioDuckingCfg>

pub fn vp_set_other_audio_ducking_cfg( &mut self, cfg: &VoiceIoOtherAudioDuckingCfg, ) -> Result

impl Output<UninitializedState>

pub fn new_apple() -> Result<Self>

pub fn new_apple_vp() -> Result<Self>

Apple voice processing unit

Note: It is reported that the echo source need to be specified as the output device. If no output device is specified, MacOS would take the default output device as echo source. See mumble-voip

pub fn allocate_resources(self) -> Result<Output<InitializedState>>

pub fn start_ts_at_zero(&self) -> Result<bool>

pub fn set_start_ts_at_zero(&mut self, val: bool) -> Result

pub fn set_output_stream_format(&mut self, val: &StreamBasicDesc) -> Result

pub fn set_input_stream_format(&mut self, val: &StreamBasicDesc) -> Result

pub fn output_device(&self) -> Result<Device>

pub fn set_output_device(&mut self, val: &Device) -> Result

Set device after enable IO

pub fn input_device(&self) -> Result<Device>

pub fn set_input_device(&mut self, val: &Device) -> Result

Set device after enable IO

impl Output<InitializedState>

pub fn input_device(&self) -> Result<Device>

pub fn deallocate_resources(self) -> Result<Output<UninitializedState>>

pub fn render<const N: usize>( &mut self, n_frames: u32, buf_list: &mut BufList<N>, bus: u32, ) -> Result

pub fn start(&mut self) -> Result

pub fn stop(&mut self) -> Result