cidre::at::audio::unit

Struct ParamId

#[repr(transparent)]
pub struct ParamId(pub u32);

Type used for audio unit parameters.

Parameters are typically used to control and set render state (for instance, filter cut-off frequency)

Tuple Fields
0: u32
Implementations

impl ParamId

The following specifies the equivalent parameterID’s for the Group scope for standard MIDI Controllers. This list is not exhaustive. It represents the parameters, and their corresponding MIDI messages, that should be supported in Group scope by MIDI capable AUs.

Group scope parameter IDs from 0 < 512 are reserved for mapping MIDI controllers.

pub const VOLUME: Self

value 0 < 128

pub const SUSTAIN: Self

value 0-63 (off), 64-127 (on)

pub const SOSTENUTO: Self

value 0-63 (off), 64-127 (on)

pub const ALL_NOTES_OFF: Self

value ignored

pub const MOD_WHEEL: Self

value 0 < 128

pub const PITCH_BEND: Self

value -8192 - 8191

pub const ALL_SOUND_OFF: Self

value ignored

pub const RESET_ALL_CONTROLLERS: Self

value ignored

pub const PAN: Self

value 0 < 128

pub const FOOT: Self

value 0 < 128

pub const CHANNEL_PRESSURE: Self

value 0 < 128

pub const KEY_PRESSURE: Self

value 0 < 128

pub const EXPRESSION: Self

value 0 < 128

pub const DATA_ENTRY: Self

value 0 < 128

pub const VOLUME_LSB: Self

value 0 < 128

pub const MOD_WHEEL_LSB: Self

value 0 < 128

pub const PAN_LSB: Self

value 0 < 128

pub const FOOT_LSB: Self

value 0 < 128

pub const EXPRESSION_LSB: Self

value 0 < 128

pub const DATA_ENTRY_LSB: Self

value 0 < 128

pub const KEY_PRESSURE_FIRST_KEY: Self

value 0 < 128

pub const KEY_PRESSURE_LAST_KEY: Self

value 0 < 128

impl ParamId

Parameters for all Panner AudioUnits

pub const GAIN: Self

Global, Linear, 0->1, 1

pub const AZIMUTH: Self

Global, Degrees, -180->180, 0

pub const ELEVATION: Self

Global, Degrees, -90->90, 0

pub const DISTANCE: Self

Global, Linear, 0->1, 1

pub const COORD_SCALE: Self

Global, Meters, 0.01->1000, 1

pub const REF_DISTANCE: Self

Global, Meters, 0.01->1000, 1

impl ParamId

Parameters for the AUSpatialMixer unit

pub const SPATIAL_MIXER_AZIMUTH: Self

Input, Degrees, -180->180, 0

pub const SPATIAL_MIXER_ELEVATION: Self

Input, Degrees, -90->90, 0

pub const SPATIAL_MIXER_DISTANCE: Self

Input, Metres, 0->10000, 0

pub const SPATIAL_MIXER_GAIN: Self

Input/Output, dB, -120->20, 0

pub const SPATIAL_MIXER_PLAYBACK_RATE: Self

Input, rate scaler 0.5 -> 2.0, 1.0

pub const SPATIAL_MIXER_ENABLE: Self

bus enable : 0 or 1, 1

pub const SPATIAL_MIXER_MIN_GAIN: Self

Minimum input gain constraint : 0.0 -> 10.0, 0.0

pub const SPATIAL_MIXER_MAX_GAIN: Self

Maximum input gain constraint : 0.0 -> 10.0, 10.0

pub const SPATIAL_MIXER_REVERB_BLEND: Self

Input, Dry/Wet equal-power blend, % 0.0 -> 100.0, 30.0

pub const SPATIAL_MIXER_GLOBAL_REVERB_GAIN: Self

Global, dB, -40.0 -> +40.0, 0.0

pub const SPATIAL_MIXER_OCCLUSION_ATTENUATION: Self

Input, Lowpass filter attenuation at 5KHz : decibels -100.0dB -> 0.0dB, 0.0dB smaller values make both direct and reverb sound more muffled; a value of 0.0 indicates no filtering Occlusion is a filter applied to the sound prior to the reverb send

pub const SPATIAL_MIXER_OBSTRUCTION_ATTENUATION: Self

Input, Lowpass filter attenuation at 5KHz : decibels -100.0dB -> 0.0dB, 0.0dB smaller values make direct sound more muffled; a value of 0.0 indicates no filtering Obstruction is a filter applied to the “direct” part of the sound (so is post reverb send)

pub const SPATIAL_MIXER_HEAD_YAW: Self

Global, Degrees, -180->180, 0

pub const SPATIAL_MIXER_HEAD_PITCH: Self

Global, Degrees, -90->90, 0

pub const SPATIAL_MIXER_HEAD_ROLL: Self

Global, Degrees, -180->180, 0

impl ParamId

Parameters for the AUMultiChannelMixer unit these are available for both desktop and iphone

pub const MULTI_CHANNEL_MIXER_VOLUME: Self

Global, Linear Gain, 0->1, 1. (the volume value can actually be any finite number, including negative.)

pub const MULTI_CHANNEL_MIXER_ENABLE: Self

Global, Boolean, 0->1, 1

pub const MULTI_CHANNEL_MIXER_PAN: Self

Global, Pan, -1->1, 0

-1 - 0 - 1, only valid when output is not mono setting kAudioUnitProperty_MatrixLevels overrides any previously set kMultiChannelMixerParam_Pan and vice versa

pub const MULTI_CHANNEL_MIXER_PRE_AVERAGE_POWER: Self

pub const MULTI_CHANNEL_MIXER_PRE_PEAK_HOLD_LEVEL: Self

pub const MULTI_CHANNEL_MIXER_POST_AVERAGE_POWER: Self

pub const MULTI_CHANNEL_MIXER_POST_PEAK_HOLD_LEVEL: Self

impl ParamId

Parameters for the AUMatrixMixer unit

pub const MATRIX_MIXER_VOLUME: Self

pub const MATRIX_MIXER_ENABLE: Self

pub const MATRIX_MIXER_PRE_AVERAGE_POWER: Self

pub const MATRIX_MIXER_PRE_PEAK_HOLD_LEVEL: Self

pub const MATRIX_MIXER_POST_AVERAGE_POWER: Self

pub const MATRIX_MIXER_POST_PEAK_HOLD_LEVEL: Self

pub const MATRIX_MIXER_PRE_AVERAGE_POWER_LINEAR: Self

pub const MATRIX_MIXER_PRE_PEAK_HOLD_LEVEL_LINEAR: Self

pub const MATRIX_MIXER_POST_AVERAGE_POWER_LINEAR: Self

pub const MATRIX_MIXER_POST_PEAK_HOLD_LEVEL_LINEAR: Self

impl ParamId

Output Units Parameters for the AudioDeviceOutput, DefaultOutputUnit, and SystemOutputUnit units

pub const HAL_OUTPUT_VOLUME: Self

impl ParamId

Parameters for the AUTimePitch, AUTimePitch (offline), AUPitch units

pub const TIME_PITCH_RATE: Self

pub const TIME_PITCH_PITCH: Self

pub const TIME_PITCH_EFFECT_BLEND: Self

only for the AUPitch unit

impl ParamId

Parameters for AUNewTimePitch

pub const NEW_TIME_PITCH_RATE: Self

rate control. Global, rate, 1/32 -> 32.0, 1.0

pub const NEW_TIME_PITCH_PITCH: Self

pitch shift in cents. Global, Cents, -2400 -> 2400, 1.0

pub const NEW_TIME_PITCH_SMOOTHNESS: Self

The generated output can be made to sound smoother by increasing the density of the processing time frames. The value is directly proportional to the CPU cost. When slowing down percussive audio, lower values may be better. Global, generic, 3.0 -> 32.0, 8.0

pub const NEW_TIME_PITCH_ENABLE_SPECTRAL_COHERENCE: Self

Spectral phase coherence is enabled through peak locking. This adds some computation cost but results in a less “phasey” or reverberant sound. Global, Boolean, 0->1, 1

pub const NEW_TIME_PITCH_ENABLE_TRANSIENT_PRESERVATION: Self

Transient preservation uses group delay to identify transients. It resets the phase at points of transients to preserve the original spectral phase relationships. It also sets the stretch factor to 1 at those points. Global, Boolean, 0->1, 1

impl ParamId

Parameters for the AUSampler unit

pub const AU_SAMPLER_GAIN: Self

Global, dB, -90->12, 0

pub const AU_SAMPLER_COARSE_TUNING: Self

Global, Semitones, -24->24, 0

pub const AU_SAMPLER_FINE_TUNING: Self

Global, Cents, -99->99, 0

pub const AU_SAMPLER_PAN: Self

Global, -1.0->1.0, 0

impl ParamId

Parameters for the AUBandpass unit

pub const BANDPASS_CENTER_FREQUENCY: Self

Global, Hz, 20->(SampleRate/2), 5000

pub const BANDPASS_BANDWIDTH: Self

Global, Cents, 100->12000, 600

impl ParamId

Parameters for the AUHipass unit

pub const HI_PASS_CUTOFF_FREQUENCY: Self

Global, Hz, 10->(SampleRate/2), 6900

pub const HI_PASS_RESONANCE: Self

Global, dB, -20->40, 0

impl ParamId

Parameters for the AULowpass unit

pub const LOW_PASS_CUTOFF_FREQUENCY: Self

pub const LOW_PASS_RESONANCE: Self

impl ParamId

Parameters for the AUHighShelfFilter unit

pub const HIGH_SHELF_CUT_OFF_FREQUENCY: Self

Global, Hz, 10000->(SampleRate/2), 10000

pub const HIGH_SHELF_GAIN: Self

Global, dB, -40->40, 0

impl ParamId

Parameters for the AULowShelfFilter unit

pub const LOW_SHELF_CUTOFF_FREQUENCY: Self

pub const LOW_SHELF_GAIN: Self

impl ParamId

Parameters for the AUParametricEQ unit

pub const PARAMETRIC_EQ_CENTER_FREQ: Self

Global, Hz, 20->(SampleRate/2), 2000

pub const PARAMETRIC_EQ_Q: Self

Global, Hz, 0.1->20, 1.0

pub const PARAMETRIC_EQ_GAIN: Self

Global, dB, -20->20, 0

impl ParamId

Parameters for the AUPeakLimiter unit

pub const LIMITER_ATTACK_TIME: Self

Global, Secs, 0.001->0.03, 0.012

pub const LIMITER_DECAY_TIME: Self

Global, Secs, 0.001->0.06, 0.024

pub const LIMITER_PRE_GAIN: Self

Global, dB, -40->40, 0

impl ParamId

Parameters for the AUDynamicsProcessor unit Note that the dynamics processor does not have fixed compression ratios. Instead, kDynamicsProcessorParam_HeadRoom adjusts the amount of compression. Lower kDynamicsProcessorParam_HeadRoom values results in higher compression. The compression ratio is automatically adjusted to not exceed kDynamicsProcessorParam_Threshold + kDynamicsProcessorParam_HeadRoom values.

pub const DYNAMICS_PROCESSOR_THRESHOLD: Self

Global, dB, -40->20, -20

pub const DYNAMICS_PROCESSOR_HEAD_ROOM: Self

Global, dB, 0.1->40.0, 5

pub const DYNAMICS_PROCESSOR_EXPANSION_RATIO: Self

Global, rate, 1->50.0, 2

pub const DYNAMICS_PROCESSOR_EXPANSION_THRESHOLD: Self

Global, dB

pub const DYNAMICS_PROCESSOR_ATTACK_TIME: Self

Global, secs, 0.0001->0.2, 0.001

pub const DYNAMICS_PROCESSOR_RELEASE_TIME: Self

Global, secs, 0.01->3, 0.05

pub const DYNAMICS_PROCESSOR_OVERALL_GAIN: Self

Global, dB, -40->40, 0

pub const DYNAMICS_PROCESSOR_COMPRESSION_AMOUNT: Self

Global, dB, read-only parameter

pub const DYNAMICS_PROCESSOR_INPUT_AMPLITUDE: Self

pub const DYNAMICS_PROCESSOR_OUTPUT_AMPLITUDE: Self

impl ParamId

Parameters for the AUVarispeed unit

pub const VARISPEED_PLAYBACK_RATE: Self

Global, Rate, 0.25 -> 4.0, 1.0

pub const VARISPEED_PLAYBACK_CENTS: Self

Global, Cents, -2400 -> 2400, 0.0

impl ParamId

Parameters for the Distortion unit

pub const DISTORTION_DELAY: Self

Global, Milliseconds, 0.1 -> 500, 0.1

pub const DISTORTION_DECAY: Self

Global, Rate, 0.1 -> 50, 1.0

pub const DISTORTION_DELAY_MIX: Self

Global, Percent, 0 -> 100, 50

pub const DISTORTION_DECIMATION: Self

Global, Percent, 0 -> 100

pub const DISTORTION_ROUNDING: Self

Global, Percent, 0 -> 100, 0

pub const DISTORTION_DECIMATION_MIX: Self

Global, Percent, 0 -> 100, 50

pub const DISTORTION_LINEAR_TERM: Self

Global, Linear Gain, 0 -> 1, 1

pub const DISTORTION_SQUARED_TERM: Self

Global, Linear Gain, 0 -> 20, 0

pub const DISTORTION_CUBIC_TERM: Self

Global, Linear Gain, 0 -> 20, 0

pub const DISTORTION_POLYNOMIAL_MIX: Self

Global, Percent, 0 -> 100, 50

pub const DISTORTION_RING_MOD_FREQ1: Self

Global, Hertz, 0.5 -> 8000, 100

pub const DISTORTION_RING_MOD_FREQ2: Self

Global, Hertz, 0.5 -> 8000, 100

pub const DISTORTION_RING_MOD_BALANCE: Self

Global, Percent, 0 -> 100, 50

pub const DISTORTION_RING_MOD_MIX: Self

Global, Percent, 0 -> 100, 0

pub const DISTORTION_SOFT_CLIP_GAIN: Self

Global, dB, -80 -> 20, -6

pub const DISTORTION_FINAL_MIX: Self

Global, Percent, 0 -> 100, 50

impl ParamId

Parameters for the AUDelay unit

pub const DELAY_WET_DRY_MIX: Self

Global, EqPow Crossfade, 0->100, 50

pub const DELAY_DELAY_TIME: Self

Global, Secs, 0->2, 1

pub const DELAY_FEEDBACK: Self

Global, Percent, -100->100, 50

pub const DELAY_LOPASS_CUTOFF: Self

Global, Hz, 10->(SampleRate/2), 15000

impl ParamId

Parameters for the AUSampleDelay unit

pub const SAMPLE_DELAY_FRAMES: Self

Global, Samples, 0->(SampleRate), 0

impl ParamId

Parameters for the AUNBandEQ unit

Note that the parameter IDs listed correspond to band 0 (zero) of the unit. The parameter IDs for higher bands can be obtained by adding the zero-indexed band number to the corresponding band 0 parameter ID up to the number of bands minus one, where the number of bands is described by the AUNBandEQ property kAUNBandEQProperty_NumberOfBands. For example, the parameter ID corresponding to the filter type of band 4 would be kAUNBandEQParam_FilterType + 3. kAUNBandEQParam_GlobalsGain is an overall gain and does not have a band.

pub const NBAND_EQ_GLOBAL_GAIN: Self

pub const NBAND_EQ_BYPASS_BAND: Self

pub const NBAND_EQ_FILTER_TYPE: Self

pub const NBAND_EQ_FREQUENCY: Self

pub const NBAND_EQ_GAIN: Self

pub const NBAND_EQ_BANDWIDTH: Self

impl ParamId

Parameters for the AURoundTripAACParam unit

pub const ROUND_TRIP_AAC_FORMAT: Self

Global, indexed : AAC, HE-AAC, HE-AACv2

pub const ROUND_TRIP_AAC_ENCODING_STRATEGY: Self

Global, indexed

pub const ROUND_TRIP_AAC_RATE_OR_QUALITY: Self

Global, indexed

impl ParamId

Parameters for the AUSoundIsolation unit

pub const SOUND_ISOLATION_WET_DRY_MIX_PERCENT: Self

pub const SOUND_ISOLATION_SOUND_TO_ISOLATE: Self

impl ParamId

Some parameters for the AUGraphicEQ unit

pub const GRAPHIC_EQ_NUMBER_OF_BANDS: Self

Global, Indexed, currently either 10 or 31

impl ParamId

Parameters for the AUMatrixReverb unit

pub const REVERB_DRY_WET_MIX: Self

Global, EqPow CrossFade, 0->100, 100

pub const REVERB_SMALL_LARGE_MIX: Self

Global, EqPow CrossFade, 0->100, 50

pub const REVERB_SMALL_SIZE: Self

Global, Secs, 0.005->0.020, 0.06

pub const REVERB_LARGE_SIZE: Self

Global, Secs, 0.4->10.0, 3.07

pub const REVERB_PRE_DELAY: Self

Global, Secs, 0.001->0.03, 0.025

pub const REVERB_LARGE_DELAY: Self

Global, Secs, 0.001->0.1, 0.035

pub const REVERB_SMALL_DENSITY: Self

Global, Genr, 0->1, 0.28

pub const REVERB_LARGE_DENSITY: Self

Global, Genr, 0->1, 0.82

pub const REVERB_LARGE_DELAY_RANGE: Self

Global, Genr, 0->1, 0.3

pub const REVERB_SMALL_BRIGHTNESS: Self

Global, Genr, 0.1->1, 0.96

pub const REVERB_LARGE_BRIGHTNESS: Self

Global, Genr, 0.1->1, 0.49

pub const REVERB_SMALL_DELAY_RANGE: Self

Global, Genr, 0->1 0.5

pub const REVERB_MODULATION_RATE: Self

Global, Hz, 0.001->2.0, 1.0

pub const REVERB_MODULATION_DEPTH: Self

Global, Genr, 0.0 -> 1.0, 0.2

impl ParamId

pub const MULTIBAND_COMPRESSOR_PREGAIN: Self

Global, dB, -40 -> 40, 0

pub const MULTIBAND_COMPRESSOR_POSTGAIN: Self

Global, dB, -40 -> 40, 0

pub const MULTIBAND_COMPRESSOR_CROSSOVER1: Self

Global, Hertz, 20 -> (SampleRate/2), 120.0

pub const MULTIBAND_COMPRESSOR_CROSSOVER2: Self

Global, Hertz, 20 -> (SampleRate/2), 700.0

pub const MULTIBAND_COMPRESSOR_CROSSOVER3: Self

Global, Hertz, 20 -> (SampleRate/2), 3000.0

pub const MULTIBAND_COMPRESSOR_THRESHOLD1: Self

Global, dB, -100.0 -> 0.0, -22.0

pub const MULTIBAND_COMPRESSOR_THRESHOLD2: Self

Global, dB, -100.0 -> 0.0, -32.0

pub const MULTIBAND_COMPRESSOR_THRESHOLD3: Self

Global, dB, -100.0 -> 0.0, -33.0

pub const MULTIBAND_COMPRESSOR_THRESHOLD4: Self

Global, dB, -100.0 -> 0.0, -36.0

pub const MULTIBAND_COMPRESSOR_HEADROOM1: Self

Global, dB, 0.1 -> 40.0, 5.0

pub const MULTIBAND_COMPRESSOR_HEADROOM2: Self

Global, dB, 0.1 -> 40.0, 12.0

pub const MULTIBAND_COMPRESSOR_HEADROOM3: Self

Global, dB, 0.1 -> 40.0, 5.0

pub const MULTIBAND_COMPRESSOR_HEADROOM4: Self

Global, dB, 0.1 -> 40.0, 7.5

pub const MULTIBAND_COMPRESSOR_ATTACK_TIME: Self

Global, Secs, 0.001 -> 0.200, 0.080

pub const MULTIBAND_COMPRESSOR_RELEASE_TIME: Self

Global, Secs, 0.010 -> 3.0, 0.120

pub const MULTIBAND_COMPRESSOR_EQ1: Self

Global, dB, -20 -> 20, 0

pub const MULTIBAND_COMPRESSOR_EQ2: Self

Global, dB, -20 -> 20, 0

pub const MULTIBAND_COMPRESSOR_EQ3: Self

Global, dB, -20 -> 20, 0

pub const MULTIBAND_COMPRESSOR_EQ4: Self

Global, dB, -20 -> 20, 0

pub const MULTIBAND_COMPRESSOR_COMPRESSION_AMOUNT1: Self

read-only parameters Global, dB, 0 -> 20

pub const MULTIBAND_COMPRESSOR_COMPRESSION_AMOUNT2: Self

Global, dB, 0 -> 20

pub const MULTIBAND_COMPRESSOR_COMPRESSION_AMOUNT3: Self

Global, dB, 0 -> 20

pub const MULTIBAND_COMPRESSOR_COMPRESSION_AMOUNT4: Self

Global, dB, 0 -> 20

pub const MULTIBAND_COMPRESSOR_INPUT_AMPLITUDE1: Self

Global, dB, -120 -> 20

pub const MULTIBAND_COMPRESSOR_INPUT_AMPLITUDE2: Self

Global, dB, -120 -> 20

pub const MULTIBAND_COMPRESSOR_INPUT_AMPLITUDE3: Self

Global, dB, -120 -> 20

pub const MULTIBAND_COMPRESSOR_INPUT_AMPLITUDE4: Self

Global, dB, -120 -> 20

pub const MULTIBAND_COMPRESSOR_OUTPUT_AMPLITUDE1: Self

Global, dB, -120 -> 20

pub const MULTIBAND_COMPRESSOR_OUTPUT_AMPLITUDE2: Self

Global, dB, -120 -> 20

pub const MULTIBAND_COMPRESSOR_OUTPUT_AMPLITUDE3: Self

Global, dB, -120 -> 20

pub const MULTIBAND_COMPRESSOR_OUTPUT_AMPLITUDE4: Self

Global, dB, -120 -> 20

impl ParamId

Parameters for the AUFilter unit

pub const MULTIBAND_FILTER_LOW_FILTER_TYPE: Self

Global, indexed, 0 -> 1, 0

pub const MULTIBAND_FILTER_LOW_FREQUENCY: Self

Global, Hertz, 10 -> (SampleRate/2), 100

pub const MULTIBAND_FILTER_LOW_GAIN: Self

Global, dB, -18 -> +18, 0

pub const MULTIBAND_FILTER_CENTER_FREQ1: Self

Global, Hertz, 10 -> (SampleRate/2), 100

pub const MULTIBAND_FILTER_CENTER_GAIN1: Self

Global, dB, -18 -> +18, 0

pub const MULTIBAND_FILTER_BANDWIDTH1: Self

Global, Octaves, 0.05 -> 3.0, 2.0

pub const MULTIBAND_FILTER_CENTER_FREQ2: Self

Global, Hertz, 10 -> (SampleRate/2), 100

pub const MULTIBAND_FILTER_CENTER_GAIN2: Self

Global, dB, -18 -> +18, 0

pub const MULTIBAND_FILTER_BANDWIDTH2: Self

Global, Octaves, 0.05 -> 3.0, 2.0

pub const MULTIBAND_FILTER_CENTER_FREQ3: Self

Global, Hertz, 10 -> (SampleRate/2), 100

pub const MULTIBAND_FILTER_CENTER_GAIN3: Self

Global, dB, -18 -> +18, 0

pub const MULTIBAND_FILTER_BANDWIDTH3: Self

Global, Octaves, 0.05 -> 3.0, 2.0

pub const MULTIBAND_FILTER_HIGH_FILTER_TYPE: Self

Global, indexed, 0 -> 1, 0

pub const MULTIBAND_FILTER_HIGH_FREQUENCY: Self

Global, Hertz, 10 -> (SampleRate/2), 100

pub const MULTIBAND_FILTER_HIGH_GAIN: Self

Global, dB, -18 -> +18, 0

impl ParamId

Parameters for AURogerBeep

pub const ROGER_BEEP_IN_GATE_THRESHOLD: Self

Global, dB, -80 -> 0, -6

pub const ROGER_BEEP_IN_GATE_THRESHOLD_TIME: Self

Global, Milliseconds, 0 -> 1000, 1000

pub const ROGER_BEEP_OUT_GATE_THRESHOLD: Self

Global, dB, -80 -> 0, -6

pub const ROGER_BEEP_OUT_GATE_THRESHOLD_TIME: Self

Global, Milliseconds, 0 -> 1000, 1000

pub const ROGER_BEEP_SENSITIVITY: Self

Global, indexed, 0 -> 2, 2

pub const ROGER_BEEP_ROGER_TYPE: Self

Global, indexed, 0 -> 2, 0

pub const ROGER_BEEP_ROGER_GAIN: Self

Global, dB, -80 -> 20, -6

impl ParamId

Parameters for the Stereo Mixer unit

pub const STEREO_MIXER_VOLUME: Self

Input/Output, Mixer Fader Curve, 0->1, 1

pub const STEREO_MIXER_PAN: Self

Input, Pan, 0->1, 0.5

pub const STEREO_MIXER_PRE_AVERAGE_POWER: Self

pub const STEREO_MIXER_PRE_PEAK_HOLD_LEVEL: Self

pub const STEREO_MIXER_POST_AVERAGE_POWER: Self

pub const STEREO_MIXER_POST_PEAK_HOLD_LEVEL: Self

impl ParamId

Parameters for the AUNetReceive unit

pub const NET_RECEIVE_STATUS: Self

Global, indexed, 0 -> 5, read only

pub const NET_RECEIVE_NUM_PARAMETERS: Self

impl ParamId

Parameters for the AUNetSend unit

pub const NET_SEND_STATUS: Self

Global, indexed, 0 -> 5, read only

pub const NET_SEND_NUM_PARAMS: Self

impl ParamId

Music Device Parameters for the DLSMusicDevice unit - defined and reported in the global scope

pub const MUSIC_DEVICE_TUNING: Self

Global, Cents, -1200, 1200, 0

pub const MUSIC_DEVICE_VOLUME: Self

Global, dB, -120->40, 0

pub const MUSIC_DEVICE_REVERB_VOLUME: Self

Global, dB, -120->40, 0

impl ParamId

Parameters for the AURandom unit

pub const RANDOM_BOUND_A: Self

pub const RANDOM_BOUND_B: Self

pub const RANDOM_CURVE: Self

impl ParamId

Parameters for the iOS reverb unit

pub const REVERB2_DRY_WET_MIX: Self

Global, CrossFade, 0->100, 100

pub const REVERB2_GAIN: Self

Global, Decibels, -20->20, 0

pub const REVERB2_MIN_DELAY_TIME: Self

Global, Secs, 0.0001->1.0, 0.008

pub const REVERB2_MAX_DELAY_TIME: Self

Global, Secs, 0.0001->1.0, 0.050

pub const REVERB2_DECAY_TIME_AT0_HZ: Self

Global, Secs, 0.001->20.0, 1.0

pub const REVERB2_DECAY_TIME_AT_NYQUIST: Self

Global, Secs, 0.001->20.0, 0.5

pub const REVERB2_RANDOMIZE_REFLECTIONS: Self

Global, Integer, 1->1000