cidre::at::audio::unit

Struct SubType

#[repr(transparent)]
pub struct SubType(pub u32);

Apple input/output audio unit sub types

Subtypes for the various input/output units that Apple ships. Input/output units add an additional notion of Start and Stop.

Tuple Fields
0: u32
Implementations

impl SubType

pub const GENERIC_OUTPUT: Self

A generic output unit provides the start/stop API, and provides the basic services to convert Linear PCM formats.

pub const VOICE_PROCESSING_IO: Self

This audio unit can do input as well as output. Bus 0 is used for the output side, bus 1 is used to get audio input (thus, on the iPhone, it works in a very similar way to the Remote I/O). This audio unit does signal processing on the incoming audio (taking out any of the audio that is played from the device at a given time from the incoming audio).

pub const HAL_OUTPUT: Self

(a.k.a. “AUHAL”) The audio unit that interfaces to any audio device. The user specifies which audio device to track. Bus 0 is used to send audio output to the device; bus 1 is used to receive audio input from the device. Available on macOS only.

pub const DEFAULT_OUTPUT: Self

A specialization of AUHAL that is used to track the user’s selection of the default device as set in the Sound Prefs. Available on macOS only.

pub const SYSTEM_OUTPUT: Self

A specialization of AUHAL that is used to track the user’s selection of the device to use for sound effects, alerts and other UI sounds. Available on macOS only.

impl SubType

Apple music instrument audio unit sub types

pub const DLS_SYNTH: Self

A multi-timbral music device that can use sample banks in either DLS or SoundFont formats. It fully supports GM-MIDI and the basic extensions of GS-MIDI.

pub const SAMPLER: Self

A mono-timbral music device which is a sampler synthesizer and supports full interactive editing of all state.

pub const MIDI_SYNTH: Self

A fully GM-compatible multi-timbral music device which is a sampler synthesizer. It can load instruments from sample banks in either DLS or SoundFont formats.

impl SubType

Apple converter audio unit sub types

pub const CONVERTER: Self

An audio unit that uses an AudioConverter to do Linear PCM conversions (sample rate, bit depth, interleaving).

pub const VARISPEED: Self

pub const DEFERRED_RENDERER: Self

pub const SPLITTER: Self

pub const MUTLI_SPLITTER: Self

pub const MERGER: Self

pub const NEW_TIME_PITCH: Self

pub const ROUND_TRIP_AAC: Self

impl SubType

Apple effect audio unit sub types

pub const PEAK_LIMITER: Self

A peak limiter

pub const DYNAMICS_PROCESSOR: Self

A dynamics compressor/expander

pub const LOW_PASS_FILTER: Self

A filter that passes frequencies below a specified cut-off frequency

pub const HIGH_PASS_FILTER: Self

A filter that passes frequencies above a specified cut-off frequency

pub const BAND_PASS_FILTER: Self

A filter that passes frequencies between a low and high cut-off frequency.

pub const HIGH_SHELF_FILTER: Self

A filter that can be used to implement a “treble” control

pub const LOW_SHELF_FILTER: Self

A filter that can be used to implement a “bass” control

pub const PARAMETRIC_EQ: Self

A parametric EQ filter

pub const DISTORTION: Self

A distortion audio unit

pub const DELAY: Self

A delay audio unit

pub const SAMPLE_DELAY: Self

A delay that is used to delay the input a specified number of samples until the output

pub const N_BAND_EQ: Self

A generalized N-band graphic EQ with specifiable filter types per-band

pub const REVERB2: Self

A lite reverb that can be used to simulate various and different spaces

pub const SOUND_ISOLATION: Self

An audio unit that can be used to isolate a specified sound type

pub const AUDIO_MIX: Self

An audio unit that supports AudioMix separate-and-remix functionality

impl SubType

Apple effect audio unit sub types (macOS only)

pub const GRAPHIC_EQ: Self

A 10 or 31 band Graphic EQ

pub const MULTI_BAND_COMPRESSOR: Self

A 4 band compressor/expander

pub const MATRIX_REVERB: Self

A reverb that can be used to simulate various and different spaces

pub const PITCH: Self

An audio unit used to change the pitch

pub const FILTER: Self

A filter unit that combines 5 different filters (low, 3 mids, high)

pub const NET_SEND: Self

An audio unit that is used in conjunction with _NetReceive to send audio across the network (or between different applications)

pub const ROGER_BEEP: Self

An audio unit that can be used to emit a short tone in gaps between speech similar to the tones used in a walkie-talkie

impl SubType

Apple mixer audio unit sub types

These are the subtypes for the various mixer units that apple ships

pub const MULTI_CHANNEL_MIXER: Self

Can have any number of inputs, with any number of channels on any input to one output bus with any number of channels.

pub const MATRIX_MIXER: Self

Any number of input and output buses with any number of channels on any bus. The mix is presented as a matrix of channels that can be controlled through input volume per channel, “cross-point” volume (a given input channel to a given output channel), output volume per channel and a global volume across the whole matrix

pub const SPATIAL_MIXER: Self

Inputs that are mono will be panned around using 3D coordinates and parameters. Stereo inputs are passed directly through to the output. A single output is presented with 2, 4, 5, 6, 7 or 8 channels. There is also a built in reverb.

impl SubType

Apple mixer audio unit sub types (macOS only)

pub const STEREO_MIXER: Self

Inputs can be mono or stereo. Single stereo output

impl SubType

Apple generator audio unit sub types

pub const NET_RECEIVE: Self

A generator unit that is paired with _NetSend to receive the audio that unit sends. It presents a custom UI so can be used in a UI context as well

pub const SCHEDULED_SOUND_PLAYER: Self

A generator unit that can be used to schedule slices of audio to be played at a specified time. The audio is scheduled using the time stamps for the render operation, and can be scheduled from any thread.

pub const AUDIO_FILE_PLAYER: Self

A generator unit that is used to play a file. It presents a custom UI so can be used in a UI context as well