cidre::at::audio::unit

Struct Type

#[repr(transparent)]
pub struct Type(pub u32);

Different types of audio units

Audio units are classified into different types, where those types perform different roles and functions.

There are some general categories of functionality that apply across different types of audio units:

Real-time usage.

The audio unit will complete its operations in less time than is represented by the render buffer. All audio units with the exception of the OfflineEffect should meet this criteria.

Real-time I/O.

Will request the same amount of audio input as it is being asked to produce for output. Effects, Panners, Mixers and MusicDevices are required to adhere to this restriction. FormatConverter’s can with some constraints be used in this situation (for instance, sample rate conversion, float-int), but the host of the audio unit is responsible for insuring this.

UI versus Programmatic usage.

UI usage covers the case of using an audio unit in a Digital Audio Workstation (DAW) with appropriate UI (for example a filter in Garage Band or Logic). Effects, Panners, MusicDevices are all expected to be usable within this context.

Programmatic usage is where an audio unit is used by a host app as part of a general signal processing chain.

For instance, a mixer audio unit can be used to take several different audio sources in a game and mix them together. Mixers, Output units are for programmatic usage. FormatConverter and Generator types are generally programmatic audio units, but if they can be used in a UI situation, they specify a custom view. The app can then use that to decide that, with appropriate constraints, the audio unit could be presented in a DAW type application. For instance, the AUConverter audio unit from apple can do sample rate conversion, etc, but has not general utility for a user in a DAW app. Apple’s Varispeed or AUTimePitch audio units can be used to change the playback rate and pitch and so could be used to good affect by a user in a DAW type environment, as well as just providing this general functionality to any program.

Tuple Fields
0: u32
Implementations

impl Type

pub const OUTPUT: Self

An output unit can be used standalone or as part of an AUGraph or crate::av::AudioEngine. Apple provides a number of output units that interface directly with an audio device.

pub const MUSIC_DEVICE: Self

A software musical instrument such as a sampler or synthesizer. They respond to MIDI and create notes, which are then controlled through parameters or MIDI control messages.

pub const MUSIC_EFFECT: Self

An audio unit that will process some x number of audio input samples to produce x number of audio output samples. The common case for an effect is to have a single input to a single output, though some effects take side-chain inputs as well. Effects can be run in “offline” contexts (such as processing a file), but they are expected to run in real-time. A delay unit or reverb is a good example of this.

pub const FORMAT_CONVERTER: Self

A format converter is a general category for audio units that can change the format (for instance, sample rate conversion) from an input to an output, as well as other, non-I/O type manipulations (like a deferred render or varispeed type of operation). As such, a format converter can ask for as much or as little audio input to produce a given output. They are still expected to complete their rendering within the time represented by the output buffer. For format converters that have some utility as an “audio effect or processor”, it is quite common to provide an offline version of this audio unit as well. For instance, Apple ships a format converter (for use in a “real-time” like situation) and an offline version (for processing audio files) of the Time Pitch and Varispeed audio units.

pub const EFFECT: Self

An audio unit that will process some x number of audio input samples to produce x number of audio output samples. The common case for an effect is to have a single input to a single output, though some effects take side-chain inputs as well. Effects can be run in “offline” contexts (such as processing a file), but they are expected to run in real-time. A delay unit or reverb is a good example of this.

pub const MIXER: Self

An audio unit that takes some number of inputs, mixing them to provide 1 or more audio outputs. A stereo mixer (mono and stereo inputs to produce one stereo output) is an example of this.

pub const PANNER: Self

A panner is a specialised effect that will pan a single audio input to a single output. Panner units are required to support a collection of standardised parameters that specify the panning coordinates (aside from whatever custom parameters the panner may provide). A surround panner is an example of this

pub const GENERATOR: Self

A generator will have no audio input, but will just produce audio output. In some ways it is similar to a MusicDevice, except that a generator provides no MIDI input, or notion of “notes”. A tone generator is a good example of this.

pub const OFFLINE_EFFECT: Self

An offline effect is used to process data from a file and is also used to publish a capability that cannot be run in real-time. For instance, the process of normalization requires seeing the entire audio input before the scalar to apply in the normalization process can be estimated. As such, offline effects also have a notion of a priming stage that can be performed before the actual rendering/processing phase is executed.

pub const MIDI_PROCESSOR: Self

Plugins of this type process MIDI input and produce MIDI output. They do not produce audio.

pub const SPEECH_SYNTHESIZER: Self

An offline audio unit that produces synthesized speech audio output.