cidre::av::audio

Trait PlayerDelegate

pub trait PlayerDelegate: Obj {
    // Provided methods
    fn audio_player_did_finish_playing(
        &mut self,
        player: &mut Player,
        successfully: bool,
    ) { ... }
    fn sel_audio_player_did_finish_playing() -> &'static Sel { ... }
    fn audio_player_decode_error_did_occur(
        &mut self,
        player: &mut Player,
        err: Option<&Error>,
    ) { ... }
    fn sel_audio_player_decode_error_did_occur() -> &'static Sel { ... }
}

Provided Methods

fn audio_player_did_finish_playing( &mut self, player: &mut Player, successfully: bool, )

fn sel_audio_player_did_finish_playing() -> &'static Sel

@selector(audioPlayerDidFinishPlaying:successfully:) but dynamic use this function to check if object responds to selector

fn audio_player_decode_error_did_occur( &mut self, player: &mut Player, err: Option<&Error>, )

fn sel_audio_player_decode_error_did_occur() -> &'static Sel

@selector(audioPlayerDecodeErrorDidOccur:error:) but dynamic use this function to check if object responds to selector

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.