cidre::av

Module capture

Re-exports

pub use device::AuthorizationStatus;

pub use device::AutoFocusRangeRestriction;

pub use device::AutoFocusSys;

pub use device::CamLensSmudgeDetectionStatus;

pub use device::CenterStageControlMode;

pub use device::CinematicVideoFocusMode;

pub use device::ColorSpace;

pub use device::ConfigLockGuard as DeviceConfigurationLockGuard;

pub use device::Device;

pub use device::DiscoverySession;

pub use device::ExposureMode;

pub use device::FlashMode;

pub use device::FocusMode;

pub use device::Format as DeviceFormat;

pub use device::FrameRateRange;

pub use device::MicMode;

pub use device::Pos as DevicePos;

pub use device::RotationCoordinator as DeviceRotationCoordinator;

pub use device::TorchMode;

pub use device::Type as DeviceType;

pub use device::VideoStabilizationMode;

pub use device::WbChromaticityValues;

pub use device::WbGains;

pub use device::WbMode;

pub use device::WbTempTintValues;

pub use device::notifications as device_notifications;

pub use input::DeviceInput;

pub use input::Input;

pub use input::Port as InputPort;

pub use input::port_notifications as input_port_notifications;

pub use output_base::DataDroppedReason;

pub use output_base::Output;

pub use session_preset::SessionPreset;

pub use session::AudioChannel;

pub use session::Connection;

pub use session::ControlsDelegate;

pub use session::ControlsDelegateImpl;

pub use session::MultiCamSession;

pub use session::Session;

pub use session::VideoOrienation;

pub use session::err_key as session_err_key;

pub use session::notifications as session_notifications;

pub use metadata_output::MetadataOutput;

pub use metadata_output::MetadataOutputObjsDelegate;

pub use metadata_output::MetadataOutputObjsDelegateImpl;

pub use video_data_output::VideoDataOutput;

pub use video_data_output::VideoDataOutputSampleBufDelegate;

pub use video_data_output::VideoDataOutputSampleBufDelegateImpl;

pub use video_preview_layer::VideoPreviewLayer;

pub use photo_output::PhotoOutput;

pub use photo_output::PhotoQualityPrioritization;

pub use photo_output::PhotoSettings;

pub use audio_data_output::AudioDataOutput;

pub use audio_data_output::AudioDataOutputSampleBufDelegate;

pub use audio_data_output::AudioDataOutputSampleBufDelegateImpl;

pub use reactions::EffectState as ReactionEffectState;

pub use reactions::ReactionType;

Modules

audio_data_output

device

input

metadata_output

output_base

photo_output

reactions

session

session_preset

video_data_output

video_preview_layer

Structs

Control

IndexPicker

Slider

SysExposureBiasSlider

A control that adjusts the exposure bias of a capture device within the system-recommended range.

SysZoomSlider

A control that adjusts the video zoom factor of a capture device within the system-recommended range.