cidre::av::capture

Module device

Modules

notifications

Structs

ConfigLockGuard

Device

DiscoverySession

Allows clients to search for devices by certain criteria.

Format

An av::CaptureDeviceFormat wraps a cm::FormatDesc and other format-related information, such as min and max framerate.

FrameRateRange

RotationCoordinator

Allows clients to monitor rotations of a given av::CaptureDevice instance and be provided the video rotation angle that should be applied for horizon-level preview and capture relative to gravity.

Type

WbChromaticityValues

WbGains

WbTempTintValues

ZoomRange

Expresses an inclusive range of supported zoom factors.

Enums

AuthorizationStatus

AutoFocusRangeRestriction

AutoFocusSys

CamLensSmudgeDetectionStatus

CenterStageControlMode

CinematicVideoFocusMode

ColorSpace

ExposureMode

FlashMode

FocusMode

MicMode

Pos

SysUi

TorchMode

VideoStabilizationMode

WbMode

AVCaptureWhiteBalanceMode

Functions

torch_max_level