cidre::av::capture::device

Struct ConfigLockGuard

pub struct ConfigLockGuard<'a> { /* private fields */ }

Implementations

impl<'a> ConfigLockGuard<'a>

AVCaptureDeviceTorch

pub unsafe fn set_torch_mode_throws(&mut self, val: TorchMode)

pub fn set_torch_mode<'ear>(&mut self, val: TorchMode) -> ExResult<'ear>

pub unsafe fn set_torch_mode_on_with_level_err<'ear>( &mut self, torch_level: f32, error: *mut Option<&'ear Error>, ) -> bool

pub fn set_torch_mode_on_with_level<'ear>( &mut self, torch_level: f32, ) -> Result<'ear>

impl<'a> ConfigLockGuard<'a>

pub fn set_video_hdr_enabled<'ear>(&mut self, val: bool) -> ExResult<'ear>

pub fn set_automatically_adjusts_video_hdr_enabled<'ear>( &mut self, val: bool, ) -> ExResult<'ear>

impl<'a> ConfigLockGuard<'a>

pub fn set_active_format(&mut self, val: &Format)

pub fn set_auto_video_frame_rate_enabled<'ear>( &mut self, val: bool, ) -> ExResult<'ear>

Availability

macos_15_0, ios_18_0, tvos_18_0, maccatalyst_18_0

pub fn set_active_video_min_frame_duration<'ear>( &mut self, val: Time, ) -> ExResult<'ear>

pub fn set_active_video_max_frame_duration<'ear>( &mut self, val: Time, ) -> ExResult<'ear>

pub fn set_center_stage_rect_of_interest<'ear>( &mut self, val: Rect, ) -> ExResult<'ear>

pub unsafe fn set_active_color_space_throws(&mut self, val: ColorSpace)

Will throw uncatchable NSInvalidArgumentsException if ColorSpace is not supported.

NOTE: I suspect bc of dispatch_sync on other queue.

pub unsafe fn set_focus_mode_throws(&mut self, mode: FocusMode)

pub fn set_focus_mode<'ear>(&mut self, mode: FocusMode) -> ExResult<'ear>

pub unsafe fn set_focus_point_of_interest_throws(&mut self, val: Point)

pub fn set_focus_point_of_interest<'ear>( &mut self, val: Point, ) -> ExResult<'ear>

The value of this property is a ‘cg::Point’ that determines the receiver’s focus point of interest, if it has one. A value of (0,0) indicates that the camera should focus on the top left corner of the image, while a value of (1,1) indicates that it should focus on the bottom right. The default value is (0.5,0.5).

Clients can observe automatic changes to the receiver’s ‘focusPointOfInterest’ by key value observing this property. Note that setting focusPointOfInterest alone does not initiate a focus operation. After setting ‘set_focus_point_of_interest’, call ‘set_focus_mode()’ to apply the new point of interest.

impl<'a> ConfigLockGuard<'a>

AVCaptureExposureMode

pub unsafe fn set_exposure_mode_throws(&mut self, val: ExposureMode)

pub fn set_exposure_mode<'ear>(&mut self, val: ExposureMode) -> ExResult<'ear>

pub unsafe fn set_exposure_point_of_interest_throws(&mut self, val: Point)

pub fn set_exposure_point_of_interest<'ear>( &mut self, val: Point, ) -> ExResult<'ear>

impl<'a> ConfigLockGuard<'a>

This impl block contains no items.

AVCaptureDeviceToneMapping

impl<'a> ConfigLockGuard<'a>

AVCaptureDeviceWhiteBalance

pub unsafe fn set_wb_mode_throws(&mut self, val: WbMode)

pub fn set_wb_mode<'ear>(&mut self, val: WbMode) -> ExResult<'ear>

impl<'a> ConfigLockGuard<'a>

AVCaptureDeviceVideoZoom

pub fn set_video_zoom_factor(&mut self, val: Float)

pub fn ramp_to_video_zoom_factor(&mut self, factor: Float, rate: f32)

pub fn cancel_video_zoom_ramp(&mut self)

impl<'a> ConfigLockGuard<'a>

CameraLensSmudgeDetection

pub unsafe fn set_cam_lens_smudge_detection_enabled<'ear>( &mut self, val: bool, detection_interval: Time, ) -> ExResult<'ear>

Availability

Not macos_26_0, ios_26_0, tvos_26_0, maccatalyst_26_0