cidre::av::capture::device

Struct Device

pub struct Device(/* private fields */);

Implementations

impl Device

pub fn retained(&self) -> R<Self>

impl Device

pub fn cls() -> &'static Class<Self>

pub fn cls_ptr() -> *const c_void

pub fn alloc() -> A<Self>

pub fn with_type_media_and_pos_ar( device_type: &Type, media_type: Option<&MediaType>, position: Pos, ) -> Option<Rar<Self>>

pub fn with_type_media_and_pos( device_type: &Type, media_type: Option<&MediaType>, position: Pos, ) -> Option<R<Self>>

pub fn devices_ar() -> Rar<Array<Self>>

pub fn devices() -> R<Array<Self>>

pub fn default_with_media_ar(media_type: &MediaType) -> Option<Rar<Self>>

pub fn default_with_media(media_type: &MediaType) -> Option<R<Self>>

pub fn with_unique_id_ar(unique_id: &String) -> Option<Rar<Self>>

pub fn with_unique_id(unique_id: &String) -> Option<R<Self>>

pub fn unique_id_ar(&self) -> Rar<String>

pub fn unique_id(&self) -> R<String>

pub fn model_id_ar(&self) -> Rar<String>

The model ID of the receiver.

The value of this property is an identifier unique to all devices of the same model. The value is persistent across device connections and disconnections, and across different systems. For example, the model ID of the camera built in to two identical iPhone models will be the same even though they are different physical devices.

pub fn model_id(&self) -> R<String>

The model ID of the receiver.

The value of this property is an identifier unique to all devices of the same model. The value is persistent across device connections and disconnections, and across different systems. For example, the model ID of the camera built in to two identical iPhone models will be the same even though they are different physical devices.

pub fn localized_name_ar(&self) -> Rar<String>

A localized human-readable name for the receiver.

This property can be used for displaying the name of a capture device in a user interface.

pub fn localized_name(&self) -> R<String>

A localized human-readable name for the receiver.

This property can be used for displaying the name of a capture device in a user interface.

pub fn manufacturer_ar(&self) -> Option<Rar<String>>

The human-readable manufacturer name for the receiver.

This property can be used to identify capture devices from a particular manufacturer. All Apple devices return “Apple Inc.”. Devices from third party manufacturers may return an empty string.

pub fn manufacturer(&self) -> Option<R<String>>

The human-readable manufacturer name for the receiver.

This property can be used to identify capture devices from a particular manufacturer. All Apple devices return “Apple Inc.”. Devices from third party manufacturers may return an empty string.

pub fn has_media_type(&self, media_type: &MediaType) -> bool

Returns whether the receiver provides media with the given media type.

‘true’ if the device outputs the given media type, ‘false’ otherwise.

pub fn formats_ar(&self) -> Rar<Array<Format>>

pub fn formats(&self) -> R<Array<Format>>

pub fn supports_preset(&self, preset: &CaptureSessionPreset) -> bool

pub fn active_format_ar(&self) -> Option<Rar<Format>>

NOTE: On audio devices active format is None

pub fn active_format(&self) -> Option<R<Format>>

NOTE: On audio devices active format is None

pub fn is_connected(&self) -> bool

Indicates whether the device is connected and available to the system.

The value of this property is a ‘bool’ indicating whether the device represented by the receiver is connected and available for use as a capture device. Clients can key value observe the value of this property to be notified when a device is no longer available. When the value of this property becomes ‘false’ for a given instance, it will not become ‘true’ again. If the same physical device again becomes available to the system, it will be represented using a new instance of ‘av::CaptureDevice’.

pub fn active_video_min_frame_duration(&self) -> Time

pub fn active_video_max_frame_duration(&self) -> Time

pub fn is_auto_video_frame_rate_enabled(&self) -> bool

pub fn sel_is_auto_video_frame_rate_enabled() -> &'static Sel

@selector(isAutoVideoFrameRateEnabled) but dynamic use this function to check if object responds to selector

pub fn config_lock(&mut self) -> Result<ConfigLockGuard<'_>, R<Error>>

pub unsafe fn lock_for_config(&mut self, error: *mut Option<R<Error>>) -> bool

pub unsafe fn unlock_for_config(&mut self)

pub unsafe fn set_active_format(&mut self, val: &Format)

pub unsafe fn set_auto_video_frame_rate_enabled_throws(&mut self, val: bool)

pub fn sel_set_auto_video_frame_rate_enabled_throws() -> &'static Sel

@selector(setAutoVideoFrameRateEnabled:) but dynamic use this function to check if object responds to selector

pub unsafe fn set_auto_video_frame_rate_enabled<'ear>( &mut self, val: bool, ) -> ExResult<'ear>

Availability

macos_15_0, ios_18_0, tvos_18_0, maccatalyst_18_0

pub unsafe fn set_active_video_min_frame_duration_throws(&mut self, val: Time)

pub unsafe fn set_active_video_min_frame_duration<'ear>( &mut self, val: Time, ) -> ExResult<'ear>

pub unsafe fn set_active_video_max_frame_duration_throws(&mut self, val: Time)

pub unsafe fn set_active_video_max_frame_duration<'ear>( &mut self, val: Time, ) -> ExResult<'ear>

impl Device

AVCaptureDeviceFlash

pub fn has_flash(&self) -> bool

Indicates whether the receiver has a flash.

pub fn is_flash_available(&self) -> bool

Indicates whether the receiver’s flash is currently available for use.

The value of this property is a bool indicating whether the receiver’s flash is currently available. The flash may become unavailable if, for example, the device overheats and needs to cool off. This property is key-value observable.

impl Device

AVCaptureDeviceTorch

pub fn has_torch(&self) -> bool

pub fn is_torch_available(&self) -> bool

Indicates whether the receiver’s torch is currently available for use.

pub fn is_torch_active(&self) -> bool

pub fn torch_level(&self) -> f32

pub fn is_torch_mode_supported(&self, val: TorchMode) -> bool

pub fn torch_mode(&self) -> TorchMode

impl Device

AVCaptureDeviceReactionEffects

pub fn reaction_effects_enabled() -> bool

pub fn sel_reaction_effects_enabled() -> &'static Sel

@selector(reactionEffectsEnabled) but dynamic use this function to check if object responds to selector

pub fn reaction_effect_gestures_enabled() -> bool

pub fn sel_reaction_effect_gestures_enabled() -> &'static Sel

@selector(reactionEffectGesturesEnabled) but dynamic use this function to check if object responds to selector

pub fn can_perform_reaction_effects(&self) -> bool

pub fn sel_can_perform_reaction_effects() -> &'static Sel

@selector(canPerformReactionEffects) but dynamic use this function to check if object responds to selector

pub fn available_reaction_types_ar(&self) -> Rar<Set<CaptureReactionType>>

Returns a list of reaction types which can be passed to perform_effect_for_reaction.

The list may differ between devices, or be affected by changes to active format, and can be key-value observed.

pub fn sel_available_reaction_types() -> &'static Sel

@selector(availableReactionTypes) but dynamic use this function to check if object responds to selector

pub fn available_reaction_types(&self) -> R<Set<CaptureReactionType>>

Returns a list of reaction types which can be passed to perform_effect_for_reaction.

The list may differ between devices, or be affected by changes to active format, and can be key-value observed.

pub fn perform_effect_for_reaction( &mut self, reaction_type: &CaptureReactionType, )

Triggers a specified reaction on the video stream.

pub fn sel_perform_effect_for_reaction() -> &'static Sel

@selector(performEffectForReaction:) but dynamic use this function to check if object responds to selector

pub fn reaction_effects_in_progress_ar( &self, ) -> Rar<Array<CaptureReactionEffectState>>

pub fn sel_reaction_effects_in_progress() -> &'static Sel

@selector(reactionEffectsInProgress) but dynamic use this function to check if object responds to selector

pub fn reaction_effects_in_progress( &self, ) -> R<Array<CaptureReactionEffectState>>

impl Device

AVCaptureDeviceBackgroundReplacement

pub fn is_background_replacement_enabled() -> bool

A class property indicating whether the user has enabled the Background Replacement feature for this application.

pub fn sel_is_background_replacement_enabled() -> &'static Sel

@selector(isBackgroundReplacementEnabled) but dynamic use this function to check if object responds to selector

pub fn is_background_replacement_active(&self) -> bool

pub fn sel_is_background_replacement_active() -> &'static Sel

@selector(isBackgroundReplacementActive) but dynamic use this function to check if object responds to selector

impl Device

AVCaptureDeviceContinuityCamera

pub fn is_continuity_camera(&self) -> bool

A property that reports YES if the receiver is a Continuity Camera.

Access this property to discover if the receiver is a Continuity Camera (external iPhone webcam).

impl Device

AVCaptureDeviceDeskViewCamera

pub fn companion_desk_view_camera_ar(&self) -> Option<Rar<Device>>

A reference to the Desk View Camera that is associated with and derived from this camera.

pub fn companion_desk_view_camera(&self) -> Option<R<Device>>

A reference to the Desk View Camera that is associated with and derived from this camera.

impl Device

Configuring HDR Settings

pub fn automatically_adjusts_video_hdr_enabled(&self) -> bool

pub unsafe fn set_automatically_adjusts_video_hdr_enabled_throws( &mut self, val: bool, )

pub fn is_video_hdr_enabled(&self) -> bool

pub unsafe fn set_video_hdr_enabled_throws(&mut self, val: bool)

The value of this property is a bool indicating whether the receiver is currently streaming high dynamic range video buffers, also known as Extended Dynamic Range (EDR). The value of this property is ignored when device.activeColorSpace is HLG BT2020 color space since HDR is effectively always on and can’t be disabled. The property may only be set if you first set automaticallyAdjustsVideoHDREnabled to bool, otherwise an [ns::GenericException] is thrown. videoHDREnabled may only be set to YES if the receiver’s activeFormat.isVideoHDRSupported property returns YES, otherwise an [ns::GenericException] is thrown. This property may be key-value observed.

Note that setting this property may cause a lengthy reconfiguration of the receiver, similar to setting a new active format or av::CaptureSession sessionPreset. If you are setting either the active format or the av::CaptureSession’s sessionPreset AND this property, you should bracket these operations with [session beginConfiguration] and [session commitConfiguration] to minimize reconfiguration time.

impl Device

AVCaptureMicrophoneMode

pub fn preferred_mic_mode() -> MicMode

Indicates the microphone mode that has been selected by the user in Control Center.

This readonly property returns the microphone mode selected by the user in Control Center. It is key-value observable.

pub fn sel_preferred_mic_mode() -> &'static Sel

@selector(preferredMicrophoneMode) but dynamic use this function to check if object responds to selector

pub fn active_mic_mode() -> MicMode

Indicates the currently active microphone mode.

This readonly property returns the currently active microphone mode, which may differ from the ‘preferred_mic_mode()’ if the application’s active audio route does not support the preferred microphone mode. This property is key-value observable.

pub fn sel_active_mic_mode() -> &'static Sel

@selector(activeMicrophoneMode) but dynamic use this function to check if object responds to selector

impl Device

pub fn center_stage_control_mode() -> CenterStageControlMode

Current mode of Center Stage control (user, app, or cooperative).

This class property determines how the Center Stage feature is controlled. When set to the default value of ‘av::CaptureCenterStageControlMode::User’, ‘set_center_stage_enabled’ may not be set programmatically and throws an NSInvalidArgumentException. In User mode, the feature may only be set by the user in Control Center. If you wish to take Center Stage control away from the user and exclusively enable / disable it programmatically, set this property to ‘av::CaptureCenterStageControlMode::App’. When under exclusive app control, Center Stage user control is disallowed (for instance, the toggle is grayed out in Control Center). If you wish to take control of Center Stage, but also cooperate with the user by listening for and appropriately reacting to their changes to the ‘center_stage_enabled’ property, set this property to ‘av::CaptureCenterStageControlMode::Cooperative’. Note that in this mode, the onus is on you, the app developer, to honor user intent and conform your ‘av::CaptureSession’ configuration to make Center Stage active (see the ‘av::CaptureDevice’ instance property ‘center_stage_active’). In cooperative mode, the ‘center_stage_enabled’ property may change at any time (such as when the user enables / disables the feature in Control Center).

pub fn set_center_stage_control_mode(val: CenterStageControlMode)

pub fn is_center_stage_active(&self) -> bool

Indicates whether Center Stage is currently active on a particular av::CaptureDevice.

This readonly property returns ‘true’ when Center Stage is currently active on the receiver. When active, the camera automatically adjusts to keep people optimally framed within the field of view. The field of view may pan, tighten or widen as needed. Certain restrictions come into play when Center Stage is active:

The device’s ‘min_available_video_zoom_factor’ and ‘max_available_video_zoom_factor’ become restricted (see av::CaptureDeviceFormat’s ‘video_min_zoom_factor_for_center_stage’ and ‘video_max_zoom_factor_for_center_stage’).

The device’s ‘active_video_min_frame_duration’ and ‘active_video_max_frame_duration’ are limited (see av::CaptureDeviceFormat’s ‘video_frame_rate_range_for_center_stage’).

Center Stage may be enabled via user control or application control, depending on the current ‘av::CaptureDevice::center_stage_control_mode()’. When ‘av::CaptureDevice::is_center_stage_enabled()’ is ‘true’, a particular ‘av::CaptureDevice’ instance may return ‘true’ for this property, depending whether it supports the feature in its current configuration. Some device features are mutually exclusive to Center Stage:

If depth data delivery is enabled on any output, such as ‘av::CaptureDepthDataOutput’, or ‘av::CapturePhotoOutput.depth_data_delivery_enabled’, Center Stage is deactivated.

If ‘geometric_distortion_correction_supported’ is ‘true’, ‘geometric_distortion_correction_enabled’ must also be ‘true’, or Center Stage is deactivated.

This property is key-value observable.

pub fn center_stage_rect_of_interest(&self) -> Rect

pub unsafe fn set_center_stage_rect_of_interest_throws(&mut self, val: Rect)

Specifies the effective region within the output pixel buffer that will be used to perform Center Stage framing.

Applications that wish to apply additional processing (such as cropping) on top of Center Stage’s output can use this property to guide Center Stage’s framing.

The rectangle’s origin is top left and is relative to the coordinate space of the output pixel buffer. The default value of this property is the value ‘cg::Rect::new(0.0, 0.0, 1.0, 1.0)’, where {0.0,0.0} represents the top left of the picture area, and {1.0,1.0} represents the bottom right on an unrotated picture. This rectangle of interest is applied prior to rotation, mirroring or scaling.

Pixels outside of this rectangle of interest will be blackened out.

Setting this property has no impact on objects specified in the metadata output.

Safety

‘set_center_stage_rect_of_interest_throws’: throws an ‘ns::ExceptionName::generic()’ if called without first obtaining exclusive access to the receiver using ‘lock_for_configuration’ ‘set_center_stage_rect_of_interest_throws’ throws an ‘ns::ExceptionName::invalid_argument()’ if none of the av::CaptureDeviceFormats supported by the receiver support CenterStage. ‘set_center_stage_rect_of_interest_throws’ throws an ‘ns::ExceptionName::invalid_argument()’ ‘::center_stage_enabled’ is ‘false’ on the av::CaptureDevice class. ‘set_center_stage_rect_of_interest_throws’ throws an ‘ns::ExceptionName::invalid_argument()’ if the provided rect of interest goes outside the normalized (0-1) coordinate space.

impl Device

AVCaptureSystemUserInterface

pub fn show_sys_ui(system_ui: SysUi)

pub fn sel_show_sys_ui() -> &'static Sel

@selector(showSystemUserInterface:) but dynamic use this function to check if object responds to selector

impl Device

AVCaptureDeviceColorSpaceSupport https://developer.apple.com/videos/play/wwdc2016/501/?time=2972

pub fn active_color_space(&self) -> ColorSpace

Indicates the receiver’s current active color space.

pub unsafe fn set_active_color_space_throws(&mut self, val: ColorSpace)

By default, an ‘av::CaptureDevice’ attached to an ‘av::CaptureSession’ is automatically configured for wide color by the ‘av::CaptureSession’ (see AVCaptureSession automaticallyConfiguresCaptureDeviceForWideColor). You may also set the active_ColorSpace manually. To prevent the AVCaptureSession from undoing your work, remember to set AVCaptureSession’s automaticallyConfiguresCaptureDeviceForWideColor property to NO. Changing the receiver’s activeColorSpace while the session is running requires a disruptive reconfiguration of the capture render pipeline. Movie captures in progress will be ended immediately; unfulfilled photo requests will be aborted; video preview will temporarily freeze. -setActiveColorSpace: throws an NSGenericException if called without first obtaining exclusive access to the receiver using -lockForConfiguration:.

impl Device

AVCaptureDevicePosition

pub fn pos(&self) -> Pos

impl Device

AVCaptureDeviceFocus

pub fn is_focus_mode_supported(&self, mode: FocusMode) -> bool

pub fn focus_mode(&self) -> FocusMode

pub fn is_focus_point_of_interest_supported(&self) -> bool

pub fn focus_point_of_intereset(&self) -> Point

pub fn is_adjusting_focus(&self) -> bool

The value of this property is a bool indicating whether the receiver’s camera focus is being automatically adjusted by means of a focus scan, because its focus mode is ‘av::CaptureFocusModeAutoFocus’ or ‘av::CaptureFocusModeContinuousAutoFocus’. Clients can observe the value of this property to determine whether the camera’s focus is stable.

pub fn minimum_focus_distance(&self) -> Integer

A property indicating the minimum focus distance.

The minimum focus distance is given in millimeters, -1 if unknown. For virtual cameras, the value reported is the smallest minimum focus distance of the auto-focus-capable cameras that it sources.

impl Device

AVCaptureExposureMode

pub fn is_exposure_mode_supported(&self, val: ExposureMode) -> bool

Returns whether the receiver supports the given exposure mode.

pub fn exposure_mode(&self) -> ExposureMode

pub fn is_exposure_point_of_interest_supported(&self) -> bool

pub fn exposure_point_of_interest(&self) -> Point

pub fn is_adjusting_exposure(&self) -> bool

impl Device

This impl block contains no items.

AVCaptureDeviceToneMapping

impl Device

AVCaptureDeviceWhiteBalance

pub fn is_wb_mode_supported(&self, mode: WbMode) -> bool

pub fn wb_mode(&self) -> WbMode

pub fn is_adjusting_wb(&self) -> bool

impl Device

AVCaptureDeviceVideoZoom

pub fn video_zoom_factor(&self) -> Float

pub fn is_ramping_video_zoom(&self) -> bool

pub unsafe fn ramp_to_video_zoom_factor_throws( &mut self, factor: Float, rate: f32, )

pub unsafe fn cancel_video_zoom_ramp_throws(&mut self)

pub fn display_video_zoom_factor_multiplier(&self) -> Float

impl Device

AVCaptureDeviceType

pub fn device_type(&self) -> &Type

impl Device

AVCaptureDeviceAuthorization

pub unsafe fn authorization_status_for_media_type_throws( media_type: &MediaType, ) -> AuthorizationStatus

pub fn authorization_status_for_media_type<'ear>( media_type: &MediaType, ) -> ExResult<'ear, AuthorizationStatus>

pub unsafe fn request_access_for_media_type_ch_throws( media_type: &MediaType, block: &mut SendBlock<fn(bool)>, )

pub fn request_access_for_media_type_ch<'ear>( media_type: &MediaType, block: &mut SendBlock<fn(bool)>, ) -> ExResult<'ear>

pub async fn request_access_for_media_type( media_type: &MediaType, ) -> Result<bool, R<Exception>>

impl Device

CameraLensSmudgeDetection

pub fn sel_set_cam_lens_smudge_detection_enabled_throws() -> &'static Sel

@selector(setCameraLensSmudgeDetectionEnabled:detectionInterval:) but dynamic use this function to check if object responds to selector

pub unsafe fn set_cam_lens_smudge_detection_enabled_throws( &mut self, val: bool, detection_interval: Time, )

pub fn sel_is_cam_lens_smudge_detection_enabled() -> &'static Sel

@selector(isCameraLensSmudgeDetectionEnabled) but dynamic use this function to check if object responds to selector

pub unsafe fn is_cam_lens_smudge_detection_enabled(&self) -> bool

pub fn sel_cam_lens_smudge_detection_interval() -> &'static Sel

@selector(cameraLensSmudgeDetectionInterval) but dynamic use this function to check if object responds to selector

pub unsafe fn cam_lens_smudge_detection_interval(&self) -> Time

pub fn sel_cam_lens_smudge_detection_status() -> &'static Sel

@selector(cameraLensSmudgeDetectionStatus) but dynamic use this function to check if object responds to selector

pub unsafe fn cam_lens_smudge_detection_status( &self, ) -> CamLensSmudgeDetectionStatus