cidre::av::capture::device

Struct Format

pub struct Format(/* private fields */);

An av::CaptureDeviceFormat wraps a cm::FormatDesc and other format-related information, such as min and max framerate.

Apple Documentation

Implementations

impl Format

pub fn retained(&self) -> R<Self>

impl Format

pub fn is_auto_video_frame_rate_supported(&self) -> bool

Indicates whether the device format supports auto video frame rate.

pub fn sel_is_auto_video_frame_rate_supported() -> &'static Sel

@selector(isAutoVideoFrameRateSupported) but dynamic use this function to check if object responds to selector

impl Format

Determining Reaction Effects Support

AVCaptureDeviceReactionEffects

pub fn reaction_effects_supported(&self) -> bool

Indicates whether the device supports reaction effects.

pub fn sel_reaction_effects_supported() -> &'static Sel

@selector(reactionEffectsSupported) but dynamic use this function to check if object responds to selector

pub fn video_frame_rate_range_for_reaction_effects_in_progress( &self, ) -> Option<&FrameRateRange>

The minimum and maximum frame rates available when a reaction effect runs.

Unlike other video effects, enabling reaction effects doesn’t limit the stream’s frame rate because most of the time the system isn’t rendering the effect. The frame rate only ramps down when the system renders a reaction on the stream.

pub fn sel_video_frame_rate_range_for_reaction_effects_in_progress() -> &'static Sel

@selector(videoFrameRateRangeForReactionEffectsInProgress) but dynamic use this function to check if object responds to selector

impl Format

Determining Supported Media Formats

pub fn media_type_ar(&self) -> Rar<MediaType>

pub fn media_type(&self) -> R<MediaType>

pub fn format_desc(&self) -> &FormatDesc

impl Format

Determining Output Support

pub fn unsupported_capture_output_classes_ar(&self) -> Rar<Array<Class<Id>>>

The list of capture output subclasses not allowed for capture with this format, if any.

pub fn unsupported_capture_output_classes(&self) -> R<Array<Class<Id>>>

The list of capture output subclasses not allowed for capture with this format, if any.

impl Format

Determining Video Capture Support

pub fn sel_is_video_hdr_supported() -> &'static Sel

@selector(isVideoHDRSupported) but dynamic use this function to check if object responds to selector

pub fn video_supported_frame_rate_ranges_ar(&self) -> Rar<Array<FrameRateRange>>

pub fn video_supported_frame_rate_ranges(&self) -> R<Array<FrameRateRange>>

impl Format

Determining Field of View

AVCaptureDeviceFormatGeometricDistortionCorrection

pub fn sel_geometric_distortion_corrected_video_fov() -> &'static Sel

@selector(geometricDistortionCorrectedVideoFieldOfView) but dynamic use this function to check if object responds to selector

impl Format

Determining Exposure

pub fn sel_min_exposure_duration() -> &'static Sel

@selector(minExposureDuration) but dynamic use this function to check if object responds to selector

pub fn sel_max_exposure_duration() -> &'static Sel

@selector(maxExposureDuration) but dynamic use this function to check if object responds to selector

pub fn sel_min_iso() -> &'static Sel

@selector(minISO) but dynamic use this function to check if object responds to selector

pub fn sel_max_iso() -> &'static Sel

@selector(maxISO) but dynamic use this function to check if object responds to selector

pub fn sel_is_global_tone_mapping_supported() -> &'static Sel

@selector(isGlobalToneMappingSupported) but dynamic use this function to check if object responds to selector

pub fn is_high_photo_quality_supported(&self) -> bool

pub fn sel_is_high_photo_quality_supported() -> &'static Sel

@selector(isHighPhotoQualitySupported) but dynamic use this function to check if object responds to selector

pub fn sel_is_highest_photo_quality_supported() -> &'static Sel

@selector(isHighestPhotoQualitySupported) but dynamic use this function to check if object responds to selector

pub fn auto_focus_sys(&self) -> AutoFocusSys

Indicating the autofocus system.

pub fn sel_auto_focus_sys() -> &'static Sel

@selector(autoFocusSystem) but dynamic use this function to check if object responds to selector

impl Format

Center Stage

AVCaptureDeviceFormatCenterStage

pub fn is_center_stage_supported(&self) -> bool

pub fn video_min_zoom_factor_for_center_stage(&self) -> Float

pub fn video_frame_rate_range_for_center_stage_ar( &self, ) -> Option<Rar<FrameRateRange>>

pub fn video_frame_rate_range_for_center_stage( &self, ) -> Option<R<FrameRateRange>>

impl Format

Portrait Effect

AVCaptureDeviceFormatPortraitEffect

pub fn is_portrait_effect_supported(&self) -> bool

Indicates whether the format supports the Portrait Effect feature.

pub fn sel_is_portrait_effect_supported() -> &'static Sel

@selector(isPortraitEffectSupported) but dynamic use this function to check if object responds to selector

pub fn video_frame_rate_range_for_portrait_effect_ar( &self, ) -> Option<Rar<FrameRateRange>>

pub fn sel_video_frame_rate_range_for_portrait_effect() -> &'static Sel

@selector(videoFrameRateRangeForPortraitEffect) but dynamic use this function to check if object responds to selector

pub fn video_frame_rate_range_for_portrait_effect( &self, ) -> Option<R<FrameRateRange>>

impl Format

AVCaptureDeviceFormatStudioLight

pub fn is_studio_light_supported(&self) -> bool

pub fn sel_is_studio_light_supported() -> &'static Sel

@selector(isStudioLightSupported) but dynamic use this function to check if object responds to selector

pub fn video_frame_rate_range_for_studio_light_ar( &self, ) -> Option<Rar<FrameRateRange>>

Devices may support a limited frame rate range when Studio Light is active. If this device format does not support Studio Light, this property returns None.

pub fn sel_video_frame_rate_range_for_studio_light() -> &'static Sel

@selector(videoFrameRateRangeForStudioLight) but dynamic use this function to check if object responds to selector

pub fn video_frame_rate_range_for_studio_light( &self, ) -> Option<R<FrameRateRange>>

Devices may support a limited frame rate range when Studio Light is active. If this device format does not support Studio Light, this property returns None.

impl Format

Determining Color Support

pub fn supported_color_spaces_ar(&self) -> Rar<Array<Number>>

The list of color spaces the format supports for image and video capture.

The ns::Number objects in this array contain av::CaptureColorSpace values. The ordering of the array is such that spaces with a narrower color gamut appear before those with wider color gamuts.

All devices and formats support capture in the sRGB color space. Some devices and formats can also capture in the P3 color space, which includes a much wider gamut of colors than the sRGB color space. (Content captured in the P3 color space is viewable on all devices. Devices without wide-color displays render P3 content as accurately as possible in the sRGB color gamut). By default, a capture session automatically enables wide-gamut capture for supported devices and capture workflows (for details, see the av::CaptureSession property automaticallyConfiguresCaptureDeviceForWideColor).

pub fn supported_color_spaces(&self) -> R<Array<Number>>

The list of color spaces the format supports for image and video capture.

The ns::Number objects in this array contain av::CaptureColorSpace values. The ordering of the array is such that spaces with a narrower color gamut appear before those with wider color gamuts.

All devices and formats support capture in the sRGB color space. Some devices and formats can also capture in the P3 color space, which includes a much wider gamut of colors than the sRGB color space. (Content captured in the P3 color space is viewable on all devices. Devices without wide-color displays render P3 content as accurately as possible in the sRGB color gamut). By default, a capture session automatically enables wide-gamut capture for supported devices and capture workflows (for details, see the av::CaptureSession property automaticallyConfiguresCaptureDeviceForWideColor).

pub fn color_space_supported(&self, val: CaptureColorSpace) -> bool

pub fn global_tone_mapping_supported(&self) -> bool

Whether the format supports global tone mapping.

impl Format

AVCaptureDeviceFormatMultiCamAdditions

pub fn sel_is_mutli_cam_supported() -> &'static Sel

@selector(isMultiCamSupported) but dynamic use this function to check if object responds to selector

impl Format

AVCaptureDeviceFormatSpatialVideoCapture

pub fn is_spatial_video_capture_supported(&self) -> bool

Returns whether or not the format supports capturing spatial video to a file.

pub fn sel_is_spatial_video_capture_supported() -> &'static Sel

@selector(isSpatialVideoCaptureSupported) but dynamic use this function to check if object responds to selector

impl Format

AVCaptureDeviceFormatBackgroundReplacement

pub fn is_background_replacement_supported(&self) -> bool

Indicates whether the format supports the Background Replacement feature.

This property returns YES if the format supports Background Replacement background replacement. See +AVCaptureDevice.backgroundReplacementEnabled.

pub fn sel_is_background_replacement_supported() -> &'static Sel

@selector(isBackgroundReplacementSupported) but dynamic use this function to check if object responds to selector

pub fn video_frame_rate_range_for_background_replacement_ar( &self, ) -> Option<Rar<FrameRateRange>>

Indicates the minimum / maximum frame rates available when background replacement is active.

Devices may support a limited frame rate range when Background Replacement is active. If this device format does not support Background Replacement, this property returns None.

pub fn sel_video_frame_rate_range_for_background_replacement() -> &'static Sel

@selector(videoFrameRateRangeForBackgroundReplacement) but dynamic use this function to check if object responds to selector

pub fn video_frame_rate_range_for_background_replacement( &self, ) -> Option<R<FrameRateRange>>

Indicates the minimum / maximum frame rates available when background replacement is active.

Devices may support a limited frame rate range when Background Replacement is active. If this device format does not support Background Replacement, this property returns None.

impl Format

AVCaptureDeviceFormatCinematicVideoSupport

pub fn sel_is_cinematic_video_capture_supported() -> &'static Sel

@selector(isCinematicVideoCaptureSupported) but dynamic use this function to check if object responds to selector

pub unsafe fn is_cinematic_video_capture_supported(&self) -> bool

pub fn sel_default_simulated_aperture() -> &'static Sel

@selector(defaultSimulatedAperture) but dynamic use this function to check if object responds to selector

pub unsafe fn default_simulated_aperture(&self) -> f32

pub fn sel_min_simulated_aperture() -> &'static Sel

@selector(minSimulatedAperture) but dynamic use this function to check if object responds to selector

pub unsafe fn min_simulated_aperture(&self) -> f32

pub fn sel_max_simulated_aperture() -> &'static Sel

@selector(maxSimulatedAperture) but dynamic use this function to check if object responds to selector

pub unsafe fn max_simulated_aperture(&self) -> f32

pub fn sel_video_min_zoom_factor_for_cinematic_video() -> &'static Sel

@selector(videoMinZoomFactorForCinematicVideo) but dynamic use this function to check if object responds to selector

pub unsafe fn video_min_zoom_factor_for_cinematic_video(&self) -> f32

pub fn sel_video_max_zoom_factor_for_cinematic_video() -> &'static Sel

@selector(videoMaxZoomFactorForCinematicVideo) but dynamic use this function to check if object responds to selector

pub unsafe fn video_max_zoom_factor_for_cinematic_video(&self) -> f32

pub fn sel_video_frame_rate_range_for_cinematic_video() -> &'static Sel

@selector(videoFrameRateRangeForCinematicVideo) but dynamic use this function to check if object responds to selector

pub unsafe fn video_frame_rate_range_for_cinematic_video_ar( &self, ) -> Option<Rar<FrameRateRange>>

pub unsafe fn video_frame_rate_range_for_cinematic_video( &self, ) -> Option<R<FrameRateRange>>

Check availability with selector1 Self::sel_video_frame_rate_range_for_cinematic_video()

impl Format

CameraLensSmudgeDetection

pub fn sel_is_cam_lens_smudge_detection_supported() -> &'static Sel

@selector(isCameraLensSmudgeDetectionSupported) but dynamic use this function to check if object responds to selector

pub unsafe fn is_cam_lens_smudge_detection_supported(&self) -> bool