cidre::av::capture

Module session

Modules

notifications

Structs

AnyControlsDelegate

AudioChannel

Represents a single channel of audio flowing through an av::CaptureSession

Connection

An object that represents a connection from a capture input to a capture output.

MultiCamSession

A subclass of av::CaptureSession which supports simultaneous capture from multiple inputs of the same media type.

Session

An object that configures capture behavior and coordinates the flow of data from input devices to capture outputs.

Enums

VideoOrienation

Constants indicating video orientation, for use with av::CaptureVideoPreviewLayer and av::CaptureConnection.

Traits

ControlsDelegate

ControlsDelegateImpl

Functions

err_key