cidre::av::metadata_object

Struct Type

pub struct Type(/* private fields */);

Implementations

impl Type

pub fn retained(&self) -> R<Self>

impl Type

pub fn human_body() -> &'static Type

pub fn human_full_body() -> &'static Type

pub fn cat_body() -> &'static Type

pub fn dog_body() -> &'static Type

pub fn salient_object() -> &'static Type

pub fn face() -> &'static Type

pub fn upc_e_code() -> &'static Type

objects generated from UPC-E codes return this constant as their type.

pub fn code_39_code() -> &'static Type

objects generated from Code 39 codes return this constant as their type.

pub fn code_39_mod_43_code() -> &'static Type

objects generated from Code 39 mod 43 codes return this constant as their type.

pub fn ean_13_code() -> &'static Type

objects generated from EAN-13 (including UPC-A) codes return this constant as their type.

pub fn ean_8_code() -> &'static Type

objects generated from EAN-8 codes return this constant as their type.

pub fn code_93_code() -> &'static Type

objects generated from Code 93 codes return this constant as their type.

pub fn code_128_code() -> &'static Type

objects generated from Code 128 codes return this constant as their type.

pub fn pdf_417_code() -> &'static Type

objects generated from PDF417 codes return this constant as their type

pub fn qr_code() -> &'static Type

objects generated from QR codes return this constant as their type.

pub fn aztec_code() -> &'static Type

objects generated from Aztec codes return this constant as their type.

pub fn interleaved_2_to_5_code() -> &'static Type

objects generated from Interleaved 2 of 5 codes return this constant as their type.

pub fn itf_14_code() -> &'static Type

objects generated from ITF14 codes return this constant as their type.

pub fn data_matrix_code() -> &'static Type

objects generated from DataMatrix codes return this constant as their type.

pub fn codebar_code() -> &'static Type

objects generated from Codabar codes return this constant as their type.

pub fn gs_1_data_bar_code() -> &'static Type

objects generated from GS1DataBar codes return this constant as their type.

pub fn gs_1_data_bar_expanded_code() -> &'static Type

objects generated from GS1DataBarExpanded codes return this constant as their type.

pub fn gs_1_data_bar_limited_code() -> &'static Type

objects generated from GS1DataBarLimited codes return this constant as their type.

pub fn micro_qr_code() -> &'static Type

objects generated from MicroQR codes return this constant as their type.

pub fn micro_pdf_417_code() -> &'static Type

objects generated from MicroPDF417 codes return this constant as their type.

pub unsafe fn cat_head() -> Option<&'static Type>

Availability

Not macos_26_0, ios_26_0, tvos_26_0, maccatalyst_26_0

pub unsafe fn dog_head() -> Option<&'static Type>

Availability

Not macos_26_0, ios_26_0, tvos_26_0, maccatalyst_26_0