cidre::av::player::item

Struct Item

pub struct Item(/* private fields */);

Implementations

impl Item

pub fn retained(&self) -> R<Self>

impl Item

pub fn cls() -> &'static Class<Self>

pub fn cls_ptr() -> *const c_void

pub fn alloc() -> A<Self>

pub fn with_url_ar(val: &Url) -> Rar<Self>

pub fn with_url(val: &Url) -> R<Self>

pub fn with_asset_ar(val: &Asset) -> Rar<Self>

pub fn with_asset(val: &Asset) -> R<Self>

pub fn with_asset_auto_loaded_keys_ar( val: &Asset, asset_keys: Option<&Array<String>>, ) -> Rar<Self>

pub fn with_asset_auto_loaded_keys( val: &Asset, asset_keys: Option<&Array<String>>, ) -> R<Self>

pub fn copy_ar(&self) -> Rar<Self>

pub fn copy(&self) -> R<Self>

pub fn status(&self) -> Status

pub fn error_ar(&self) -> Option<Rar<Error>>

If the receiver’s status is AVPlayerItemStatusFailed, this describes the error that caused the failure.

The value of this property is an NSError that describes what caused the receiver to no longer be able to be played. If the receiver’s status is not AVPlayerItemStatusFailed, the value of this property is nil.

pub fn error(&self) -> Option<R<Error>>

If the receiver’s status is AVPlayerItemStatusFailed, this describes the error that caused the failure.

The value of this property is an NSError that describes what caused the receiver to no longer be able to be played. If the receiver’s status is not AVPlayerItemStatusFailed, the value of this property is nil.

impl Item

AVPlayerItemInspection

pub fn asset_ar(&self) -> Rar<Asset>

Accessor for underlying av::Asset.

pub fn asset(&self) -> R<Asset>

Accessor for underlying av::Asset.

pub fn tracks_ar(&self) -> Rar<Array<PlayerItemTrack>>

pub fn tracks(&self) -> R<Array<PlayerItemTrack>>

pub fn duration(&self) -> Time

pub fn presentation_size(&self) -> Size

pub fn auto_loaded_asset_keys_ar(&self) -> Rar<Array<String>>

pub fn auto_loaded_asset_keys(&self) -> R<Array<String>>

impl Item

AVPlayerItemRateAndSteppingSupport

pub fn can_play_fast_forward(&self) -> bool

pub fn can_play_slow_forward(&self) -> bool

pub fn can_play_reverse(&self) -> bool

pub fn can_play_slow_reverse(&self) -> bool

pub fn can_play_fast_reverse(&self) -> bool

pub fn can_step_forward(&self) -> bool

pub fn can_step_backward(&self) -> bool

pub fn configured_time_offset_from_live(&self) -> bool

pub fn set_configured_time_offset_from_live(&mut self, val: Time)

pub fn recommended_time_offset_from_live(&self) -> Time

pub fn automatically_preserves_time_offset_from_live(&self) -> bool

pub fn set_automatically_preserves_time_offset_from_live(&mut self, val: bool)

impl Item

AVPlayerItemTimeControl

pub fn current_time(&self) -> Time

pub fn forward_playback_end_time(&self) -> Time

pub fn set_forward_playback_end_time(&mut self, val: Time)

pub fn reverse_playback_end_time(&self) -> Time

pub fn set_reverse_playback_end_time(&mut self, val: Time)

pub fn seekable_time_ranges_ar(&self) -> Rar<Array<Value>>

pub fn seekable_time_ranges(&self) -> R<Array<Value>>

pub unsafe fn seek_to_time_ch_throws( &mut self, val: Time, block: Option<&mut SendBlock<fn(finished: bool)>>, )

pub fn seek_to_time_ch<'ear>( &mut self, val: Time, block: Option<&mut SendBlock<fn(finished: bool)>>, ) -> ExResult<'ear>

pub unsafe fn seek_to_time_with_tolerance_ch_throws( &mut self, val: Time, tolerance_befor: Time, tolerance_after: Time, block: Option<&mut SendBlock<fn(finished: bool)>>, )

pub fn seek_to_time_with_tolerance_ch<'ear>( &mut self, val: Time, tolerance_befor: Time, tolerance_after: Time, block: Option<&mut SendBlock<fn(finished: bool)>>, ) -> ExResult<'ear>

pub fn cancel_pending_seeks(&mut self)

pub fn current_date_ar(&self) -> Option<Rar<Date>>

pub fn current_date(&self) -> Option<R<Date>>

pub fn step_by_count(&mut self, steps_count: Integer)

pub fn timebase(&self) -> &Timebase

pub fn timebase_mut(&mut self) -> &mut Timebase

impl Item

This impl block contains no items.

AVPlayerItemVisualPresentation

impl Item

AVPlayerItemAudioProcessing

pub fn audio_time_pitch_algorithm_ar( &self, ) -> Option<Rar<AudioTimePitchAlgorithm>>

pub fn audio_time_pitch_algorithm(&self) -> Option<R<AudioTimePitchAlgorithm>>

pub fn set_audio_time_pitch_algorithm( &mut self, val: Option<&AudioTimePitchAlgorithm>, )

pub fn allowed_audio_spatialization_formats(&self) -> AudioSpatializationFormats

pub fn set_allowed_audio_spatialization_formats( &mut self, val: AudioSpatializationFormats, )

pub fn audio_mix_ar(&self) -> Option<Rar<AudioMix>>

pub fn audio_mix(&self) -> Option<R<AudioMix>>

pub fn set_audio_mix(&mut self, val: Option<&AudioMix>)

impl Item

AVPlayerItemPlayability

pub fn loaded_time_ranges_ar(&self) -> Rar<Array<Value>>

pub fn loaded_time_ranges(&self) -> R<Array<Value>>

pub fn is_playback_likely_to_keep_up(&self) -> bool

pub fn is_playback_buffer_full(&self) -> bool

pub fn is_playback_buffer_empty(&self) -> bool

pub fn can_use_network_resources_for_live_streaming_while_paused(&self) -> bool

pub fn set_can_use_network_resources_for_live_streaming_while_paused( &mut self, val: bool, )

pub fn preferred_forward_buf_duration(&self) -> TimeInterval

pub fn set_preferred_forward_buf_duration(&mut self, val: TimeInterval)

impl Item

AVPlayerItemOutputs

pub fn add_output(&mut self, output: &PlayerItemOutput)

Adds the specified instance of AVPlayerItemOutput to the receiver’s collection of outputs.

pub fn remove_output(&mut self, output: &PlayerItemOutput)

Removes the specified instance of AVPlayerItemOutput from the receiver’s collection of outputs.

pub fn outputs_ar(&self) -> Rar<Array<PlayerItemOutput>>

The collection of associated outputs.

pub fn outputs(&self) -> R<Array<PlayerItemOutput>>

The collection of associated outputs.