cidre::ax

Module action

Many UiElements have a set of actions that they can perform. Actions are designed to be simple. Actions roughly correspond to things you could do with a single click of the mouse on the UiElement. Buttons and menu items, for example, have a single action: push or pick, respectively. A scroll bar has several actions: page up, page down, up one line, down one line.

Functions

cancel

Simulate a Cancel action, such as hitting the Cancel button.

confirm

Simulate pressing Return in the UiElement, such as a text field.

decrement

Decrement the value of the UiElement.

increment

Increment the value of the UiElement.

pick

press

Simulate clicking the UiElement, such as a button.

raise

show_alternate_ui

Show alternate or hidden UI.

show_default_ui

Show default UI.

show_menu