cidre::ax

Module attr

Functions

allowed_values

A cf::Array ref of whatever type the element uses for its kAXValueAttribute.

alternate_ui_visible

ampm_field

cancel_button

children

children_in_nav_order

clear_button

close_button

column_count

column_header_ui_elements

column_index_range

column_title

column_titles

columns

contents

critical_value

day_field

decrement_button

default_button

desc

disclosed_by_row

disclosed_rows

disclosing

disclosure_level

document

edited

element_busy

enabled

expanded

extras_menu_bar

filename

focused

focused_app

focused_ui_element

focused_window

frame

frontmost

full_screen_button

grow_area

handles

header

help

A localized, human-readable cf::String ref that offers help content for an element.

hidden

horizontal_scroll_bar

horizontal_unit_desc

horizontal_units

hour_field

id

increment_button

incrementor

index

insertion_point_line_num

is_app_running

is_editable

label_ui_elements

label_value

linked_ui_elements

main

main_window

marker_type

marker_type_desc

marker_ui_elements

matte_content_ui_element

matte_hole

max_value

menu_bar

menu_item_cmd_char

menu_item_cmd_glyph

menu_item_cmd_modifiers

menu_item_cmd_virtual_key

menu_item_mark_char

menu_item_primary_ui_element

min_value

minimize_button

minimized

minute_field

modal

month_field

next_contents

num_chars

ordered_by_row

orientation

overflow_button

parent

placeholder_value

pos

previous_contents

proxy

role

Identifies the basic type of an element.

role_desc

row

row_count

row_header_ui_elements

row_index_range

rows

search_button

second_field

selected

selected_cells

selected_children

selected_columns

selected_rows

selected_text

selected_text_range

selected_text_ranges

serves_as_title_for_ui_elements

shared_character_range

shared_focus_elements

shared_text_ui_elements

shown_menu_ui_element

size

splitters

subrole

More specifically identifies the type of an element beyond the basic type provided by kAXRoleAttribute.

tabs

text

title

title_ui_element

toolbar_button

top_level_ui_element

unit_desc

units

url

value

value_desc

value_inc

value_wraps

vertical_scroll_bar

vertical_unit_desc

vertical_units

visible_cells

visible_character_range

visible_children

visible_columns

visible_rows

visible_text

warning_value

window

windows

year_field

zoom_button