cidre::ax

Module err

Constants

ACTION_UNSUPPORTED

The action is not supported by the ax::UiElement ref.

API_DISABLED

The accessibility API is disabled (as when, for example, the user deselects “Enable access for assistive devices” in Universal Access Preferences).

ATTR_UNSUPPORTED

The attribute is not supported by the ax::UiElement ref.

CANNOT_COMPLETE

The function cannot complete because messaging failed in some way or because the application with which the function is communicating is busy or unresponsive.

FAILURE

A system error occurred, such as the failure to allocate an object.

ILLEGAL_ARGUMENT

An illegal argument was passed to the function

INVALID_UI_ELEMENT

The ax::UiElement ref passed to the function is invalid.

INVALID_UI_ELEMENT_OBSERVER

The ax::Observer ref passed to the function is not a valid observer.

NOTIFICATION_ALREADY_REGISTERED

This notification has already been registered for.

NOTIFICATION_NOT_REGISTERED

Indicates that a notification is not registered yet.

NOTIFICATION_UNSUPPORTED

The notification is not supported by the ax::UiElement ref.

NOT_ENOUGH_PRECISION

Not enough precision.

NOT_IMPLEMENTED

Indicates that the function or method is not implemented (this can be returned if a process does not support the accessibility API).

NO_VALUE

The requested value or AXUIElementRef does not exist.

PARAMETERIZED_ATTR_UNSUPPORTED

The parameterized attribute is not supported by the ax::UiElement ref.