cidre::blocks

Struct Block

pub struct Block<Sig, Attr = NoEsc>(/* private fields */);

Implementations

impl<Sig, Attr> Block<Sig, Attr>

pub const fn with_fn0<R>( func: extern "C" fn(*const c_void) -> R, ) -> StaticBlock<Sig>

pub const fn with_fn1<A, R>( func: extern "C" fn(*const c_void, A) -> R, ) -> StaticBlock<Sig>

pub const fn with_fn2<A, B, R>( func: extern "C" fn(*const c_void, A, B) -> R, ) -> StaticBlock<Sig>

pub const fn with_fn3<A, B, C, R>( func: extern "C" fn(*const c_void, A, B, C) -> R, ) -> StaticBlock<Sig>

pub const fn with_fn4<A, B, C, D, R>( func: extern "C" fn(*const c_void, A, B, C, D) -> R, ) -> StaticBlock<Sig>

pub const fn with_fn5<A, B, C, D, E, R>( func: extern "C" fn(*const c_void, A, B, C, D, E) -> R, ) -> StaticBlock<Sig>

pub const fn with_fn6<A, B, C, D, E, F, R>( func: extern "C" fn(*const c_void, A, B, C, D, E, F) -> R, ) -> StaticBlock<Sig>

impl<Sig> Block<Sig, NoEsc>

pub const unsafe fn stack0<R, Closure>( closure: &mut Closure, ) -> StackBlock<'_, Closure, Sig>
where Sig: Fn() -> R, for<'c> Closure: FnMut() -> R,

pub const unsafe fn stack1<A, R, Closure>( closure: &mut Closure, ) -> StackBlock<'_, Closure, Sig>
where Sig: Fn(A) -> R, for<'c> Closure: FnMut(A) -> R,

pub const unsafe fn stack2<A, B, R, Closure>( closure: &mut Closure, ) -> StackBlock<'_, Closure, Sig>
where Sig: Fn(A, B) -> R, for<'c> Closure: FnMut(A, B) -> R,

pub const unsafe fn stack3<A, B, C, R, Closure>( closure: &mut Closure, ) -> StackBlock<'_, Closure, Sig>
where Sig: Fn(A, B, C) -> R, for<'c> Closure: FnMut(A, B, C) -> R,

pub const unsafe fn stack4<A, B, C, D, R, Closure>( closure: &mut Closure, ) -> StackBlock<'_, Closure, Sig>
where Sig: Fn(A, B, C, D) -> R, for<'c> Closure: FnMut(A, B, C, D) -> R,

pub const unsafe fn stack5<A, B, C, D, E, R, Closure>( closure: &mut Closure, ) -> StackBlock<'_, Closure, Sig>
where Sig: Fn(A, B, C, D, E) -> R, for<'c> Closure: FnMut(A, B, C, D, E) -> R,

pub const unsafe fn stack6<A, B, C, D, E, F, R, Closure>( closure: &mut Closure, ) -> StackBlock<'_, Closure, Sig>
where Sig: Fn(A, B, C, D, E, F) -> R, for<'c> Closure: FnMut(A, B, C, D, E, F) -> R,

pub fn new0<R, Closure>(closure: Closure) -> R<Self>
where Sig: Fn() -> R, for<'c> Closure: FnMut() -> R,

pub fn new1<A, R, Closure>(closure: Closure) -> R<Self>
where Sig: Fn(A) -> R, for<'c> Closure: FnMut(A) -> R,

pub fn new2<A, B, R, Closure>(closure: Closure) -> R<Self>
where Sig: Fn(A, B) -> R, for<'c> Closure: FnMut(A, B) -> R,

pub fn new3<A, B, C, R, Closure>(closure: Closure) -> R<Self>
where Sig: Fn(A, B, C) -> R, for<'c> Closure: FnMut(A, B, C) -> R,

pub fn new4<A, B, C, D, R, Closure>(closure: Closure) -> R<Self>
where Sig: Fn(A, B, C, D) -> R, for<'c> Closure: FnMut(A, B, C, D) -> R,

pub fn new5<A, B, C, D, E, R, Closure>(closure: Closure) -> R<Self>
where Sig: Fn(A, B, C, D, E) -> R, for<'c> Closure: FnMut(A, B, C, D, E) -> R,

pub fn new6<A, B, C, D, E, F, R, Closure>(closure: Closure) -> R<Self>
where Sig: Fn(A, B, C, D, E, F) -> R, for<'c> Closure: FnMut(A, B, C, D, E, F) -> R,

impl<Sig> Block<Sig, Esc>

pub fn new0<R, Closure>(closure: Closure) -> R<Self>
where Sig: Fn() -> R, for<'c> Closure: FnMut() -> R + 'static,

pub fn new1<A, R, Closure>(closure: Closure) -> R<Self>
where Sig: Fn(A) -> R, for<'c> Closure: FnMut(A) -> R + 'static,

pub fn new2<A, B, R, Closure>(closure: Closure) -> R<Self>
where Sig: Fn(A, B) -> R, for<'c> Closure: FnMut(A, B) -> R + 'static,

pub fn new3<A, B, C, R, Closure>(closure: Closure) -> R<Self>
where Sig: Fn(A, B, C) -> R, for<'c> Closure: FnMut(A, B, C) -> R + 'static,

pub fn new4<A, B, C, D, R, Closure>(closure: Closure) -> R<Self>
where Sig: Fn(A, B, C, D) -> R, for<'c> Closure: FnMut(A, B, C, D) -> R + 'static,

pub fn new5<A, B, C, D, E, R, Closure>(closure: Closure) -> R<Self>
where Sig: Fn(A, B, C, D, E) -> R, for<'c> Closure: FnMut(A, B, C, D, E) -> R + 'static,

pub fn new6<A, B, C, D, E, F, R, Closure>(closure: Closure) -> R<Self>
where Sig: Fn(A, B, C, D, E, F) -> R, for<'c> Closure: FnMut(A, B, C, D, E, F) -> R + 'static,

pub fn as_noesc_mut(&mut self) -> &mut Block<Sig, NoEsc>

impl<Sig> Block<Sig, Send>

pub fn new0<R, Closure>(closure: Closure) -> R<Self>
where Sig: Fn() -> R, for<'c> Closure: FnMut() -> R + 'static + MarkerSend,

pub fn new1<A, R, Closure>(closure: Closure) -> R<Self>
where Sig: Fn(A) -> R, for<'c> Closure: FnMut(A) -> R + 'static + MarkerSend,

pub fn new2<A, B, R, Closure>(closure: Closure) -> R<Self>
where Sig: Fn(A, B) -> R, for<'c> Closure: FnMut(A, B) -> R + 'static + MarkerSend,

pub fn new3<A, B, C, R, Closure>(closure: Closure) -> R<Self>
where Sig: Fn(A, B, C) -> R, for<'c> Closure: FnMut(A, B, C) -> R + 'static + MarkerSend,

pub fn new4<A, B, C, D, R, Closure>(closure: Closure) -> R<Self>
where Sig: Fn(A, B, C, D) -> R, for<'c> Closure: FnMut(A, B, C, D) -> R + 'static + MarkerSend,

pub fn new5<A, B, C, D, E, R, Closure>(closure: Closure) -> R<Self>
where Sig: Fn(A, B, C, D, E) -> R, for<'c> Closure: FnMut(A, B, C, D, E) -> R + 'static + MarkerSend,

pub fn new6<A, B, C, D, E, F, R, Closure>(closure: Closure) -> R<Self>
where Sig: Fn(A, B, C, D, E, F) -> R, for<'c> Closure: FnMut(A, B, C, D, E, F) -> R + 'static + MarkerSend,

pub fn as_esc_mut(&mut self) -> &mut Block<Sig, Esc>

pub fn as_noesc_mut(&mut self) -> &mut Block<Sig, NoEsc>

impl<Sig> Block<Sig, Sync>

pub fn new0<R, Closure>(closure: Closure) -> R<Self>
where Sig: Fn() -> R, for<'c> Closure: FnMut() -> R + 'static + MarkerSync,

pub fn new1<A, R, Closure>(closure: Closure) -> R<Self>
where Sig: Fn(A) -> R, for<'c> Closure: FnMut(A) -> R + 'static + MarkerSync,

pub fn new2<A, B, R, Closure>(closure: Closure) -> R<Self>
where Sig: Fn(A, B) -> R, for<'c> Closure: FnMut(A, B) -> R + 'static + MarkerSync,

pub fn new3<A, B, C, R, Closure>(closure: Closure) -> R<Self>
where Sig: Fn(A, B, C) -> R, for<'c> Closure: FnMut(A, B, C) -> R + 'static + MarkerSync,

pub fn new4<A, B, C, D, R, Closure>(closure: Closure) -> R<Self>
where Sig: Fn(A, B, C, D) -> R, for<'c> Closure: FnMut(A, B, C, D) -> R + 'static + MarkerSync,

pub fn new5<A, B, C, D, E, R, Closure>(closure: Closure) -> R<Self>
where Sig: Fn(A, B, C, D, E) -> R, for<'c> Closure: FnMut(A, B, C, D, E) -> R + 'static + MarkerSync,

pub fn new6<A, B, C, D, E, F, R, Closure>(closure: Closure) -> R<Self>
where Sig: Fn(A, B, C, D, E, F) -> R, for<'c> Closure: FnMut(A, B, C, D, E, F) -> R + 'static + MarkerSync,

pub fn as_send_mut(&mut self) -> &mut Block<Sig, Send>

pub fn as_esc_mut(&mut self) -> &mut Block<Sig, Esc>

pub fn as_noesc_mut(&mut self) -> &mut Block<Sig, NoEsc>

impl<R, Attr> Block<fn() -> R, Attr>

pub fn call(&mut self) -> R

impl<A, R, Attr> Block<fn(A) -> R, Attr>

pub fn call(&mut self, a: A) -> R

impl<A, B, R, Attr> Block<fn(A, B) -> R, Attr>

pub fn call(&mut self, a: A, b: B) -> R

impl<A, B, C, R, Attr> Block<fn(A, B, C) -> R, Attr>

pub fn call(&mut self, a: A, b: B, c: C) -> R

impl<A, B, C, D, R, Attr> Block<fn(A, B, C, D) -> R, Attr>

pub fn call(&mut self, a: A, b: B, c: C, d: D) -> R

impl<A, B, C, D, E, R, Attr> Block<fn(A, B, C, D, E) -> R, Attr>

pub fn call(&mut self, a: A, b: B, c: C, d: D, e: E) -> R

impl<A, B, C, D, E, F, R, Attr> Block<fn(A, B, C, D, E, F) -> R, Attr>

pub fn call(&mut self, a: A, b: B, c: C, d: D, e: E, f: F) -> R

impl<A, B, C, D, E, F, G, R, Attr> Block<fn(A, B, C, D, E, F, G) -> R, Attr>

pub fn call(&mut self, a: A, b: B, c: C, d: D, e: E, f: F, g: G) -> R

impl<A, B, C, D, E, F, G, H, R, Attr> Block<fn(A, B, C, D, E, F, G, H) -> R, Attr>

pub fn call(&mut self, a: A, b: B, c: C, d: D, e: E, f: F, g: G, h: H) -> R

impl<A, B, C, D, E, F, G, H, I, R, Attr> Block<fn(A, B, C, D, E, F, G, H, I) -> R, Attr>

pub fn call( &mut self, a: A, b: B, c: C, d: D, e: E, f: F, g: G, h: H, i: I, ) -> R

impl<A, B, C, D, E, F, G, H, I, J, R, Attr> Block<fn(A, B, C, D, E, F, G, H, I, J) -> R, Attr>

pub fn call( &mut self, a: A, b: B, c: C, d: D, e: E, f: F, g: G, h: H, i: I, j: J, ) -> R

impl<A, B, C, D, E, F, G, H, I, J, K, R, Attr> Block<fn(A, B, C, D, E, F, G, H, I, J, K) -> R, Attr>

pub fn call( &mut self, a: A, b: B, c: C, d: D, e: E, f: F, g: G, h: H, i: I, j: J, k: K, ) -> R

impl<A, Attr> Block<fn(a: Option<&A>), Attr>

pub fn handle(&mut self, a: Option<&A>)

impl<A, B, Attr> Block<fn(a: Option<&A>, b: Option<&B>), Attr>

pub fn handle(&mut self, a: Option<&A>, b: Option<&B>)