cidre::ca

Struct MetalLayer

pub struct MetalLayer(/* private fields */);

Implementations

impl MetalLayer

pub fn retained(&self) -> R<Self>

impl MetalLayer

pub fn cls() -> &'static Class<Self>

pub fn cls_ptr() -> *const c_void

pub fn alloc() -> A<Self>

pub fn new() -> R<MetalLayer>

shortcut to Self::alloc().init()

impl MetalLayer

pub fn device(&self) -> Option<&Device>

pub fn preferred_device(&self) -> Option<&Device>

pub fn set_device(&mut self, val: Option<&Device>)

pub fn next_drawable_ar(&self) -> Option<Rar<AnyMetalDrawable>>

pub fn next_drawable(&self) -> Option<R<AnyMetalDrawable>>

pub fn pixel_format(&self) -> PixelFormat

This property controls the pixel format of the mtl::Texture objects.

pub fn set_pixel_format(&mut self, val: PixelFormat)

The default value is MTLPixelFormatBGRA8Unorm.

You must use one of the following formats: MTLPixelFormatBGRA8Unorm MTLPixelFormatBGRA8Unorm_sRGB MTLPixelFormatRGBA16Float MTLPixelFormatRGB10A2Unorm (macOS only) MTLPixelFormatBGR10A2Unorm (macOS only) MTLPixelFormatBGRA10_XR MTLPixelFormatBGRA10_XR_sRGB MTLPixelFormatBGR10_XR MTLPixelFormatBGR10_XR_sRGB

pub fn framebuffer_only(&self) -> bool

pub fn set_framebuffer_only(&self, val: bool)

pub fn drawable_size(&self) -> Size

pub fn set_drawable_size(&mut self, val: Size)

pub fn maximum_drawable_count(&self) -> usize

The number of Metal drawables in the resource pool managed by Core Animation.

pub fn set_maximum_drawable_count(&self, val: usize)

The number of Metal drawables in the resource pool managed by Core Animation.

You can set this value to 2 or 3 only; if you pass a different value, Core Animation ignores the value and throws an exception.

The default value is 3.

pub fn presents_with_transaction(&self) -> bool

pub fn set_presents_with_transaction(&mut self, val: bool)

pub fn colorspace(&self) -> Option<&ColorSpace>

pub fn set_colorspace(&mut self, val: Option<&ColorSpace>)

pub fn wants_extended_dynamic_range_content(&self) -> bool

pub fn sel_wants_extended_dynamic_range_content() -> &'static Sel

@selector(wantsExtendedDynamicRangeContent) but dynamic use this function to check if object responds to selector

pub fn set_wants_extended_dynamic_range_content(&mut self, val: bool)

pub fn sel_set_wants_extended_dynamic_range_content() -> &'static Sel

@selector(setWantsExtendedDynamicRangeContent:) but dynamic use this function to check if object responds to selector

pub fn edr_metadata_ar(&self) -> Option<Rar<EdrMetadata>>

pub fn sel_edr_metadata() -> &'static Sel

@selector(EDRMetadata) but dynamic use this function to check if object responds to selector

pub fn edr_metadata(&self) -> Option<R<EdrMetadata>>

pub fn set_edr_metadata(&mut self, val: Option<&EdrMetadata>)

pub fn sel_set_edr_metadata() -> &'static Sel

@selector(setEDRMetadata:) but dynamic use this function to check if object responds to selector

pub fn display_sync_enabled(&self) -> bool

pub fn sel_display_sync_enabled() -> &'static Sel

@selector(displaySyncEnabled) but dynamic use this function to check if object responds to selector

pub fn set_display_sync_enabled(&mut self, val: bool)

pub fn sel_set_display_sync_enabled() -> &'static Sel

@selector(setDisplaySyncEnabled:) but dynamic use this function to check if object responds to selector

pub fn allows_next_drawable_timeout(&self) -> bool

pub fn sel_allows_next_drawable_timeout() -> &'static Sel

@selector(allowsNextDrawableTimeout) but dynamic use this function to check if object responds to selector

pub fn set_allows_next_drawable_timeout(&mut self, val: bool)

pub fn sel_set_allows_next_drawable_timeout() -> &'static Sel

@selector(setAllowsNextDrawableTimeout:) but dynamic use this function to check if object responds to selector

pub fn developer_hud_props_ar(&self) -> Option<Rar<Dictionary<String, Id>>>

pub fn sel_developer_hud_props() -> &'static Sel

@selector(developerHUDProperties) but dynamic use this function to check if object responds to selector

pub fn developer_hud_props(&self) -> Option<R<Dictionary<String, Id>>>

pub fn set_developer_hud_props(&mut self, val: Option<&Dictionary<String, Id>>)

pub fn sel_set_developer_hud_props() -> &'static Sel

@selector(setDeveloperHUDProperties:) but dynamic use this function to check if object responds to selector

pub fn sel_residency_set() -> &'static Sel

@selector(residencySet) but dynamic use this function to check if object responds to selector

pub unsafe fn residency_set_ar(&self) -> Rar<ResidencySet>

pub unsafe fn residency_set(&self) -> R<ResidencySet>

Check availability with selector Self::sel_residency_set()