cidre::ca

Trait LayerDelegate

pub trait LayerDelegate: Obj {
    // Provided methods
    fn display_layer(&mut self, layer: &mut Layer) { ... }
    fn sel_display_layer() -> &'static Sel { ... }
    fn draw_layer(&mut self, layer: &mut Layer, context: &mut Context) { ... }
    fn sel_draw_layer() -> &'static Sel { ... }
    fn layer_will_draw(&mut self, layer: &mut Layer) { ... }
    fn sel_layer_will_draw() -> &'static Sel { ... }
    fn layout_sublayers_of_layer(&mut self, layer: &mut Layer) { ... }
    fn sel_layout_sublayers_of_layer() -> &'static Sel { ... }
    fn action_for_layer_for_key_ar(
        &mut self,
        layer: &mut Layer,
        key: &String,
    ) -> Option<Rar<AnyAction>> { ... }
    fn action_for_layer_for_key(
        &mut self,
        layer: &mut Layer,
        key: &String,
    ) -> Option<R<AnyAction>> { ... }
    fn sel_action_for_layer_for_key() -> &'static Sel { ... }
}

Provided Methods

fn display_layer(&mut self, layer: &mut Layer)

fn sel_display_layer() -> &'static Sel

@selector(displayLayer:) but dynamic use this function to check if object responds to selector

fn draw_layer(&mut self, layer: &mut Layer, context: &mut Context)

fn sel_draw_layer() -> &'static Sel

@selector(drawLayer:inContext:) but dynamic use this function to check if object responds to selector

fn layer_will_draw(&mut self, layer: &mut Layer)

fn sel_layer_will_draw() -> &'static Sel

@selector(layerWillDraw:) but dynamic use this function to check if object responds to selector

fn layout_sublayers_of_layer(&mut self, layer: &mut Layer)

fn sel_layout_sublayers_of_layer() -> &'static Sel

@selector(layoutSublayersOfLayer:) but dynamic use this function to check if object responds to selector

fn action_for_layer_for_key_ar( &mut self, layer: &mut Layer, key: &String, ) -> Option<Rar<AnyAction>>

fn action_for_layer_for_key( &mut self, layer: &mut Layer, key: &String, ) -> Option<R<AnyAction>>

fn sel_action_for_layer_for_key() -> &'static Sel

@selector(actionForLayer:forKey:) but dynamic use this function to check if object responds to selector

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.