cidre::cat

Module audio

Modules

err

These are the error codes returned from the APIs found through Core Audio related frameworks.

Structs

Buf

A structure to hold a buffer of audio data.

BufList

A variable length array of AudioBuffer structures.

BufListCursor

BufListN

ChannelBitmap

These constants are for use in the mChannelBitmap field of an AudioChannelLayout structure

ChannelCoordinateIndex

Constants for indexing the mCoordinates array in an AudioChannelDescription structure.

ChannelDesc

This structure describes a single channel.

ChannelFlags

ChannelLabel

A tag identifying how the channel is to be used.

ChannelLayout

ChannelLayoutTag

Some channel abbreviations used below: L - left R - right C - center Ls - left surround Rs - right surround Cs - center surround Rls - rear left surround Rrs - rear right surround Lw - left wide Rw - right wide Lsd - left surround direct Rsd - right surround direct Lc - left center Rc - right center Ts - top surround Vhl - vertical height left Vhc - vertical height center Vhr - vertical height right Ltm - left top middle Rtm - right top middle Ltr - left top rear Ctr - center top rear Rtr - right top rear Lt - left matrix total. for matrix encoded stereo. Rt - right matrix total. for matrix encoded stereo.

ClassDesc

This structure is used to describe codecs installed on the system.

Format

A four char code indicating the general kind of data in the stream.

FormatFlags

FormatListItem

Mpeg4Object

SMPTETime

SMPTETimeFlags

SMPTETimeType

SessionErrorCode

StreamBasicDesc

This structure encapsulates all the information for describing the basic format properties of a stream of audio data.

StreamPacketDesc

This structure describes the packet layout of a buffer of data where the size of each packet may not be the same or where there is extraneous data between packets.

TimeStamp

TimeStampFlags

ValueRange

This structure holds a pair of numbers that represent a continuous range of values.

Constants

STREAM_ANY_RATE

The format can use any sample rate. Note that this constant can only appear in listings of supported formats. It will never appear in a current format.

Type Aliases

Integer

SessionId

UInteger