cidre::cat::audio

Struct ChannelLabel

#[repr(transparent)]
pub struct ChannelLabel(pub u32);

A tag identifying how the channel is to be used.

Tuple Fields
0: u32
Implementations

impl ChannelLabel

pub const UNKNOWN: Self

unknown or unspecified other use

pub const UNUSED: Self

channel is present, but has no intended use or destination

pub const USE_COORDINATES: Self

channel is described by the mCoordinates fields.

pub const LEFT: Self

pub const RIGHT: Self

pub const CENTER: Self

pub const LFE_SCREEN: Self

pub const LEFT_SURROUND: Self

pub const RIGHT_SURROUND: Self

pub const LEFT_CENTER: Self

pub const RIGHT_CENTER: Self

pub const CENTER_SURROUND: Self

WAVE: “Back Center” or plain “Rear Surround”

pub const LEFT_SURROUND_DIRECT: Self

pub const RIGHT_SURROUND_DIRECT: Self

pub const TOP_CENTER_SURROUND: Self

pub const VERTICAL_HEIGHT_LEFT: Self

WAVE: “Top Front Left

pub const VERTICAL_HEIGHT_CENTER: Self

WAVE: “Top Front Center”

pub const VERTICAL_HEIGHT_RIGHT: Self

WAVE: “Top Front Right”

pub const TOP_BACK_LEFT: Self

pub const TOP_BACK_CENTER: Self

pub const TOP_BACK_RIGHT: Self

pub const REAR_SURROUND_LEFT: Self

pub const REAR_SURROUND_RIGHT: Self

pub const LEFT_WIDE: Self

pub const RIGHT_WIDE: Self

pub const LFE2: Self

pub const LEFT_TOTAL: Self

matrix encoded 4 channels

pub const RIGHT_TOTAL: Self

matrix encoded 4 channels

pub const HEARING_IMPAIRED: Self

pub const NARRATION: Self

pub const MONO: Self

pub const DIALOG_CENTRIC_MIX: Self

pub const CENTER_SURROUND_DIRECT: Self

back center, non diffuse

pub const HAPTIC: Self

pub const LEFT_TOP_FRONT: Self = Self::VERTICAL_HEIGHT_LEFT

pub const CENTER_TOP_FRONT: Self = Self::VERTICAL_HEIGHT_CENTER

pub const RIGHT_TOP_FRONT: Self = Self::VERTICAL_HEIGHT_RIGHT

pub const LEFT_TOP_MIDDLE: Self

pub const CENTER_TOP_MIDDLE: Self = Self::TOP_CENTER_SURROUND

pub const RIGHT_TOP_MIDDLE: Self

pub const LEFT_TOP_REAR: Self

pub const CENTER_TOP_REAR: Self

pub const RIGHT_TOP_REAR: Self

pub const AMBISONIC_W: Self

pub const AMBISONIC_X: Self

pub const AMBISONIC_Y: Self

pub const AMBISONIC_Z: Self

pub const MS_MID: Self

pub const MS_SIDE: Self

pub const XY_X: Self

pub const XY_Y: Self

pub const BINAURAL_LEFT: Self

pub const BINAURAL_RIGHT: Self

pub const HEADPHONES_LEFT: Self

pub const HEADPHONES_RIGHT: Self

pub const CLICK_TRACK: Self

pub const FOREIGN_LANGUAGE: Self

pub const DISCRETE: Self

pub const DISCRETE_0: Self

pub const DISCRETE_1: Self

pub const DISCRETE_2: Self

pub const DISCRETE_3: Self

pub const DISCRETE_4: Self

pub const DISCRETE_5: Self

pub const DISCRETE_6: Self

pub const DISCRETE_7: Self

pub const DISCRETE_8: Self

pub const DISCRETE_9: Self

pub const DISCRETE_10: Self

pub const DISCRETE_11: Self

pub const DISCRETE_12: Self

pub const DISCRETE_13: Self

pub const DISCRETE_14: Self

pub const DISCRETE_15: Self

pub const DISCRETE_65535: Self

pub const HOA_ACN: Self

pub const HOA_ACN_0: Self

pub const HOA_ACN_1: Self

pub const HOA_ACN_2: Self

pub const HOA_ACN_3: Self

pub const HOA_ACN_4: Self

pub const HOA_ACN_5: Self

pub const HOA_ACN_6: Self

pub const HOA_ACN_7: Self

pub const HOA_ACN_8: Self

pub const HOA_ACN_9: Self

pub const HOA_ACN_10: Self

pub const HOA_ACN_11: Self

pub const HOA_ACN_12: Self

pub const HOA_ACN_13: Self

pub const HOA_ACN_14: Self

pub const HOA_ACN_15: Self

pub const HOA_ACN_65024: Self

pub const HOA_SN3D: Self = Self::HOA_ACN_0

pub const HOA_N3D: Self

pub const BEGIN_RESERVED: Self

pub const END_RESERVED: Self