cidre::cat::audio

Struct ChannelLayoutTag

#[repr(transparent)]
pub struct ChannelLayoutTag(pub u32);

Some channel abbreviations used below: L - left R - right C - center Ls - left surround Rs - right surround Cs - center surround Rls - rear left surround Rrs - rear right surround Lw - left wide Rw - right wide Lsd - left surround direct Rsd - right surround direct Lc - left center Rc - right center Ts - top surround Vhl - vertical height left Vhc - vertical height center Vhr - vertical height right Ltm - left top middle Rtm - right top middle Ltr - left top rear Ctr - center top rear Rtr - right top rear Lt - left matrix total. for matrix encoded stereo. Rt - right matrix total. for matrix encoded stereo.

Tuple Fields
0: u32
Implementations

impl ChannelLayoutTag

These constants are used in the mChannelLayoutTag field of an AudioChannelLayout structure.

pub const USE_CHANNEL_DESCRIPTIONS: Self

use the array of AudioChannelDescriptions to define the mapping.

pub const USE_CHANNEL_BITMAP: Self

use the bitmap to define the mapping.

pub const MONO: Self

a standard mono stream

pub const STEREO: Self

a standard stereo stream (L R) - implied playback

pub const STEREO_HEADPHONES: Self

a standard stereo stream (L R) - implied headphone playback

pub const MATRIX_STEREO: Self

a matrix encoded stereo stream (Lt, Rt)

pub const MID_SIDE: Self

mid/side recording

pub const XY: Self

coincident mic pair (often 2 figure 8’s)

pub const BINAURAL: Self

binaural stereo (left, right)

pub const AMBISONIC_B_FORMAT: Self

W, X, Y, Z

pub const QUADRAPHONIC: Self

L R Ls Rs – 90 degree speaker separation

pub const PENTAGONAL: Self

L R Ls Rs C – 72 degree speaker separation

pub const HEXAGONAL: Self

L R Ls Rs C Cs – 60 degree speaker separation

pub const OCTAGONAL: Self

L R Ls Rs C Cs Lw Rw – 45 degree speaker separation

pub const CUBE: Self

left, right, rear left, rear right top left, top right, top rear left, top rear right

pub const MPEG_1_0: Self = Self::MONO

MPEG defined layouts C

pub const MPEG_2_0: Self = Self::STEREO

L R

pub const MPEG_3_0_A: Self

L R C

pub const MPEG_3_0_B: Self

C L R

pub const MPEG_4_0_A: Self

L R C Cs

pub const MPEG_4_0_B: Self

C L R Cs

pub const MPEG_5_0_A: Self

L R C Ls Rs

pub const MPEG_5_0_B: Self

L R Ls Rs C

pub const MPEG_5_0_C: Self

L C R Ls Rs

pub const MPEG_5_0_D: Self

C L R Ls Rs

pub const MPEG_5_1_A: Self

L R C LFE Ls Rs

pub const MPEG_5_1_B: Self

L R Ls Rs C LFE

pub const MPEG_5_1_C: Self

L C R Ls Rs LFE

pub const MPEG_5_1_D: Self

C L R Ls Rs LFE

pub const MPEG_6_1_A: Self

L R C LFE Ls Rs Cs

pub const MPEG_7_1_A: Self

L R C LFE Ls Rs Lc Rc

pub const MPEG_7_1_B: Self

C Lc Rc L R Ls Rs LFE (doc: IS-13818-7 MPEG2-AAC Table 3.1)

pub const MPEG_7_1_C: Self

L R C LFE Ls Rs Rls Rrs

pub const EMAGIC_DEFAULT_7_1: Self

L R Ls Rs C LFE Lc Rc

pub const SMPTE_DTV: Self

L R C LFE Ls Rs Lt Rt (kAudioChannelLayoutTag_ITU_5_1 plus a matrix encoded stereo mix)

pub const ITU_1_0: Self = Self::MONO

C

pub const ITU_2_0: Self = Self::STEREO

L R

pub const ITU_2_1: Self

L R Cs

pub const ITU_2_2: Self

L R Ls Rs

pub const ITU_3_0: Self = Self::MPEG_3_0_A

L R C

pub const ITU_3_1: Self = Self::MPEG_4_0_A

L R C Cs

pub const ITU_3_2: Self = Self::MPEG_5_0_A

< L R C Ls Rs

pub const ITU_3_2_1: Self = Self::MPEG_5_1_A

L R C LFE Ls Rs

pub const ITU_3_4_1: Self = Self::MPEG_7_1_C

L R C LFE Ls Rs Rls Rrs

pub const DVD_0: Self = Self::MONO

C (mono)

pub const DVD_1: Self = Self::STEREO

L R

pub const DVD_2: Self = Self::ITU_2_1

L R Cs

pub const DVD_3: Self = Self::ITU_2_2

L R Ls Rs

pub const DVD_4: Self

L R LFE

pub const DVD_5: Self

L R LFE Cs

pub const DVD_6: Self

L R LFE Ls Rs

pub const DVD_7: Self = Self::MPEG_3_0_A

L R C

pub const DVD_8: Self = Self::MPEG_4_0_A

L R C Cs

pub const DVD_9: Self = Self::MPEG_5_0_A

L R C Ls Rs

pub const DVD_10: Self

L R C LFE

pub const DVD_11: Self

L R C LFE Cs

pub const DVD_12: Self = Self::MPEG_5_1_A

L R C LFE Ls Rs

pub const DVD_13: Self = Self::DVD_8

L R C Cs

pub const DVD_14: Self = Self::DVD_9

L R C Ls Rs

pub const DVD_15: Self = Self::DVD_10

L R C LFE

pub const DVD_16: Self = Self::DVD_11

L R C LFE Cs

pub const DVD_17: Self = Self::DVD_12

L R C LFE Ls Rs

pub const DVD_18: Self

L R Ls Rs LFE

pub const DVD_19: Self = Self::MPEG_5_0_B

L R Ls Rs C

pub const DVD_20: Self = Self::MPEG_5_1_B

L R Ls Rs C LFE

pub const AUDIO_UNIT_4: Self = Self::QUADRAPHONIC

pub const AUDIO_UNIT_5: Self = Self::PENTAGONAL

pub const AUDIO_UNIT_6: Self = Self::HEXAGONAL

pub const AUDIO_UNIT_8: Self = Self::OCTAGONAL

pub const AUDIO_UNIT_5_0: Self = Self::MPEG_5_0_B

L R Ls Rs C

pub const AUDIO_UNIT_6_0: Self

L R Ls Rs C Cs

pub const AUDIO_UNIT_7_0: Self

L R Ls Rs C Rls Rrs

pub const AUDIO_UNIT_7_0_FRONT: Self

L R Ls Rs C Lc Rc

pub const AUDIO_UNIT_5_1: Self = Self::MPEG_5_1_A

L R C LFE Ls Rs

pub const AUDIO_UNIT_6_1: Self = Self::MPEG_6_1_A

L R C LFE Ls Rs Cs

pub const AUDIO_UNIT_7_1: Self = Self::MPEG_7_1_C

L R C LFE Ls Rs Rls Rrs

pub const AUDIO_UNIT_7_1_FRONT: Self = Self::MPEG_7_1_A

L R C LFE Ls Rs Lc Rc

pub const AAC_3_0: Self = Self::MPEG_3_0_B

pub const AAC_QUADRAPHONIC: Self = Self::QUADRAPHONIC

C L R

pub const AAC_4_0: Self = Self::MPEG_4_0_B

L R Ls Rs

pub const AAC_5_0: Self = Self::MPEG_5_0_D

C L R Cs

pub const AAC_5_1: Self = Self::MPEG_5_1_D

C L R Ls Rs

pub const AAC_6_0: Self

C L R Ls Rs Lfe

pub const AAC_6_1: Self

C L R Ls Rs Cs

pub const AAC_7_0: Self

C L R Ls Rs Cs Lfe

pub const AAC_7_1: Self = Self::MPEG_7_1_B

C L R Ls Rs Rls Rrs

pub const AAC_7_1_B: Self

C Lc Rc L R Ls Rs Lfe

pub const AAC_7_1_C: Self

C L R Ls Rs Rls Rrs LFE

pub const AAC_OCTAGONAL: Self

C L R Ls Rs LFE Vhl Vhr

pub const TMH_10_2_STD: Self

C L R Ls Rs Rls Rrs Cs L R C Vhc Lsd Rsd Ls Rs Vhl Vhr Lw Rw Csd Cs LFE1 LFE2

pub const TMH_10_2_FULL: Self

TMH_10_2_std plus: Lc Rc HI VI Haptic

pub const AC3_1_0_1: Self

C LFE

pub const AC3_3_0: Self

L C R

pub const AC3_3_1: Self

L C R Cs

pub const AC3_3_0_1: Self

L C R LFE

pub const AC3_2_1_1: Self

L R Cs LFE

pub const AC3_3_1_1: Self

L C R Cs LFE

pub const EAC_6_0_A: Self

L C R Ls Rs Cs

pub const EAC_7_0_A: Self

L C R Ls Rs Rls Rrs

pub const EAC3_6_1_A: Self

L C R Ls Rs LFE Cs

pub const EAC3_6_1_B: Self

L C R Ls Rs LFE Ts

pub const EAC3_6_1_C: Self

L C R Ls Rs LFE Vhc

pub const EAC3_7_1_A: Self

L C R Ls Rs LFE Rls Rrs

pub const EAC3_7_1_B: Self

L C R Ls Rs LFE Lc Rc

pub const EAC3_7_1_C: Self

L C R Ls Rs LFE Lsd Rsd

pub const EAC3_7_1_D: Self

L C R Ls Rs LFE Lw Rw

pub const EAC3_7_1_E: Self

L C R Ls Rs LFE Vhl Vhr

pub const EAC3_7_1_F: Self

L C R Ls Rs LFE Cs Ts

pub const EAC3_7_1_G: Self

L C R Ls Rs LFE Cs Vhc

pub const EAC3_7_1_H: Self

L C R Ls Rs LFE Ts Vhc

pub const DTS_3_1: Self

C L R LFE

pub const DTS_4_1: Self

C L R Cs LFE

pub const DTS_6_0_A: Self

Lc Rc L R Ls Rs

pub const DTS_6_0_B: Self

C L R Rls Rrs Ts

pub const DTS_6_0_C: Self

C Cs L R Rls Rrs

pub const DTS_6_1_A: Self

Lc Rc L R Ls Rs LFE

pub const DTS_6_1_B: Self

C L R Rls Rrs Ts LFE

pub const DTS_6_1_C: Self

C Cs L R Rls Rrs LFE

pub const DTS_7_0: Self

Lc C Rc L R Ls Rs

pub const DTS_7_1: Self

Lc C Rc L R Ls Rs LFE

pub const DTS_8_0_A: Self

Lc Rc L R Ls Rs Rls Rrs

pub const DTS_8_0_B: Self

Lc C Rc L R Ls Cs Rs

pub const DTS_8_1_A: Self

Lc Rc L R Ls Rs Rls Rrs LFE

pub const DTS_8_1_B: Self

Lc C Rc L R Ls Cs Rs LFE

pub const DTS_6_1_D: Self

C L R Ls Rs LFE Cs

pub const WAVE_2_1: Self = Self::DVD_4

3 channels, L R LFE

pub const WAVE_3_0: Self = Self::MPEG_3_0_A

3 channels, L R C

pub const WAVE_4_0_A: Self = Self::ITU_2_2

4 channels, L R Ls Rs

pub const WAVE_4_0_B: Self

4 channels, L R Rls Rrs

pub const WAVE_5_0_A: Self = Self::MPEG_5_0_A

5 channels, L R C Ls Rs

pub const WAVE_5_0_B: Self

5 channels, L R C Rls Rrs

pub const WAVE_5_1_A: Self = Self::MPEG_5_1_A

6 channels, L R C LFE Ls Rs

pub const WAVE_5_1_B: Self

6 channels, L R C LFE Rls Rrs

pub const WAVE_6_1: Self

7 channels, L R C LFE Cs Ls Rs

pub const WAVE_7_1: Self

8 channels, L R C LFE Rls Rrs Ls Rs

pub const HOA_ACN_SN3D: Self

Higher Order Ambisonics, Ambisonics Channel Number, SN3D normalization needs to be ORed with the actual number of channels (not the HOA order)

pub const HOA_ACN_N3D: Self

Higher Order Ambisonics, Ambisonics Channel Number, N3D normalization needs to be ORed with the actual number of channels (not the HOA order)

pub const ATMOS_5_1_2: Self

L R C LFE Ls Rs Ltm Rtm

pub const ATMOS_5_1_4: Self

L R C LFE Ls Rs Vhl Vhr Ltr Rtr

pub const ATMOS_7_1_2: Self

L R C LFE Ls Rs Rls Rrs Ltm Rtm

pub const ATMOS_7_1_4: Self

L R C LFE Ls Rs Rls Rrs Vhl Vhr Ltr Rtr

pub const ATMOS_9_1_6: Self

L R C LFE Ls Rs Rls Rrs Lw Rw Vhl Vhr Ltm Rtm Ltr Rtr

pub const LOGIC_MONO: Self = Self::MONO

C

pub const LOGIC_STEREO: Self = Self::STEREO

L R

pub const LOGIC_QUADRAPHONIC: Self = Self::QUADRAPHONIC

L R Ls Rs

pub const LOGIC_4_0_A: Self = Self::MPEG_4_0_A

L R C Cs

pub const LOGIC_4_0_B: Self = Self::MPEG_4_0_B

C L R Cs

pub const LOGIC_4_0_C: Self

L R Cs C

pub const LOGIC_5_0_A: Self = Self::MPEG_5_0_A

L R C Ls Rs

pub const LOGIC_5_0_B: Self = Self::MPEG_5_0_B

L R Ls Rs C

pub const LOGIC_5_0_C: Self = Self::MPEG_5_0_C

L C R Ls Rs

pub const LOGIC_5_0_D: Self = Self::MPEG_5_0_D

C L R Ls Rs

pub const LOGIC_5_1_A: Self = Self::MPEG_5_1_A

L R C LFE Ls Rs

pub const LOGIC_5_1_B: Self = Self::MPEG_5_1_B

L R Ls Rs C LFE

pub const LOGIC_5_1_C: Self = Self::MPEG_5_1_C

L C R Ls Rs LFE

pub const LOGIC_5_1_D: Self = Self::MPEG_5_1_D

C L R Ls Rs LFE

pub const LOGIC_6_0_A: Self = Self::AAC_6_0

C L R Ls Rs Cs

pub const LOGIC_6_0_B: Self

L R Ls Rs Cs C

pub const LOGIC_6_0_C: Self = Self::AUDIO_UNIT_6_0

L R Ls Rs C Cs

pub const LOGIC_6_1_A: Self = Self::AAC_6_1

C L R Ls Rs Cs LFE

pub const LOGIC_6_1_B: Self

L R Ls Rs Cs C LFE

pub const LOGIC_6_1_C: Self = Self::MPEG_6_1_A

L R C LFE Ls Rs Cs

pub const LOGIC_6_1_D: Self

L C R Ls Cs Rs LFE

pub const LOGIC_7_1_A: Self = Self::AUDIO_UNIT_7_1

L R C LFE Ls Rs Rls Rrs

pub const LOGIC_7_1_B: Self

L R Ls Rs Rls Rrs C LFE

pub const LOGIC_7_1_C: Self = Self::MPEG_7_1_C

L R C LFE Ls Rs Rls Rrs

pub const LOGIC_7_1_SDDS_A: Self = Self::MPEG_7_1_A

L R C LFE Ls Rs Lc Rc

pub const LOGIC_7_1_SDDS_B: Self = Self::MPEG_7_1_B

C Lc Rc L R Ls Rs LFE

pub const LOGIC_7_1_SDDS_C: Self = Self::EMAGIC_DEFAULT_7_1

L R Ls Rs C LFE Lc Rc

pub const LOGIC_ATMOS_5_1_2: Self = Self::ATMOS_5_1_2

L R C LFE Ls Rs Ltm Rtm

pub const LOGIC_ATMOS_5_1_4: Self = Self::ATMOS_5_1_4

L R C LFE Ls Rs Vhl Vhr Ltr Rtr

pub const LOGIC_ATMOS_7_1_2: Self = Self::ATMOS_7_1_2

L R C LFE Ls Rs Rls Rrs Ltm Rtm

pub const LOGIC_ATMOS_7_1_4_A: Self = Self::ATMOS_7_1_4

L R C LFE Ls Rs Rls Rrs Vhl Vhr Ltr Rtr

pub const LOGIC_ATMOS_7_1_4_B: Self

L R Rls Rrs Ls Rs C LFE Vhl Vhr Ltr Rtr

pub const LOGIC_ATMOS_7_1_6: Self

L R Rls Rrs Ls Rs C LFE Vhl Vhr Ltm Rtm Ltr Rtr

pub const DISCRETE_IN_ORDER: Self

needs to be ORed with the actual number of channels

pub const BEGIN_RESERVED: Self

Channel layout tag values in this range are reserved for internal use

pub const END_RESERVED: Self

Channel layout tag values in this range are reserved for internal use

pub const UNKNOWN: Self

needs to be ORed with the actual number of channels

pub const fn number_of_channels(&self) -> u32

The low 16 bits of an AudioChannelLayoutTag gives the number of channels except for kAudioChannelLayoutTag_UseChannelDescriptions and kAudioChannelLayoutTag_UseChannelBitmap