cidre::cat::audio

Struct Format

#[repr(transparent)]
pub struct Format(pub u32);

A four char code indicating the general kind of data in the stream.

Tuple Fields
0: u32
Implementations

impl Format

The AudioFormatIDs used to identify individual formats of audio data.

pub const LINEAR_PCM: Self

Linear PCM, uses the standard flags.

pub const AC3: Self

AC-3, has no flags.

pub const _60958AC3: Self

AC-3 packaged for transport over an IEC 60958 compliant digital audio interface. Uses the standard flags.

pub const APPLE_IMA4: Self

Apples implementation of IMA 4:1 ADPCM, has no flags.

pub const MPEG4_AAC: Self

MPEG-4 Low Complexity AAC audio object, has no flags.

pub const MPEG4_CELP: Self

MPEG-4 CELP audio object, has no flags.

pub const MPEG4_HVXC: Self

MPEG-4 HVXC audio object, has no flags.

pub const MPEG4_TWIN_VQ: Self

MPEG-4 TwinVQ audio object type, has no flags.

pub const MACE3: Self

MACE 3:1, has no flags.

pub const MACE6: Self

MACE 6:1, has no flags.

pub const U_LAW: Self

µLaw 2:1, has no flags.

pub const A_LAW: Self

aLaw 2:1, has no flags.

pub const Q_DESIGN: Self

QDesign music, has no flags

pub const Q_DESIGN2: Self

QDesign2 music, has no flags

pub const QUALCOMM: Self

QUALCOMM PureVoice, has no flags

pub const MPEGLAYER1: Self

MPEG-1/2, Layer 1 audio, has no flags

pub const MPEGLAYER2: Self

MPEG-1/2, Layer 2 audio, has no flags

pub const MPEGLAYER3: Self

MPEG-1/2, Layer 3 audio, has no flags

pub const TIME_CODE: Self

A stream of IOAudioTimeStamps, uses the IOAudioTimeStamp flags (see IOKit/audio/IOAudioTypes.h).

pub const MIDI_STREAM: Self

A stream of MIDIPacketLists where the time stamps in the MIDIPacketList are sample offsets in the stream. The mSampleRate field is used to describe how time is passed in this kind of stream and an AudioUnit that receives or generates this stream can use this sample rate, the number of frames it is rendering and the sample offsets within the MIDIPacketList to define the time for any MIDI event within this list. It has no flags.

pub const PARAMETER_VALUE_STREAM: Self

A “side-chain” of Float32 data that can be fed or generated by an AudioUnit and is used to send a high density of parameter value control information. An AU will typically run a ParameterValueStream at either the sample rate of the AudioUnit’s audio data, or some integer divisor of this (say a half or a third of the sample rate of the audio). The Sample Rate of the ASBD describes this relationship. It has no flags.

pub const APPLE_LOSSLESS: Self

Apple Lossless, the flags indicate the bit depth of the source material.

pub const MPEG4_AAC_HE: Self

MPEG-4 High Efficiency AAC audio object, has no flags.

pub const MPEG4_AAC_LD: Self

MPEG-4 AAC Low Delay audio object, has no flags.

pub const MPEG4_AAC_ELD: Self

MPEG-4 AAC Enhanced Low Delay audio object, has no flags. This is the formatID of the base layer without the SBR extension. See also kAudioFormatMPEG4AAC_ELD_SBR

pub const MPEG4_AAC_ELD_SBR: Self

MPEG-4 AAC Enhanced Low Delay audio object with SBR extension layer, has no flags.

pub const MPEG4_AAC_ELD_V2: Self

pub const MPEG4_AAC_HE_V2: Self

MPEG-4 High Efficiency AAC Version 2 audio object, has no flags.
note: https://lists.apple.com/archives/coreaudio-api/2011/Mar/msg00176.html frames_per_packet should be 2048

pub const MPEG4_AAC_SPATIAL: Self

MPEG-4 Spatial Audio audio object, has no flags.

pub const MPEG_D_USAC: Self

MPEG-D Unified Speech and Audio Coding, has no flags.

pub const AMR: Self

The AMR Narrow Band speech codec.

pub const AMR_WB: Self

The AMR Wide Band speech codec.

pub const AUDIBLE: Self

The format used for Audible audio books. It has no flags.

pub const LBC: Self

The iLBC narrow band speech codec. It has no flags.

pub const DVI_INTEL_IMA: Self

DVI/Intel IMA ADPCM - ACM code 17.

pub const MICROSOFT_GSM: Self

Microsoft GSM 6.10 - ACM code 49.

pub const AES3: Self

This format is defined by AES3-2003, and adopted into MXF and MPEG-2 containers and SDTI transport streams with SMPTE specs 302M-2002 and 331M-2000. It has no flags.

pub const ENHANCED_AC3: Self

Enhanced AC-3, has no flags.

pub const FLAC: Self

Free Lossless Audio Codec, the flags indicate the bit depth of the source material.

pub const OPUS: Self

Opus codec, has no flags.

pub const APAC: Self

Apple Positional Audio Codec, has no flags.

pub fn to_ns_number(self) -> &'static Number