cidre::cat::audio

Struct FormatFlags

#[repr(transparent)]
pub struct FormatFlags(pub u32);

Tuple Fields
0: u32
Implementations

impl FormatFlags

pub fn is_empty(&self) -> bool

pub fn any(&self, mask: Self) -> bool

pub fn contains(&self, val: Self) -> bool

pub fn insert(&mut self, val: Self)

pub fn remove(&mut self, val: Self)

pub fn set(&mut self, val: Self, on: bool)

impl FormatFlags

ios app audio - IS_BIG_ENDIAN | IS_SIGNED_INTEGER | IS_PACKED mic - IS_SIGNED_INTEGER | IS_PACKED These are the standard AudioFormatFlags for use in the mFormatFlags field of the AudioStreamBasicDescription structure. Typically, when an ASBD is being used, the fields describe the complete layout of the sample data in the buffers that are represented by this description - where typically those buffers are represented by an AudioBuffer that is contained in an AudioBufferList.

However, when an ASBD has the kAudioFormatFlagIsNonInterleaved flag, the AudioBufferList has a different structure and semantic. In this case, the ASBD fields will describe the format of ONE of the AudioBuffers that are contained in the list, AND each AudioBuffer in the list is determined to have a single (mono) channel of audio data. Then, the ASBD’s mChannelsPerFrame will indicate the total number of AudioBuffers that are contained within the AudioBufferList - where each buffer contains one channel. This is used primarily with the AudioUnit (and AudioConverter) representation of this list - and won’t be found in the AudioHardware usage of this structure.

pub const IS_FLOAT: Self

Set for floating point, clear for integer.

pub const IS_BIG_ENDIAN: Self

Set for big endian, clear for little endian.

pub const IS_SIGNED_INTEGER: Self

Set for signed integer, clear for unsigned integer. This is only valid if AudioFormatFlags::IS_FLOAT is clear.

pub const IS_PACKED: Self

Set if the sample bits occupy the entire available bits for the channel, clear if they are high or low aligned within the channel. Note that even if this flag is clear, it is implied that this flag is set if the AudioStreamBasicDescription is filled out such that the fields have the following relationship: ((bits_per_sample / 8) * channels_per_frame) == bytes_per_frame

pub const IS_ALIGNED_HIGH: Self

Set if the sample bits are placed into the high bits of the channel, clear for low bit placement. This is only valid if kAudioFormatFlagIsPacked is clear.

pub const IS_NON_INTERLEAVED: Self

Set if the samples for each channel are located contiguously and the channels are layed out end to end, clear if the samples for each frame are layed out contiguously and the frames layed out end to end.

pub const IS_NON_MIXABLE: Self

Set to indicate when a format is non-mixable. Note that this flag is only used when interacting with the HAL’s stream format information. It is not a valid flag for any other uses.

pub const ALL_CLEAR: Self

Set if all the flags would be clear in order to preserve 0 as the wild card value.

pub const LINEAR_PCM_IS_FLOAT: Self = Self::IS_FLOAT

pub const LINEAR_PCM_IS_BIG_ENDIAN: Self = Self::IS_BIG_ENDIAN

pub const LINEAR_PCM_IS_SIGNED_INTEGER: Self = Self::IS_SIGNED_INTEGER

pub const LINEAR_PCM_IS_PACKED: Self = Self::IS_PACKED

pub const LINEAR_PCM_IS_ALIGNED_HIGH: Self = Self::IS_ALIGNED_HIGH

pub const LINEAR_PCM_IS_NON_INTERLEAVED: Self = Self::IS_NON_INTERLEAVED

pub const LINEAR_PCM_IS_NON_MIXABLE: Self = Self::IS_NON_MIXABLE

pub const NATIVE_ENDIAN: Self

pub const NATIVE_FLOAT_PACKED: Self

pub const LINEAR_PCM_SAMPLE_FRACTION_SHIFT: Self

The linear PCM flags contain a 6-bit bitfield indicating that an integer format is to be interpreted as fixed point. The value indicates the number of bits are used to represent the fractional portion of each sample value. This constant indicates the bit position (counting from the right) of the bitfield in mFormatFlags.

pub const LINEAR_PCM_SAMPLE_FRACTION_MASK: Self

number_fractional_bits = (mFormatFlags & kLinearPCMFormatFlagsSampleFractionMask) >> kLinearPCMFormatFlagsSampleFractionShift

pub const LINEAR_PCM_ARE_ALL_CLEAR: Self = Self::ALL_CLEAR

pub const APPLE_LOSSLESS_16_BIT_SOURCE_DATA: Self

This flag is set for Apple Lossless data that was sourced from 16 bit native endian signed integer data.

pub const APPLE_LOSSLESS_20_BIT_SOURCE_DATA: Self

This flag is set for Apple Lossless data that was sourced from 20 bit native endian signed integer data aligned high in 24 bits.

pub const APPLE_LOSSLESS_24_BIT_SOURCE_DATA: Self

his flag is set for Apple Lossless data that was sourced from 24 bit native endian signed integer data.

pub const APPLE_LOSSLESS_32_BIT_SOURCE_DATA: Self

This flag is set for Apple Lossless data that was sourced from 32 bit native endian signed integer data.