cidre::cf

Module base

Structs

Allocator

AllocatorContext

A structure that defines the context or operating environment for an allocator (cf::Allocator) object. Every Core Foundation allocator object must have a context defined for it.

Null

OptionFlags

Plist

PlistMutabilityOpts

Range

Enums

ComparisonResult

PlistFormat

Constants

NOT_FOUND

Functions

type_id_desc⚠

Type Aliases

AllocatorAllocateCb

AllocatorCopyDescCb

AllocatorDealloacteCb

AllocatorPreferredSizeCb

AllocatorRealloacteCb

AllocatorReleaseCb

AllocatorRetainCb

ComparatorFn

HashCode

Index

TypeId