cidre::cf::run_loop

Struct RunLoop

pub struct RunLoop(/* private fields */);

Implementations

impl RunLoop

pub fn retained(&self) -> R<Self>

impl RunLoop

pub fn run()

pub fn current() -> &'static RunLoop

pub fn main() -> &'static RunLoop

pub fn stop(&self)

pub fn current_mode(&self) -> Option<R<Mode>>

pub fn all_modes(&self) -> R<ArrayOf<Mode>>

pub fn contains_src(&self, source: &Src, mode: &Mode) -> bool

pub fn add_src(&self, source: &Src, mode: &Mode)

pub fn remove_src(&self, source: &Src, mode: &Mode)

pub fn contains_observer(&self, observer: &Observer, mode: &Mode) -> bool

pub fn add_observer(&self, observer: &Observer, mode: &Mode)

pub fn remove_observer(&self, observer: &Observer, mode: &Mode)

pub fn contains_timer(&self, timer: &Timer, mode: &Mode) -> bool

pub fn add_timer(&self, timer: &Timer, mode: &Mode)

pub fn remove_timer(&self, timer: &Timer, mode: &Mode)

pub fn run_in_mode( mode: &Mode, seconds: TimeInterval, return_after_source_handled: bool, ) -> RunResult

Run loops can be run recursively. You can call Mode from within any run loop callout and create nested run loop activations on the current thread’s call stack. You are not restricted in which modes you can run from within a callout. You can create another run loop activation running in any available run loop mode, including any modes already running higher in the call stack.

The run loop exits with the following return values under the indicated conditions: - Result::Finished. The run loop mode mode has no sources or timers. - Result::Stopped. The run loop was stopped with CFRunLoopStop. - Result::TimedOut. The time interval seconds passed. - Result::HandledSource. A source was processed. This exit condition only applies when returnAfterSourceHandled is true.

You must not specify the kCFRunLoopCommonModes constant for the mode parameter. Run loops always run in a specific mode. You specify the common modes only when configuring a run-loop observer and only in situations where you want that observer to run in more than one mode.

pub fn is_waiting(&self) -> bool

pub fn wakeup(&self)