cidre::cg::color_space

Module names

Functions

acescg_linear

adobe_rgb_1998

dci_p3

display_p3

The Display P3 color space, created by Apple.

display_p3_hlg

display_p3_pq

extended_display_p3

extended_gray

extended_itur_2020

extended_linear_display_p3

extended_linear_gray

extended_linear_itur_2020

extended_linear_srgb

extended_srgb

generic_cmyk

generic_gray

The name of the “Generic” Gray color space. It is a legacy color space so use it only if you definitely know that you need it. If you need just a reasonable monochrome color space, please use Generic Gray Gamma 2.2, i.e. use generic_gray_gamma_2_2() name instead of generic_gray().

generic_gray_gamma_2_2

generic_lab

generic_rgb

generic_rgb_linear

generic_xyz

itur_709

itur_709_pq

itur_2020

itur_2020_srgb_gamma

itur_2100_hlg

itur_2100_pq

linear_display_p3

linear_gray

linear_srgb

linerar_itur_2020

rommrgb

srgb

The standard Red Green Blue (sRGB) color space.