cidre::cg

Module event_types

Structs

Event

EventField

EventFlags

EventSrc

EventTapOpts

EventType

Enums

EventSrcState

EventTapLocation

Constants that specify possible tapping points for events.

EventTapPlacement

Constants that specify where a new event tap is inserted into the list of active event taps.

GesturePhase

Constants that specify gesture phases.

MomentumScrollPhase

Constants that specify momentum scroll phases.

MouseButton

Constants that specify buttons on a one, two, or three-button mouse.

ScrollEventUnit

Constants that specify the unit of measurement for a scrolling event.

ScrollPhase

Constants that specify scroll phases.

Type Aliases

EventMask

EventSrcKeyboardType

EventTapCb

The callback is passed a proxy for the tap, the event type, the incoming event, and the user-defined data specified when the event tap was created. The function should return the (possibly modified) passed-in event, a newly constructed event, or None if the event is to be deleted.

EventTapProxy