cidre::cg::event_types

Struct EventField

#[repr(transparent)]
pub struct EventField(pub u32);

Tuple Fields
0: u32
Implementations

impl EventField

pub const MOUSE_EVENT_NUMBER: Self

Key to access an integer field that contains the mouse button event number. Matching mouse-down and mouse-up events will have the same event number.

pub const MOUSE_EVENT_CLICK_STATE: Self

Key to access an integer field that contains the mouse button click state. A click state of 1 represents a single click. A click state of 2 represents a double-click. A click state of 3 represents a triple-click.

pub const MOUSE_EVENT_PRESSURE: Self

Key to access a double field that contains the mouse button pressure. The pressure value may range from 0 to 1, with 0 representing the mouse being up. This value is commonly set by tablet pens mimicking a mouse.

pub const MOUSE_EVENT_BUTTON_NUMBER: Self

Key to access an integer field that contains the mouse button number.

pub const MOUSE_EVENT_DELTA_X: Self

Key to access an integer field that contains the horizontal mouse delta since the last mouse movement event.

pub const MOUSE_EVENT_DELTA_Y: Self

Key to access an integer field that contains the vertical mouse delta since the last mouse movement event.

pub const MOUSE_EVENT_INSTANT_MOUSER: Self

Key to access an integer field. The value is non-zero if the event should be ignored by the Inkwell subsystem.

pub const MOUSE_EVENT_SUBTYPE: Self

Key to access an integer field that encodes the mouse event subtype as a `kCFNumberIntType’

pub const KEYBOARD_EVENT_AUTOREPEAT: Self

Key to access an integer field, non-zero when this is an autorepeat of a key-down, and zero otherwise.

pub const KEYBOARD_EVENT_KEYCODE: Self

Key to access an integer field that contains the virtual keycode of the key-down or key-up event.

pub const KEYBOARD_EVENT_KEYBOARD_TYPE: Self

Key to access an integer field that contains the keyboard type identifier.

pub const SCROLL_WHEEL_EVENT_DELTA_AXIS1: Self

Key to access an integer field that contains scrolling data. This field typically contains the change in vertical position since the last scrolling event from a Mighty Mouse scroller or a single-wheel mouse scroller.

pub const SCROLL_WHEEL_EVENT_DELTA_AXIS2: Self

Key to access an integer field that contains scrolling data. This field typically contains the change in horizontal position since the last scrolling event from a Mighty Mouse scroller.

pub const SCROLL_WHEEL_EVENT_DELTA_AXIS3: Self

This field is not used.

pub const SCROLL_WHEEL_EVENT_FIXED_PT_DELTA_AXIS1: Self

Key to access a field that contains scrolling data. The scrolling data represents a line-based or pixel-based change in vertical position since the last scrolling event from a Mighty Mouse scroller or a single-wheel mouse scroller. The scrolling data uses a fixed-point 16.16 signed integer format. If this key is passed to CGEventGetDoubleValueField, the fixed-point value is converted to a double value.

pub const SCROLL_WHEEL_EVENT_FIXED_PT_DELTA_AXIS2: Self

Key to access a field that contains scrolling data. The scrolling data represents a line-based or pixel-based change in horizontal position since the last scrolling event from a Mighty Mouse scroller. The scrolling data uses a fixed-point 16.16 signed integer format. If this key is passed to CGEventGetDoubleValueField, the fixed-point value is converted to a double value.

pub const SCROLL_WHEEL_EVENT_FIXED_PT_DELTA_AXIS3: Self

This field is not used.

pub const SCROLL_WHEEL_EVENT_POINT_DELTA_AXIS1: Self

Key to access an integer field that contains pixel-based scrolling data. The scrolling data represents the change in vertical position since the last scrolling event from a Mighty Mouse scroller or a single-wheel mouse scroller.

pub const SCROLL_WHEEL_EVENT_POINT_DELTA_AXIS2: Self

Key to access an integer field that contains pixel-based scrolling data. The scrolling data represents the change in horizontal position since the last scrolling event from a Mighty Mouse scroller.

pub const SCROLL_WHEEL_EVENT_POINT_DELTA_AXIS3: Self

This field is not used.

pub const SCROLL_WHEEL_EVENT_SCROLL_PHASE: Self

Key to access an integer field that contains the scroll phase.

pub const SCROLL_WHEEL_EVENT_SCROLL_COUNT: Self

Key to access an integer field that contains the scroll count.

pub const SCROLL_WHEEL_EVENT_MOMENTUM_PHASE: Self

Key to access an integer field that contains the momentum phase.

pub const SCROLL_WHEEL_EVENT_INSTANT_MOUSER: Self

Key to access an integer field that indicates whether the event should be ignored by the Inkwell subsystem. If the value is non-zero, the event should be ignored.

pub const TABLET_EVENT_POINT_X: Self

Key to access an integer field that contains the absolute X coordinate in tablet space at full tablet resolution.

pub const TABLET_EVENT_POINT_Y: Self

Key to access an integer field that contains the absolute Y coordinate in tablet space at full tablet resolution.

pub const TABLET_EVENT_POINT_Z: Self

Key to access an integer field that contains the absolute Z coordinate in tablet space at full tablet resolution.

pub const TABLET_EVENT_POINT_BUTTONS: Self

Key to access an integer field that contains the tablet button state. Bit 0 is the first button, and a set bit represents a closed or pressed button. Up to 16 buttons are supported.

pub const TABLET_EVENT_POINT_PRESSURE: Self

Key to access a double field that contains the tablet pen pressure. A value of 0.0 represents no pressure, and 1.0 represents maximum pressure.

pub const TABLET_EVENT_TILT_X: Self

Key to access a double field that contains the horizontal tablet pen tilt. A value of 0 represents no tilt, and 1 represents maximum tilt.

pub const TABLET_EVENT_TILT_Y: Self

Key to access a double field that contains the vertical tablet pen tilt. A value of 0 represents no tilt, and 1 represents maximum tilt.

pub const TABLET_EVENT_ROTATION: Self

Key to access a double field that contains the tablet pen rotation.

pub const TABLET_EVENT_TANGENTIAL_PRESSURE: Self

Key to access a double field that contains the tangential pressure on the device. A value of 0.0 represents no pressure, and 1.0 represents maximum pressure.

pub const TABLET_EVENT_DEVICE_ID: Self

Key to access an integer field that contains the system-assigned unique device ID.

pub const TABLET_EVENT_VENDOR_1: Self

Key to access an integer field that contains a vendor-specified value.

pub const TABLET_EVENT_VENDOR_2: Self

Key to access an integer field that contains a vendor-specified value.

pub const TABLET_EVENT_VENDOR_3: Self

Key to access an integer field that contains a vendor-specified value.

pub const TABLET_PROXIMITY_EVENT_VENDOR_ID: Self

Key to access an integer field that contains the vendor-defined ID, typically the USB vendor ID.

pub const TABLET_PROXIMITY_EVENT_TABLET_ID: Self

Key to access an integer field that contains the vendor-defined tablet ID, typically the USB product ID.

pub const TABLET_PROXIMITY_EVENT_POINTER_ID: Self

Key to access an integer field that contains the vendor-defined ID of the pointing device.

pub const TABLET_PROXIMITY_EVENT_DEVICE_ID: Self

Key to access an integer field that contains the system-assigned device ID.

pub const TABLET_PROXIMITY_EVENT_SYSTEM_TABLET_ID: Self

Key to access an integer field that contains the system-assigned unique tablet ID.

pub const TABLET_PROXIMITY_EVENT_VENDOR_POINTER_TYPE: Self

Key to access an integer field that contains the vendor-assigned pointer type.

pub const TABLET_PROXIMITY_EVENT_VENDOR_POINTER_SERIAL_NUMBER: Self

Key to access an integer field that contains the vendor-defined pointer serial number.

pub const TABLET_PROXIMITY_EVENT_VENDOR_UNIQUE_ID: Self

Key to access an integer field that contains the vendor-defined unique ID.

pub const TABLET_PROXIMITY_EVENT_CAPABILITY_MASK: Self

Key to access an integer field that contains the device capabilities mask.

pub const TABLET_PROXIMITY_EVENT_POINTER_TYPE: Self

Key to access an integer field that contains the pointer type.

pub const TABLET_PROXIMITY_EVENT_ENTER_PROXIMITY: Self

Key to access an integer field that indicates whether the pen is in proximity to the tablet. The value is non-zero if the pen is in proximity to the tablet and zero when leaving the tablet.

pub const EVENT_TARGET_PROCESS_SERIAL_NUMBER: Self

Key to access a field that contains the event target process serial number. The value is a 64-bit value.

pub const EVENT_TARGET_UNIX_PROCESS_ID: Self

Key to access a field that contains the event target Unix process ID.

pub const EVENT_SOURCE_UNIX_PROCESS_ID: Self

Key to access a field that contains the event source Unix process ID.

pub const EVENT_SOURCE_USER_DATA: Self

Key to access a field that contains the event source user-supplied data, up to 64 bits.

pub const EVENT_SOURCE_USER_ID: Self

Key to access a field that contains the event source Unix effective UID.

pub const EVENT_SOURCE_GROUP_ID: Self

Key to access a field that contains the event source Unix effective GID.

pub const EVENT_SOURCE_STATE_ID: Self

Key to access a field that contains the event source state ID used to create this event.

pub const SCROLL_WHEEL_EVENT_IS_CONTINUOUS: Self

Key to access an integer field that indicates whether a scrolling event contains continuous, pixel-based scrolling data. The value is non-zero when the scrolling data is pixel-based and zero when the scrolling data is line-based.

pub const MOUSE_EVENT_WINDOW_UNDER_MOUSE_POINTER: Self

Key to access an integer field that contains the window under the mouse pointer.

pub const MOUSE_EVENT_WINDOW_UNDER_MOUSE_POINTER_THAT_CAN_HANDLE_THIS_EVENT: Self

Key to access an integer field that contains the window under the mouse pointer that can handle this event.

pub const EVENT_UNACCELERATED_POINTER_MOVEMENT_X: Self

Key to access an integer field that contains unaccelerated pointer movement along the X axis.

pub const EVENT_UNACCELERATED_POINTER_MOVEMENT_Y: Self

Key to access an integer field that contains unaccelerated pointer movement along the Y axis.

pub const SCROLL_WHEEL_EVENT_MOMENTUM_OPTION_PHASE: Self

Key to access an integer field that contains the optional momentum phase.

pub const SCROLL_WHEEL_EVENT_ACCELERATED_DELTA_AXIS1: Self

Key to access an integer field that contains the accelerated vertical scroll delta.

pub const SCROLL_WHEEL_EVENT_ACCELERATED_DELTA_AXIS2: Self

Key to access an integer field that contains the accelerated horizontal scroll delta.

pub const SCROLL_WHEEL_EVENT_RAW_DELTA_AXIS1: Self

Key to access an integer field that contains the raw vertical scroll delta.

pub const SCROLL_WHEEL_EVENT_RAW_DELTA_AXIS2: Self

Key to access an integer field that contains the raw horizontal scroll delta.