cidre::cg::image_properties

Module gif_keys

Functions

canvas_pixel_height

The height of the main image, in pixels.

canvas_pixel_width

The width of the main image, in pixels.

delay_time

The number of seconds to wait before displaying the next image in an animated sequence, clamped to a minimum of 100 milliseconds.

frame_info_array

An array of dictionaries that contain timing information for the image sequence.

has_global_color_map

A Boolean value that indicates whether the GIF has a global color map.

image_color_map

The image color map.

loop_count

The number of times to repeat an animated sequence.

unclamped_delay_time

The number of seconds to wait before displaying the next image in an animated sequence.