cidre::cg

Struct Font

pub struct Font(/* private fields */);

Implementations

impl Font

pub fn retained(&self) -> R<Self>

impl Font

pub fn type_id() -> TypeId

pub fn with_name(name: &String) -> Option<R<Self>>

pub fn with_variations( &self, variations: Option<&Dictionary>, ) -> Option<R<Self>>

pub fn nglyphs(&self) -> usize

pub fn units_per_em(&self) -> i32

pub fn post_script_name(&self) -> R<String>

pub fn full_name(&self) -> R<String>

pub fn ascent(&self) -> i32

The ascent is the maximum distance above the baseline of glyphs in a font. The value is specified in glyph space units.

pub fn descent(&self) -> i32

The descent is the maximum distance below the baseline of glyphs in a font. The value is specified in glyph space units.

pub fn leading(&self) -> i32

The leading is the spacing between consecutive lines of text in a font. The value is specified in glyph space units.

pub fn cap_height(&self) -> i32

The cap height is the distance above the baseline of the top of flat capital letters of glyphs in a font. The value is specified in glyph space units.

pub fn x_height(&self) -> i32

The x-height is the distance above the baseline of the top of flat, non-ascending lowercase letters (such as “x”) of glyphs in a font. The value is specified in glyph space units.

pub fn bbox(&self) -> Rect

The font bounding box is the union of all of the bounding boxes for all the glyphs in a font. The value is specified in glyph space units.

pub fn italic_angle(&self) -> Float

pub fn stem_v(&self) -> Float

The thickness of the dominant vertical stems of glyphs in font. The value is specified in glyph space units.

pub fn variation_axes(&self) -> Option<R<Array>>

An array of the variation axis dictionaries for font. Each variation axis dictionary contains values for the variation axis keys listed below. This function returns None if font doesn’t support variations.

pub fn variations(&self) -> Option<R<Array>>

The variation specification dictionary from font. This dictionary contains keys corresponding the variation axis names of the font. Each key is a variation axis name; the value for each key is the value specified for that particular variation axis represented as a cf::Number. This function returns None if font doesn’t support variations.

pub fn glyph_advances(&self, glyphs: &[Glyph]) -> Result<Vec<i32>, ()>

pub fn glyph_advances_u16(&self, glyphs: &[u16]) -> Result<Vec<i32>, ()>

pub fn glyph_bboxes(&self, glyphs: &[Glyph]) -> Result<Vec<Rect>, ()>

pub fn glyph_bboxes_u16(&self, glyphs: &[u16]) -> Result<Vec<Rect>, ()>

pub fn glyph_with_name(&self, name: &String) -> Glyph

pub fn name_for_glyph(&self, glyph: Glyph) -> Option<R<String>>

pub fn name_for_glyph_u16(&self, glyph: u16) -> Option<R<String>>

pub fn can_create_post_script_subset( &self, format: FontPostScriptFormat, ) -> bool