cidre::cg

Module window_keys

Functions

alpha

The alpha fade of the window.

backing_location_video_memory

Whether the window backing store is in video memory.

bounds

The bounds of the window in screen space.

is_on_screen

Whether the window is ordered on screen.

layer

The window layer number of the window.

memory_usage

An estimate of the memory in bytes currently used by the window.

name

The name of the window.

number

The window ID, a unique value within the user session representing the window.

owner_name

The name of the application process which owns the window.

owner_pid

The process ID of the process that owns the window.

sharing_state

The sharing state of the window.

store_type

The backing store type of the window.

workspaceDeprecated

The workspace ID of the workspace associated with the window.