cidre

Module ci

Core Image

Structs

BarcodeDesc

Context

An evaluation context for rendering image processing results and performing image analysis.

Format

Pixel data formats for image input, output, and processing.

Image

A representation of an image to be processed or produced by Core Image filters.

ImageOpt

Vec