cidre

Module cm

Core Media

Re-exports

pub use buffer_queue::Buf;

pub use buffer_queue::BufBoolCb;

pub use buffer_queue::BufBoolHandler;

pub use buffer_queue::BufCompareCb;

pub use buffer_queue::BufCompareHandler;

pub use buffer_queue::BufHandlers;

pub use buffer_queue::BufQueue;

pub use buffer_queue::BufQueueOf;

pub use buffer_queue::BufSizeCb;

pub use buffer_queue::BufSizeHandler;

pub use buffer_queue::BufTimeCb;

pub use buffer_queue::BufTimeHandler;

pub use buffer_queue::SampleBufQueue;

pub use buffer_queue::err as buf_queue_err;

pub use sample_buffer::BlockBufAudioBufList;

pub use sample_buffer::BlockBufAudioBufListN;

pub use sample_buffer::Flags as SampleBufFlags;

pub use sample_buffer::SampleBuf;

pub use sample_buffer::SampleTimingInfo;

pub use sample_buffer::err as sample_buf_err;

pub use attachment::Bearer as AttachBearer;

pub use attachment::Mode as AttachMode;

pub use block_buffer::BlockBuf;

pub use block_buffer::Flags as BlockBufFlags;

pub use block_buffer::err as block_buf_err;

pub use sync::Clock;

pub use sync::ClockOrTimebase;

pub use sync::Timebase;

pub use memory_pool::MemPool;

pub use memory_pool::keys as mem_pool_keys;

pub use simple_queue::SimpleQueue;

pub use simple_queue::err as simple_queue_err;

Modules

attachment

audio_device_clock

block_buffer

buffer_queue

format_desc_bridge_err

io

memory_pool

sample_buffer

simple_queue

sync

tag_collection_err

Errors returned from the cm::TagCollection routines.

tag_err

Errors returned from the cm::Tag routines

tagged_buf_group_err

The os::Errors returned from the CMTaggedBufferGroup routines.

Structs

FormatDesc

FormatDescExtKey

ImageDescFlavor

LogTransferFn

MediaType

PersistentTrackId

PixelFormat

SoundDescFlavor

Tag

TagCollection

TagCollectionMut

TaggedBufGroup

TextDescFlavor

Time

TimeFlags

TimeMapping

TimeRange

VideoCodec

VideoDimensions

Enums

MuxedStreamType

Muxed media format/subtype.

TagCategory

An enum indicating the available cm::TagCategory identifiers that can be used to distinguish the tag from other kinds.

TagDType

TaggedBufGroupFormatType

The subtypes of cm::TaggedBufGroup media type.

TimeRoundingMethod

Constants

TIME_SCALE_MAX

Note: cm::TIME_SCALE_MAX is NOT a good choice of timescale for movie files.
(Recommended timescales for movie files range from 600 to 90000.)

Functions

swap_be_image_desc_to_host

Converts a ImageDescription data structure from big-endian to host-endian in place.

swap_be_sound_desc_to_host

Converts a SoundDescription data structure from big-endian to host-endian in place.

swap_host_image_desc_to_be

Converts an ImageDescription data structure from host-endian to big-endian in place.

swap_host_sound_desc_to_be

Converts a SoundDescription data structure from host-endian to big-endian in place.

Type Aliases

AudioFormatDesc

ItemCount

ItemIndex

MuxedFormatDesc

TaggedBufGroupFormatDesc

TimeEpoch

The epoch is typically 0, but you can use a different value — for example to denote a particular iteration of a loop.

TimeScale

Timescales must be positive.

TimeValue

An integer time value.

VideoFormatDesc