cidre::cm::io::hardware

Module err

The err module constants unique to the DAL.

These are the error constants that are unique to the DAL. Note that the DAL’s functions can and will return other codes that are not listed here. While these constants give a general idea of what might have gone wrong during the execution of an API call.

Constants

DEVICE_PERMISSIONS

The requested operation can’t be completed because the process doesn’t have permission.

DEVICE_UNSUPPORTED_FORMAT

The ‘cm::io::Stream’ doesn’t support the requested format.

HW_BAD_DEVICE

The ‘cm::io::ObjectId’ passed to the function doesn’t map to a valid cm::io::Device.

HW_BAD_OBJECT

The ‘cm::io::ObjectId’ passed to the function doesn’t map to a valid ‘cm::io::Object’.

HW_BAD_PROP_SIZE

An improperly sized buffer was provided when accessing the data of a property.

HW_BAD_STREAM

The ‘cm::io::StreamId’ passed to the function doesn’t map to a valid ‘cm::io::Stream’.

HW_ILLEGAL_OPERATION

The requested operation couldn’t be completed.

HW_NOT_RUNNING

The function call requires that the hardware be running but it isn’t.

HW_NOT_STOPPED

The function call requires that the hardware be stopped but it isn’t.

HW_SUSPENDED_BY_SYSTEM

The function call failed because because access been suspended by the system.

HW_UNKNOWN_PROP

The CMIOObject doesn’t know about the property at the given address.

HW_UNSPECIFIED

The function call failed while doing something that doesn’t provide any error messages.

HW_UNSUPPORTED_OPERATION

The ‘cm::io::Object’ doesn’t support the requested operation.