cidre::cm::sample_buffer

Module buf_attach_keys

use get_attachment or dictionary_of_attachments

Functions

camera_intrinsic_matrix

Indicates the 3x3 camera intrinsic matrix applied to the current sample buffer.

display_empty_media_immediately

Tells that the empty marker should be dequeued immediately regardless of its timestamp.

drain_after_decoding

cf::Boolean

dropped_frame_reason

Indicates the reason the current video frame was dropped.

dropped_frame_reason_info

Indicates additional information regarding the dropped video frame.

empty_media

Marks an intentionally empty interval in the sequence of samples.

ends_previous_sample_duration

Indicates that sample buffer’s decode timestamp may be used to define the previous sample buffer’s duration.

fill_discontinuities_with_silence

Fill the difference between discontiguous sample buffers with silence

force_key_frame

Indicates that the current or next video sample buffer should be forced to be encoded as a key frame.

permanent_empty_media

Marks the end of the sequence of samples.

post_notification_when_consumed

cf::Dictionary (client-defined)

reset_decoder_before_decoding

cf::Boolean

resume_output

‘cf::Number’ (ResumeTag)

reverse

Indicates that the decoded contents of the sample buffer should be reversed.

sample_reference_byte_offset

Indicates the byte offset at which the sample data begins.

sample_reference_url

Indicates the URL where the sample data is.

still_image_lens_stabilization_info

Indicates information about the lens stabilization applied to the current still image buffer.

transition_id

Marks a transition from one source of buffers (eg. song) to another

trim_duration_at_end

The duration that should be removed at the end of the sample buffer, after decoding.

trim_duration_at_start

The duration that should be removed at the beginning of the sample buffer, after decoding.