cidre::cm::sample_buffer

Module err

Constants

ALLOC_FAILED

An allocation failed.

ALREADY_HAS_DATA_BUFFER

Attempt was made to set a dataBuffer on a cm::SampleBuffer that already has one.

ARRAY_TOO_SMALL

Output array was not large enough for the array being requested.

BUFFER_HAS_NO_SAMPLE_SIZES

Attempt to get sample size information when there was none.

BUFFER_HAS_NO_SAMPLE_TIMING_INFO

Attempt to get sample timing information when there was none.

BUFFER_NOT_READY

Buffer could not be made ready.

CANNOT_SUBDIVIDE

Sample buffer does not contain sample sizes. This can happen when the samples in the buffer are non-contiguous (eg. non-interleaved audio, where the channel values for a single sample are scattered through the buffer).

DATA_CANCELED

the sample buffer’s data loading operation was canceled.

DATA_FAILED

the sample buffer’s data loading operation failed (generic error).

INVALIDATED

the sample buffer was invalidated.

INVALID_ENTRY_COUNT

Timing info or size array entry count was not 0, 1, or numSamples.

INVALID_MEDIA_FORMAT

The format of the given media does not match the given format description (eg. a format description paired with a cv::ImageBuffer that fails cm::VideoFormatDescriptionMatchesImageBuffer).

INVALID_MEDIA_TYPE_FOR_OPERATION

the media type specified by a format description is not valid for the given operation (eg. a cm::SampleBuffer with a non-audio format description passed to cm::SampleBufferGetAudioStreamPacketDescriptionsPtr).

INVALID_SAMPLE_DATA

Buffer contains bad data. Only returned by cm::SampleBuffer functions that inspect its sample data.

REQUIRED_PARAMETER_MISSING

NULL or 0 was passed for a required parameter.

SAMPLE_INDEX_OUT_OF_RANGE

Sample index was not between 0 and numSamples-1, inclusive.

SAMPLE_TIMING_INFO_INVALID

buffer unexpectedly contains a non-numeric sample timing info