cidre::cm::sample_buffer

Struct SampleBuf

pub struct SampleBuf(/* private fields */);

Implementations

impl SampleBuf

pub fn tagged_bufs(&self) -> Option<&TaggedBufGroup>

Availability

macos_14_0, ios_17_0, tvos_17_0, watchos_10_0, visionos_1_0

impl SampleBuf

pub fn retained(&self) -> R<Self>

impl SampleBuf

An object that contains zero or more media samples of a uniform media type

Sample buffers are Core Foundation objects that the system uses to move media sample data through the media pipeline. An instance of cm::SampleBuf contains zero or more compressed (or uncompressed) samples of a particular media type and contains one of the following:

A cm::BlockBuf of one or more media samples

A cm::ImageBuf, a reference to the format description for the stream of cm::SampleBuf-s, size and timing information for each of the contained media samples, and both buffer-level and sample-level attachments

A sample buffer can contain both sample-level and buffer-level attachments. Each individual sample in a buffer may provide attachments that include information such as timestamps and video frame dependencies.

It’s possible for a sample buffer to describe samples it doesn’t yet contain. For example, some media services may have access to sample size, timing, and format information before they read the data. Such services may create sample buffers with that information and insert them into queues early, and attach (or fill) the buffer of media data later, when it becomes ready. Sample buffers have the concept of data-readiness, which means you can test, set, and force them to become ready “now.” It’s also possible for a sample buffer to contain nothing but a special buffer-level attachment that describes a media stream event (for example, “discontinuity: drain and reset decoder before processing the next cm::SampleBuffer”).

CMSampleBuffer

pub fn data_is_ready(&self) -> bool

Returns whether or not a cm::SampleBuf’s data is ready.

Example:

use cidre::{cf, cm};

let res = cm::SampleBuf::new(
    None,
    true,
    None,
);

let buf = res.unwrap();
assert!(buf.data_is_ready());

pub fn set_data_ready(&mut self)

pub fn new( data_buf: Option<&BlockBuf>, data_ready: bool, format_description: Option<&FormatDesc>, ) -> Result<R<SampleBuf>>

use cidre::{cf, cm};

let res = cm::SampleBuf::new(
    None,
    true,
    None,
);

let buf = res.unwrap();
buf.show();

pub unsafe fn create_in( allocator: Option<&Allocator>, data_buffer: Option<&BlockBuf>, data_ready: bool, make_data_ready_cb: Option<&SampleBufMakeDataReadyCb>, make_data_ready_refcon: *const c_void, format_description: Option<&FormatDesc>, num_samples: ItemCount, num_samples_timing_entries: ItemCount, sample_timing_array: *const SampleTimingInfo, num_sample_size_entries: ItemCount, sample_size_array: *const usize, sample_buffer_out: *mut Option<R<SampleBuf>>, ) -> Result

CMSampleBufferCreate

pub fn with_image_buf_in( allocator: Option<&Allocator>, image_buf: &ImageBuf, data_ready: bool, make_data_ready_cb: Option<&SampleBufMakeDataReadyCb>, make_data_ready_refcon: *const c_void, format_description: &FormatDesc, sample_timing: &SampleTimingInfo, ) -> Result<R<SampleBuf>>

pub fn with_image_buf( image_buffer: &ImageBuf, data_ready: bool, make_data_ready_cb: Option<&SampleBufMakeDataReadyCb>, make_data_ready_refcon: *const c_void, format_description: &FormatDesc, sample_timing: &SampleTimingInfo, ) -> Result<R<SampleBuf>>

pub fn image_buf(&self) -> Option<&ImageBuf>

pub fn image_buf_mut(&mut self) -> Option<&mut ImageBuf>

pub fn data_buf(&self) -> Option<&BlockBuf>

pub fn set_data_buf(&mut self, val: &BlockBuf) -> Result

pub fn duration(&self) -> Time

Returns the output duration of a sample buffer.

pub fn output_duration(&self) -> Time

Returns the output duration of a sample buffer.

pub fn pts(&self) -> Time

Returns the presentation timestamp that’s the earliest numerically of all the samples in a sample buffer.

pub fn dts(&self) -> Time

Returns the decode timestamp that’s the earliest numerically of all the samples in a sample buffer.

pub fn output_pts(&self) -> Time

Returns the output presentation timestamp of the CMSampleBuffer.

pub fn output_dts(&self) -> Time

pub fn set_output_pts(&self, val: Time)

pub fn timing_info(&self, sample_index: ItemIndex) -> Result<SampleTimingInfo>

pub fn sample_size(&self, sample_index: ItemIndex) -> usize

Returns the size in bytes of a specified sample in a ‘cm::SampleBuf’.

Size in bytes of the specified sample in the ‘cm::SampleBuf’.

pub fn total_sample_size(&self) -> usize

Returns the total size in bytes of sample data in a ‘cm::SampleBuf’.

Total size in bytes of sample data in the cm::SampleBuffer. If there are no sample sizes in this ‘cm::SampleBuf’, a size of 0 will be returned.

pub fn format_desc(&self) -> Option<&FormatDesc>

pub fn attaches( &self, create_if_necessary: bool, ) -> Option<&ArrayOf<DictionaryOf<String, Plist>>>

Returns a reference to a cm::SampleBuf’s immutable array of mutable sample attachments dictionaries (one dictionary per sample in the ‘cm::SampleBuf’)

pub fn attaches_mut( &mut self, create_if_necessary: bool, ) -> Option<&mut ArrayOf<DictionaryOfMut<String, Plist>>>

pub fn is_key_frame(&self) -> bool

pub unsafe fn contains_not_sync(&self) -> bool

pub fn is_valid(&self) -> bool

Returns a value that indicates whether a sample buffer is valid.

pub fn invalidate(&self) -> Result

Makes the sample buffer invalid, calling any installed invalidation callback.

An invalid sample buffer cannot be used – all accessors will return kCMSampleBufferError_Invalidated. It is not a good idea to do this to a sample buffer that another module may be accessing concurrently. Example of use: the invalidation callback could cancel pending I/O.

pub fn make_data_ready(&self) -> Result

Makes a cm::SampleBuf’s data ready, by calling the client’s cm::SampleBufferMakeDataReadyCallback.

pub fn copy_pcm_data_into_audio_buf_list<const N: usize>( &self, frame_offset: i32, num_frames: i32, buffer_list: &mut BufList<N>, ) -> Result

Copies PCM audio data from the given cm::SampleBuf into a pre-populated audio::BufList. The audio::BufList must contain the same number of channels and its data buffers must be sized to hold the specified number of frames. This API is specific to audio format sample buffers, and will return kCMSampleBufferError_InvalidMediaTypeForOperation if called with a non-audio sample buffer. It will return an error if the cm::SampleBuffer does not contain PCM audio data or if its dataBuffer is not ready.

pub fn audio_stream_packet_descs(&self) -> Result<Option<&[StreamPacketDesc]>>

Returns a pointer to (and size of) a constant array of AudioStreamPacketDescriptions for the variable bytes per packet or variable frames per packet audio data in the provided cm::SampleBuffer. The pointer will remain valid as long as the sbuf continues to be retained. Constant bitrate, constant frames-per-packet audio yields a return value of noErr and no packet descriptions. This API is specific to audio format sample buffers, and will return kCMSampleBufferError_InvalidMediaTypeForOperation if called with a non-audio sample buffer.

pub fn audio_buf_list<const N: usize>(&self) -> Result<BlockBufAudioBufList<N>>

pub fn audio_buf_list_n<'a>( &self, list: &'a mut AudioBufListN, ) -> Result<BlockBufAudioBufListN<'a>>

pub fn audio_buf_list_in<const N: usize>( &self, flags: Flags, block_buffer_structure_allocator: Option<&Allocator>, block_buffer_allocator: Option<&Allocator>, ) -> Result<BlockBufAudioBufList<N>>

Creates an audio::BufList containing the data from the cm::SampleBuf, and a cm::BlockBuf which references (and manages the lifetime of) the data in that audio::BufList. The data may or may not be copied, depending on the contiguity and 16-byte alignment of the cm::SampleBuffer’s data. The buffers placed in the audio::BufList are guaranteed to be contiguous. The buffers in the audio::BufferList will be 16-byte-aligned if kCMSampleBufferFlag_AudioBufferList_Assure16ByteAlignment is passed in.

pub fn num_samples(&self) -> Index