cidre::cm

Struct FormatDesc

pub struct FormatDesc(/* private fields */);

Implementations

impl FormatDesc

pub fn retained(&self) -> R<Self>

impl FormatDesc

pub fn type_id() -> TypeId

pub fn media_type(&self) -> MediaType

pub fn media_sub_type(&self) -> FourCharCode

pub fn exts(&self) -> Option<&DictionaryOf<String, Plist>>

pub fn ext<'a>(&'a self, key: &FormatDescExtKey) -> Option<&'a Plist>

pub fn original_compression_session_settings( &self, ) -> Option<&DictionaryOf<String, Plist>>

pub fn ext_atoms(&self) -> Option<&DictionaryOf<String, Plist>>

pub fn avcc(&self) -> Option<Vec<u8>>

pub fn hvcc(&self) -> Option<Vec<u8>>

pub fn verbatim_sample_desc(&self) -> Option<&Data>

pub fn verbatim_iso_sample_entry(&self) -> Option<&Data>

pub fn create_in( media_type: MediaType, media_sub_type: FourCharCode, extensions: Option<&DictionaryOf<FormatDescExtKey, Type>>, format_description_out: *mut Option<R<FormatDesc>>, allocator: Option<&Allocator>, ) -> Result

pub fn new( media_type: MediaType, media_sub_type: FourCharCode, extensions: Option<&DictionaryOf<FormatDescExtKey, Type>>, ) -> Result<R<Self>>

use cidre::{cm, mac_types::FourCharCode};

let desc = cm::FormatDesc::new(
    cm::MediaType::VIDEO,
    FourCharCode::from_be_bytes(*b"avc1"),
    None
).unwrap();

impl FormatDesc

pub fn video( codec: VideoCodec, width: i32, height: i32, extensions: Option<&DictionaryOf<FormatDescExtKey, Type>>, ) -> Result<R<Self>>

use cidre::cm;

let desc = cm::VideoFormatDesc::video(cm::VideoCodec::H264, 1920, 1080, None).unwrap();

pub fn create_video_in( codec: VideoCodec, width: i32, height: i32, extensions: Option<&DictionaryOf<FormatDescExtKey, Type>>, format_description_out: *mut Option<R<VideoFormatDesc>>, allocator: Option<&Allocator>, ) -> Result

pub fn dims(&self) -> VideoDimensions

Returns the dimensions (in encoded pixels)

This does not take into account pixel aspect ratio or clean aperture tags.

pub fn matches_image_buf(&self, image_buffer: &ImageBuf) -> bool

pub fn tag_collections(&self) -> Result<R<ArrayOf<TagCollection>>>

Availability

macos_14_0, ios_17_0, tvos_17_0, watchos_10_0, visionos_1_0

pub fn with_image_buf_in( image_buffer: &ImageBuf, allocator: Option<&Allocator>, ) -> Result<R<VideoFormatDesc>>

pub fn with_image_buf(image_buffer: &ImageBuf) -> Result<R<VideoFormatDesc>>

pub fn with_h264_param_sets<const N: usize>( pointers: &[*const u8; N], sizes: &[usize; N], nal_unit_header_length: i32, ) -> Result<R<VideoFormatDesc>>

pub fn with_h264_param_sets_in<const N: usize>( pointers: &[*const u8; N], sizes: &[usize; N], nal_unit_header_length: i32, allocator: Option<&Allocator>, ) -> Result<R<VideoFormatDesc>>

pub fn with_hevc_param_sets( count: usize, pointers: &[*const u8], sizes: &[usize], nal_unit_header_length: i32, extensions: Option<&DictionaryOf<FormatDescExtKey, Type>>, ) -> Result<R<VideoFormatDesc>>

pub fn with_hevc_param_sets_in( count: usize, pointers: &[*const u8], sizes: &[usize], nal_unit_header_length: i32, extensions: Option<&DictionaryOf<FormatDescExtKey, Type>>, allocator: Option<&Allocator>, ) -> Result<R<VideoFormatDesc>>

pub fn h264_params_count_and_header_len(&self) -> Result<(usize, i32)>

pub unsafe fn hevc_param_set_at_index( &self, parameter_set_index: usize, parameter_set_pointer_out: *mut *const u8, parameter_set_size_out: *mut usize, parameter_set_count_out: *mut usize, nal_unit_header_length_out: *mut i32, ) -> Result

pub fn hevc_params_count_and_header_len(&self) -> Result<(usize, i32)>

pub fn h264_param_set_at(&self, index: usize) -> Result<&[u8]>

pub fn hevc_param_set_at(&self, index: usize) -> Result<&[u8]>

pub fn common_image_buf_ext_keys() -> &'static ArrayOf<FormatDescExtKey>

Returns an array of the keys that are used both as cm::VideoFormatDesc extensions and cv::ImageBuf attachments and attributes.

impl FormatDesc

pub fn with_asbd(asbd: &StreamBasicDesc) -> Result<R<Self>>

pub fn audio_in( asbd: &StreamBasicDesc, layout_size: usize, layout: Option<&AudioChannelLayout<1>>, magic_cookie_size: usize, magic_cookie: Option<&c_void>, extensions: Option<&DictionaryOf<FormatDescExtKey, Type>>, format_description_out: *mut Option<R<Self>>, allocator: Option<&Allocator>, ) -> Result

pub fn stream_basic_desc(&self) -> Option<&StreamBasicDesc>

Returns a read-only pointer to the audio stream description in an audio format description.

This API is specific to audio format descriptions, and returns None if used with a non-audio format descriptions.

impl FormatDesc

pub fn as_be_image_desc_cm_buf_in( &self, string_encoding: StringEncoding, flavor: Option<&ImageDescFlavor>, allocator: Option<&Allocator>, ) -> Result<R<BlockBuf>>

Copies the contents of a cm::VideoFormatDescription to a cm::BlockBuffer in big-endian byte ordering.

pub fn as_be_image_desc_cm_buf( &self, flavor: Option<&ImageDescFlavor>, ) -> Result<R<BlockBuf>>

pub fn from_be_image_desc_buf_in( image_description_block_buffer: &BlockBuf, string_encoding: StringEncoding, flavor: Option<&ImageDescFlavor>, allocator: Option<&Allocator>, ) -> Result<R<Self>>

pub fn from_be_image_desc_buf( image_description_block_buffer: &BlockBuf, flavor: Option<&ImageDescFlavor>, ) -> Result<R<Self>>

pub fn from_be_image_desc_data_in( data: &[u8], string_encoding: StringEncoding, flavor: Option<&ImageDescFlavor>, allocator: Option<&Allocator>, ) -> Result<R<Self>>

pub fn from_be_image_desc_data( data: &[u8], string_encoding: StringEncoding, flavor: Option<&ImageDescFlavor>, ) -> Result<R<Self>>

impl FormatDesc

pub fn as_be_sound_desc_cm_buf_in( &self, flavor: Option<&SoundDescFlavor>, allocator: Option<&Allocator>, ) -> Result<R<BlockBuf>>

Copies the contents of a CMAudioFormatDescription to a cm::BlockBuf in big-endian byte ordering.

Note that the dataRefIndex field of the SampleDescription is intentionally filled with garbage values (0xFFFF). The caller must overwrite these values with a valid dataRefIndex if writing the SampleDescription to a QuickTime/ISO file.

pub fn as_be_sound_desc_cm_buf( &self, flavor: Option<&SoundDescFlavor>, ) -> Result<R<BlockBuf>>

pub fn from_be_sound_desc_data_in( data: &[u8], flavor: Option<&SoundDescFlavor>, allocator: Option<&Allocator>, ) -> Result<R<Self>>

pub fn from_be_sound_desc_data( data: &[u8], flavor: Option<&SoundDescFlavor>, ) -> Result<R<Self>>

pub fn from_be_sound_desc_buf_in( buffer: &BlockBuf, flavor: Option<&SoundDescFlavor>, allocator: Option<&Allocator>, ) -> Result<R<Self>>

pub fn from_be_sound_desc_buf( buffer: &BlockBuf, flavor: Option<&SoundDescFlavor>, ) -> Result<R<Self>>