cidre::cm::sync

Struct Clock

pub struct Clock(/* private fields */);

A timing source object.

A clock represents a source of time information: generally, a piece of hardware that measures the passage of time. One example of a clock is the host time clock, accessible via cm::Clock::host_time_clock(). It measures time using the CPU system clock, which on Mac OS X is mach_absolute_time(). Every audio device can also be considered a clock since the audio samples that it outputs or inputs each have a defined duration (eg, 1/48000 of a second for 48 kHz audio).

cm::Clocks are read-only: they cannot be stopped or started, and the current time cannot be set. A CMClock has one primary function, cm::Clock::get_time, which tells what time it is now. Additionally, the cm::Sync infrastructure monitors relative drift between cm::Clocks.

Implementations

impl Clock

pub fn retained(&self) -> R<Self>

impl Clock

pub fn get_type_id() -> TypeId

pub fn host_time_clock() -> &'static Clock

Returns a reference to the singleton clock logically identified with host time.

pub fn time(&self) -> Time

pub fn convert_host_time_to_sys_units(host_time: Time) -> u64

Converts a host time from CMTime to the host time’s native units.

This function performs a scale conversion, not a clock conversion. It can be more accurate than CMTimeConvertScale because the system units may have a non-integer timescale. On Mac OS X, this function converts to the units of mach_absolute_time.

pub fn make_host_time_from_sys_units(host_time: u64) -> Time

Converts a host time from native units to cm::Time.

pub fn anchor_time( &self, clock_time_out: &mut Time, ref_clock_time_out: &mut Time, ) -> Result

pub fn might_dirft(&self, other: &Clock) -> bool

Indicates whether it is possible for two clocks to drift relative to each other.

pub fn invalidate(&mut self)

Makes the clock stop functioning.

After invalidation, the clock will return errors from all APIs. This should only be called by the “owner” of the clock, who knows (for example) that some piece of hardware has gone away, and the clock will no longer work (and might even crash).

impl Clock

pub fn with_default_audio_output() -> Result<R<Self>>

Unified api with iOS

pub fn with_audio_device_uid_in( device_uid: Option<&String>, allocator: Option<&Allocator>, ) -> Result<Option<R<Self>>>

pub fn with_audio_device_uid(device_uid: Option<&String>) -> Result<R<Self>>

pub fn with_audio_device_in( audio_device: Device, allocator: Option<&Allocator>, ) -> Result<Option<R<Self>>>

pub fn with_audio_device(audio_device: Device) -> Result<R<Self>>

pub fn set_audio_device_uid(&mut self, device_uid: Option<&String>) -> Result

pub fn set_audio_device(&mut self, audio_device: Device) -> Result

pub fn audio_device_uid(&self) -> Result<Option<&String>>

pub fn audio_device(&self) -> Result<Device>

pub fn is_tracking_default_device(&self) -> Result<bool>