cidre::cm::sync

Struct Timebase

pub struct Timebase(/* private fields */);

Models a timeline under application control.

A timebase represents a timeline that clients can control by setting the rate and time. Each timebase has either a source clock or a source timebase (previously referred to as a master clock or master timebase). The rate of the timebase is expressed relative to its source. When a timebase has rate 0.0, its time is fixed and does not change as its source’s time changes. When a timebase has rate 1.0, its time increases one second as its source’s time increases by one second. When a timebase has rate 2.0, its time increases two seconds as its source’s time increases by one second. When a timebase has rate -1.0, its time decreases one second as its source’s time increases by one second.

If a timebase has a source timebase, the source timebase’s rate is a factor in determining the timebase’s effective rate. In fact, a timebase’s effective rate is defined as the product of its rate, its source timebase’s rate, its source timebase’s source timebase’s rate, and so on up to the ultimate source clock. This is the rate at which the timebase’s time changes relative to the ultimate source clock.

Implementations

impl Timebase

pub fn retained(&self) -> R<Self>

impl Timebase

pub const VERY_LONG_CF_TIME_INTERVAL: TimeInterval = 8073216000f64

quite a while

pub const FAR_FUTURE_CF_ABS_TIME: AbsTime = 8073216000f64

quite a while from 2001

pub fn get_type_id() -> TypeId

pub fn with_src_clock_in( src_clock: &Clock, allocator: Option<&Allocator>, ) -> Result<R<Self>>

pub fn with_src_clock(src_clock: &Clock) -> Result<R<Self>>

Creates a timebase driven by the given clock.

The timebase will initially have rate zero and time zero. Pass cm::Clock::host_time_clock() for source_clock to have the host time clock drive this timebase.

pub fn with_src_timebase_in( src_timebase: &Timebase, allocator: Option<&Allocator>, ) -> Result<R<Self>>

pub fn with_src_timebase(src_timebase: &Timebase) -> Result<R<Self>>

Creates a timebase driven by the given source timebase.

The timebase will initially have rate zero and time zero.

pub fn src_timebase(&self) -> Option<R<Timebase>>

Returns the immediate source timebase of a timebase.

Returns None if the timebase actually has a source clock instead of a source timebase.

pub fn src_clock(&self) -> Option<R<Clock>>

Returns the immediate source clock of a timebase.

Returns None if the timebase actually has a source timebase instead of a source clock.

pub fn src(&self) -> R<ClockOrTimebase>

Returns the immediate source (either timebase or clock) of a timebase.

pub fn ultimate_src_clock(&self) -> R<Clock>

Returns the source clock that is the source of all of a timebase’s source timebases.

pub fn set_src_clock(&mut self, val: &Clock) -> Result

Sets the source clock of a timebase.

The timebase will stop receiving timing information from its current source clock or source timebase, and will begin receiving timing information from the new source clock. Prior to the change a cm::Timebase::notification_src_will_change will be posted. When the change has completed, a cm::Timebase::notification_src_did_change notification will be posted.

pub fn set_src_timebase(&mut self, val: &Timebase) -> Result

Sets the source timebase of a timebase.

The timebase will stop receiving timing information from its current source clock or source timebase, and will begin receiving timing information from the new source timebase. Prior to the change a cm::Timebase::notification_src_will_change will be posted. When the change has completed, a cm::Timebase::notification_src_did_change notification will be posted.

pub fn notification_src_will_change() -> &'static NotificationName

pub fn notification_src_did_change() -> &'static NotificationName

pub fn notification_effective_rate_changed() -> &'static NotificationName

pub fn notification_time_jumped() -> &'static NotificationName

pub fn notification_key_event_time() -> &'static String

pub fn time(&self) -> Time

Retrieves the current time from a timebase.

pub fn time_with_scale( &self, timescale: TimeScale, method: TimeRoundingMethod, ) -> Time

Retrieves the current time from a timebase in the specified timescale.

pub fn set_time(&mut self, val: Time) -> Result

Sets the current time of a timebase.

pub fn set_achor_time( &mut self, timebase_time: Time, immediate_src_time: Time, ) -> Result

Sets the time of a timebase at a particular source time.

cm::Timebase::time’s results will be interpolated from that anchor time.

pub fn rate(&self) -> f64

Retrieves the current rate of a timebase.

This is the rate relative to its immediate source clock or timebase.
For example, if a timebase is running at twice the rate of its source, its rate is 2.0.

pub fn time_and_rate(&self, time_out: &mut Time, rate_out: &mut f64) -> Result

Retrieves the current time and rate of a timebase.

You can use this function to take a consistent snapshot of the two values, avoiding possible inconsistencies due to external changes between retrieval of time and rate.

pub fn set_rate(&mut self, val: f64) -> Result

Sets the rate of a timebase.

pub fn set_rate_and_anchor_time( &mut self, rate: f64, timebase_time: Time, immediate_src_time: Time, ) -> Result

Sets the time of a timebase at a particular source time, and changes the rate at exactly that time.

pub fn effective_rate(&self) -> f64

Gets the effective rate of a timebase (which combines its rate with the rates of all its source timebases).

pub fn add_timer(&mut self, timer: &RunLoopTimer, run_loop: &RunLoop) -> Result

Adds the timer to the list of timers managed by the timebase.

The timer must be a repeating run loop timer (with a very long interval at least as long as cm::Timebase::VERY_LONG_CF_TIME_INTERVAL), attached to a runloop.
The timebase will retain the timer, and will maintain its “NextFireDate” according to the cm::Time set using CMTimebaseSetTimerNextFireTime. Until the first call to CMTimebaseSetTimerNextFireTime, the “NextFireDate” will be set far, far in the future. The runloop that timer is attached to must be passed in and the timebase will retain that runloop. The retained runloop will be used to call cf::RunLoop::wake_up() any time the timebase modifies the timer’s fire date.

pub fn remove_timer(&mut self, timer: &RunLoopTimer) -> Result

Removes the timer from the list of timers managed by the timebase.

The timebase will no longer maintain the timer’s “NextFireDate”. If the timer is invalidated, the timebase will eventually remove it from its list and release it even if this function is not called.

pub fn set_timer_next_fire_time( &self, timer: &mut RunLoopTimer, fire_time: Time, ) -> Result

Sets the cm::Time on the timebase’s timeline at which the timer should next be fired.

The timer must be on the list of timers managed by the timebase. The timebase will continue to update the timer’s “NextFireDate” according to time jumps and effective rate changes. If fireTime is not numeric, or if the timebase is not moving, the “NextFireDate” will be set to a date far, far in the future.

IMPORTANT NOTE: Due to the way that cf::RunLoopTimers are implemented, if a timer passes through a state in which it is due to fire, it may fire even if its rescheduled before the runloop runs again. Clients should take care to avoid temporarily scheduling timers in the past. For example, set the timebase’s rate or time before you set the timer’s next fire time, if you are doing both at once. (If setting the timebase’s rate or time might put the timer’s fire time in the past, you may need to set the fire time to cm::Time::invalid() across the timebase change.)

pub fn set_timer_to_fire_immediately(&self, timer: &mut RunLoopTimer) -> Result

Sets the timer to fire immediately once, overriding any previous cm::Timebase::set_timer_next_fire_time call.

pub fn add_timer_dispatch_src(&mut self, timer_src: &TimerSrc) -> Result

Adds the timer dispatch source to the list of timers managed by the timebase.

pub fn remove_timer_dispatch_src(&mut self, timer_src: &TimerSrc) -> Result

Removes the timer dispatch source from the list of timers managed by the timebase.

The timebase will no longer maintain the timer source’s start time. If the timer source is cancelled, the timebase will eventually remove it from its list and release it even if this function is not called.

pub fn set_timer_dispatch_src_next_fire_time( &mut self, timer_src: &TimerSrc, fire_time: Time, ) -> Result

Sets the cm::Time on the timebase’s timeline at which the timer dispatch source should next be fired.

The timer source must be on the list of timers managed by the timebase. The timebase will continue to update the timer dispatch source’s start time according to time jumps and effective rate changes. If fireTime is not numeric, or if the timebase is not moving, the start time will be set to DISPATCH_TIME_FOREVER.

IMPORTANT NOTE: Due to the way that timer dispatch sources are implemented, if a timer passes through a state in which it is due to fire, it may fire even if its rescheduled before the event handler is run. Clients should take care to avoid temporarily scheduling timers in the past. For example, set the timebase’s rate or time before you set the timer’s next fire time, if you are doing both at once. (If setting the timebase’s rate or time might put the timer’s fire time in the past, you may need to set the fire time to kCMTimeInvalid across the timebase change.)

pub fn set_timer_dispatch_src_to_fire_immediately( &mut self, timer_src: &TimerSrc, ) -> Result

Sets the timer dispatch source to fire immediately once, overriding any previous cm::Timebase::set_timer_dispatch_src_next_fire_time call.

pub fn notification_barrier(&self) -> Result

Requests that the timebase wait until it is not posting any notifications.