cidre::cm

Type Alias AudioFormatDesc

pub type AudioFormatDesc = FormatDesc;

Aliased Type

pub struct AudioFormatDesc(/* private fields */);

Implementations

impl AudioFormatDesc

pub fn with_asbd(asbd: &StreamBasicDesc) -> Result<R<Self>>

pub fn audio_in( asbd: &StreamBasicDesc, layout_size: usize, layout: Option<&AudioChannelLayout<1>>, magic_cookie_size: usize, magic_cookie: Option<&c_void>, extensions: Option<&DictionaryOf<FormatDescExtKey, Type>>, format_description_out: *mut Option<R<Self>>, allocator: Option<&Allocator>, ) -> Result

pub fn stream_basic_desc(&self) -> Option<&StreamBasicDesc>

Returns a read-only pointer to the audio stream description in an audio format description.

This API is specific to audio format descriptions, and returns None if used with a non-audio format descriptions.

impl AudioFormatDesc

pub fn as_be_sound_desc_cm_buf_in( &self, flavor: Option<&SoundDescFlavor>, allocator: Option<&Allocator>, ) -> Result<R<BlockBuf>>

Copies the contents of a CMAudioFormatDescription to a cm::BlockBuf in big-endian byte ordering.

Note that the dataRefIndex field of the SampleDescription is intentionally filled with garbage values (0xFFFF). The caller must overwrite these values with a valid dataRefIndex if writing the SampleDescription to a QuickTime/ISO file.

pub fn as_be_sound_desc_cm_buf( &self, flavor: Option<&SoundDescFlavor>, ) -> Result<R<BlockBuf>>

pub fn from_be_sound_desc_data_in( data: &[u8], flavor: Option<&SoundDescFlavor>, allocator: Option<&Allocator>, ) -> Result<R<Self>>

pub fn from_be_sound_desc_data( data: &[u8], flavor: Option<&SoundDescFlavor>, ) -> Result<R<Self>>

pub fn from_be_sound_desc_buf_in( buffer: &BlockBuf, flavor: Option<&SoundDescFlavor>, allocator: Option<&Allocator>, ) -> Result<R<Self>>

pub fn from_be_sound_desc_buf( buffer: &BlockBuf, flavor: Option<&SoundDescFlavor>, ) -> Result<R<Self>>