cidre::cm

Type Alias VideoFormatDesc

pub type VideoFormatDesc = FormatDesc;

Aliased Type

pub struct VideoFormatDesc(/* private fields */);

Implementations

impl VideoFormatDesc

pub fn video( codec: VideoCodec, width: i32, height: i32, extensions: Option<&DictionaryOf<FormatDescExtKey, Type>>, ) -> Result<R<Self>>

use cidre::cm;

let desc = cm::VideoFormatDesc::video(cm::VideoCodec::H264, 1920, 1080, None).unwrap();

pub fn create_video_in( codec: VideoCodec, width: i32, height: i32, extensions: Option<&DictionaryOf<FormatDescExtKey, Type>>, format_description_out: *mut Option<R<VideoFormatDesc>>, allocator: Option<&Allocator>, ) -> Result

pub fn dims(&self) -> VideoDimensions

Returns the dimensions (in encoded pixels)

This does not take into account pixel aspect ratio or clean aperture tags.

pub fn matches_image_buf(&self, image_buffer: &ImageBuf) -> bool

pub fn tag_collections(&self) -> Result<R<ArrayOf<TagCollection>>>

Availability

macos_14_0, ios_17_0, tvos_17_0, watchos_10_0, visionos_1_0

pub fn with_image_buf_in( image_buffer: &ImageBuf, allocator: Option<&Allocator>, ) -> Result<R<VideoFormatDesc>>

pub fn with_image_buf(image_buffer: &ImageBuf) -> Result<R<VideoFormatDesc>>

pub fn with_h264_param_sets<const N: usize>( pointers: &[*const u8; N], sizes: &[usize; N], nal_unit_header_length: i32, ) -> Result<R<VideoFormatDesc>>

pub fn with_h264_param_sets_in<const N: usize>( pointers: &[*const u8; N], sizes: &[usize; N], nal_unit_header_length: i32, allocator: Option<&Allocator>, ) -> Result<R<VideoFormatDesc>>

pub fn with_hevc_param_sets( count: usize, pointers: &[*const u8], sizes: &[usize], nal_unit_header_length: i32, extensions: Option<&DictionaryOf<FormatDescExtKey, Type>>, ) -> Result<R<VideoFormatDesc>>

pub fn with_hevc_param_sets_in( count: usize, pointers: &[*const u8], sizes: &[usize], nal_unit_header_length: i32, extensions: Option<&DictionaryOf<FormatDescExtKey, Type>>, allocator: Option<&Allocator>, ) -> Result<R<VideoFormatDesc>>

pub fn h264_params_count_and_header_len(&self) -> Result<(usize, i32)>

pub unsafe fn hevc_param_set_at_index( &self, parameter_set_index: usize, parameter_set_pointer_out: *mut *const u8, parameter_set_size_out: *mut usize, parameter_set_count_out: *mut usize, nal_unit_header_length_out: *mut i32, ) -> Result

pub fn hevc_params_count_and_header_len(&self) -> Result<(usize, i32)>

pub fn h264_param_set_at(&self, index: usize) -> Result<&[u8]>

pub fn hevc_param_set_at(&self, index: usize) -> Result<&[u8]>

pub fn common_image_buf_ext_keys() -> &'static ArrayOf<FormatDescExtKey>

Returns an array of the keys that are used both as cm::VideoFormatDesc extensions and cv::ImageBuf attachments and attributes.

impl VideoFormatDesc

pub fn as_be_image_desc_cm_buf_in( &self, string_encoding: StringEncoding, flavor: Option<&ImageDescFlavor>, allocator: Option<&Allocator>, ) -> Result<R<BlockBuf>>

Copies the contents of a cm::VideoFormatDescription to a cm::BlockBuffer in big-endian byte ordering.

pub fn as_be_image_desc_cm_buf( &self, flavor: Option<&ImageDescFlavor>, ) -> Result<R<BlockBuf>>

pub fn from_be_image_desc_buf_in( image_description_block_buffer: &BlockBuf, string_encoding: StringEncoding, flavor: Option<&ImageDescFlavor>, allocator: Option<&Allocator>, ) -> Result<R<Self>>

pub fn from_be_image_desc_buf( image_description_block_buffer: &BlockBuf, flavor: Option<&ImageDescFlavor>, ) -> Result<R<Self>>

pub fn from_be_image_desc_data_in( data: &[u8], string_encoding: StringEncoding, flavor: Option<&ImageDescFlavor>, allocator: Option<&Allocator>, ) -> Result<R<Self>>

pub fn from_be_image_desc_data( data: &[u8], string_encoding: StringEncoding, flavor: Option<&ImageDescFlavor>, ) -> Result<R<Self>>