cidre

Module core_audio

Re-exports

pub use hardware::AggregateDevice;

pub use hardware::Device;

pub use hardware::DeviceIoBlock;

pub use hardware::DeviceIoProc;

pub use hardware::DeviceIoProcId;

pub use hardware::Process;

pub use hardware::PropListenerBlock;

pub use hardware::PropListenerFn;

pub use hardware::Stream;

pub use hardware::StreamDir;

pub use hardware::System;

pub use hardware::aggregate_device_keys;

pub use hardware::device_start;

pub use hardware::sub_device_keys;

pub use hardware_tapping::Tap;

pub use hardware_tapping::TapGuard;

Modules

hardware

hardware_err

The error constants unique to the HAL.

hardware_tapping

Structs

Class

DeviceTransportType

Transport Type IDs

Obj

PropAddr

PropElement

PropScope

PropSelector

StreamRangedDesc

StreamTerminalType

TapDesc

This class describes a tap object that contains an input stream. The input stream is a mix of all of the specified processes output audio.

Enums

TapMuteBehavior