cidre::core_audio::hardware

Module aggregate_device_keys

Functions

clock_device

The key used in a CFDictionary that describes the composition of an AudioAggregateDevice. The value for this key is a CFString that contains the UID for the clock device that is the time source for the AudioAggregateDevice. If the aggregate device includes both a main audio device and a clock device, the clock device will control the time base.

is_private

The key used in a CFDictionary that describes the composition of an AudioAggregateDevice. The value for this key is a CFNumber where a value of 0 means that the AudioAggregateDevice is to be published to the entire system and a value of 1 means that the AudioAggregateDevice is private to the process that created it. Note that a private AudioAggregateDevice is not persistent across launches of the process that created it. Note that if this key is not present, it implies that the AudioAggregateDevice is published to the entire system.

is_stacked

The key used in a CFDictionary that describes the composition of an AudioAggregateDevice. The value for this key is a CFNumber where a value of 0 means that the sub-devices of the AudioAggregateDevice are arranged such that the output streams are all fed the same data.

main_sub_device

The key used in a cf::Dictionary that describes the composition of an AudioAggregateDevice. The value for this key is a cf::String that contains the UID for the sub-device that is the time source for the AudioAggregateDevice.

name

The key used in a CFDictionary that describes the composition of an AudioAggregateDevice. The value for this key is a CFString that contains the human readable name of the AudioAggregateDevice.

sub_device_list

The key used in a CFDictionary that describes the composition of an AudioAggregateDevice. The value for this key is a CFArray of CFDictionaries that describe each sub-device in the AudioAggregateDevice. The keys for this CFDictionary are defined in the AudioSubDevice section.

tap_auto_start

The key used in a CFDictionary that describes the composition of an AudioAggregateDevice. The value for this key is a CFNumber where a non-zero
value indicates that this aggregate device’s start should wait for the first tap that receives audio. When this key is used, calling AudioDeviceStart with the aggregate device will wait until a tapped process begins receiving its
first audio from any tapped applications. The composition must also include the private key so that the aggregate is private to the process that created it.

tap_list

The key used in a CFDictionary that describes the Tap composition of an AudioAggregateDevice. The value for this key is a CFArray of CFDictionaries that describe each tap in the AudioAggregateDevice. The keys for this CFDictionary are defined in the AudioTap section.

uid

The key used in a CFDictionary that describes the composition of an AudioAggregateDevice. The value for this key is a CFString that contains the UID of the AudioAggregateDevice.