cidre::core_audio::hardware

Struct Device

#[repr(transparent)]
pub struct Device(pub Obj);

Tuple Fields
0: Obj
Implementations

impl Device

pub fn with_uid(uid: &String) -> Result<Self>

pub fn is_unknown(&self) -> bool

pub fn uid(&self) -> Result<R<String>>

pub fn is_alive(&self) -> Result<bool>

pub fn is_running(&self) -> Result<bool>

pub fn nominal_sample_rate(&self) -> Result<f64>

pub fn set_nominal_sample_rate(&mut self, val: f64) -> Result

pub fn actual_sample_rate(&self) -> Result<f64>

A f64 that indicates the current actual sample rate of the AudioDevice as measured by its time stamps.

pub fn asbd(&self, scope: PropScope) -> Result<AudioStreamBasicDesc>

pub fn input_asbd(&self) -> Result<AudioStreamBasicDesc>

pub fn output_asbd(&self) -> Result<AudioStreamBasicDesc>

pub fn available_nominal_sample_rates(&self) -> Result<Vec<ValueRange>>

pub fn stream_cfg(&self, scope: PropScope) -> Result<AudioBufListN>

pub fn input_stream_cfg(&self) -> Result<AudioBufListN>

pub fn output_stream_cfg(&self) -> Result<AudioBufListN>

pub fn streams(&self) -> Result<Vec<Stream>>

pub fn transport_type(&self) -> Result<DeviceTransportType>

pub fn clock_uid(&self) -> Result<R<String>>

A cf::String that contains the UID for the AudioClockDevice that is currently serving as the main time base of the device.

pub fn is_process_muted(&self, scope: PropScope) -> Result<bool>

Indicates that the current process’s audio will be zeroed out by the system.

Note that this property does not apply to aggregate devices, just real, physical devices.

pub fn is_process_input_muted(&self) -> Result<bool>

pub fn is_process_output_muted(&self) -> Result<bool>

pub fn set_process_muted(&self, scope: PropScope, val: bool) -> Result

pub fn set_process_input_muted(&self, val: bool) -> Result

pub fn set_process_output_muted(&self, val: bool) -> Result

pub fn create_io_proc_id<const IN: usize, const ON: usize, T>( &self, proc: DeviceIoProc<IN, ON, T>, client_data: Option<&mut T>, ) -> Result<DeviceIoProcId>

pub fn create_io_proc_id_with_block<const IN: usize, const ON: usize>( &self, dispatch_queue: Option<&Queue>, block: &mut DeviceIoBlock<IN, ON>, ) -> Result<DeviceIoProcId>

pub fn buf_frame_size(&self) -> Result<u32>

pub fn set_buf_frame_size(&mut self, val: u32) -> Result

pub fn buf_frame_size_range(&self) -> Result<ValueRange>

pub fn uses_variable_buf_sizes(&self) -> Result<bool>

pub fn is_vad_enabled(&self) -> Result<bool>

Voice activity detection can be used with input audio and has echo cancellation. Detection works when a process mute is used, but not with hardware mute.

pub fn set_vad_enabled(&self, val: bool) -> Result

pub fn vad_state(&self) -> Result<bool>

pub fn input_volume_scalar(&self) -> Result<f32>

pub fn set_input_volume_scalar(&mut self, val: f32) -> Result

pub fn input_volume_db(&self) -> Result<f32>

pub fn set_input_volume_db(&mut self, val: f32) -> Result

pub fn input_volume_db_range(&self) -> Result<ValueRange>

pub fn input_volume_scalar_to_db(&self, val: &mut f32) -> Result

pub fn input_volume_db_to_scalar(&self, val: &mut f32) -> Result

impl Device

pub fn duck( &self, in_ducked_level: f32, in_start_time: Option<&AudioTimeStamp>, in_ramp_duration: f32, ) -> Result