cidre::core_audio

Module hardware_err

The error constants unique to the HAL.

These are the error constants that are unique to the HAL. Note that the HAL’s functions can and will return other codes that are not listed here. While these constants give a general idea of what might have gone wrong during the execution of an API call, if an API call returns anything other than os::Status::NO_ERR it is to be viewed as the same failure regardless of what constant is actually returned.

Constants

BAD_DEVICE

The AudioObjectID passed to the function doesn’t map to a valid AudioDevice.

BAD_OBJ

The AudioObjectID passed to the function doesn’t map to a valid AudioObject.

BAD_PROP_SIZE

An improperly sized buffer was provided when accessing the data of a property.

BAD_STREAM

The AudioObjectID passed to the function doesn’t map to a valid AudioStream.

ILLEGAL_OP

The requested operation couldn’t be completed.

NOT_READY

The AudioObject isn’t ready to do the requested operation.

NOT_RUNNING

The function call requires that the hardware be running but it isn’t.

PERMISSIONS

The requested operation can’t be completed because the process doesn’t have permission.

UNKNOWN_PROP

The AudioObject doesn’t know about the property at the given address.

UNSPECIFIED

The function call failed while doing something that doesn’t provide any error messages.

UNSUPPORTED_FORMAT

The AudioStream doesn’t support the requested format.

UNSUPPORTED_OP

The AudioObject doesn’t support the requested operation.