cidre::core_audio

Struct PropSelector

#[repr(transparent)]
pub struct PropSelector(pub u32);

Tuple Fields
0: u32
Implementations

impl PropSelector

pub const fn global_addr(self) -> PropAddr

pub const fn input_addr(self) -> PropAddr

pub const fn output_addr(self) -> PropAddr

pub const fn addr(self, scope: PropScope, element: PropElement) -> PropAddr

impl PropSelector

pub const WILDCARD: Self

pub const BASE_CLASS: Self

An AudioClassId that identifies the class from which the class of the AudioObject is derived. This value must always be one of the standard classes.

pub const CLASS: Self

An AudioClassId that identifies the class of the AudioObject.

pub const OWNER: Self

An AudioObjectId that identifies the the AudioObject that owns the given AudioObject. Note that all AudioObjects are owned by some other AudioObject. The only exception is the AudioSystemObject, for which the value of this property is AudioObject::UNKNOWN.

pub const NAME: Self

A cf::String that contains the human readable name of the object. The caller is responsible for releasing the returned cf::Object.

pub const MODEL_NAME: Self

A cf::String that contains the human readable model name of the object. The model name differs from kAudioObjectPropertyName in that two objects of the same model will have the same value for this property but may have different values for kAudioObjectPropertyName.

pub const MANUFACTURER: Self

A cf::String that contains the human readable name of the manufacturer of the hardware the AudioObject is a part of. The caller is responsible for releasing the returned CFObject.

pub const ELEMENT_NAME: Self

A cf::String that contains a human readable name for the given element in the given scope. The caller is responsible for releasing the returned CFObject.

pub const ELEMENT_CATEGORY_NAME: Self

A cf::String that contains a human readable name for the category of the given element in the given scope. The caller is responsible for releasing the returned CFObject.

pub const ELEMENT_NUMBER_NAME: Self

A cf::String that contains a human readable name for the number of the given element in the given scope. The caller is responsible for releasing the returned CFObject.

pub const OWNED_OBJS: Self

An array of AudioObjectIDs that represent all the AudioObjects owned by the given object. The qualifier is an array of AudioClassIDs. If it is non-empty, the returned array of AudioObjectIDs will only refer to objects whose class is in the qualifier array or whose is a subclass of one in the qualifier array.

pub const IDENTIFY: Self

A u32 where a value of one indicates that the object’s hardware is drawing attention to itself, typically by flashing or lighting up its front panel display. A value of 0 indicates that this function is turned off. This makes it easy for a user to associate the physical hardware with its representation in an application. Typically, this property is only supported by AudioDevices and AudioBoxes.

pub const SERIAL_NUMBER: Self

A cf::String that contains the human readable serial number for the object. This property will typically be implemented by AudioBox and AudioDevice objects. Note that the serial number is not defined to be unique in the same way that an AudioBox’s or AudioDevice’s UID property are defined. This is purely an informational value. The caller is responsible for releasing the returned CFObject.

pub const FIRMWARE_VERSION: Self

A cf::String that contains the human readable firmware version for the object. This property will typically be implemented by AudioBox and AudioDevice objects. Note that this is purely an informational value. The caller is responsible for releasing the returned CFObject.

impl PropSelector

AudioPlugIn Properties

pub const PLUG_IN_BUNDLE_ID: Self

pub const PLUG_IN_DEVICE_LIST: Self

pub const PLUG_IN_TRANSLATE_UID_TO_DEVICE: Self

pub const PLUG_IN_BOX_LIST: Self

pub const PLUG_IN_TRANSLATE_UID_TO_BOX: Self

pub const PLUG_IN_CLOCK_DEVICE_LIST: Self

pub const PLUG_IN_TRANSLATE_UID_TO_CLOCK_DEVICE: Self

impl PropSelector

AudioTransportManager Properties

pub const TRANSPORT_MANAGER_END_POINT_LIST: Self

pub const TRANSPORT_MANAGER_TRANSLATE_UID_TO_END_POINT: Self

pub const TRANSPORT_MANAGER_TRANSPORT_TYPE: Self

impl PropSelector

AudioObjectPropertySelector values provided by the AudioBox class.

pub const BOX_UID: Self

pub const BOX_TRANSPORT_TYPE: Self

pub const BOX_HAS_AUDIO: Self

pub const BOX_HAS_VIDEO: Self

pub const BOX_HAS_MIDI: Self

pub const BOX_IS_PROTECTED: Self

pub const BOX_ACQUIRED: Self

pub const BOX_ACQUISITION_FAILED: Self

pub const BOX_DEVICE_LIST: Self

pub const BOX_CLOCK_DEVICE_LIST: Self

impl PropSelector

Values provided by the AudioDevice class

pub const DEVICE_CONFIGURATION_APP: Self

A cf::String that contains the bundle ID for an application that provides a GUI for configuring the AudioDevice.

By default, the value of this property is the bundle ID for Audio MIDI Setup. The caller is responsible for releasing the returned cf::Obj.

pub const DEVICE_UID: Self

A cf::String that contains a persistent identifier for the AudioDevice. An AudioDevice’s UID is persistent across boots. The content of the UID string is a black box and may contain information that is unique to a particular instance of an AudioDevice’s hardware or unique to the CPU. Therefore they are not suitable for passing between CPUs or for identifying similar models of hardware. The caller is responsible for releasing the returned cf::Obj.

pub const DEVICE_MODEL_UID: Self

A cf::String that contains a persistent identifier for the model of an AudioDevice. The identifier is unique such that the identifier from two AudioDevices are equal if and only if the two AudioDevices are the exact same model from the same manufacturer. Further, the identifier has to be the same no matter on what machine the AudioDevice appears. The caller is responsible for releasing the returned cf::Obj.

pub const DEVICE_TRANSPORT_TYPE: Self

A u32 whose value indicates how the AudioDevice is connected to the CPU. Constants for some of the values for this property can be found in the enum in the AudioDevice Constants section of this file.

pub const DEVICE_RELATED_DEVICES: Self

An array of AudioDeviceIDs for devices related to the AudioDevice. For IOAudio-based devices, AudioDevices are related if they share the same IOAudioDevice object.

pub const DEVICE_CLOCK_DOMAIN: Self

A u32 whose value indicates the clock domain to which this AudioDevice belongs. AudioDevices that have the same value for this property are able to be synchronized in hardware. However, a value of 0 indicates that the clock domain for the device is unspecified and should be assumed to be separate from every other device’s clock domain, even if they have the value of 0 as their clock domain as well.

pub const DEVICE_IS_ALIVE: Self

A u32 where a value of 1 means the device is ready and available and 0 means the device is unusable and will most likely go away shortly.

pub const DEVICE_IS_RUNNING: Self

A u32 where a value of 0 means the AudioDevice is not performing IO and a value of 1 means that it is. Note that the device can be running even if there are no active IOProcs such as by calling AudioDeviceStart() and passing a NULL IOProc. Note that the notification for this property is usually sent from the AudioDevice’s IO thread.

pub const DEVICE_CAN_BE_DEFAULT_DEVICE: Self

A u32 where 1 means that the AudioDevice is a possible selection for kAudioHardwarePropertyDefaultInputDevice or kAudioHardwarePropertyDefaultOutputDevice depending on the scope.

pub const DEVICE_CAN_BE_DEFAULT_SYS_DEVICE: Self

A u32 where 1 means that the AudioDevice is a possible selection for kAudioHardwarePropertyDefaultSystemOutputDevice.

pub const DEVICE_LATENCY: Self

A u32 containing the number of frames of latency in the AudioDevice. Note that input and output latency may differ. Further, the AudioDevice’s AudioStreams may have additional latency so they should be queried as well. If both the device and the stream say they have latency, then the total latency for the stream is the device latency summed with the stream latency.

pub const DEVICE_STREAMS: Self

An array of AudioStreamIDs that represent the AudioStreams of the AudioDevice. Note that if a notification is received for this property, any cached AudioStreamIDs for the device become invalid and need to be re-fetched.

pub const CTRL_LIST: Self

An array of AudioObjectIDs that represent the AudioControls of the AudioDevice. Note that if a notification is received for this property, any cached AudioObjectIDs for the device become invalid and need to be re-fetched.

pub const DEVICE_SAFETY_OFFSET: Self

A u32 whose value indicates the number for frames in ahead (for output) or behind (for input the current hardware position that is safe to do IO.

pub const DEVICE_NOMINAL_SAMPLE_RATE: Self

A f64 that indicates the current nominal sample rate of the AudioDevice.

pub const DEVICE_AVAILABLE_NOMINAL_SAMPLE_RATES: Self

An array of AudioValueRange structs that indicates the valid ranges for the nominal sample rate of the AudioDevice.

pub const DEVICE_ICON: Self

A cf::Url that indicates an image file that can be used to represent the device visually. The caller is responsible for releasing the returned cf::Obj.

pub const DEVICE_IS_HIDDEN: Self

A u32 where a non-zero value indicates that the device is not included in the normal list of devices provided by kAudioHardwarePropertyDevices nor can it be the default device. Hidden devices can only be discovered by knowing their UID and using kAudioHardwarePropertyDeviceForUID.

pub const DEVICE_PREFERRED_CHANNELS_FOR_STEREO: Self

An array of two u32s, the first for the left channel, the second for the right channel, that indicate the channel numbers to use for stereo IO on the device. The value of this property can be different for input and output and there are no restrictions on the channel numbers that can be used.

pub const DEVICE_PREFERRED_CHANNEL_LAYOUT: Self

An AudioChannelLayout that indicates how each channel of the AudioDevice should be used.

impl PropSelector

AudioObjectPropertySelector values provided by the AudioStream class.

pub const STREAM_IS_ACTIVE: Self

A u32 where a non-zero value indicates that the stream is enabled and doing IO.

pub const STREAM_DIRECTION: Self

A u32 where a value of 0 means that this AudioStream is an output stream and a value of 1 means that it is an input stream.

pub const STREAM_TERMINAL_TYPE: Self

A u32 whose value describes the general kind of functionality attached to the AudioStream.

pub const STREAM_STARTING_CHANNEL: Self

A u32 that specifies the first element in the owning device that corresponds to element one of this stream.

pub const STREAM_LATENCY: Self = Self::DEVICE_LATENCY

A u32 containing the number of frames of latency in the AudioStream. Note that the owning AudioDevice may have additional latency so it should be queried as well. If both the device and the stream say they have latency, then the total latency for the stream is the device latency summed with the stream latency.

pub const STREAM_VIRTUAL_FORMAT: Self

An AudioStreamBasicDescription that describes the current data format for the AudioStream. The virtual format refers to the data format in which all IOProcs for the owning AudioDevice will perform IO transactions.

pub const STREAM_AVAILABLE_VIRTUAL_FORMATS: Self

An array of AudioStreamRangedDescriptions that describe the available data formats for the AudioStream. The virtual format refers to the data format in which all IOProcs for the owning AudioDevice will perform IO transactions.

pub const STREAM_PHYSICAL_FORMAT: Self

An AudioStreamBasicDescription that describes the current data format for the AudioStream. The physical format refers to the data format in which the hardware for the owning AudioDevice performs its IO transactions.

pub const STREAM_AVAILABLE_PHYSICAL_FORMATS: Self

An array of AudioStreamRangedDescriptions that describe the available data formats for the AudioStream. The physical format refers to the data format in which the hardware for the owning AudioDevice performs its IO transactions.

impl PropSelector

ca::PropSelector values provided by the AudioClockDevice class.

The AudioClockDevice class is a subclass of the AudioObject class. The class has just the global scope, kAudioObjectPropertyScopeGlobal, and only a main element.

pub const CLOCK_DEVICE_UID: Self

A CFString that contains a persistent identifier for the AudioClockDevice. An AudioClockDevice’s UID is persistent across boots. The content of the UID string is a black box and may contain information that is unique to a particular instance of an clock’s hardware or unique to the CPU. Therefore they are not suitable for passing between CPUs or for identifying similar models of hardware. The caller is responsible for releasing the returned CFObject.

pub const CLOCK_DEVICE_TRANSPORT_TYPE: Self

A u32 whose value indicates how the AudioClockDevice is connected to the CPU. Constants for some of the values for this property can be found in the enum in the AudioDevice Constants section of this file.

pub const CLOCK_DEVICE_DOMAIN: Self

A u32 whose value indicates the clock domain to which this AudioClockDevice belongs. AudioClockDevices and AudioDevices that have the same value for this property are able to be synchronized in hardware. However, a value of 0 indicates that the clock domain for the device is unspecified and should be assumed to be separate from every other device’s clock domain, even if they have the value of 0 as their clock domain as well.

pub const CLOCK_DEVICE_IS_ALIVE: Self

A UInt32 where a value of 1 means the device is ready and available and 0 means the device is usable and will most likely go away shortly.

pub const CLOCK_DEVICE_IS_RUNNING: Self

A UInt32 where a value of 0 means the AudioClockDevice is not providing times and a value of 1 means that it is. Note that the notification for this property is usually sent from the AudioClockDevice’s IO thread.

pub const CLOCK_DEVICE_LATENCY: Self

A u32 containing the number of frames of latency in the AudioClockDevice.

pub const CLOCK_DEVICE_CONTROL_LIST: Self

An array of AudioObjectIDs that represent the AudioControls of the AudioClockDevice. Note that if a notification is received for this property, any cached AudioObjectIDs for the device become invalid and need to be re-fetched.

pub const CLOCK_DEVICE_NOMINAL_SAMPLE_RATE: Self

A Float64 that indicates the current nominal sample rate of the AudioClockDevice.

pub const CLOCK_DEVICE_AVAILABLE_NOMINAL_SAMPLE_RATES: Self

An array of AudioValueRange structs that indicates the valid ranges for the nominal sample rate of the AudioClockDevice.

impl PropSelector

Processes AudioObjectPropertySelector values provided by the Process class.

pub const PROCESS_PID: Self

A pid_t indicating the process ID associated with the process.

pub const PROCESS_BUNDLE_ID: Self

A cf::String that contains the bundle ID of the process. The caller is responsible for releasing the returned cf::Object.

pub const PROCESS_DEVICES: Self

An array of AudioObjectIds that represent the devices currently used by the process for input or used by the process for output. The scope will select the input or output device list.

pub const PROCESS_IS_RUNNING: Self

A u32 where a value of 0 indicates that there is not audio IO in progress in the process, and a value of 1 indicates that there is audio IO in progress in the process. Note that audio IO may in progress even if no input or output streams are active.

pub const PROCESS_IS_RUNNING_INPUT: Self

A u32 where a value of 0 indicates that the process is not running any IO or there is not any active input streams, and a value of 1 indicates that the process is running IO and there is at least one active input stream.

pub const PROCESS_IS_RUNNING_OUTPUT: Self

A u32 where a value of 0 indicates that the process is not running any IO or there is not any active output streams, and a value of 1 indicates that the process is running IO and there is at least one active output stream.

impl PropSelector

AudioSystemObject Properties

pub const HW_PROCESS_INPUT_MUTE: Self

A u32 where a non-zero value indicates that all data coming into the process for all devices will be silent. A value of 0 indicates that input data will be received normally.

pub const HW_DEVICES: Self

An array of the AudioObjectIds that represent all the devices currently available to the system.

pub const HW_DEFAULT_INPUT_DEVICE: Self

The AudioObjectId of the default input AudioDevice.

pub const HW_DEFAULT_OUTPUT_DEVICE: Self

The AudioObjectId of the default output AudioDevice.

pub const HW_DEFAULT_SYS_OUTPUT_DEVICE: Self

The AudioObjectId of the output AudioDevice to use for system related sound from the alert sound to digital call progress.

pub const HW_TRANSLATE_UID_TO_DEVICE: Self

This property fetches the AudioObjectId that corresponds to the AudioDevice that has the given UID. The UID is passed in via the qualifier as a cf::String while the AudioObjectId for the AudioDevice is returned to the caller as the property’s data. Note that an error is not returned if the UID doesn’t refer to any AudioDevices. Rather, this property will return kAudioObjectUnknown as the value of the property.

pub const HW_PROCESS_OBJ_LIST: Self

An array of AudioObjectIds that represent the Process objects for all client processes currently connected to the system.

pub const HW_TRANSLATE_PID_TO_PROCESS_OBJ: Self

This property fetches the AudioObjectID that corresponds to the Process object that has the given PID. The PID is passed in via the qualifier as a pid_t while the AudioObjectID for the Process is returned to the caller as the property’s data. Note that an error is not returned if the PID doesn’t refer to any Process. Rather, this property will return kAudioObjectUnknown as the value of the property.

pub const HW_CLOCK_DEVICE_LIST: Self

An array of AudioObjectIDs that represent all the AudioClockDevice objects currently provided by the system.

pub const HW_TAP_LIST: Self

An array of AudioObjectIDs that represent the Tap objects on the system.

impl PropSelector

AudioAggregateDevice Properties

pub const AGGREGATE_DEVICE_FULL_SUB_DEVICE_LIST: Self

A CFArray of CFStrings that contain the UIDs of all the devices, active or inactive, contained in the AudioAggregateDevice. The order of the items in the array is significant and is used to determine the order of the streams of the AudioAggregateDevice. The caller is responsible for releasing the returned CFObject.

pub const AGGREGATE_DEVICE_ACTIVE_SUB_DEVICE_LIST: Self

An array of AudioObjectIDs for all the active sub-devices in the aggregate device.

pub const AGGREGATE_DEVICE_COMPOSITION: Self

A CFDictionary that describes the composition of the AudioAggregateDevice. The keys for this CFDicitionary are defined in the AudioAggregateDevice Constants section. The caller is responsible for releasing the returned CFObject.

pub const AGGREGATE_DEVICE_MAIN_SUB_DEVICE: Self

A CFString that contains the UID for the AudioDevice that is currently serving as the time base of the aggregate device. The caller is responsible for releasing the returned CFObject.

pub const AGGREGATE_DEVICE_CLOCK_DEVICE: Self

A CFString that contains the UID for the AudioClockDevice that is currently serving as the time base of the aggregate device. If the aggregate device includes both a main audio device and a clock device, the clock device will control the time base. Setting this property will enable drift correction for all subdevices in the aggregate device. The caller is responsible for releasing the returned CFObject.

pub const AGGREGATE_DEVICE_TAP_LIST: Self

A CFArray of CFStrings that contain the UUIDs of all the tap objects in the contained in the AudioAggregateDevice.

pub const AGGREGATE_DEVICE_SUB_TAP_LIST: Self

An array of AudioObjectIDs for all the active sub-taps in the aggregate device.

impl PropSelector

AudioDevice Properties

pub const DEVICE_PLUG_IN: Self

An os::Status that contains any error codes generated by loading the IOAudio driver plug-in for the AudioDevice or kAudioHardwareNoError if the plug-in loaded successfully. This property only exists for IOAudio-based AudioDevices whose driver has specified a plug-in to load.

pub const DEVICE_HAS_CHANGED: Self

The type of this property is a u32, but its value has no meaning. This property exists so that clients can listen to it and be told when the configuration of the AudioDevice has changed in ways that cannot otherwise be conveyed through other notifications. In response to this notification, clients should re-evaluate everything they need to know about the device, particularly the layout and values of the controls.

pub const DEVICE_IS_RUNNING_SOMEWHERE: Self

u32 where 1 means that the AudioDevice is running in at least one process on the system and 0 means that it isn’t running at all.

pub const DEVICE_PROCESSOR_OVERLOAD: Self

A u32 where the value has no meaning. This property exists so that clients can be notified when the AudioDevice detects that an IO cycle has run past its deadline. Note that the notification for this property is usually sent from the AudioDevice’s IO thread.

pub const DEVICE_IO_STOPPED_ABNORMALLY: Self

A u32 where the value has no meaning. This property exists so that clients can be notified when IO on the device has stopped outside of the normal mechanisms. This typically comes up when IO is stopped after AudioDeviceStart has returned successfully but prior to the notification for kAudioDevicePropertyIsRunning being sent.

pub const DEVICE_HOG_MODE: Self

A pid_t indicating the process that currently owns exclusive access to the AudioDevice or a value of -1 indicating that the device is currently available to all processes. If the AudioDevice is in a non-mixable mode, the HAL will automatically take hog mode on behalf of the first process to start an IOProc.

Note that when setting this property, the value passed in is ignored. If another process owns exclusive access, that remains unchanged. If the current process owns exclusive access, it is released and made available to all processes again. If no process has exclusive access (meaning the current value is -1), this process gains ownership of exclusive access. On return, the pid_t pointed to by inPropertyData will contain the new value of the property.

pub const DEVICE_BUF_FRAME_SIZE: Self

A u32 whose value indicates the number of frames in the IO buffers.

pub const DEVICE_BUF_FRAME_SIZE_RANGE: Self

An AudioValueRange indicating the minimum and maximum values, inclusive, for kAudioDevicePropertyBufferFrameSize.

pub const DEVICE_USES_VARIABLE_BUF_FRAME_SIZES: Self

A u32 that, if implemented by a device, indicates that the sizes of the buffers passed to an IOProc will vary by a small amount. The value of this property will indicate the largest buffer that will be passed and kAudioDevicePropertyBufferFrameSize will indicate the smallest buffer that will get passed to the IOProc. The usage of this property is narrowed to only allow for devices whose buffer sizes vary by small amounts greater than kAudioDevicePropertyBufferFrameSize. It is not intended to be a license for devices to be able to send buffers however they please. Rather, it is intended to allow for hardware whose natural rhythms lead to this necessity.

pub const DEVICE_IO_CYCLE_USAGE: Self

A f32 whose range is from 0 to 1. This value indicates how much of the client portion of the IO cycle the process will use. The client portion of the IO cycle is the portion of the cycle in which the device calls the IOProcs so this property does not the apply to the duration of the entire cycle.

pub const DEVICE_STREAM_CFG: Self

This property returns the stream configuration of the device in an AudioBufListN (with the buffer pointers set to NULL) which describes the list of streams and the number of channels in each stream. This corresponds to what will be passed into the IOProc.

pub const DEVICE_IO_PROC_STREAM_USAGE: Self

An AudioHardwareIOProcStreamUsage structure which details the stream usage of a given IO proc. If a stream is marked as not being used, the given IOProc will see a corresponding NULL buffer pointer in the AudioBufferList passed to its IO proc. Note that the number of streams detailed in the AudioHardwareIOProcStreamUsage must include all the streams of that direction on the device. Also, when getting the value of the property, one must fill out the mIOProc field of the AudioHardwareIOProcStreamUsage with the address of the of the IOProc whose stream usage is to be retrieved.

pub const DEVICE_ACTUAL_SAMPLE_RATE: Self

A f64 that indicates the current actual sample rate of the AudioDevice as measured by its time stamps.

pub const DEVICE_CLOCK_DEVICE: Self

A cf::String that contains the UID for the AudioClockDevice that is currently serving as the main time base of the device. The caller is responsible for releasing the returned cf::String.

pub const DEVICE_IO_THREAD_OS_WORKGROUP: Self

An os_workgroup_t that represents the thread workgroup the AudioDevice’s IO thread belongs to. The caller is responsible for releasing the returned object.

pub const DEVICE_PROCESS_MUTE: Self

A u32 where a non-zero value indicates that the current process’s audio will be zeroed out by the system. Note that this property does not apply to aggregate devices, just real, physical devices.

impl PropSelector

AudioDevice Properties Implemented via AudioControl objects

AudioObjectPropertySelector values for AudioDevice properties that are implemented by AudioControl objects.

These properties are also accessible by locating the AudioControl object attached to the AudioDevice and using that object to access the properties of the control.

pub const DEVICE_JACK_IS_CONNECTED: Self

A u32 where a value of 0 means that there isn’t anything plugged into the jack associated withe given element and scope. This property is implemented by an AudioJackControl, a subclass of AudioBooleanControl.

pub const DEVICE_VOLUME_SCALAR: Self

A f32 that represents the value of the volume control. The range is between 0.0 and 1.0 (inclusive). Note that the set of all Float32 values between 0.0 and 1.0 inclusive is much larger than the set of actual values that the hardware can select. This means that the Float32 range has a many to one mapping with the underlying hardware values. As such, setting a scalar value will result in the control taking on the value nearest to what was set. This property is implemented by an AudioControl object that is a subclass of AudioVolumeControl.

pub const DEVICE_VOLUME_DECIBELS: Self

A f32 that represents the value of the volume control in dB. Note that the set of all f32 values in the dB range for the control is much larger than the set of actual values that the hardware can select. This means that the f32 range has a many to one mapping with the underlying hardware values. As such, setting a dB value will result in the control taking on the value nearest to what was set. This property is implemented by an AudioControl object that is a subclass of AudioVolumeControl.

pub const DEVICE_VOLUME_RANGE_DECIBELS: Self

An AudioValueRange that contains the minimum and maximum dB values the control can have. This property is implemented by an AudioControl object that is a subclass of AudioVolumeControl.

pub const DEVICE_VOLUME_SCALAR_TO_DECIBELS: Self

A f32 that on input contains a scalar volume value for the and on exit contains the equivalent dB value. This property is implemented by an AudioControl object that is a subclass of AudioVolumeControl.

pub const DEVICE_VOLUME_DECIBELS_TO_SCALAR: Self

A f32 that on input contains a dB volume value for the and on exit contains the equivalent scalar value. This property is implemented by AudioControl object that is a subclass of AudioVolumeControl.

pub const DEVICE_STEREO_PAN: Self

A f32 where 0.0 is full left, 1.0 is full right, and 0.5 is center. This property is implemented by an AudioControl object that is a subclass of AudioStereoPanControl.

pub const DEVICE_STEREO_PAN_CHANNELS: Self

An array of two u32s that indicate which elements of the owning object the signal is being panned between. This property is implemented by an AudioControl object that is a subclass of AudioStereoPanControl.

pub const DEVICE_MUTE: Self

A u32 where a value of 1 means that mute is enabled making that element inaudible. The property is implemented by an AudioControl object that is a subclass of AudioMuteControl.

pub const DEVICE_SOLO: Self

A u32 where a value of 1 means that just that element is audible and the other elements are inaudible. The property is implemented by an AudioControl object that is a subclass of AudioSoloControl.

pub const DEVICE_PHANTOM_POWER: Self

A u32 where a value of 1 means that the AudioDevice has enabled phantom power for the given element. The property is implemented by an AudioControl object that is a subclass of AudioPhantomPowerControl.

pub const DEVICE_PHASE_INVERT: Self

A u32 where a value of 1 means that phase of the signal for the given element has been flipped 180 degrees. The property is implemented by an AudioControl object that is a subclass of AudioPhaseInvertControl.

pub const DEVICE_CLIP_LIGHT: Self

A u32 where a value of 1 means that the signal for the element has exceeded the sample range. Once a clip light is turned on, it is to stay on until either the value of the control is set to false or the current IO session stops and a new IO session starts. The property is implemented by an AudioControl object that is a subclass of AudioClipLightControl.

pub const DEVICE_TALKBACK: Self

A u32 where a value of 1 means that the talkback channel is enabled. The property is implemented by an AudioControl object that is a subclass of AudioTalkbackControl.

pub const DEVICE_LISTENBACK: Self

A u32 where a value of 1 means that the listenback channel is enabled. The property is implemented by an AudioControl object that is a subclass of AudioListenbackControl.

pub const DEVICE_DATA_SRC: Self

An array of u32s whose values are the item IDs for the currently selected data sources. This property is implemented by an AudioControl object that is a subclass of AudioDataSourceControl.

pub const DEVICE_DATA_SRCS: Self

An array of u32s that are represent all the IDs of all the data sources currently available. This property is implemented by an AudioControl object that is a subclass of AudioDataSourceControl.

pub const DEVICE_DATA_SRC_NAME_FOR_IDCF_STR: Self

This property translates the given data source item ID into a human readable name using an AudioValueTranslation structure. The input data is the u32 containing the item ID to translated and the output data is a cf::String. The caller is responsible for releasing the returned cf::Object. This property is implemented by an AudioControl object that is a subclass of AudioDataSourceControl.

pub const DEVICE_DATA_SRC_KIND_FOR_ID: Self

This property returns a u32 that identifies the kind of data source the item ID refers to using an AudioValueTranslation structure. The input data is the u32 containing the item ID and the output data is the u32.

pub const DEVICE_CLOCK_SRC: Self

An array of u32s whose values are the item IDs for the currently selected clock sources. This property is implemented by an AudioControl object that is a subclass of AudioClockControl.

pub const DEVICE_CLOCK_SRCS: Self

An array of u32s that are represent all the IDs of all the clock sources currently available. This property is implemented by an AudioControl object that is a subclass of AudioClockControl.

pub const DEVICE_CLOCK_SRC_NAME_FOR_IDCF_STR: Self

This property translates the given clock source item ID into a human readable name using an AudioValueTranslation structure. The input data is the u32 containing the item ID to translated and the output data is a cf::String. The caller is responsible for releasing the returned cf::Object. This property is implemented by an AudioControl object that is a subclass of AudioClockControl.

pub const DEVICE_CLOCK_SRC_KIND_FOR_ID: Self

This property returns a u32 that identifies the kind of clock source the item ID refers to using an AudioValueTranslation structure. The input data is the u32 containing the item ID and the output data is the u32.

pub const DEVICE_PLAY_THRU: Self

A u32 where a value of 0 means that play through is off and a value of 1 means that it is on. This property is implemented by an AudioControl object that is a subclass of AudioMuteControl. Further, the control that implements this property is only available through kAudioDevicePropertyScopePlayThrough.

pub const DEVICE_PLAY_THRU_SOLO: Self

A u32 where a value of 1 means that just that play through element is audible and the other elements are inaudible. The property is implemented by an AudioControl object that is a subclass of AudioSoloControl. Further, the control that implements this property is only available through kAudioDevicePropertyScopePlayThrough.

pub const DEVICE_PLAY_THRU_VOLUME_SCALAR: Self

A f32 that represents the value of the volume control. The range is between 0.0 and 1.0 (inclusive). Note that the set of all f32 values between 0.0 and 1.0 inclusive is much larger than the set of actual values that the hardware can select. This means that the f32 range has a many to one mapping with the underlying hardware values. As such, setting a scalar value will result in the control taking on the value nearest to what was set. This property is implemented by an AudioControl object that is a subclass of AudioVolumeControl.Further, the control that implements this property is only available through kAudioDevicePropertyScopePlayThrough.

pub const DEVICE_PLAY_THRU_VOLUME_DECIBELS: Self

A f32 that represents the value of the volume control in dB. Note that the set of all f32 values in the dB range for the control is much larger than the set of actual values that the hardware can select. This means that the f32 range has a many to one mapping with the underlying hardware values. As such, setting a dB value will result in the control taking on the value nearest to what was set. This property is implemented by an AudioControl object that is a subclass of AudioVolumeControl. Further, the control that implements this property is only available through kAudioDevicePropertyScopePlayThrough.

pub const DEVICE_PLAY_THRU_VOLUME_RANGE_DECIBELS: Self

An AudioValueRange that contains the minimum and maximum dB values the control can have. This property is implemented by an AudioControl object that is a subclass of AudioVolumeControl. Further, the control that implements this property is only available through kAudioDevicePropertyScopePlayThrough.

pub const DEVICE_PLAY_THRU_VOLUME_SCALAR_TO_DECIBELS: Self

A f32 that on input contains a scalar volume value for the and on exit contains the equivalent dB value. This property is implemented by an AudioControl object that is a subclass of AudioVolumeControl. Further, the control that implements this property is only available through kAudioDevicePropertyScopePlayThrough.

pub const DEVICE_PLAY_THRU_VOLUME_DECIBELS_TO_SCALAR: Self

A f32 that on input contains a dB volume value for the and on exit contains the equivalent scalar value. This property is implemented by an AudioControl object that is a subclass of AudioVolumeControl. Further, the control that implements this property is only available through kAudioDevicePropertyScopePlayThrough.

pub const DEVICE_PLAY_THRU_STEREO_PAN: Self

A f32 where 0.0 is full left, 1.0 is full right, and 0.5 is center. This property is implemented by an AudioControl object that is a subclass of AudioStereoPanControl. Further, the control that implements this property is only available through kAudioDevicePropertyScopePlayThrough.

pub const DEVICE_PLAY_THRU_STEREO_PAN_CHANNELS: Self

An array of two u32s that indicate which elements of the owning object the signal is being panned between. This property is implemented by an AudioControl object that is a subclass of AudioStereoPanControl. Further, the control that implements this property is only available through kAudioDevicePropertyScopePlayThrough.

pub const DEVICE_PLAY_THRU_DST: Self

An array of u32s whose values are the item IDs for the currently selected play through data destinations. This property is implemented by an AudioControl object that is a subclass of AudioDataDestinationControl. Further, the control that implements this property is only available through kAudioDevicePropertyScopePlayThrough.

pub const DEVICE_PLAY_THRU_DSTS: Self

An array of u32s that are represent all the IDs of all the play through data destinations currently available. This property is implemented by an AudioControl object that is a subclass of AudioDataDestinationControl. Further, the control that implements this property is only available through kAudioDevicePropertyScopePlayThrough.

pub const DEVICE_PLAY_THRU_DST_NAME_FOR_IDCF_STR: Self

This property translates the given play through data destination item ID into a human readable name using an AudioValueTranslation structure. The input data is the u32 containing the item ID to translated and the output data is a cf::String. The caller is responsible for releasing the returned cf::Object. This property is implemented by an AudioControl object that is a subclass of AudioDataDestinationControl. Further, the control that implements this property is only available through kAudioDevicePropertyScopePlayThrough.

pub const DEVICE_CHANNEL_NOMINAL_LINE_LEVEL: Self

An array of u32s whose values are the item IDs for the currently selected nominal line levels. This property is implemented by an AudioControl object that is a subclass of AudioLineLevelControl.

pub const DEVICE_CHANNEL_NOMINAL_LINE_LEVELS: Self

An array of u32s that represent all the IDs of all the nominal line levels currently available. This property is implemented by an AudioControl object that is a subclass of AudioLineLevelControl.

pub const DEVICE_CHANNEL_NOMINAL_LINE_LEVEL_NAME_FOR_IDCF_STR: Self

This property translates the given nominal line level item ID into a human readable name using an AudioValueTranslation structure. The input data is the u32 containing the item ID to be translated and the output data is a cf::String. The caller is responsible for releasing the returned cf::Object. This property is implemented by an AudioControl object that is a subclass of AudioLineLevelControl.

pub const DEVICE_HIGH_PASS_FILTER_SETTING: Self

An array of u32s whose values are the item IDs for the currently selected high pass filter setting. This property is implemented by an AudioControl object that is a subclass of AudioHighPassFilterControl.

pub const DEVICE_HIGH_PASS_FILTER_SETTINGS: Self

An array of u32s that represent all the IDs of all the high pass filter settings currently available. This property is implemented by an AudioControl object that is a subclass of AudioHighPassFilterControl.

pub const DEVICE_HIGH_PASS_FILTER_SETTING_NAME_FOR_IDCF_STR: Self

This property translates the given high pass filter setting item ID into a human readable name using an AudioValueTranslation structure. The input data is the u32 containing the item ID to be translated and the output data is a cf::String. The caller is responsible for releasing the returned cf::Object. This property is implemented by an AudioControl object that is a subclass of AudioHighPassFilterControl.

pub const DEVICE_SUB_VOLUME_SCALAR: Self

A f32 that represents the value of the LFE volume control. The range is between 0.0 and 1.0 (inclusive). Note that the set of all f32 values between 0.0 and 1.0 inclusive is much larger than the set of actual values that the hardware can select. This means that the f32 range has a many to one mapping with the underlying hardware values. As such, setting a scalar value will result in the control taking on the value nearest to what was set. This property is implemented by an AudioControl object that is a subclass of AudioLFEVolumeControl.

pub const DEVICE_SUB_VOLUME_DECIBELS: Self

A f32 that represents the value of the LFE volume control in dB. Note that the set of all f32 values in the dB range for the control is much larger than the set of actual values that the hardware can select. This means that the f32 range has a many to one mapping with the underlying hardware values. As such, setting a dB value will result in the control taking on the value nearest to what was set. This property is implemented by an AudioControl object that is a subclass of AudioLFE VolumeControl.

pub const DEVICE_SUB_VOLUME_RANGE_DECIBELS: Self

An AudioValueRange that contains the minimum and maximum dB values the control can have. This property is implemented by an AudioControl object that is a subclass of AudioLFEVolumeControl.

pub const DEVICE_SUB_VOLUME_SCALAR_TO_DECIBELS: Self

A f32 that on input contains a scalar volume value for the and on exit contains the equivalent dB value. This property is implemented by an AudioControl object that is a subclass of AudioLFEVolumeControl.

pub const DEVICE_SUB_VOLUME_DECIBELS_TO_SCALAR: Self

A f32 that on input contains a dB volume value for the and on exit contains the equivalent scalar value. This property is implemented by an AudioControl object that is a subclass of AudioLFEVolumeControl.

pub const DEVICE_SUB_MUTE: Self

A u32 where a value of 1 means that mute is enabled making the LFE on

that element inaudible. The property is implemented by an AudioControl object that is a subclass of AudioLFEMuteControl.

pub const DEVICE_VOICE_ACTIVITY_DETECTION_ENABLE: Self

A u32 where 0 disables voice activity detection process and non-zero enables it.

Voice activity detection can be used with input audio and has echo cancellation. Detection works when a process mute is used, but not with hardware mute.

pub const DEVICE_VOICE_ACTIVITY_DETECTION_STATE: Self

A read-only u32 where 0 indicates no voice currently detected and 1 indicates voice.

Used in conjunction with VOICE_ACTIVITY_DETECTION_ENABLE. A client would normally register to listen to this property for changes and then query the state rather than continuously poll the value. NOTE: If input audio is not active/runnning or the voice activity detection is disabled, then it is not analyzed and this will provide 0.

impl PropSelector

AudioObjectPropertySelector values provided by the Tap Object class. The Tap class is a subclass of the AudioObject class. the class has just the global scope, kAudioObjectPropertyScopeGlobal, and only a master element.

pub const TAP_UID: Self

A cf::String that contains a persistent identifier for the Tap. A Taps UID persists until the tap is destroyed. The caller is responsible for releasing the returned cf::Object.

pub const TAP_DESCRIPTION: Self

The ca::TapDesc used to initially create this tap. This property can be used to modify and set the description of an existing tap.

pub const TAP_FORMAT: Self

An AudioStreamBasicDescription that describes the current data format for the tap. This is the format of that data that will be accessible in any aggregate device that contains the tap.