cidre::ct

Struct Font

pub struct Font(/* private fields */);

Implementations

impl Font

pub fn retained(&self) -> R<Self>

impl Font

pub fn type_id() -> TypeId

pub fn with_name_size_matrix( name: &String, size: Float, matrix: Option<&AffineTransform>, ) -> R<Self>

pub fn with_name_size(name: &String, size: Float) -> R<Self>

pub fn with_name_matrix(name: &String, matrix: &AffineTransform) -> R<Self>

pub fn with_ui_type_size_lang( ui_type: UIFontType, size: Float, language: Option<&String>, ) -> Option<R<Self>>

pub fn with_ui_type_size(ui_type: UIFontType, size: Float) -> Option<R<Self>>

pub fn size(&self) -> Float

pub fn matrix(&self) -> AffineTransform

pub fn symbolic_traits(&self) -> FontSymbolicTraits

pub fn glyphs_for_characters( &self, charactes: &[UniChar], glyphs: &mut [Glyph], ) -> Result<(), ()>

pub fn bounding_rect_for_glyphs( &self, orientation: FontOrientation, glyphs: &[Glyph], ) -> Rect

pub fn bounding_rects_for_glyphs( &self, orientation: FontOrientation, glyphs: &[Glyph], bounding_rects: &mut [Rect], ) -> Rect

pub fn optical_bounds_for_glyphs( &self, glyphs: &[Glyph], bounding_rects: &mut [Rect], ) -> Rect

pub fn advance_for_glyphs( &self, orientation: FontOrientation, glyphs: &[Glyph], ) -> f64

This function returns the summed glyph advance of an array of glyphs.

pub fn advances_for_glyphs( &self, orientation: FontOrientation, glyphs: &[Glyph], advances: &mut [Size], ) -> f64

This function returns the summed glyph advance of an array of glyphs. Individual glyph advances are passed back via the advances parameter. These are the ideal metrics for each glyph scaled and transformed in font space.

pub fn path_for_glyph( &self, glyph: Glyph, matrix: Option<&AffineTransform>, ) -> Option<R<Path>>

pub fn name_for_glyph(&self, glyph: Glyph) -> Option<R<String>>

pub fn glyph_with_name(&self, name: &String) -> Glyph

pub fn ascent(&self) -> Float

pub fn descent(&self) -> Float

pub fn leading(&self) -> Float

Returns the scaled font leading metric.

pub fn underline_pos(&self) -> Float

pub fn underline_thickness(&self) -> Float

pub fn slant_angle(&self) -> Float

pub fn cap_height(&self) -> Float

Returns the font underline position metric scaled based on the point size and matrix of the font reference.

pub fn x_height(&self) -> Float

This function returns the font X height metric scaled based on the point size and matrix of the font reference

pub fn units_per_em(&self) -> u32

pub fn supported_languages(&self) -> R<ArrayOf<String>>

impl Font

pub fn as_ns(&self) -> &Font