cidre::cv::buffer

Struct Buf

pub struct Buf(/* private fields */);

Implementations

impl Buf

pub fn retained(&self) -> R<Self>

impl Buf

pub fn attach<'a>( &'a self, key: &String, attachment_mode: AttachMode, ) -> Option<&'a Type>

pub fn set_attach( &mut self, key: &String, val: &Type, attachment_mode: AttachMode, )

pub fn remove_attach(&mut self, key: &String)

pub fn remove_all_attaches(&mut self)

pub fn set_attaches( &mut self, attachments: &Dictionary, attachment_mode: AttachMode, )

pub fn propagate_attaches(&self, destination_buffer: &mut Buf)

pub fn copy_attaches( &self, attachment_mode: AttachMode, ) -> Option<R<Dictionary>>

pub fn copy_attach( &self, key: &String, attachment_mode: AttachMode, ) -> Option<R<Type>>

pub fn has_attach(&self, key: &String) -> bool

pub fn propagated_attaches_key() -> &'static String

Useful with the CoreVideo pool and texture cache APIs so that you can specify an initial set of default buffer attachments to automatically be attached to the buffer when it is created.

Availability

macos_10_4, ios_4_0, visionos_1_0

pub fn not_propagated_attaches_key() -> &'static String

Useful with the CoreVideo pool and texture cache APIs so that you can specify an initial set of default buffer attachments to automatically be attached to the buffer when it is created.

Availability

macos_10_4, ios_4_0, visionos_1_0

pub fn movie_time_key() -> &'static String

Availability

macos_10_4, ios_4_0, visionos_1_0

pub fn time_value_key() -> &'static String

Availability

macos_10_4, ios_4_0, visionos_1_0

pub fn time_scale_key() -> &'static String

Availability

macos_10_4, ios_4_0, visionos_1_0

impl Buf

pub fn encoded_size(&self) -> Size

Returns the full encoded dimensions of a [‘cv::ImageBuf’]. For example, for an NTSC DV frame this would be 720x480

Example:

use cidre::{cv, cg};

let pixel_buf = cv::PixelBuf::new(200, 100, cv::PixelFormat::_32_BGRA, None).unwrap();

let size = pixel_buf.encoded_size();
assert_eq!(cg::Size { width: 200.0, height: 100.0 }, size);

pub fn display_size(&self) -> Size

Returns the nominal output display size, in square pixels, of a Core Video image buffer.

use cidre::{cv, cg};

let pixel_buf = cv::PixelBuf::new(200, 100, cv::PixelFormat::_32_BGRA, None).unwrap();

let display_size = pixel_buf.display_size();
assert_eq!(cg::Size { width: 200.0, height: 100.0}, display_size);

pub fn clean_rect(&self) -> Rect

Returns the source rectangle of a Core Video image buffer that represents the clean aperture of the buffer in encoded pixels.

use cidre::{cv, cg};

let pixel_buf = cv::PixelBuf::new(200, 100, cv::PixelFormat::_32_BGRA, None).unwrap();

let rect = pixel_buf.clean_rect();
assert_eq!(cg::Rect { origin: cg::Point::zero(), size: cg::Size { width: 200.0, height: 100.0 }}, rect);

pub fn is_flipped(&self) -> bool

Returns a Boolean value indicating whether the image is flipped vertically.

use cidre::{cv, cg};

let pixel_buf = cv::PixelBuf::new(200, 100, cv::PixelFormat::_32_BGRA, None).unwrap();

assert_eq!(true, pixel_buf.is_flipped());

pub fn color_space(&self) -> Option<&ColorSpace>

use cidre::{cv, cg};

let buf = cv::PixelBuf::new(200, 100, cv::PixelFormat::_32_BGRA, None).unwrap();
assert!(buf.color_space().is_none());

pub fn color_space_from_attaches( attachments: &Dictionary, ) -> Option<R<ColorSpace>>

impl Buf

pub fn type_id() -> TypeId

pub fn width(&self) -> usize

Width in pixels.

pub fn height(&self) -> usize

Height in pixels.

pub fn pixel_format(&self) -> PixelFormat

Returns the PixelFormat of the PixelBuf.

pub fn plane_count(&self) -> usize

pub fn plane_width(&self, plane_index: usize) -> usize

pub fn plane_height(&self, plane_index: usize) -> usize

pub fn plane_base_address(&self, plane_index: usize) -> *const u8

pub fn plane_bytes_per_row(&self, plane_index: usize) -> usize

pub fn new( width: usize, height: usize, pixel_format_type: PixelFormat, pixel_buf_attrs: Option<&Dictionary>, ) -> Result<R<PixelBuf>>

use cidre::{cv, cg};

let pixel_buf = cv::PixelBuf::new(200, 100, cv::PixelFormat::_32_BGRA, None).unwrap();

assert_eq!(200, pixel_buf.width());
assert_eq!(100, pixel_buf.height());
assert_eq!(cv::PixelFormat::_32_BGRA, pixel_buf.pixel_format());
assert_eq!(0, pixel_buf.plane_count());
assert_eq!(cv::PixelBuf::type_id(), pixel_buf.get_type_id());

pub fn with_bytes( width: usize, height: usize, base_address: *mut c_void, bytes_per_row: usize, release_callback: ReleaseCallback, release_ref_con: *mut c_void, pixel_format_type: PixelFormat, pixel_buf_attrs: Option<&Dictionary>, ) -> Result<R<PixelBuf>>

pub fn create_in( width: usize, height: usize, pixel_format_type: PixelFormat, pixel_buf_attrs: Option<&Dictionary>, pixel_buf_out: *mut Option<R<PixelBuf>>, allocator: Option<&Allocator>, ) -> Return

pub fn create_with_bytes_in( width: usize, height: usize, pixel_format_type: PixelFormat, base_address: *mut c_void, bytes_per_row: usize, release_callback: ReleaseCallback, release_ref_con: *mut c_void, pixel_buf_attrs: Option<&Dictionary>, pixel_buf_out: *mut Option<R<PixelBuf>>, allocator: Option<&Allocator>, ) -> Return

pub unsafe fn lock_base_addr(&mut self, flags: LockFlags) -> Return

pub unsafe fn unlock_lock_base_addr(&mut self, flags: LockFlags) -> Return

pub fn base_address_lock( &mut self, flags: LockFlags, ) -> Result<BaseAddrLockGuard<'_>>

pub fn io_surf(&self) -> Option<&Surf>

pub fn io_surf_mut(&mut self) -> Option<&mut Surf>

pub fn with_io_surf( surface: &Surf, pixel_buffer_attributes: Option<&Dictionary>, ) -> Result<R<Self>>

pub fn with_io_surf_in( surface: &Surf, pixel_buf_attrs: Option<&Dictionary>, allocator: Option<&Allocator>, ) -> Result<R<Self>>

Call to create a single cv::PixelBuf for a passed-in ‘io::Surf’.

impl Buf

pub fn texture(&self) -> Option<&Texture>

pub fn is_texture_flipped(&self) -> bool

pub fn clean_tex_coords( &self, lower_left: &mut [f32; 2], lower_right: &mut [f32; 2], upper_right: &mut [f32; 2], upper_left: &mut [f32; 2], )

Returns convenient normalized texture coordinates for the part of the image that should be displayed

This function automatically takes into account whether or not the texture is flipped.