cidre::cv

Module err

Constants

ALLOC_FAILED

The allocation for a buffer or buffer pool failed. Most likely because of lack of resources.

DISPLAY_LINK_ALREADY_RUNNING

The cv::DisplayLink is already started and running.

DISPLAY_LINK_CALLBACKS_NOT_SET

The output callback is not set.

DISPLAY_LINK_NOT_RUNNING

The cv::DisplayLink has not been started.

ERROR

FIRST

Placeholder to mark the beginning of the range of cv::err codes.

INVALID_ARG

At least one of the arguments passed in is not valid. Either out of range or the wrong type.

INVALID_DISPLAY

A cv::DisplayLink cannot be created for the given DisplayRef.

INVALID_PIXEL_BUF_ATTRS

A cv::Buffer cannot be created with the given attributes.

INVALID_PIXEL_FORMAT

The requested pixelformat is not supported for the cv::Buf type.

INVALID_POOL_ATTRS

A cv::BufPool cannot be created with the given attributes.

INVALID_SIZE

The requested size (most likely too big) is not supported for the cv::Buf type.

LAST

Placeholder to mark the end of the range of cv::err codes.

PIXEL_BUF_NOT_METAL_COMPATIBLE

The Buffer cannot be used with Metal as either its size, pixelformat or attributes are not supported by Metal.

PIXEL_BUF_NOT_OPEN_GL_COMPATIBLE

The Buffer cannot be used with OpenGL as either its size, pixelformat or attributes are not supported by OpenGL.

POOL_ALLOC_FAILED

The allocation for the buffer pool failed. Most likely because of lack of resources. Check if your parameters are in range.

RETRY

A scan hasn’t completely traversed the CVBufferPool due to a concurrent operation. The client can retry the scan.

UNSUPPORTED

WOULD_EXCEED_ALLOCATION_THRESHOLD

The allocation request failed because it would have exceeded a specified allocation threshold (see kCVPixelBufferPoolAllocationThresholdKey).