cidre::cv::image_buf_attachment

Module display_mask_rect

Functions

left_edge_points⚠

Availability

rect_left⚠

Specifies the horizontal pixel offset of the rectangle from the left of the bounding raster. The left offset value is a cf::Number of unsigned 16-bit integer that is less than the reference raster width value.

rect_top⚠

Specifies the vertical pixel offset of the rectangle from the top of the bounding raster. The top offset value is a cf::Number of unsigned 16-bit integer that is less than the reference raster height value.

rect_width⚠

Specifies the width of the rectangle starting at rectangle’s left offset toward the rectangle’s right edge. The width value is a cf::Number of unsigned 16-bit integer.

reference_raster_height⚠

Specifies the height in pixels of the 2D coordinate system to define the rectangle. 0,0 origin is the top-left. The raster height value is a cf::Number of unsigned 16-bit integer. Usually matches the height of the video or the output device.

reference_raster_width⚠

Specifies the width in pixels of the 2D coordinate system to define the rectangle. 0,0 origin is the top-left. The raster width value is a cf::Number of unsigned 16-bit integer. Usually matches the width of the video or the output device.

right_edge_points⚠

Availability