cidre::cv::pixel_buffer

Struct PixelFormat

#[repr(transparent)]
pub struct PixelFormat(pub Type);

CoreVideo pixel format type constants.

CoreVideo does not provide support for all of these formats; this list just defines their names.

Tuple Fields
0: Type
Implementations

impl PixelFormat

pub const _1_MONOCHROME: Self

1 bit indexed

pub const _2_INDEXED: Self

2 bit indexed

pub const _4_INDEXED: Self

4 bit indexed

pub const _8_INDEXED: Self

8 bit indexed

pub const _1_INDEXED_GREY_WHITE_IS_ZERO: Self

1 bit indexed gray, white is zero

pub const _2_INDEXED_GREY_WHITE_IS_ZERO: Self

2 bit indexed gray, white is zero

pub const _4_INDEXED_GREY_WHITE_IS_ZERO: Self

4 bit indexed gray, white is zero

pub const _8_INDEXED_GREY_WHITE_IS_ZERO: Self

8 bit indexed gray, white is zero

pub const _16_BE_555: Self

16 bit BE RGB 555

pub const _16_LE_555: Self

16 bit LE RGB 555

pub const _16_LE_5551: Self

16 bit LE RGB 5551

pub const _16_BE_565: Self

16 bit BE RGB 565

pub const _16_LE_565: Self

16 bit LE RGB 565

pub const _24_RGB: Self

24 bit RGB

pub const _24_BGR: Self

24 bit BGR

pub const _32_ARGB: Self

32 bit ARGB

pub const _32_BGRA: Self

32 bit BGRA

pub const _32_ABGR: Self

32 bit ABGR

pub const _32_RGBA: Self

32 bit RGBA

pub const _64_ARGB: Self

64 bit ARGB, 16-bit big-endian samples

pub const _64_RGBALE: Self

64 bit RGBA, 16-bit little-endian full-range (0-65535) samples

pub const _30_RGB: Self

30 bit RGB, 10-bit big-endian samples, 2 unused padding bits (at least significant end).

pub const _30_RGB_R210: Self

30 bit RGB, 10-bit big-endian samples, 2 unused padding bits (at most significant end), video-range (64-940).

pub const _422_YP_CB_CR_8: Self

Component Y’CbCr 8-bit 4:2:2, ordered Cb Y’0 Cr Y’1

pub const _2VUY: Self = Self::_422_YP_CB_CR_8

pub const _4444_YP_CB_CR_A_8: Self

Component Y’CbCrA 8-bit 4:4:4:4, ordered Cb Y’ Cr A

pub const _4444_YP_CB_CR_A_8_R: Self

Component Y’CbCrA 8-bit 4:4:4:4, rendering format. full range alpha, zero biased YUV, ordered A Y’ Cb Cr

pub const _4444_A_YP_CB_CR_8: Self

Component Y’CbCrA 8-bit 4:4:4:4, ordered A Y’ Cb Cr, full range alpha, video range Y’CbCr.

pub const _4444_A_YP_CB_CR_16: Self

Component Y’CbCrA 16-bit 4:4:4:4, ordered A Y’ Cb Cr, full range alpha, video range Y’CbCr, 16-bit little-endian samples.

pub const _4444_A_YP_CB_CR_FLOAT: Self

Component AY’CbCr single precision floating-point 4:4:4:4

pub const _444_YP_CB_CR_8: Self

Component Y’CbCr 8-bit 4:4:4, ordered Cr Y’ Cb, video range Y’CbCr

pub const _422_YP_CB_CR_16: Self

Component Y’CbCr 10,12,14,16-bit 4:2:2

pub const _422_YP_CB_CR_10: Self

Component Y’CbCr 10-bit 4:2:2

pub const _444_YP_CB_CR_10: Self

Component Y’CbCr 10-bit 4:4:4

pub const _420_YP_CB_CR_8_PLANAR_FULL_RANGE: Self

Planar Component Y’CbCr 8-bit 4:2:0, full range. baseAddr points to a big-endian CVPlanarPixelBufferInfo_YCbCrPlanar struct

pub const _422_YP_CB_CR_4_A_8_BI_PLANAR: Self

First plane: Video-range Component Y’CbCr 8-bit 4:2:2, ordered Cb Y’0 Cr Y’1; second plane: alpha 8-bit 0-255

pub const _420_YP_CB_CR_8_BI_PLANAR_VIDEO_RANGE: Self

Bi-Planar Component Y’CbCr 8-bit 4:2:0, video-range (luma=[16,235] chroma=[16,240]).

pub const _420V: Self = Self::_420_YP_CB_CR_8_BI_PLANAR_VIDEO_RANGE

pub const _420_YP_CB_CR_8_BI_PLANAR_FULL_RANGE: Self

Bi-Planar Component Y’CbCr 8-bit 4:2:0, full-range (luma=[0,255] chroma=[1,255]). baseAddr points to a big-endian

pub const _420F: Self = Self::_420_YP_CB_CR_8_BI_PLANAR_FULL_RANGE

pub const _420_YP_CB_CR_10_BI_PLANAR_VIDEO_RANGE: Self

2 plane YCbCr10 4:2:0, each 10 bits in the MSBs of 16bits, video-range (luma=[64,940] chroma=[64,960])

pub const _422_YP_CB_CR_10_BI_PLANAR_VIDEO_RANGE: Self

2 plane YCbCr10 4:2:2, each 10 bits in the MSBs of 16bits, video-range (luma=[64,940] chroma=[64,960])

pub const _444_YP_CB_CR_10_BI_PLANAR_VIDEO_RANGE: Self

2 plane YCbCr10 4:4:4, each 10 bits in the MSBs of 16bits, video-range (luma=[64,940] chroma=[64,960])

pub const _420_YP_CB_CR_10_BI_PLANAR_FULL_RANGE: Self

2 plane YCbCr10 4:2:0, each 10 bits in the MSBs of 16bits, full-range (Y range 0-1023)

pub const _422_YP_CB_CR_10_BI_PLANAR_FULL_RANGE: Self

2 plane YCbCr10 4:2:2, each 10 bits in the MSBs of 16bits, full-range (Y range 0-1023)

pub const _444_YP_CB_CR_10_BI_PLANAR_FULL_RANGE: Self

2 plane YCbCr10 4:4:4, each 10 bits in the MSBs of 16bits, full-range (Y range 0-1023)

pub const ARGB_2101010_LE_PACKED: Self

little-endian ARGB2101010 full-range ARGB

pub const ONE_COMPONENT_8: Self

pub const ONE_COMPONENT_H16: Self

pub const ONE_COMPONENT_F32: Self

pub const TWO_COMPONENT_H16: Self

16 bit two component IEEE half-precision float, 16-bit little-endian samples

pub const TWO_COMPONENT_F32: Self

32 bit two component IEEE float, 32-bit little-endian samples

pub const _64_RGBA_HALF: Self

64 bit RGBA IEEE half-precision float, 16-bit little-endian samples

pub const _128_RGBA_FLOAT: Self

128 bit RGBA IEEE float, 32-bit little-endian samples

pub fn from_cf_number(number: &Number) -> Self

pub fn to_desc(&self) -> Option<R<Dictionary>>

pub fn to_cf_number(&self) -> &'static Number

pub fn to_ns_number(&self) -> &'static Number

pub fn is_compressed_avaliable(&self) -> bool

Checks if a compressed pixel format is supported on the current platform.

impl PixelFormat

Lossless-Compressed Pixel Formats

The following pixel formats can be used to reduce the memory bandwidth involved in large-scale pixel data flow, which can have benefits for battery life and thermal efficiency.

They work by dividing pixel buffers into fixed-width, fixed-height, fixed-byte-size blocks. Hardware units (video codecs, GPU, ISP, etc.) attempt to write a compressed encoding for each block using a lossless algorithm. If a block of pixels is successfully encoded using fewer bytes than the uncompressed pixel data, the hardware unit does not need to write as many bytes for that pixel block. If the encoding is unsuccessful, the uncompressed pixel data is written, filling the whole pixel block. Each compressed pixel buffer has a separate area of metadata recording the encoding choices for each pixel block.

Padding bits are eliminated, so for example, 10-bit-per-component lossless-compressed pixel buffers are slightly smaller than their uncompressed equivalents. For pixel formats with no padding, the lossless-compressed pixel buffers are slightly larger due to the metadata.

Important caveats:

Some devices do not support these pixel formats at all. Before using one of these pixel formats, call Self::is_compressed_avaliable() to check that it is available on the current device.

On different devices, the concrete details of these formats may be different.

On different devices, the degree and details of support by hardware units (video codecs, GPU, ISP, etc.) may be different.

Do not ship code that reads the contents of lossless-compressed pixel buffers directly with the CPU, or which saves or transfers it to other devices, as this code will break with future hardware.

The bandwidth benefits of these formats are generally outweighed by the cost of buffer copies to convert to uncompressed pixel formats, so if you find that you need to perform a buffer copy to covert for CPU usage, it’s likely that you would have been better served by using the equivalent uncompressed pixel formats in the first place.

pub const LOSSLESS_32_BGRA: Self

Lossless-compressed form of ‘cv::PixelFormat::_32BGRA’

pub const LOSSLESS_64_RGBA_HALF: Self

Lossless-compressed form of ‘cv::PixelFormat::_64_RGBA_HALF’. No CVPlanarPixelBufferInfo struct.

pub const LOSSLESS_420_YP_CB_CR_8_BI_PLANAR_VIDEO_RANGE: Self

Lossless-compressed form of ‘cv::PixelFormat::_420_YP_CB_CR_8_BI_PLANAR_VIDEO_RANGE’.

pub const LOSSLESS_420V: Self = Self::LOSSLESS_420_YP_CB_CR_8_BI_PLANAR_VIDEO_RANGE

pub const LOSSLESS_420_YP_CB_CR_8_BI_PLANAR_FULL_RANGE: Self

Lossless-compressed form of ‘cv::PixelFormat::_420_YP_CB_CR_8_BI_PLANAR_FULL_RANGE’

pub const LOSSLESS_420F: Self = Self::LOSSLESS_420_YP_CB_CR_8_BI_PLANAR_FULL_RANGE

pub const LOSSLESS_420_YP_CB_CR_10_PACKED_BI_PLANAR_VIDEO_RANGE: Self

Lossless-compressed-packed form of ‘cv::PixelFormat::_420_YP_CB_CR_10_BI_PLANAR_VIDEO_RANGE’. Format is compressed-packed with no padding bits between pixels.

pub const LOSSLESS_422_YP_CB_CR_10_PACKED_BI_PLANAR_VIDEO_RANGE: Self

Lossless-compressed form of ‘cv::PixelFormat::_422_YP_CB_CR_10_BI_PLANAR_VIDEO_RANGE’. Format is compressed-packed with no padding bits between pixels.

impl PixelFormat

Lossy-Compressed Pixel Formats

The following pixel formats can be used to reduce memory bandwidth and memory footprint involved in large-scale pixel data flow, which can have benefits for battery life and thermal efficiency.

Similar to lossless pixel formats, they work by dividing pixel buffers into fixed-width, fixed-height, fixed-byte-size blocks. Pixel buffers allocated using lossy formats have reduced memory footprint than their lossless equivalents; this reduced footprint may or may not result in loss of quality depending on the content of the individual block. Hardware units (video codecs, GPU, ISP, etc.) attempt to write a compressed encoding for each block using either a lossless or lossy algorithm. If a block of pixels is successfully encoded within its pre-defined memory footprint, then the lossless alogrithm is applied; if the encoded block of pixels exceeds the pre-defined memory footprint then the lossy algorithm is applied. Each compressed pixel buffer has a separate area of metadata recording the encoding choices for each pixel block.

Usefull links:

https://developer.apple.com/documentation/technotes/tn3104-recording-video-in-apple-prores

https://developer.apple.com/documentation/technotes/tn3121-selecting-a-pixel-format-for-an-avcapturevideodataoutput

pub const LOSSY_32_BGRA: Self

Lossy-compressed form of cv::PixelFormat::_32_BGRA.

pub const LOSSY_420_YP_CB_CR_8_BI_PLANAR_VIDEO_RANGE: Self

Lossy-compressed form of cv::PixelFormat::_420_YP_CB_CR_8_BI_PLANAR_VIDEO_RANGE.

pub const LOSSY_420V: Self = Self::LOSSY_420_YP_CB_CR_8_BI_PLANAR_VIDEO_RANGE

pub const LOSSY_420_YP_CB_CR_8_BI_PLANAR_FULL_RANGE: Self

Lossy-compressed form of cv::PixelFormat::_420_YP_CB_CR_8_BI_PLANAR_FULL_RANGE.

pub const LOSSY_420F: Self = Self::LOSSY_420_YP_CB_CR_8_BI_PLANAR_FULL_RANGE

pub const LOSSY_420_YP_CB_CR_10_PACKED_BI_PLANAR_VIDEO_RANGE: Self

Lossy-compressed form of cv::PixelFormat::_420_YP_CB_CR_10_PACKED_BI_PLANAR_VIDEO_RANGE.

pub const LOSSY_422_YP_CB_CR_10_PACKED_BI_PLANAR_VIDEO_RANGE: Self

Lossy-compressed form of cv::PixelFormat::_422_YP_CB_CR_10_PACKED_BI_PLANAR_VIDEO_RANGE. Format is compressed-packed with no padding bits between pixels.