cidre::cv

Type Alias ImageBuf

pub type ImageBuf = Buf;

Aliased Type

pub struct ImageBuf(/* private fields */);

Implementations

impl ImageBuf

pub fn encoded_size(&self) -> Size

Returns the full encoded dimensions of a [‘cv::ImageBuf’]. For example, for an NTSC DV frame this would be 720x480

Example:

use cidre::{cv, cg};

let pixel_buf = cv::PixelBuf::new(200, 100, cv::PixelFormat::_32_BGRA, None).unwrap();

let size = pixel_buf.encoded_size();
assert_eq!(cg::Size { width: 200.0, height: 100.0 }, size);

pub fn display_size(&self) -> Size

Returns the nominal output display size, in square pixels, of a Core Video image buffer.

use cidre::{cv, cg};

let pixel_buf = cv::PixelBuf::new(200, 100, cv::PixelFormat::_32_BGRA, None).unwrap();

let display_size = pixel_buf.display_size();
assert_eq!(cg::Size { width: 200.0, height: 100.0}, display_size);

pub fn clean_rect(&self) -> Rect

Returns the source rectangle of a Core Video image buffer that represents the clean aperture of the buffer in encoded pixels.

use cidre::{cv, cg};

let pixel_buf = cv::PixelBuf::new(200, 100, cv::PixelFormat::_32_BGRA, None).unwrap();

let rect = pixel_buf.clean_rect();
assert_eq!(cg::Rect { origin: cg::Point::zero(), size: cg::Size { width: 200.0, height: 100.0 }}, rect);

pub fn is_flipped(&self) -> bool

Returns a Boolean value indicating whether the image is flipped vertically.

use cidre::{cv, cg};

let pixel_buf = cv::PixelBuf::new(200, 100, cv::PixelFormat::_32_BGRA, None).unwrap();

assert_eq!(true, pixel_buf.is_flipped());

pub fn color_space(&self) -> Option<&ColorSpace>

use cidre::{cv, cg};

let buf = cv::PixelBuf::new(200, 100, cv::PixelFormat::_32_BGRA, None).unwrap();
assert!(buf.color_space().is_none());

pub fn color_space_from_attaches( attachments: &Dictionary, ) -> Option<R<ColorSpace>>