cidre

Module dispatch

Grand Central Dispatch

Re-exports

pub use time::Time;

pub use time::WallTime;

pub use queue::Attr;

pub use queue::AutoreleaseFrequency;

pub use queue::Concurrent as ConcurrentQueue;

pub use queue::Global as GlobalQueue;

pub use queue::Main as MainQueue;

pub use queue::Priority as QueuePriority;

pub use queue::QosClass;

pub use queue::Queue;

pub use data::Applier as DataApplier;

pub use data::Data;

pub use source::MachRecvFlags as SourceMachRecvFlags;

pub use source::MachSendFlags as SourceMachSendFlags;

pub use source::MemoryPressureFlags as SourceMemoryPressureFlags;

pub use source::ProcFlags as SourceProcFlags;

pub use source::Src;

pub use source::TimerFlags as SourceTimerFlags;

pub use source::TimerSrc;

pub use source::Type as SourceType;

pub use source::TypeDataAdd as SourceDataAdd;

pub use work_item::WorkItem;

pub use block::Flags as BlockFlags;

Modules

block

data

queue

source

time

work_item

Structs

Group

Object

Semaphore

Functions

main

This function “parks” the main thread and waits for blocks to be submitted to the main queue. Applications that call UIApplicationMain (iOS), NSApplicationMain (macOS), or CFRunLoopRun on the main thread must not call dispatch::main().

Type Aliases

Block

Fn

Functions that take Function<T> type as a parameter also take a pointer to contextual data that you provide. When your dispatch function is called, the pointer to that contextual data is passed as the parameter to the function. The pointer to the contextual data is passed unmodified to your function and it is your responsibility to ensure that the pointer is valid.