cidre::io::surface

Module key

The following list of properties are used with the cf::Dictionary passed to io::Surface::create

Functions

alloc_size

cf::Number of the total allocation size of the buffer including all planes.
Defaults to BufferHeight * BytesPerRow if not specified. Must be specified for dimensionless buffers.

bytes_per_element

cf::Number for the total number of bytes in an element.

bytes_per_row

cache_mode

color_space

content_headroom

cf::Number representation of the content headroom, which is defined as the ratio of nominal peak luminance (“peak white”) to nominal diffuse luminance (“reference white” or “diffuse white”).

element_height

cf::Number for how many pixels high each element is.

element_width

cf::Number for how many pixels wide each element is.

height

cf::Number for the height of the io::Surface buffer in pixels. Required for planar io::Surfaces

icc_profile

offset

pixel_format

pixel_size_casting_allowed

plane_base

plane_bits_per_element

plane_bytes_per_element

plane_bytes_per_row

plane_component_bit_depths

plane_component_bit_offsets

plane_element_height

plane_element_width

plane_height

plane_info

plane_offset

plane_size

plane_width

width

cf::Number for the width of the io::Surface buffer in pixels. Required for planar io::Surfaces