cidre::mach::kern_return

Module err

Constants

ABORTED

The operation was aborted. Ipc code will catch this and reflect it as a message error.

ALREADY_IN_SET

The receive right is already a member of the portset.

ALREADY_WAITING

A thread is attempting to wait for an event for which there is already a waiting thread.

CODESIGN_ERROR

During a page fault, indicates that the page was rejected as a result of a signature check.

DEFAULT_SET

An attempt was made to destroy the default processor set.

DENIED

Denied by security policy

EXCEPTION_PROTECTED

An attempt was made to fetch an exception port that is protected, or to abort a thread while processing a protected exception.

FAILURE

The function could not be performed. A catch-all.

INSUFFICIENT_BUF_SIZE

The provided buffer is of insufficient size for the requested data.

INVALID_ADDRESS

Specified address is not currently valid.

INVALID_ARGUMENT

The function requested was not applicable to this type of argument, or an argument is invalid

INVALID_CAPABILITY

The supplied (port) capability is improper.

INVALID_HOST

Target host isn’t actually a host.

INVALID_KC

The KC on which the function is operating is invalid

INVALID_LEDGER

A ledger was required but not supplied.

INVALID_MEMORY_CONTROL

The port was not a memory cache control port.

INVALID_NAME

The name doesn’t denote a right in the task.

INVALID_OBJECT

The external memory manager failed to initialize the memory object.

INVALID_POLICY

The specified scheduling policy is not currently enabled for the processor set.

INVALID_PROCESSOR_SET

An argument applied to assert processor set privilege was not a processor set control port.

INVALID_RIGHT

The name denotes a right, but not an appropriate right.

INVALID_SECURITY

An argument supplied to assert security privilege was not a host security port.

INVALID_TASK

Target task isn’t an active task.

INVALID_VALUE

A blatant range error.

LOCK_OWNED

The lock is already owned by another thread

LOCK_OWNED_SELF

The lock is already owned by the calling thread

LOCK_SET_DESTROYED

Lock set has been destroyed and is no longer available.

LOCK_UNSTABLE

The thread holding the lock terminated before releasing the lock

MEMORY_DATA_MOVED

A page was requested of a memory manager via memory_object_data_request for an object using a MEMORY_OBJECT_COPY_CALL strategy, with the VM_PROT_WANTS_COPY flag being used to specify that the page desired is for a copy of the object, and the memory manager has detected the page was pushed into a copy of the object while the kernel was walking the shadow chain from the copy to the object. This error code is delivered via memory_object_data_error and is handled by the kernel (it forces the kernel to restart the fault). It will not be seen by users.

MEMORY_ERROR

During a page fault, the memory object indicated that the data could not be returned. This failure may be temporary; future attempts to access this same data may succeed, as defined by the memory object.

MEMORY_FAILURE

During a page fault, the target address refers to a memory object that has been destroyed. This failure is permanent.

MEMORY_PRESENT

An attempt was made to supply “precious” data for memory that is already present in a memory object.

MEMORY_RESTART_COPY

A strategic copy was attempted of an object upon which a quicker copy is now possible. The caller should retry the copy using vm_object_copy_quickly. This error code is seen only by the kernel.

MISSING_KC

The KC on which the function is operating is missing

NAME_EXISTS

The name already denotes a right in the task.

NODE_DOWN

Remote node down or inaccessible.

NOT_DEPRESSED

thread_depress_abort was called on a thread which was not currently depressed.

NOT_FOUND

A search or query operation did not return a result

NOT_IN_SET

The receive right is not a member of a port set.

NOT_RECEIVER

The task in question does not hold receive rights for the port argument.

NOT_SUPPORTED

Empty thread activation (No thread linked to it)

NOT_WAITING

A signalled thread was not actually waiting.

NO_ACCESS

Bogus access restriction.

NO_SPACE

The address range specified is already in use, or no address range of the size specified could be found.

OPERATION_TIMED_OUT

Some thread-oriented operation (semaphore_wait) timed out

POLICY_LIMIT

The specified scheduling attributes exceed the thread’s limits.

POLICY_STATIC

The requested property cannot be changed at this time.

PROTECTION_FAILURE

Specified memory is valid, but does not permit the required forms of access.

RESOURCE_SHORTAGE

A system resource could not be allocated to fulfill this request. This failure may not be permanent.

RETURN_MAX

Maximum return value allowable

RIGHT_EXISTS

The task already has send or receive rights for the port under another name.

RPC_CONTINUE_ORPHAN

Allow an orphaned activation to continue executing.

RPC_SERVER_TERMINATED

Return from RPC indicating the target server was terminated before it successfully replied

RPC_TERMINATE_ORPHAN

Terminate an orphaned activation.

SEMAPHORE_DESTROYED

Semaphore has been destroyed and is no longer available.

TERMINATED

Object has been terminated and is no longer available

UREFS_OVERFLOW

Operation would overflow limit on user-references.