cidre::mach::message

Module err

Constants

IPC_KERNEL

Kernel resource shortage handling an IPC capability.

IPC_SPACE

No room in IPC name space for another capability name.

MASK

RCV_BODY_ERROR

Error receiving message body. See special bits.

RCV_HEADER_ERROR

Error receiving message header. See special bits.

RCV_INTERRUPTED

Software interrupt.

RCV_INVALID_ARGUMENTS

invalid receive arguments, receive has not started

RCV_INVALID_DATA

Bogus message buffer for inline data.

RCV_INVALID_NAME

Bogus name for receive port/port-set.

RCV_INVALID_NOTIFY

Bogus notify port argument.

RCV_INVALID_REPLY

invalid reply port used in a STRICT_REPLY message

RCV_INVALID_TRAILER

trailer type or number of trailer elements not supported

RCV_INVALID_TYPE

Invalid msg-type specification in scatter list.

RCV_IN_PROGRESS

Thread is waiting for receive. (Internal use only.)

RCV_IN_PROGRESS_TIMED

Waiting for receive with timeout. (Internal use only.)

RCV_IN_SET

compatibility: no longer a returned error

RCV_PORT_CHANGED

compatibility: no longer a returned error

RCV_PORT_DIED

Port/set was sent away/died during receive.

RCV_SCATTER_SMALL

Out-of-line overwrite region is not large enough

RCV_TIMED_OUT

Didn’t get a message within the timeout value.

RCV_TOO_LARGE

Message buffer is not large enough for inline data.

SEND_AUX_TOO_SMALL

Message auxiliary data is too small

SEND_INTERRUPTED

Software interrupt.

SEND_INVALID_CONTEXT

The sending thread context did not match the context on the dest port

SEND_INVALID_DATA

Bogus in-line data.

SEND_INVALID_DEST

Bogus destination port.

SEND_INVALID_HEADER

A field in the header had a bad value.

SEND_INVALID_MEMORY

Invalid out-of-line memory pointer.

SEND_INVALID_NOTIFY

Bogus notify port argument.

SEND_INVALID_OPTIONS

Send options are invalid.

SEND_INVALID_REPLY

Bogus reply port.

SEND_INVALID_RIGHT

Bogus port rights in the message body.

SEND_INVALID_RT_OOL_SIZE

Compatibility: no longer a returned error

SEND_INVALID_TRAILER

The trailer to be sent does not match kernel format.

SEND_INVALID_TYPE

Invalid msg-type specification.

SEND_INVALID_VOUCHER

Bogus voucher port.

SEND_IN_PROGRESS

Thread is waiting to send. (Internal use only.)

SEND_MSG_FILTERED

Message send was rejected by message filter

SEND_NO_BUFFER

No message buffer is available.

SEND_NO_GRANT_DEST

The destination port doesn’t accept ports in body

SEND_SEND_AUX_TOO_LARGE

Message auxiliary data is too large

SEND_SEND_MSG_TOO_SMALL

Data doesn’t contain a complete message.

SEND_TIMED_OUT

Message not sent before timeout expired.

SEND_TOO_LARGE

Send is too large for port

VM_KERNEL

Kernel resource shortage handling out-of-line memory.

VM_SPACE

No room in VM address space for out-of-line memory.