cidre::mc

Trait SessionDelegate

pub trait SessionDelegate: Obj {
    // Provided methods
    fn session_peer_did_change_state(
        &mut self,
        session: &Session,
        peer: &PeerId,
        state: SessionState,
    ) { ... }
    fn session_did_receive_data_from_peer(
        &mut self,
        session: &Session,
        data: &Data,
        from_peer: &PeerId,
    ) { ... }
    fn session_did_receive_stream(
        &mut self,
        session: &Session,
        stream: &InputStream,
        stream_name: &String,
        from_peer: &PeerId,
    ) { ... }
    fn session_did_start_receiving_resource(
        &mut self,
        session: &Session,
        resource_name: &String,
        from_peer: &PeerId,
        progress: &Progress,
    ) { ... }
    fn session_did_finish_receiving_resource(
        &mut self,
        session: &Session,
        resource_name: &String,
        from_peer: &PeerId,
        at_url: Option<&Url>,
        err: Option<&Error>,
    ) { ... }
    fn session_did_receive_cert(
        &mut self,
        session: &Session,
        cert: Option<&Array<Id>>,
        from_peer: &PeerId,
        handler: &mut Block<fn(bool)>,
    ) { ... }
    fn sel_session_did_receive_cert() -> &'static Sel { ... }
}

Provided Methods

fn session_peer_did_change_state( &mut self, session: &Session, peer: &PeerId, state: SessionState, )

Remote peer changed state.

fn session_did_receive_data_from_peer( &mut self, session: &Session, data: &Data, from_peer: &PeerId, )

Received data from remote peer.

fn session_did_receive_stream( &mut self, session: &Session, stream: &InputStream, stream_name: &String, from_peer: &PeerId, )

Received a byte stream from remote peer.

fn session_did_start_receiving_resource( &mut self, session: &Session, resource_name: &String, from_peer: &PeerId, progress: &Progress, )

Start receiving a resource from remote peer.

fn session_did_finish_receiving_resource( &mut self, session: &Session, resource_name: &String, from_peer: &PeerId, at_url: Option<&Url>, err: Option<&Error>, )

Finished receiving a resource from remote peer and saved the content in a temporary location - the app is responsible for moving the file to a permanent location within its sandbox.

fn session_did_receive_cert( &mut self, session: &Session, cert: Option<&Array<Id>>, from_peer: &PeerId, handler: &mut Block<fn(bool)>, )

fn sel_session_did_receive_cert() -> &'static Sel

@selector(session:didReceiveCertificate:fromPeer:certificateHandler:) but dynamic use this function to check if object responds to selector

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.