cidre::mlc

Enum ActivationType

#[repr(i32)]
pub enum ActivationType {

    None = 0,
    ReLU = 1,
    Linear = 2,
    Sigmoid = 3,
    HardSigmoid = 4,
    TanH = 5,
    Absolute = 6,
    SoftPlus = 7,
    SoftSign = 8,
    ELU = 9,
    ReLUN = 10,
    LogSigmoid = 11,
    SELU = 12,
    CELU = 13,
    HardShrink = 14,
    SoftShrink = 15,
    TanHShrink = 16,
    Threshold = 17,
    GELU = 18,
    HardSwish = 19,
    Clamp = 20,

}

An activation type that you specify for an activation descriptor.

Variants

None = 0

ReLU = 1

The ReLU activation type.

This activation type implements the following function:

f(x) = x >= 0 ? x : a * x`

Linear = 2

The linear activation type.

This activation type implements the following function:

f(x) = a * x + b

Sigmoid = 3

The sigmoid activation type. This activation type implements the following function:

f(x) = 1 / (1 + e⁻ˣ)

HardSigmoid = 4

The hard sigmoid activation type.

This activation type implements the following function:

f(x) = clamp((x * a) + b, 0, 1)

TanH = 5

The hyperbolic tangent (TanH) activation type. This activation type implements the following function:

f(x) = a * tanh(b * x)

Absolute = 6

The absolute activation type.

This activation type implements the following function:

f(x) = fabs(x)

SoftPlus = 7

The parametric soft plus activation type.

This activation type implements the following function:

f(x) = a * log(1 + e^(b * x))

SoftSign = 8

The parametric soft sign activation type.

This activation type implements the following function:

f(x) = x / (1 + abs(x))

ELU = 9

The parametric ELU activation type. This activation type implements the following function:

f(x) = x >= 0 ? x : a * (exp(x) - 1)

ReLUN = 10

The ReLUN activation type.

This activation type implements the following function:

f(x) = min((x >= 0 ? x : a * x), b)

LogSigmoid = 11

The log sigmoid activation type.

This activation type implements the following function:

f(x) = log(1 / (1 + exp(-x)))

SELU = 12

The SELU activation type.

This activation type implements the following function:

f(x) = scale * (max(0, x) + min(0, α * (exp(x) − 1)))

where:

α = 1.6732632423543772848170429916717
scale = 1.0507009873554804934193349852946

CELU = 13

The CELU activation type.

This activation type implements the following function:

f(x) = max(0, x) + min(0, a * (exp(x / a) − 1))

HardShrink = 14

The hard shrink activation type.

This activation type implements the following function:

f(x) = x, if x > a or x < −a, else 0

SoftShrink = 15

The soft shrink activation type.

This activation type implements the following function:

f(x) = x - a, if x > a, x + a, if x < −a, else 0

TanHShrink = 16

The hyperbolic tangent (TanH) shrink activation type. This activation type implements the following function:

f(x) = x - tanh(x)

Threshold = 17

The threshold activation type.

This activation type implements the following function:

 f(x) = x, if x > a, else b

GELU = 18

The GELU activation type. This activation type implements the following function:

f(x) = x * CDF(x)

HardSwish = 19

The hardswish activation type.

This activation type implements the following function:

f(x) = 0, if x <= -3
f(x) = x, if x >= +3
f(x) = x * (x + 3)/6, otherwise

Clamp = 20

The clamp activation type. This activation type implements the following function:

 f(x) = min(max(x, a), b)

Implementations

impl ActivationType

pub fn debug_desc(self) -> &'static String

Returns a textual description of the arithmetic operation, suitable for debugging