cidre::mlc

Enum ArithmeticOp

#[repr(i32)]
pub enum ArithmeticOp {

    Add = 0,
    Subtract = 1,
    Multiply = 2,
    Divide = 3,
    Floor = 4,
    Round = 5,
    Ceil = 6,
    Sqrt = 7,
    RSqrt = 8,
    Sin = 9,
    Cos = 10,
    Tan = 11,
    ASin = 12,
    ACos = 13,
    ATan = 14,
    SinH = 15,
    CosH = 16,
    TanH = 17,
    ASinH = 18,
    ACosH = 19,
    ATanH = 20,
    Pow = 21,
    Exp = 22,
    Exp2 = 23,
    Log = 24,
    Log2 = 25,
    MultiplyNoNaN = 26,
    DivideNoNaN = 27,
    Min = 28,
    Max = 29,

}

Variants

Add = 0

An operation that calculates the elementwise sum of its two inputs.

Subtract = 1

An operation that calculates the elementwise difference of its two inputs.

Multiply = 2

An operation that calculates the elementwise product of its two inputs.

Divide = 3

An operation that calculates the elementwise division of its two inputs.

Floor = 4

An operation that calculates the elementwise floor of its two inputs.

Round = 5

An operation that calculates the elementwise round of its inputs.

Ceil = 6

An operation that calculates the elementwise ceiling of its inputs.

Sqrt = 7

An operation that calculates the elementwise square root of its inputs.

RSqrt = 8

An operation that calculates the elementwise reciprocal of the square root of its inputs.

Sin = 9

An operation that calculates the elementwise sine of its inputs.

Cos = 10

An operation that calculates the elementwise cosine of its inputs.

Tan = 11

An operation that calculates the elementwise tangent of its inputs.

ASin = 12

An operation that calculates the elementwise inverse sine of its inputs.

ACos = 13

An operation that calculates the elementwise inverse cosine of its inputs.

ATan = 14

An operation that calculates the elementwise inverse tangent of its inputs.

SinH = 15

An operation that calculates the elementwise hyperbolic sine of its inputs.

CosH = 16

An operation that calculates the elementwise hyperbolic cosine of its inputs.

TanH = 17

An operation that calculates the elementwise hyperbolic tangent of its inputs.

ASinH = 18

An operation that calculates the elementwise inverse hyperbolic sine of its inputs.

ACosH = 19

An operation that calculates the elementwise inverse hyperbolic cosine of its inputs.

ATanH = 20

An operation that calculates the elementwise inverse hyperbolic tangent of its inputs.

Pow = 21

An operation that calculates the elementwise first input raised to the power of its second input.

Exp = 22

An operation that calculates the elementwise result of e raised to the power of its input.

Exp2 = 23

An operation that calculates the elementwise result of 2 raised to the power of its input.

Log = 24

An operation that calculates the elementwise natural logarithm of its input.

Log2 = 25

An operation that calculates the elementwise base 2 logarithm of its input.

MultiplyNoNaN = 26

An operation that calculates the elementwise product of its two inputs. Returns 0 if y in x * y is zero, even if x is NaN or INF

DivideNoNaN = 27

An operations that calculates the elementwise division of its two inputs. Returns 0 if the denominator is 0.

Min = 28

An operation that calculates the elementwise min of two inputs.

Max = 29

An operations that calculates the elementwise max of two inputs.

Implementations

impl ArithmeticOp

pub fn debug_desc(self) -> &'static String