cidre::mps

Enum DType

#[repr(u32)]
pub enum DType {

    Invalid = 0,
    F32 = 268_435_488,
    F16 = 268_435_472,
    ComplexF32 = 285_212_736,
    ComplexF16 = 285_212_704,
    I8 = 536_870_920,
    I16 = 536_870_928,
    I32 = 536_870_944,
    I64 = 536_870_976,
    U8 = 8,
    U16 = 16,
    U32 = 32,
    U64 = 64,
    Bool = 2_147_483_656,
    UNorm1 = 1_073_741_825,
    UNorm8 = 1_073_741_832,

}

A value to specify a type of data.

Variants

Invalid = 0

F32 = 268_435_488

32-bit floating point (single-precision).

F16 = 268_435_472

16-bit floating point (half-precision). (IEEE-754-2008 float16 exchange format)

ComplexF32 = 285_212_736

Complex number composed of two 32-bit floating point numbers (single-precision).

ComplexF16 = 285_212_704

Complex number composed of two 16-bit floating point numbers (half-precision). (IEEE-754-2008 float16 exchange format)

I8 = 536_870_920

Signed 8-bit integer.

I16 = 536_870_928

Signed 16-bit integer.

I32 = 536_870_944

I64 = 536_870_976

U8 = 8

Unsigned 8-bit integer. Not normalized

U16 = 16

Unsigned 16-bit integer. Not normalized

U32 = 32

Unsigned 32-bit integer. Not normalized

U64 = 64

Bool = 2_147_483_656

UNorm1 = 1_073_741_825

Unsigned 1-bit normalized value.

UNorm8 = 1_073_741_832

Unsigned 8-bit normalized value.

Implementations

impl DType

pub const FLOAT_BIT: u32 = 268_435_456u32

A common bit for all floating point data types. Zero for integer types

pub const COMPLEX_BIT: u32 = 16_777_216u32

A common bit for all complex point data types. Zero for integer types

pub const SIGNED_BIT: u32 = 536_870_912u32

pub const ALTERNATE_ENCODING_BIT: u32 = 2_147_483_648u32

pub const NORMALIZED_BIT: u32 = 1_073_741_824u32

If set, the value of the shall be interpreted as value / UNORM_TYPE_MAX Normalized values have range [0, 1.0] if unsigned and [-1,1] if signed. SNORM_TYPE_MIN is interpreted as SNORM_TYPE_MIN+1 per standard Metal rules.

pub fn size_of(&self) -> usize

A utility function to get the size of an MPSDataType. This implementation relies upon the specific bit pattern used to encode the type.