cidre::mps

Module graph

Structs

CompilationDesc

Conv2dOpDesc

Device

Executable

ExecutableExecutionDesc

ExecutionDesc

Graph

GraphType

Op

ShapedType

SingleGateRnnDesc

Tensor

TensorData

VariableOp

Enums

DeviceType

ExecutionStage

Execution events that can be used with shared events

Optimization

OptimizationProfile

Opts

Options to be utilized by the graph

PaddingMode

Tensor Padding mode

PaddingStyle

ReductionMode

Reduction Mode

ResizeMode

ResizeNearestRoundingMode

RnnActivation

TensorNamedDataLayout

Type Aliases

CompilationCompletionHandler

CompletionHandler

ExecutableCompletionHandler

ExecutableScheduledHandler

ScheduledHandler

TensorDataDictionary

A dictionary of tensors and correspondiing tensorData for them

TensorShapedTypeDictionary

A dictionary of tensors and correspondiing shapes for them