cidre::mps::graph

Struct Graph

pub struct Graph(/* private fields */);

Implementations

impl Graph

pub fn placeholder_with_shape_ar( &self, shape: Option<&Shape>, data_type: DType, name: Option<&String>, ) -> Rar<Tensor>

pub fn placeholder_with_shape( &self, shape: Option<&Shape>, data_type: DType, name: Option<&String>, ) -> R<Tensor>

pub fn constant_with_data_shape_data_type_ar( &self, data: &Data, shape: &Shape, data_type: DType, ) -> Rar<Tensor>

pub fn constant_with_data_shape_data_type( &self, data: &Data, shape: &Shape, data_type: DType, ) -> R<Tensor>

pub fn constant_ar(&self, scalar: f64, data_type: DType) -> Rar<Tensor>

pub fn constant(&self, scalar: f64, data_type: DType) -> R<Tensor>

pub fn constant_shape_ar( &self, scalar: f64, shape: &Shape, data_type: DType, ) -> Rar<Tensor>

pub fn constant_shape( &self, scalar: f64, shape: &Shape, data_type: DType, ) -> R<Tensor>

impl Graph

pub fn conv_2d_ar( &self, source: &Tensor, weights: &Tensor, descriptor: &Conv2dOpDesc, name: Option<&String>, ) -> Rar<Tensor>

pub fn conv_2d( &self, source: &Tensor, weights: &Tensor, descriptor: &Conv2dOpDesc, name: Option<&String>, ) -> R<Tensor>

impl Graph

pub fn resize_ar( &self, tensor: &Tensor, size: &Shape, mode: ResizeMode, center_result: bool, align_corners: bool, layout: TensorNamedDataLayout, name: Option<&String>, ) -> Rar<Tensor>

pub fn resize( &self, tensor: &Tensor, size: &Shape, mode: ResizeMode, center_result: bool, align_corners: bool, layout: TensorNamedDataLayout, name: Option<&String>, ) -> R<Tensor>

impl Graph

pub fn relu_ar(&self, tensor: &Tensor, name: Option<&String>) -> Rar<Tensor>

pub fn relu(&self, tensor: &Tensor, name: Option<&String>) -> R<Tensor>

pub fn sigmoid_ar(&self, tensor: &Tensor, name: Option<&String>) -> Rar<Tensor>

pub fn sigmoid(&self, tensor: &Tensor, name: Option<&String>) -> R<Tensor>

pub fn soft_max_ar( &self, tensor: &Tensor, axis: Integer, name: Option<&String>, ) -> Rar<Tensor>

pub fn soft_max( &self, tensor: &Tensor, axis: Integer, name: Option<&String>, ) -> R<Tensor>

impl Graph

pub fn add_ar( &self, primary: &Tensor, secondary: &Tensor, name: Option<&String>, ) -> Rar<Tensor>

pub fn add( &self, primary: &Tensor, secondary: &Tensor, name: Option<&String>, ) -> R<Tensor>

pub fn clamp_ar( &self, tensor: &Tensor, min: &Tensor, max: &Tensor, name: Option<&String>, ) -> Rar<Tensor>

pub fn clamp( &self, tensor: &Tensor, min: &Tensor, max: &Tensor, name: Option<&String>, ) -> R<Tensor>

pub fn mul_ar( &self, primary: &Tensor, secondary: &Tensor, name: Option<&String>, ) -> Rar<Tensor>

Create Multiply op and return the result tensor, it supports broadcasting as well

resultTensor = primary * secondary

pub fn mul( &self, primary: &Tensor, secondary: &Tensor, name: Option<&String>, ) -> R<Tensor>

Create Multiply op and return the result tensor, it supports broadcasting as well

resultTensor = primary * secondary

pub fn round_ar(&self, tensor: &Tensor, name: Option<&String>) -> Rar<Tensor>

pub fn round(&self, tensor: &Tensor, name: Option<&String>) -> R<Tensor>

pub fn square_root_ar( &self, tensor: &Tensor, name: Option<&String>, ) -> Rar<Tensor>

pub fn square_root(&self, tensor: &Tensor, name: Option<&String>) -> R<Tensor>

pub fn sub_ar( &self, primary: &Tensor, secondary: &Tensor, name: Option<&String>, ) -> Rar<Tensor>

pub fn sub( &self, primary: &Tensor, secondary: &Tensor, name: Option<&String>, ) -> R<Tensor>

pub fn div_ar( &self, primary: &Tensor, secondary: &Tensor, name: Option<&String>, ) -> Rar<Tensor>

pub fn div( &self, primary: &Tensor, secondary: &Tensor, name: Option<&String>, ) -> R<Tensor>

pub fn tanh_ar(&self, tensor: &Tensor, name: Option<&String>) -> Rar<Tensor>

pub fn tanh(&self, tensor: &Tensor, name: Option<&String>) -> R<Tensor>

pub fn erf_ar(&self, tensor: &Tensor, name: Option<&String>) -> Rar<Tensor>

pub fn erf(&self, tensor: &Tensor, name: Option<&String>) -> R<Tensor>

pub fn cos_ar(&self, tensor: &Tensor, name: Option<&String>) -> Rar<Tensor>

pub fn cos(&self, tensor: &Tensor, name: Option<&String>) -> R<Tensor>

pub fn sin_ar(&self, tensor: &Tensor, name: Option<&String>) -> Rar<Tensor>

pub fn sin(&self, tensor: &Tensor, name: Option<&String>) -> R<Tensor>

impl Graph

pub fn gather_ar( &self, updates: &Tensor, indices: &Tensor, axis: UInteger, batch_dimensions: UInteger, name: Option<&String>, ) -> Rar<Tensor>

pub fn gather( &self, updates: &Tensor, indices: &Tensor, axis: UInteger, batch_dimensions: UInteger, name: Option<&String>, ) -> R<Tensor>

pub fn gather_along_axis_ar( &self, axis: Integer, updates: &Tensor, indices: &Tensor, name: Option<&String>, ) -> Rar<Tensor>

pub fn gather_along_axis( &self, axis: Integer, updates: &Tensor, indices: &Tensor, name: Option<&String>, ) -> R<Tensor>

impl Graph

pub fn mat_mul_ar( &self, primary: &Tensor, secondary: &Tensor, name: Option<&String>, ) -> Rar<Tensor>

pub fn mat_mul( &self, primary: &Tensor, secondary: &Tensor, name: Option<&String>, ) -> R<Tensor>

impl Graph

pub fn mean_ar( &self, tensor: &Tensor, axes: &Array<Number>, name: Option<&String>, ) -> Rar<Tensor>

pub fn mean( &self, tensor: &Tensor, axes: &Array<Number>, name: Option<&String>, ) -> R<Tensor>

pub fn variance_ar( &self, tensor: &Tensor, axes: &Array<Number>, name: Option<&String>, ) -> Rar<Tensor>

pub fn variance( &self, tensor: &Tensor, axes: &Array<Number>, name: Option<&String>, ) -> R<Tensor>

pub fn variance_mean_ar( &self, tensor: &Tensor, mean: &Tensor, axes: &Array<Number>, name: Option<&String>, ) -> Rar<Tensor>

pub fn variance_mean( &self, tensor: &Tensor, mean: &Tensor, axes: &Array<Number>, name: Option<&String>, ) -> R<Tensor>

pub fn normalize_ar( &self, tensor: &Tensor, mean: &Tensor, variance: &Tensor, gamma: Option<&Tensor>, beta: Option<&Tensor>, epsilon: f32, name: Option<&String>, ) -> Rar<Tensor>

pub fn normalize( &self, tensor: &Tensor, mean: &Tensor, variance: &Tensor, gamma: Option<&Tensor>, beta: Option<&Tensor>, epsilon: f32, name: Option<&String>, ) -> R<Tensor>

impl Graph

pub fn broadcast_ar( &self, tensor: &Tensor, shape: &Shape, name: Option<&String>, ) -> Rar<Tensor>

pub fn broadcast( &self, tensor: &Tensor, shape: &Shape, name: Option<&String>, ) -> R<Tensor>

pub fn concat( &self, tensors: &[&Tensor], dimension: Integer, name: Option<&String>, ) -> R<Tensor>

pub fn concat_tensors_ar( &self, tensors: &Array<Tensor>, dimension: Integer, name: Option<&String>, ) -> Rar<Tensor>

pub fn concat_tensors( &self, tensors: &Array<Tensor>, dimension: Integer, name: Option<&String>, ) -> R<Tensor>

pub fn expand_dims_ar( &self, tensor: &Tensor, axis: isize, name: Option<&String>, ) -> Rar<Tensor>

pub fn expand_dims( &self, tensor: &Tensor, axis: isize, name: Option<&String>, ) -> R<Tensor>

pub fn expand_dims_axes_ar( &self, tensor: &Tensor, axes: &Array<Number>, name: Option<&String>, ) -> Rar<Tensor>

pub fn expand_dims_axes( &self, tensor: &Tensor, axes: &Array<Number>, name: Option<&String>, ) -> R<Tensor>

pub fn reshape_ar( &self, tensor: &Tensor, shape: &Shape, name: Option<&String>, ) -> Rar<Tensor>

pub fn reshape( &self, tensor: &Tensor, shape: &Shape, name: Option<&String>, ) -> R<Tensor>

pub fn transpose_with_dimension_ar( &self, tensor: &Tensor, dimention: UInteger, with_dimention: UInteger, name: Option<&String>, ) -> Rar<Tensor>

pub fn transpose_with_dimension( &self, tensor: &Tensor, dimention: UInteger, with_dimention: UInteger, name: Option<&String>, ) -> R<Tensor>

pub fn cast_ar( &self, tensor: &Tensor, to_type: DType, name: Option<&String>, ) -> Rar<Tensor>

pub fn cast( &self, tensor: &Tensor, to_type: DType, name: Option<&String>, ) -> R<Tensor>

pub fn slice_tensor_ar( &self, tensor: &Tensor, dimension: UInteger, start: Integer, length: Integer, name: Option<&String>, ) -> Rar<Tensor>

pub fn slice_tensor( &self, tensor: &Tensor, dimension: UInteger, start: Integer, length: Integer, name: Option<&String>, ) -> R<Tensor>

impl Graph

pub fn single_gate_rnn_ar( &self, source: &Tensor, recurrent_weight: &Tensor, input_weight: Option<&Tensor>, bias: Option<&Tensor>, init_state: Option<&Tensor>, mask: Option<&Tensor>, descriptor: &SingleGateRnnDesc, name: Option<&String>, ) -> Rar<Array<Tensor>>

pub fn single_gate_rnn( &self, source: &Tensor, recurrent_weight: &Tensor, input_weight: Option<&Tensor>, bias: Option<&Tensor>, init_state: Option<&Tensor>, mask: Option<&Tensor>, descriptor: &SingleGateRnnDesc, name: Option<&String>, ) -> R<Array<Tensor>>

pub fn single_gate_rnn_grad_ar( &self, source: &Tensor, recurrent_weight: &Tensor, source_gradient: &Tensor, z_state: &Tensor, stage_gradient: Option<&Tensor>, input_weight: Option<&Tensor>, bias: Option<&Tensor>, init_state: Option<&Tensor>, mask: Option<&Tensor>, descriptor: &SingleGateRnnDesc, name: Option<&String>, ) -> Rar<Array<Tensor>>

pub fn single_gate_rnn_grad( &self, source: &Tensor, recurrent_weight: &Tensor, source_gradient: &Tensor, z_state: &Tensor, stage_gradient: Option<&Tensor>, input_weight: Option<&Tensor>, bias: Option<&Tensor>, init_state: Option<&Tensor>, mask: Option<&Tensor>, descriptor: &SingleGateRnnDesc, name: Option<&String>, ) -> R<Array<Tensor>>

pub fn lstm_ar( &self, source: &Tensor, recurrent_weight: &Tensor, input_weight: Option<&Tensor>, bias: Option<&Tensor>, init_state: Option<&Tensor>, init_cell: Option<&Tensor>, mask: Option<&Tensor>, peephole: Option<&Tensor>, descriptor: &LstmDesc, name: Option<&String>, ) -> Rar<Array<Tensor>>

pub fn lstm( &self, source: &Tensor, recurrent_weight: &Tensor, input_weight: Option<&Tensor>, bias: Option<&Tensor>, init_state: Option<&Tensor>, init_cell: Option<&Tensor>, mask: Option<&Tensor>, peephole: Option<&Tensor>, descriptor: &LstmDesc, name: Option<&String>, ) -> R<Array<Tensor>>

pub fn lstm_grad_ar( &self, source: &Tensor, recurrent_weight: &Tensor, source_gradient: &Tensor, z_state: &Tensor, cell_output_fwd: &Tensor, state_gradient: Option<&Tensor>, cell_gradient: Option<&Tensor>, input_weight: Option<&Tensor>, bias: Option<&Tensor>, init_state: Option<&Tensor>, init_cell: Option<&Tensor>, mask: Option<&Tensor>, peephole: Option<&Tensor>, descriptor: &LstmDesc, name: Option<&String>, ) -> Rar<Array<Tensor>>

pub fn lstm_grad( &self, source: &Tensor, recurrent_weight: &Tensor, source_gradient: &Tensor, z_state: &Tensor, cell_output_fwd: &Tensor, state_gradient: Option<&Tensor>, cell_gradient: Option<&Tensor>, input_weight: Option<&Tensor>, bias: Option<&Tensor>, init_state: Option<&Tensor>, init_cell: Option<&Tensor>, mask: Option<&Tensor>, peephole: Option<&Tensor>, descriptor: &LstmDesc, name: Option<&String>, ) -> R<Array<Tensor>>

pub fn gru_ar( &self, source: &Tensor, recurrent_weight: &Tensor, input_weight: Option<&Tensor>, bias: Option<&Tensor>, init_state: Option<&Tensor>, mask: Option<&Tensor>, secondary_bias: Option<&Tensor>, descriptor: &GruDescr, name: Option<&String>, ) -> Rar<Array<Tensor>>

pub fn gru( &self, source: &Tensor, recurrent_weight: &Tensor, input_weight: Option<&Tensor>, bias: Option<&Tensor>, init_state: Option<&Tensor>, mask: Option<&Tensor>, secondary_bias: Option<&Tensor>, descriptor: &GruDescr, name: Option<&String>, ) -> R<Array<Tensor>>

pub fn gru_grad_ar( &self, source: &Tensor, recurrent_weight: &Tensor, source_gradient: &Tensor, z_state: &Tensor, output_fwd: &Tensor, state_gradient: Option<&Tensor>, input_weight: Option<&Tensor>, bias: Option<&Tensor>, init_state: Option<&Tensor>, mask: Option<&Tensor>, secondary_bias: Option<&Tensor>, descriptor: &GruDescr, name: Option<&String>, ) -> Rar<Array<Tensor>>

pub fn gru_grad( &self, source: &Tensor, recurrent_weight: &Tensor, source_gradient: &Tensor, z_state: &Tensor, output_fwd: &Tensor, state_gradient: Option<&Tensor>, input_weight: Option<&Tensor>, bias: Option<&Tensor>, init_state: Option<&Tensor>, mask: Option<&Tensor>, secondary_bias: Option<&Tensor>, descriptor: &GruDescr, name: Option<&String>, ) -> R<Array<Tensor>>

impl Graph

pub fn retained(&self) -> R<Self>

impl Graph

pub fn cls() -> &'static Class<Self>

pub fn cls_ptr() -> *const c_void

pub fn alloc() -> A<Self>

pub fn new() -> R<Graph>

shortcut to Self::alloc().init()

impl Graph

pub fn opts(&self) -> Opts

pub fn set_opts(&mut self, value: Opts)

pub fn compile_with_device_ar( &self, device: Option<&Device>, feeds: &TensorShapedTypeDictionary, target_tensors: &Array<Tensor>, target_operations: Option<&Array<Op>>, compilation_descriptor: Option<&CompilationDesc>, ) -> Rar<Executable>

pub fn sel_compile_with_device() -> &'static Sel

@selector(compileWithDevice:feeds:targetTensors:targetOperations:compilationDescriptor:) but dynamic use this function to check if object responds to selector

pub fn compile_with_device( &self, device: Option<&Device>, feeds: &TensorShapedTypeDictionary, target_tensors: &Array<Tensor>, target_operations: Option<&Array<Op>>, compilation_descriptor: Option<&CompilationDesc>, ) -> R<Executable>

pub fn run_with_feeds_ar( &self, feeds: &TensorDataDictionary, target_tensors: &Array<Tensor>, target_operations: Option<&Array<Op>>, ) -> Rar<TensorDataDictionary>

pub fn run_with_feeds( &self, feeds: &TensorDataDictionary, target_tensors: &Array<Tensor>, target_operations: Option<&Array<Op>>, ) -> R<TensorDataDictionary>

pub fn run_with_cmd_queue_ar( &self, command_queue: &CmdQueue, feeds: &TensorDataDictionary, target_tensors: &Array<Tensor>, target_operations: Option<&Array<Op>>, ) -> Rar<TensorDataDictionary>

pub fn run_with_cmd_queue( &self, command_queue: &CmdQueue, feeds: &TensorDataDictionary, target_tensors: &Array<Tensor>, target_operations: Option<&Array<Op>>, ) -> R<TensorDataDictionary>

pub fn run_async_with_feeds_ar( &self, feeds: &TensorDataDictionary, target_tensors: &Array<Tensor>, target_operations: Option<&Array<Op>>, execution_descriptor: Option<&ExecutionDesc>, ) -> Rar<TensorDataDictionary>

pub fn run_async_with_feeds( &self, feeds: &TensorDataDictionary, target_tensors: &Array<Tensor>, target_operations: Option<&Array<Op>>, execution_descriptor: Option<&ExecutionDesc>, ) -> R<TensorDataDictionary>

pub fn run_async_with_cmd_queue_ar( &self, command_queue: &CmdQueue, feeds: &TensorDataDictionary, target_tensors: &Array<Tensor>, target_operations: Option<&Array<Op>>, execution_descriptor: Option<&ExecutionDesc>, ) -> Rar<TensorDataDictionary>

pub fn run_async_with_cmd_queue( &self, command_queue: &CmdQueue, feeds: &TensorDataDictionary, target_tensors: &Array<Tensor>, target_operations: Option<&Array<Op>>, execution_descriptor: Option<&ExecutionDesc>, ) -> R<TensorDataDictionary>