cidre

Module mtl

Metal

Re-exports

pub use acceleration_structure_types::AxisAlignedBoundingBox;

pub use acceleration_structure_types::ComponentTransform;

pub use acceleration_structure_types::PackedF32Quaternion;

pub use acceleration_structure_types::PackedF32x3;

pub use acceleration_structure_types::PackedF32x4x3;

pub use crate::mtl4::render_pipeline::LogicalToPhysicalColorAttachMap;

pub use crate::mtl4::render_pipeline::LogicalToPhysicalColorAttachMappingState;

Modules

acceleration_structure_types

Structs

AccelerationStruct

AccelerationStructDesc

AccelerationStructGeometryDesc

Allocation

AnyDrawable

AnyFnHandle

ArgEncoder

ArrayType

Attr

An object that describes an attribute defined in the stage-in argument for a shader.

BarrierScope

Describes the types of resources that the a barrier operates on

BaseType

Binding

BlitCmdEncoder

BlitOpt

BlitPassDesc

BlitPassSampleBufAttachDesc

BlitPassSampleBufAttachDescArray

Buf

A typeless allocation accessible by both the CPU and the GPU (MTLDevice) or by only the GPU when the storage mode is MTLResourceStorageModePrivate.

CaptureDesc

CaptureManager

CaptureScope

ClearColor

CmdBuf

A serial list of commands for the device to execute.

CmdEncoder

An encoder that writes GPU commands into a command buffer.

CmdQueue

A serial queue of command buffers to be executed by the device.

CompileOpts

Compilation settings for a Metal shader library.

ComputeCmdEncoder

ComputePassDesc

ComputePassSampleBufAttachDesc

ComputePassSampleBufAttachDescArray

ComputePipelineDesc

ComputePipelineReflection

ComputePipelineState

Counter

CounterResultStageUtilization

CounterResultStatistic

CounterResultTimestamp

A timestamp value from a GPU at a particular point in time during an operation, typically at the beginning or ending of a render stage.

CounterSampleBuf

A specialized memory buffer that stores a GPU’s counter set data.

CounterSampleBufDesc

A group of properties that configures the counter sample buffers you create with it.

CounterSet

DepthStencilDesc

DepthStencilState

A depth and stencil state object that specifies the depth and stencil configuration and operations used in a render pass.

Device

DrawIndexedPrimitivesIndirectArgs

DrawPatchIndirectArgs

DrawPrimitivesIndirectArgs

Event

Fence

An object that can capture, track, and manage resource dependencies across command encoders.

Fn

An object that represents a public shader function in a Metal library.

FnConstValues

FnDesc

FnOpts

FnStitchedLibDesc

FnStitchingAttr

FnStitchingAttrAlwaysInline

FnStitchingFnNode

FnStitchingGraph

FnStitchingInputNode

FnStitchingNode

Heap

HeapDesc

IndirectCmBufExecutionRange

IndirectCmdBuf

A command buffer containing reusable commands, encoded either on the CPU or GPU.

IndirectCmdBufDesc

Describes the limits and features that can be used in an indirect command

IndirectCmdType

IndirectComputeCmd

IndirectRenderCmd

IntersectionFnTable

IntersectionFnTableDesc

Lib

MeshRenderPipelineDesc

MotionBoundingBoxGeometryDesc

Origin

Identify a pixel in an image. MTLOrigin is ususally used as the upper-left corner of a region of a texture.

ParallelRenderCmdEncoder

PipelineBufDesc

PipelineBufDescArray

PipelineOpt

PointerType

QuadTessellationFactorsHalf

Region

Identify a region in an image or texture.

RenderCmdEncoder

An interface that encodes a render pass into a command buffer, including all its draw calls and configuration.

RenderPassAttachDesc

RenderPassColorAttachDesc

RenderPassColorAttachDescArray

RenderPassDepthAttachDesc

RenderPassDesc

RenderPassStencilAttachDesc

RenderPipelineColorAttachDesc

RenderPipelineColorAttachDescArray

RenderPipelineDesc

RenderPipelineFnsDesc

RenderPipelineReflection

RenderPipelineState

RenderStages

Res

ResId

Handle of the GPU resource suitable for storing in an Argument Buffer

ResOpts

ResUsage

Describes how a resource will be used by a shader through an argument buffer

ResidencySet

ResidencySetDesc

SamplePos

Identify a sample within a pixel. Origin is top-left with a range [0,1) for both x and y.

SamplerDesc

SamplerState

ScissorRect

SharedEvent

An object you use to synchronize access to Metal resources across multiple CPUs, GPUs, and processes.

SharedEventHandle

mtl::SharedEventHandle objects may be passed between processes via XPC connections and then used to recreate a mtl::SharedEvent via an existing mtl::Device.

SharedEventListener

This class provides a simple interface for handling the dispatching of mtl::SharedEvent notifications from Metal.

SharedTextureHandle

Size

A set of dimensions to declare the size of an object, such as an image, texture, threadgroup, or grid.

SizeAlign

Stages

StencilDesc

StructMember

StructType

Tensor

TensorDesc

TensorExtents

TensorUsage

Texture

A resource that holds formatted image data.

TextureDesc

An object that you use to configure new Metal texture objects.

TextureRefType

TextureSwizzleChannels

TextureUsage

TileRenderPipelineColorAttachDesc

TileRenderPipelineColorAttachDescArray

TileRenderPipelineDesc

TriangleTessellationFactorsHalf

VertexAmplificationViewMapping

VertexAttr

An object that represents an attribute of a vertex function.

VertexAttrDesc

VertexAttrDescArray

VertexBufLayoutDesc

VertexBufLayoutDescArray

VertexDesc

Viewport

A 3D rectangular region for the viewport clipping.

VisibleFnTable

A table of shader functions visible to your app that you can pass into compute commands to customize the behavior of a shader.

VisibleFnTableDesc

A specification of how to create a visible function table.

Enums

AccelerationStructInstanceOpts

AccelerationStructUsage

ArgAccess

ArgBufsTier

ArgType

AttrFormat

BlendFactor

BlendOp

CaptureDst

The destination where you want the GPU trace to be captured to.

CmdBufErr

CmdBufStatus

Reports the current stage in the lifetime of MTLCommandBuffer, as it proceeds to enqueued, committed, scheduled, and completed.

ColorWriteMask

CompareFn

CompileSymbolVisibility

CpuCacheMode

Describes what CPU cache mode is used for the CPU’s mapping of a texture resource.

CullMode

The mode that determines whether to perform culling and which type of primitive to cull.

DType

DepthClipMode

DispatchType

FnType

GpuFamily

HazardTrackingMode

Describes how hazard tracking is performed.

HeapType

IndexType

IntersectionFnSignature

LangVersion

LibError

LibOptimizationLevel

LibType

LoadAction

MathFloatingPointFns

An enum to indicate the default math functions for single precision floating-point

MathMode

An enum to indicate if the compiler can perform optimizations for floating-point arithmetic that may violate the IEEE 754 standard

MatrixLayout

MotionBorderMode

Mutability

Specifies whether a buffer will be modified between the time it is bound and a compute or render pipeline is executed in a draw or dispatch.

PatchType

PixelFormat

Primitive

The geometric primitive type for drawing commands.

PrimitiveTopologyClass

PurgableState

Options for set_purgeable call.

ReadWriteTextureTier

SamplerAddrMode

SamplerBorderColor

SamplerMinMagFilter

SamplerMipFilter

ShaderValidation

SparsePageSize

SparseTextureMappingMode

StencilOp

StorageMode

Describes location and CPU mapping of MTLTexture.

StoreAction

Types of actions performed for an attachment at the end of a rendering pass.

StoreActionOpts

TensorDType

The possible data types for the elements of a tensor.

TessellationControlPointIndexType

TessellationFactorFormat

TessellationFactorStepFn

TessellationPartitionMode

TextureCompression

TextureSwizzle

TextureType

Describes the dimensionality of each image, and if multiple images are arranged into an array or cube.

TriangleFillMode

VertexFormat

VertexStepFn

VisibilityResultMode

VisibilityResultType

Winding

The vertex winding rule that determines a front-facing primitive.

Constants

CPU_CACHE_MODE_MASK

CPU_CACHE_MODE_SHIFT

HAZARD_TRACKING_MODE_MASK

HAZARD_TRACKING_MODE_SHIFT

STORAGE_MODE_MASK

STORAGE_MODE_SHIFT

TENSOR_MAX_RANK

The largest rank a tensor can have.

Traits

Drawable

A displayable resource that can be rendered or written to.

FnHandle

Type Aliases

CmdBufHandler

Coordinate2d

A floating point coordinate in an abstract 2D space.

DrawablePresentedHandler

LibErrorDomain

SharedEventNotificationBlock

A block of code invoked after a shareable event’s signal value equals or exceeds a given value.

Timestamp