cidre::mtl

Enum PixelFormat

#[repr(usize)]
pub enum PixelFormat {

    Invalid = 0,
    A8UNorm = 1,
    R8UNorm = 10,
    R8UNormSrgb = 11,
    R8SNorm = 12,
    R8UInt = 13,
    R8SInt = 14,
    R16UNorm = 20,
    R16SNorm = 22,
    R16UInt = 23,
    R16SInt = 24,
    R16Float = 25,
    Rg8UNorm = 30,
    Rg8UNormSrgb = 31,
    Rg8SNorm = 32,
    Rg8UInt = 33,
    Rg8SInt = 34,
    B5G6R5UNorm = 40,
    A1Bgr5UNorm = 41,
    Abgr4UNorm = 42,
    Bgr5A1UNorm = 43,
    R32UInt = 53,
    R32SInt = 54,
    R32Float = 55,
    Rg16UNorm = 60,
    Rg16SNorm = 62,
    Rg16UInt = 63,
    Rg16SInt = 64,
    Rg16Float = 65,
    Rgba8UNorm = 70,
    Rgba8UNormSrgb = 71,
    Rgba8SNorm = 72,
    Rgba8UInt = 73,
    Rgba8SInt = 74,
    Bgra8UNorm = 80,
    Bgra8UNormSrgb = 81,
    Rgb10A2UNorm = 90,
    Rgb10A2UInt = 91,
    Rg11B10Float = 92,
    Rgb9E5Float = 93,
    Bgr10A2UNorm = 94,
    Bgr10Xr = 554,
    Bgr10XrSrgb = 555,
    Rg32UInt = 103,
    Rg32SInt = 104,
    Rg32Float = 105,
    Rgba16UNorm = 110,
    Rgba16SNorm = 112,
    Rgba16UInt = 113,
    Rgba16SInt = 114,
    Rgba16Float = 115,
    Bgra10Xr = 552,
    Bgra10XrSrgb = 553,
    Rgba32UInt = 123,
    Rgba32SInt = 124,
    Rgba32Float = 125,
    Bc1Rgba = 130,
    Bc1RgbaSrgb = 131,
    Bc2Rgba = 132,
    Bc2RgbaSrgb = 133,
    Bc3Rgba = 134,
    Bc3RgbaSrgb = 135,
    Bc6HRgbFloat = 150,
    Bc6HRgbUFloat = 151,
    Bc7RgbaUNorm = 152,
    Bc7RgbaUNormSrgb = 153,
    Astc4x4Srgb = 186,
    Astc5x4Srgb = 187,
    Astc5x5Srgb = 188,
    Astc6x5Srgb = 189,
    Astc6x6Srgb = 190,
    Astc8x5Srgb = 192,
    Astc8x6Srgb = 193,
    Astc8x8Srgb = 194,
    Astc10x5Srgb = 195,
    Astc10x6Srgb = 196,
    Astc10x8Srgb = 197,
    Astc10x10Srgb = 198,
    Astc12x10Srgb = 199,
    Astc12x12Srgb = 200,
    Astc4x4Ldr = 204,
    Astc5x4Ldr = 205,
    Astc5x5Ldr = 206,
    Astc6x5Ldr = 207,
    Astc6x6Ldr = 208,
    Astc8x5Ldr = 210,
    Astc8x6Ldr = 211,
    Astc8x8Ldr = 212,
    Astc10x5Ldr = 213,
    Astc10x6Ldr = 214,
    Astc10x8Ldr = 215,
    Astc10x10Ldr = 216,
    Astc12x10Ldr = 217,
    Astc12x12Ldr = 218,
    Astc4x4Hdr = 222,
    Astc5x4Hdr = 223,
    Astc5x5Hdr = 224,
    Astc6x5Hdr = 225,
    Astc6x6Hdr = 226,
    Astc8x5Hdr = 228,
    Astc8x6Hdr = 229,
    Astc8x8Hdr = 230,
    Astc10x5Hdr = 231,
    Astc10x6Hdr = 232,
    Astc10x8Hdr = 233,
    Astc10x10Hdr = 234,
    Astc12x10Hdr = 235,
    Asrc12x12Hdr = 236,
    Gbgr422 = 240,
    Bgrg422 = 241,
    Depth16Unorm = 250,
    Depth32Float = 252,
    Stencil8 = 253,
    Depth24UnormStencil8 = 255,
    Depth32FloatStencil8 = 260,
    X32Stencil8 = 261,
    X24Stencil8 = 262,
    Unspecialized = 263,

}

Variants

Invalid = 0

The default value of the pixel format for the mtl::RenderPipelineState. You cannot create a texture with this value.

A8UNorm = 1

Ordinary format with one 8-bit normalized unsigned integer component

R8UNorm = 10

Ordinary format with one 8-bit normalized unsigned integer component

R8UNormSrgb = 11

Ordinary format with one 8-bit normalized unsigned integer component with conversion between sRGB and linear space.

R8SNorm = 12

Ordinary format with one 8-bit normalized signed integer component.

R8UInt = 13

Ordinary format with one 8-bit unsigned integer component.

R8SInt = 14

Ordinary format with one 8-bit signed integer component.

R16UNorm = 20

Ordinary format with one 16-bit normalized unsigned integer component.

R16SNorm = 22

Ordinary format with one 16-bit normalized signed integer component.

R16UInt = 23

Ordinary format with one 16-bit unsigned integer component.

R16SInt = 24

Ordinary format with one 16-bit signed integer component.

R16Float = 25

Ordinary format with one 16-bit floating-point component.

Rg8UNorm = 30

Ordinary format with two 8-bit normalized unsigned integer components.

Rg8UNormSrgb = 31

Ordinary format with two 8-bit normalized unsigned integer components with conversion between sRGB and linear space.

Rg8SNorm = 32

Ordinary format with two 8-bit normalized signed integer components.

Rg8UInt = 33

Ordinary format with two 8-bit unsigned integer components.

Rg8SInt = 34

Ordinary format with two 8-bit signed integer components.

B5G6R5UNorm = 40

Packed 16-bit format with normalized unsigned integer color components: 5 bits for blue, 6 bits for green, 5 bits for red, packed into 16 bits.

A1Bgr5UNorm = 41

Packed 16-bit format with normalized unsigned integer color components: 5 bits each for BGR and 1 for alpha, packed into 16 bits.

Abgr4UNorm = 42

Packed 16-bit format with normalized unsigned integer color components: 4 bits each for ABGR, packed into 16 bits.

Bgr5A1UNorm = 43

Packed 16-bit format with normalized unsigned integer color components: 5 bits each for BGR and 1 for alpha, packed into 16 bits.

R32UInt = 53

Ordinary format with one 32-bit unsigned integer component.

R32SInt = 54

Ordinary format with one 32-bit signed integer component.

R32Float = 55

Ordinary format with one 32-bit floating-point component.

Rg16UNorm = 60

Ordinary format with two 16-bit normalized unsigned integer components.

Rg16SNorm = 62

Ordinary format with two 16-bit normalized signed integer components.

Rg16UInt = 63

Ordinary format with two 16-bit unsigned integer components.

Rg16SInt = 64

Ordinary format with two 16-bit signed integer components.

Rg16Float = 65

Ordinary format with two 16-bit floating-point components.

Rgba8UNorm = 70

Ordinary format with four 8-bit normalized unsigned integer components in RGBA order.

Rgba8UNormSrgb = 71

Ordinary format with four 8-bit normalized unsigned integer components in RGBA order with conversion between sRGB and linear space.

Rgba8SNorm = 72

Ordinary format with four 8-bit normalized signed integer components in RGBA order.

Rgba8UInt = 73

Ordinary format with four 8-bit unsigned integer components in RGBA order.

Rgba8SInt = 74

Ordinary format with four 8-bit signed integer components in RGBA order.

Bgra8UNorm = 80

Ordinary format with four 8-bit normalized unsigned integer components in BGRA order.

Bgra8UNormSrgb = 81

Ordinary format with four 8-bit normalized unsigned integer components in BGRA order with conversion between sRGB and linear space.

Rgb10A2UNorm = 90

A 32-bit packed pixel format with four normalized unsigned integer components: 10-bit red, 10-bit green, 10-bit blue, and 2-bit alpha.

Rgb10A2UInt = 91

A 32-bit packed pixel format with four unsigned integer components: 10-bit red, 10-bit green, 10-bit blue, and 2-bit alpha.

Pixel data is stored in red, green, blue, and alpha order, from least significant bit to most significant bit.

Rg11B10Float = 92

32-bit format with floating-point color components, 11 bits each for red and green and 10 bits for blue.

Rgb9E5Float = 93

Packed 32-bit format with floating-point color components: 9 bits each for RGB and 5 bits for an exponent shared by RGB, packed into 32 bits.

Bgr10A2UNorm = 94

A 32-bit packed pixel format with four normalized unsigned integer components: 10-bit blue, 10-bit green, 10-bit red, and 2-bit alpha.

Bgr10Xr = 554

A 32-bit extended range pixel format with three fixed-point components: 10-bit blue, 10-bit green, and 10-bit red.

Pixel data is stored in blue, green, and red order, from least significant bit to most significant bit. Bits 30 and 31 are used as padding, and their value is 0.

The blue, green, and red components are linearly encoded, and their values range from -0.752941 to 1.25098.

To display wide color values on devices with wide color displays, you can set this pixel format on the colorPixelFormat property of an ‘mtk::View’ or the pixel_format property of a crate::ca::MetalLayer.

Bgr10XrSrgb = 555

A 32-bit extended range pixel format with sRGB conversion and three fixed-point components: 10-bit blue, 10-bit green, and 10-bit red.

Pixel data is stored in blue, green, and red order, from least significant bit to most significant bit. Bits 30 and 31 are used as padding, and their value is 0.

The blue, green, and red components are gamma encoded, and their values range from -0.5271 to 1.66894, before gamma expansion.

To display wide color values on devices with wide color displays, you can set this pixel format on the colorPixelFormat property of an ‘mtk::View’ or the pixel_format property of a crate::ca::MetalLayer. You must also use specify an extended sRGB color space.

Rg32UInt = 103

Ordinary format with two 32-bit unsigned integer components.

Rg32SInt = 104

Ordinary format with two 32-bit signed integer components.

Rg32Float = 105

Ordinary format with two 32-bit floating-point components.

Rgba16UNorm = 110

Ordinary format with four 16-bit normalized unsigned integer components in RGBA order.

Rgba16SNorm = 112

Ordinary format with four 16-bit normalized signed integer components in RGBA order.

Rgba16UInt = 113

Ordinary format with four 16-bit unsigned integer components in RGBA order.

Rgba16SInt = 114

Ordinary format with four 16-bit signed integer components in RGBA order.

Rgba16Float = 115

Ordinary format with four 16-bit floating-point components in RGBA order.

Bgra10Xr = 552

A 64-bit extended range pixel format with four fixed-point components: 10-bit blue, 10-bit green, 10-bit red, and 10-bit alpha.

Pixel data is stored in blue, green, red, and alpha order, from least significant bit to most significant bit. Each component is stored in a 16-bit chunk that’s arranged as follows:

The 10 left-most bits store the component’s data.

The 6 right-most bits are used as padding, and their value is 0.

The blue, green, and red components are linearly encoded, and their values range from -0.752941 to 1.25098. The alpha component is always clamped to a [0.0, 1.0] range in sampling, rendering, and writing operations, despite supporting values outside this range.

To display wide color values on devices with wide color displays, you can set this pixel format on the colorPixelFormat property of an ‘mtk::View’ or the pixelFormat property of a crate::ca::MetalLayer.

Bgra10XrSrgb = 553

A 64-bit extended range pixel format with sRGB conversion and four fixed-point components: 10-bit blue, 10-bit green, 10-bit red, and 10-bit alpha.

Pixel data is stored in blue, green, red, and alpha order, from least significant bit to most significant bit. Each component is stored in a 16-bit chunk that’s arranged as follows:

The 10 left-most bits store the component’s data.

The 6 right-most bits are used as padding, and their value is 0.

The blue, green, and red components are gamma encoded, and their values range from -0.5271 to 1.66894, before gamma expansion. The alpha component is always clamped to a [0.0, 1.0] range in sampling, rendering, and writing operations, despite supporting values outside this range.

To display wide color values on devices with wide color displays, you can set this pixel format on the colorPixelFormat property of an ‘mtk::View’ or the pixelFormat property of a crate::ca::MetalLayer. You must also use specify an extended sRGB color space.

Rgba32UInt = 123

Ordinary format with four 32-bit unsigned integer components in RGBA order.

Rgba32SInt = 124

Ordinary format with four 32-bit signed integer components in RGBA order.

Rgba32Float = 125

Ordinary format with four 32-bit floating-point components in RGBA order.

Bc1Rgba = 130

Compressed format with two 16-bit color components and one 32-bit descriptor component.

Bc1RgbaSrgb = 131

Compressed format with two 16-bit color components and one 32-bit descriptor component, with conversion between sRGB and linear space.

Bc2Rgba = 132

Compressed format with two 64-bit chunks.

The first chunk contains two 8-bit alpha components and one 48-bit descriptor component. The second chunk contains two 16-bit color components and one 32-bit descriptor component.

Bc2RgbaSrgb = 133

Compressed format with two 64-bit chunks, with conversion between sRGB and linear space.

The first chunk contains two 8-bit alpha components and one 48-bit descriptor component. The second chunk contains two 16-bit color components and one 32-bit descriptor component.

Bc3Rgba = 134

Compressed format with two 64-bit chunks.

The first chunk contains two 8-bit alpha components and one 48-bit descriptor component. The second chunk contains two 16-bit color components and one 32-bit descriptor component.

Bc3RgbaSrgb = 135

Compressed format with two 64-bit chunks, with conversion between sRGB and linear space.

The first chunk contains two 8-bit alpha components and one 48-bit descriptor component. The second chunk contains two 16-bit color components and one 32-bit descriptor component.

Bc6HRgbFloat = 150

Compressed format with four floating-point components.

Bc6HRgbUFloat = 151

Compressed format with four unsigned floating-point components.

Bc7RgbaUNorm = 152

Compressed format with four normalized unsigned integer components.

Bc7RgbaUNormSrgb = 153

Compressed format with four normalized unsigned integer components, with conversion between sRGB and linear space.

Astc4x4Srgb = 186

ASTC-compressed format with low-dynamic-range content, conversion between sRGB and linear space, a block width of 4, and a block height of 4.

Astc5x4Srgb = 187

ASTC-compressed format with low-dynamic-range content, conversion between sRGB and linear space, a block width of 5, and a block height of 4.

Astc5x5Srgb = 188

ASTC-compressed format with low-dynamic-range content, conversion between sRGB and linear space, a block width of 5, and a block height of 5.

Astc6x5Srgb = 189

ASTC-compressed format with low-dynamic-range content, conversion between sRGB and linear space, a block width of 6, and a block height of 5.

Astc6x6Srgb = 190

ASTC-compressed format with low-dynamic-range content, conversion between sRGB and linear space, a block width of 6, and a block height of 6.

Astc8x5Srgb = 192

ASTC-compressed format with low-dynamic-range content, conversion between sRGB and linear space, a block width of 8, and a block height of 5.

Astc8x6Srgb = 193

ASTC-compressed format with low-dynamic-range content, conversion between sRGB and linear space, a block width of 8, and a block height of 6.

Astc8x8Srgb = 194

ASTC-compressed format with low-dynamic-range content, conversion between sRGB and linear space, a block width of 8, and a block height of 8.

Astc10x5Srgb = 195

Astc10x6Srgb = 196

Astc10x8Srgb = 197

Astc10x10Srgb = 198

Astc12x10Srgb = 199

Astc12x12Srgb = 200

Astc4x4Ldr = 204

Astc5x4Ldr = 205

Astc5x5Ldr = 206

Astc6x5Ldr = 207

Astc6x6Ldr = 208

Astc8x5Ldr = 210

Astc8x6Ldr = 211

Astc8x8Ldr = 212

Astc10x5Ldr = 213

Astc10x6Ldr = 214

Astc10x8Ldr = 215

Astc10x10Ldr = 216

Astc12x10Ldr = 217

Astc12x12Ldr = 218

Astc4x4Hdr = 222

Astc5x4Hdr = 223

Astc5x5Hdr = 224

Astc6x5Hdr = 225

Astc6x6Hdr = 226

Astc8x5Hdr = 228

Astc8x6Hdr = 229

Astc8x8Hdr = 230

Astc10x5Hdr = 231

Astc10x6Hdr = 232

Astc10x8Hdr = 233

Astc10x10Hdr = 234

Astc12x10Hdr = 235

Asrc12x12Hdr = 236

Gbgr422 = 240

A pixel format where the red and green channels are subsampled horizontally. Two pixels are stored in 32 bits, with shared red and blue values, and unique green values.

This format is equivalent to YUY2, YUYV, yuvs, or GL_RGB_422_APPLE/GL_UNSIGNED_SHORT_8_8_REV_APPLE. The component order, from lowest addressed byte to highest, is Y0, Cb, Y1, Cr. There is no implicit colorspace conversion from YUV to RGB, the shader will receive (Cr, Y, Cb, 1). 422 textures must have a width that is a multiple of 2, and can only be used for 2D non-mipmap textures. When sampling, ClampToEdge is the only usable wrap mode.

This format is a compressed format with a block size of 2x1 in a 32-bit block.

Bgrg422 = 241

A pixel format where the red and green channels are subsampled horizontally. Two pixels are stored in 32 bits, with shared red and blue values, and unique green values.

This format is equivalent to UYVY, 2vuy, or GL_RGB_422_APPLE/GL_UNSIGNED_SHORT_8_8_APPLE. The component order, from lowest addressed byte to highest, is Cb, Y0, Cr, Y1. There is no implicit colorspace conversion from YUV to RGB, the shader will receive (Cr, Y, Cb, 1). 422 textures must have a width that is a multiple of 2, and can only be used for 2D non-mipmap textures. When sampling, ClampToEdge is the only usable wrap mode.

This format is a compressed format with a block size of 2x1 in a 32-bit block.

Depth16Unorm = 250

A pixel format with a 16-bit normalized unsigned integer component, used for a depth render target.

Depth32Float = 252

A pixel format with one 32-bit floating-point component, used for a depth render target.

Stencil8 = 253

A pixel format with an 8-bit unsigned integer component, used for a stencil render target.

Depth24UnormStencil8 = 255

A 32-bit combined depth and stencil pixel format with a 24-bit normalized unsigned integer for depth and an 8-bit unsigned integer for stencil.

Depth32FloatStencil8 = 260

A 40-bit combined depth and stencil pixel format with a 32-bit floating-point value for depth and an 8-bit unsigned integer for stencil.

When using this format, some Metal device objects allocate 64-bits per pixel.

X32Stencil8 = 261

A stencil pixel format used to read the stencil value from a texture with a combined 32-bit depth and 8-bit stencil value.

You can’t directly read the stencil value of a texture with the mtl::PixelFormat::Depth32FloatStencil8 format. To read stencil values from a texture with the mtl::PixelFormat::Depth32FloatStencil8 format, create a texture view of that texture using the mtl::PixelFormat::X32Stencil8 format, and sample the texture view instead.

X24Stencil8 = 262

A stencil pixel format used to read the stencil value from a texture with a combined 24-bit depth and 8-bit stencil value.

You can’t directly read the stencil value of a texture with the mtl::PixelFormat::Depth24UnormStencil8 format. To read stencil values from a texture with the mtl::PixelFormat::Depth24UnormStencil8 format, create a texture view of that texture using the mtl::PixelFormat::X24Stencil8 format, and sample the texture view instead.

Unspecialized = 263