cidre::nl

Struct Lang

pub struct Lang(/* private fields */);

Implementations

impl Lang

pub fn retained(&self) -> R<Self>

impl Lang

pub fn with_string(string: &String) -> &Self

pub fn undetermined() -> &'static Self

The unique identifier string for a language the Natural Language framework doesn’t recognize

pub fn amharic() -> &'static Self

The unique identifier string for the Amharic language.

pub fn arabic() -> &'static Self

The unique identifier string for the Arabic language.

pub fn armenian() -> &'static Self

The unique identifier string for the Armenian language.

pub fn bengali() -> &'static Self

The unique identifier string for the Bengali language.

pub fn bulgarian() -> &'static Self

The unique identifier string for the Bulgarian language.

pub fn burmese() -> &'static Self

The unique identifier string for the Burmese language.

pub fn catalan() -> &'static Self

The unique identifier string for the Catalan language.

pub fn cherokee() -> &'static Self

The unique identifier string for the Cherokee language.

pub fn croatian() -> &'static Self

The unique identifier string for the Croatian language.

pub fn czech() -> &'static Self

The unique identifier string for the Czech language.

pub fn danish() -> &'static Self

The unique identifier string for the Danish language.

pub fn dutch() -> &'static Self

The unique identifier string for the Dutch language.

pub fn english() -> &'static Self

The unique identifier string for the English language.

pub fn finnish() -> &'static Self

The unique identifier string for the Finnish language.

pub fn french() -> &'static Self

The unique identifier string for the French language.

pub fn georgian() -> &'static Self

The unique identifier string for the Georgian language.

pub fn german() -> &'static Self

The unique identifier string for the German language.

pub fn greek() -> &'static Self

The unique identifier string for the Greek language.

pub fn gujarati() -> &'static Self

The unique identifier string for the Gujarati language.

pub fn hebrew() -> &'static Self

The unique identifier string for the Hebrew language.

pub fn hindi() -> &'static Self

The unique identifier string for the Hindi language.

pub fn hungarian() -> &'static Self

The unique identifier string for the Hungarian language.

pub fn icelandic() -> &'static Self

The unique identifier string for the Icelandic language.

pub fn indonesian() -> &'static Self

The unique identifier string for the Indonesian language.

pub fn italian() -> &'static Self

The unique identifier string for the Italian language.

pub fn japanese() -> &'static Self

The unique identifier string for the Japanese language.

pub fn kannada() -> &'static Self

The unique identifier string for the Kannada language.

pub fn khmer() -> &'static Self

The unique identifier string for the Khmer language.

pub fn korean() -> &'static Self

The unique identifier string for the Korean language.

pub fn lao() -> &'static Self

The unique identifier string for the Lao language.

pub fn malay() -> &'static Self

The unique identifier string for the Malay language.

pub fn malayalam() -> &'static Self

The unique identifier string for the Malayalam language.

pub fn marathi() -> &'static Self

The unique identifier string for the Marathi language.

pub fn mongolian() -> &'static Self

The unique identifier string for the Mongolian language.

pub fn norwegian() -> &'static Self

The unique identifier string for the Norwegian language.

pub fn oriya() -> &'static Self

The unique identifier string for the Oriya language.

pub fn persian() -> &'static Self

The unique identifier string for the Persian language.

pub fn polish() -> &'static Self

The unique identifier string for the Polish language.

pub fn portuguese() -> &'static Self

The unique identifier string for the Portuguese language.

pub fn punjabi() -> &'static Self

The unique identifier string for the Punjabi language.

pub fn romanian() -> &'static Self

The unique identifier string for the Romanian language.

pub fn russian() -> &'static Self

The unique identifier string for the Russian language.

pub fn simplified_chinese() -> &'static Self

The unique identifier string for the Simplified Chinese language.

pub fn sinhalese() -> &'static Self

The unique identifier string for the Sinhalese language.

pub fn slovak() -> &'static Self

The unique identifier string for the Slovak language.

pub fn spanish() -> &'static Self

The unique identifier string for the Spanish language.

pub fn swedish() -> &'static Self

The unique identifier string for the Swedish language.

pub fn tamil() -> &'static Self

The unique identifier string for the Tamil language.

pub fn telugu() -> &'static Self

The unique identifier string for the Telugu language.

pub fn thai() -> &'static Self

The unique identifier string for the Thai language.

pub fn tibetan() -> &'static Self

The unique identifier string for the Tibetan language.

pub fn traditional_chinese() -> &'static Self

The unique identifier string for the Traditional Chinese language.

pub fn turkish() -> &'static Self

The unique identifier string for the Turkish language.

pub fn ukrainian() -> &'static Self

The unique identifier string for the Ukrainian language.

pub fn urdu() -> &'static Self

The unique identifier string for the Urdu language.

pub fn vietnamese() -> &'static Self

The unique identifier string for the Vietnamese language.

pub fn kazakh() -> &'static Self

The unique identifier string for the Kazakh language.