cidre

Module notify

Darwin notifications

Structs

Error

RegFlags

Flag bits used for notification registration.

Status

Token

Functions

cancel

Cancel notification delivery and free resources associated with a token.

check

Check if any notifications have been posted for a token.

is_valid_token

Determine whether a token is currently registered.

post

Post a notification for a name.

register_check

Create a token for use with Token::check.

register_dispatch

register_dispatch_block

register_file_descriptor

Request notification delivery by a write to a file descriptor.

register_mach_port

Request notification delivery by Mach message.

register_signal

Request notification delivery by UNIX signal.

resume

Resume notification delivery for a token.

set_state

Set the 64-bit state value associated with a token.

state

Get the 64-bit state value associated with a token.

suspend

Suspend notification delivery for a token.

Type Aliases

Handler

Result