cidre

Module ns

Foundation

Re-exports

pub use objc::Class;

pub use objc::Id;

pub use objc::Sel;

pub use objc::ns;

pub use objc_runtime::ComparisonResult;

pub use objc_runtime::ExceptionName;

pub use objc_runtime::NOT_FOUND;

pub use objc_runtime::class_from_ns_string;

pub use objc_runtime::protocol_from_ns_string;

pub use objc_runtime::selector_from_ns_string;

pub use objc_runtime::string_from_class;

pub use objc_runtime::string_from_protocol;

pub use objc_runtime::string_from_selector;

pub use objc_runtime::Comparator;

pub use exception::Exception;

pub use exception::UncaughtExceptionHandler;

pub use exception::set_uncaught_exception_handler;

pub use exception::try_catch;

pub use exception::try_catch_err;

pub use exception::uncaught_exception_handler;

pub use objc::ns::Integer;

pub use objc::ns::UInteger;

Modules

app_notifications

collection_view

collection_view_element_kind

color

error_user_info_keys

UserInfo Keys

exception

objc_runtime

pasteboard_name

pasteboard_reading_option_key

pasteboard_types

running_application

toolbar

toolbar_item

window_notifications

workspace

xpc

Macros

arr

Creates an ns::Array from inline arguments.

log

str

Structs

AnimationCtx

AnyCollectionViewDataSrc

AnyCollectionViewDelegate

AnyCollectionViewPrefetching

AnyGestureRecognizerDelegate

AnySplitViewDelegate

App

Array

Objective-C NSArray wrapper.

ArrayMut

Objective-C NSMutableArray wrapper.

AttrString

AttrStringFormatCtxKey

AttrStringKey

AttrStringMut

AutoresizingMaskOpts

BlockOp

Bundle

Button

CachedUrlResponse

Cell

CidreObserver

Coder

CollectionView

CollectionViewDiffableDataSrc

CollectionViewItem

CollectionViewLayout

CollectionViewLayoutAttrs

CollectionViewScrollPos

CollectionViewTransitionLayout

Color

ColorSpace

Control

ControlStateValue

Data

DataMut

DataReadOpts

DataSearchOpts

DataWriteOpts

Date

DateFormatter

Dictionary

DictionaryMut

DiffableDataSrcSnapshot

DirEnumOpts

DockTile

EdgeInsets

EnumerationOpts

Error

ErrorDomain

Event

EventButtonMask

EventMask

EventModifierFlags

EventPhase

EventSubtype

EventSwipeTrackingOpts

FeIterator

FeState

FileAttrKey

FileAttrType

FileManager

FileManagerItemReplacementOpts

FileManagerUnmountOpts

FileProtectionType

Font

FontManager

FontTraitMask

Formatter

GestureRecognizer

HttpUrlResponse

Image

ImageSymbolCfg

IndexPath

IndexSet

IndexSetMut

InlinePresentationIntent

InputStream

Iso8601DateFormatOpts

Iso8601DateFormatter

KeyedArchiver

KeyedUnarchiver

KvChangeKey

KvoOpts

LayoutPriority

LineBreakStrategy

Locale

MachPort

MatchFlags

MatchOpts

Menu

Notification

NotificationCenter

NotificationGuard

NotificationName

Null

Number

Observer

Op

OpQueue

OrderedCollectionChange

OrderedCollectionDiff

OrderedCollectionDiffCalcOpts

OutputStream

Pasteboard

PasteboardContentsOpts

Port

Predicate

ProcessInfo

Progress

Range

Regex

RegexOpts

Responder

RunLoop

RunLoopMode

RunningApp

Screen

SearchPathDomainMask

Set

SetMut

Sound

SplitView

SplitViewController

SplitViewItem

StatusBar

String

StringMut

Text

TextAttachment

TextCheckingResult

TextCheckingType

TextField

TextView

Thread

Timer

TitlebarAccessoryViewController

Toolbar

ToolbarId

ToolbarItem

ToolbarItemId

ToolbarItemVisibilityPriority

Touch

TouchPhase

TouchTypeMask

UiItemId

Url

UrlCache

UrlRequest

UrlRequestMut

UrlResKey

UrlResponse

UrlSession

UrlSessionCfg

UrlSessionDataTask

UrlSessionDownloadTask

UrlSessionStreamTask

UrlSessionTask

UrlSessionTaskPriority

UrlSessionUploadTask

UrlSessionWebSocketMessage

UrlSessionWebSocketTask

UserActivity

UserActivityPersistentId

UserDefaults

Uuid

Value

View

ViewController

VolumeEnumOpts

Window

WindowCollectionBehavior

WindowController

WindowDepth

WindowLevel

WindowOcclusionState

WindowStyleMask

Workspace

WorkspaceAuthorization

WorkspaceOpenCfg

Zone

Enums

BezelStyle

ButtonType

CellImagePos

CellType

CollectionChangeType

CollectionViewDropOp

CollectionViewItemHighlightState

CollectionViewScrollDir

ColorSysEffect

ControlSize

ControlTint

DateFormatterStyle

DecodingFailurePolicy

Describes the action an ns::Coder should take when it encounters decode failures (e.g. corrupt data) for non-TopLevel decodes.

EventGestureAxis

EventType

ExErr

Special Exception or Error type some API can still throw exception even if they return Error

GestureRecognizerState

ImageScaling

NSImageScaling

KvChange

KvSetMutationKind

LayoutAttr

LineBreakMode

NSLineBreakMode

OrderedCollectionDiffInverseOpts

PasteboardAccessBehavior

PointingDeviceType

ProcessInfoThermalState

SearchPathDirectory

SplitViewItemBehavior

SplitViewItemCollapseBehavior

StreamStatus

StringEncoding

TextFieldBezelStyle

NSTextFieldBezelStyle

TitlebarSeparatorStyle

ToolbarDisplayMode

ToolbarItemStyle

Visual styles for the toolbar item.

TouchType

UrlCacheStoragePolicy

UrlRelationship

UrlRequestAttribution

enum is used to indicate whether the user or developer specified the URL.

UrlRequestCachePolicy

UrlRequestNetworkServiceType

UrlSessionMultipathServiceType

defines constants that can be used to specify the multipath service type to associate an NSURLSession. The multipath service type determines whether multipath TCP should be attempted and the conditions for creating and switching between subflows. Using these service types requires the appropriate entitlement. Any connection attempt will fail if the process does not have the required entitlement. A primary interface is a generally less expensive interface in terms of both cost and power (such as WiFi or ethernet). A secondary interface is more expensive (such as 3G or LTE).

UrlSessionTaskState

UrlSessionWebSocketCloseCode

UrlSessionWebSocketMessageType

WindowOrderingMode

WindowTitleVisibility

WindowToolbarStyle

WorkspaceAuthorizationType

Constants

TIME_INTERVAL_SINCE_1970

Traits

AnimatablePropContainer

AppDelegate

AppDelegateImpl

CollectionViewDataSrc

CollectionViewDataSrcImpl

CollectionViewDelegate

CollectionViewDelegateImpl

CollectionViewElement

CollectionViewPrefetching

CollectionViewPrefetchingImpl

CollectionViewSectionHeaderView

Copying

CopyingImpl

CopyingMut

CopyingMutImpl

FastEnum

GestureRecognizerDelegate

GestureRecognizerDelegateImpl

KvObserverRegistration

KvObserverRegistrationImpl

KvObserving

MachPortDelegate

MachPortDelegateImpl

SplitViewDelegate

SplitViewDelegateImpl

TextFieldDelegate

TextFieldDelegateImpl

ToolbarDelegate

ToolbarDelegateImpl

ToolbarItemValidation

ToolbarItemValidationImpl

UiItemIdentification

UserActivityDelegate

UserActivityDelegateImpl

WindowDelegate

WindowDelegateImpl

Functions

full_user_name

home_dir

home_dir_for_user

home_dir_for_user_ar

if_err

if_false

if_none

log_string

search_path_for_dirs_in_domains

search_path_for_dirs_in_domains_ar

tmp_dir

user_name

Type Aliases

CollectionViewDiffableDataSrcItemProvider

CollectionViewDiffableDataSrcSupplementaryViewProvider

CollectionViewElementKind

ExResult

OsVersion

PasteboardDetectionPattern

PasteboardMetadataType

PasteboardName

PasteboardReadingOptionKey

PasteboardType

Point

Rect

Result

Size

TimeInterval