cidre::ns::collection_view

Trait CollectionViewDataSrc

pub trait CollectionViewDataSrc: Obj {
    // Provided methods
    fn collection_view_items_n_in_section(
        &mut self,
        collection_view: &mut CollectionView,
        section: Integer,
    ) -> Integer { ... }
    fn collection_view_item_for_represented_obj_at_ar(
        &mut self,
        collection_view: &mut CollectionView,
        index_path: &IndexPath,
    ) -> Rar<CollectionViewItem> { ... }
    fn collection_view_item_for_represented_obj_at(
        &mut self,
        collection_view: &mut CollectionView,
        index_path: &IndexPath,
    ) -> R<CollectionViewItem> { ... }
    fn collection_view_sections_n(
        &mut self,
        collection_view: &mut CollectionView,
    ) -> Integer { ... }
    fn sel_collection_view_sections_n() -> &'static Sel { ... }
    fn collection_view_view_for_supplementary_element_of_kind_at_index_path_ar(
        &mut self,
        collection_view: &mut CollectionView,
        kind: &CollectionViewElementKind,
        index_path: &IndexPath,
    ) -> Rar<View> { ... }
    fn collection_view_view_for_supplementary_element_of_kind_at_index_path(
        &mut self,
        collection_view: &mut CollectionView,
        kind: &CollectionViewElementKind,
        index_path: &IndexPath,
    ) -> R<View> { ... }
    fn sel_collection_view_view_for_supplementary_element_of_kind_at_index_path(    ) -> &'static Sel { ... }
}

Provided Methods

fn collection_view_items_n_in_section( &mut self, collection_view: &mut CollectionView, section: Integer, ) -> Integer

fn collection_view_item_for_represented_obj_at_ar( &mut self, collection_view: &mut CollectionView, index_path: &IndexPath, ) -> Rar<CollectionViewItem>

fn collection_view_item_for_represented_obj_at( &mut self, collection_view: &mut CollectionView, index_path: &IndexPath, ) -> R<CollectionViewItem>

fn collection_view_sections_n( &mut self, collection_view: &mut CollectionView, ) -> Integer

fn sel_collection_view_sections_n() -> &'static Sel

@selector(numberOfSectionsInCollectionView:) but dynamic use this function to check if object responds to selector

fn collection_view_view_for_supplementary_element_of_kind_at_index_path_ar( &mut self, collection_view: &mut CollectionView, kind: &CollectionViewElementKind, index_path: &IndexPath, ) -> Rar<View>

fn collection_view_view_for_supplementary_element_of_kind_at_index_path( &mut self, collection_view: &mut CollectionView, kind: &CollectionViewElementKind, index_path: &IndexPath, ) -> R<View>

fn sel_collection_view_view_for_supplementary_element_of_kind_at_index_path() -> &'static Sel

@selector(collectionView:viewForSupplementaryElementOfKind:atIndexPath:) but dynamic use this function to check if object responds to selector

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.