cidre::ns::collection_view

Trait CollectionViewDelegate

pub trait CollectionViewDelegate: Obj {

    // Provided methods
    fn collection_view_can_drag_items_at_index_paths_with_event(
        &mut self,
        collection_view: &mut CollectionView,
        index_paths: &Set<IndexPath>,
        event: &Event,
    ) -> bool { ... }
    fn sel_collection_view_can_drag_items_at_index_paths_with_event(    ) -> &'static Sel { ... }
    fn collection_view_can_drag_items_at_indexes_with_event(
        &mut self,
        collection_view: &mut CollectionView,
        indexes: &IndexSet,
        event: &Event,
    ) -> bool { ... }
    fn sel_collection_view_can_drag_items_at_indexes_with_event(    ) -> &'static Sel { ... }
    fn collection_view_write_items_at_index_paths_to_pasteboard(
        &mut self,
        collection_view: &mut CollectionView,
        index_paths: &Set<IndexPath>,
        pasteboard: &Id,
    ) -> bool { ... }
    fn sel_collection_view_write_items_at_index_paths_to_pasteboard(    ) -> &'static Sel { ... }
    fn collection_view_write_items_at_indexes_to_pasteboard(
        &mut self,
        collection_view: &mut CollectionView,
        indexes: &IndexSet,
        pasteboard: &Id,
    ) -> bool { ... }
    fn sel_collection_view_write_items_at_indexes_to_pasteboard(    ) -> &'static Sel { ... }
    fn collection_view_names_of_promised_files_dropped_at_destination_for_dragged_items_at_index_paths_ar(
        &mut self,
        collection_view: &mut CollectionView,
        drop_url: &Url,
        index_paths: &Set<IndexPath>,
    ) -> Rar<Array<String>> { ... }
    fn collection_view_names_of_promised_files_dropped_at_destination_for_dragged_items_at_index_paths(
        &mut self,
        collection_view: &mut CollectionView,
        drop_url: &Url,
        index_paths: &Set<IndexPath>,
    ) -> R<Array<String>> { ... }
    fn sel_collection_view_names_of_promised_files_dropped_at_destination_for_dragged_items_at_index_paths(    ) -> &'static Sel { ... }
    fn collection_view_names_of_promised_files_dropped_at_destination_for_dragged_items_at_indexes_ar(
        &mut self,
        collection_view: &mut CollectionView,
        drop_url: &Url,
        indexes: &IndexSet,
    ) -> Rar<Array<String>> { ... }
    fn collection_view_names_of_promised_files_dropped_at_destination_for_dragged_items_at_indexes(
        &mut self,
        collection_view: &mut CollectionView,
        drop_url: &Url,
        indexes: &IndexSet,
    ) -> R<Array<String>> { ... }
    fn sel_collection_view_names_of_promised_files_dropped_at_destination_for_dragged_items_at_indexes(    ) -> &'static Sel { ... }
    fn collection_view_dragging_image_for_items_at_index_paths_with_event_offset_ar(
        &mut self,
        collection_view: &mut CollectionView,
        index_paths: &Set<IndexPath>,
        event: &Event,
        drag_image_offset: *mut Point,
    ) -> Rar<Image> { ... }
    fn collection_view_dragging_image_for_items_at_index_paths_with_event_offset(
        &mut self,
        collection_view: &mut CollectionView,
        index_paths: &Set<IndexPath>,
        event: &Event,
        drag_image_offset: *mut Point,
    ) -> R<Image> { ... }
    fn sel_collection_view_dragging_image_for_items_at_index_paths_with_event_offset(    ) -> &'static Sel { ... }
    fn collection_view_dragging_image_for_items_at_indexes_with_event_offset_ar(
        &mut self,
        collection_view: &mut CollectionView,
        indexes: &IndexSet,
        event: &Event,
        drag_image_offset: *mut Point,
    ) -> Rar<Image> { ... }
    fn collection_view_dragging_image_for_items_at_indexes_with_event_offset(
        &mut self,
        collection_view: &mut CollectionView,
        indexes: &IndexSet,
        event: &Event,
        drag_image_offset: *mut Point,
    ) -> R<Image> { ... }
    fn sel_collection_view_dragging_image_for_items_at_indexes_with_event_offset(    ) -> &'static Sel { ... }
    fn collection_view_validate_drop_proposed_index_path_drop_operation(
        &mut self,
        collection_view: &mut CollectionView,
        dragging_info: &Id,
        proposed_drop_index_path: *mut *mut IndexPath,
        proposed_drop_operation: *mut CollectionViewDropOp,
    ) -> UInteger { ... }
    fn sel_collection_view_validate_drop_proposed_index_path_drop_operation(    ) -> &'static Sel { ... }
    fn collection_view_validate_drop_proposed_index_drop_operation(
        &mut self,
        collection_view: &mut CollectionView,
        dragging_info: &Id,
        proposed_drop_index: *mut Integer,
        proposed_drop_operation: *mut CollectionViewDropOp,
    ) -> UInteger { ... }
    fn sel_collection_view_validate_drop_proposed_index_drop_operation(    ) -> &'static Sel { ... }
    fn collection_view_accept_drop_index_path_drop_operation(
        &mut self,
        collection_view: &mut CollectionView,
        dragging_info: &Id,
        index_path: &IndexPath,
        drop_operation: CollectionViewDropOp,
    ) -> bool { ... }
    fn sel_collection_view_accept_drop_index_path_drop_operation(    ) -> &'static Sel { ... }
    fn collection_view_accept_drop_index_drop_operation(
        &mut self,
        collection_view: &mut CollectionView,
        dragging_info: &Id,
        index: Integer,
        drop_operation: CollectionViewDropOp,
    ) -> bool { ... }
    fn sel_collection_view_accept_drop_index_drop_operation() -> &'static Sel { ... }
    fn collection_view_pasteboard_writer_for_item_at_index_path_ar(
        &mut self,
        collection_view: &mut CollectionView,
        index_path: &IndexPath,
    ) -> Option<Rar<Id>> { ... }
    fn collection_view_pasteboard_writer_for_item_at_index_path(
        &mut self,
        collection_view: &mut CollectionView,
        index_path: &IndexPath,
    ) -> Option<R<Id>> { ... }
    fn sel_collection_view_pasteboard_writer_for_item_at_index_path(    ) -> &'static Sel { ... }
    fn collection_view_pasteboard_writer_for_item_at_index_ar(
        &mut self,
        collection_view: &mut CollectionView,
        index: usize,
    ) -> Option<Rar<Id>> { ... }
    fn collection_view_pasteboard_writer_for_item_at_index(
        &mut self,
        collection_view: &mut CollectionView,
        index: usize,
    ) -> Option<R<Id>> { ... }
    fn sel_collection_view_pasteboard_writer_for_item_at_index() -> &'static Sel { ... }
    fn collection_view_dragging_session_will_begin_at_point_for_items_at_index_paths(
        &mut self,
        collection_view: &mut CollectionView,
        session: &Id,
        screen_point: Point,
        index_paths: &Set<IndexPath>,
    ) { ... }
    fn sel_collection_view_dragging_session_will_begin_at_point_for_items_at_index_paths(    ) -> &'static Sel { ... }
    fn collection_view_dragging_session_will_begin_at_point_for_items_at_indexes(
        &mut self,
        collection_view: &mut CollectionView,
        session: &Id,
        screen_point: Point,
        indexes: &IndexSet,
    ) { ... }
    fn sel_collection_view_dragging_session_will_begin_at_point_for_items_at_indexes(    ) -> &'static Sel { ... }
    fn collection_view_dragging_session_ended_at_point_drag_operation(
        &mut self,
        collection_view: &mut CollectionView,
        session: &Id,
        screen_point: Point,
        operation: UInteger,
    ) { ... }
    fn sel_collection_view_dragging_session_ended_at_point_drag_operation(    ) -> &'static Sel { ... }
    fn collection_view_update_dragging_items_for_drag(
        &mut self,
        collection_view: &mut CollectionView,
        dragging_info: &Id,
    ) { ... }
    fn sel_collection_view_update_dragging_items_for_drag() -> &'static Sel { ... }
    fn collection_view_should_change_items_at_index_paths_to_highlight_state_ar(
        &mut self,
        collection_view: &mut CollectionView,
        index_paths: &Set<IndexPath>,
        highlight_state: CollectionViewItemHighlightState,
    ) -> Rar<Set<IndexPath>> { ... }
    fn collection_view_should_change_items_at_index_paths_to_highlight_state(
        &mut self,
        collection_view: &mut CollectionView,
        index_paths: &Set<IndexPath>,
        highlight_state: CollectionViewItemHighlightState,
    ) -> R<Set<IndexPath>> { ... }
    fn sel_collection_view_should_change_items_at_index_paths_to_highlight_state(    ) -> &'static Sel { ... }
    fn collection_view_did_change_items_at_index_paths_to_highlight_state(
        &mut self,
        collection_view: &mut CollectionView,
        index_paths: &Set<IndexPath>,
        highlight_state: CollectionViewItemHighlightState,
    ) { ... }
    fn sel_collection_view_did_change_items_at_index_paths_to_highlight_state(    ) -> &'static Sel { ... }
    fn collection_view_should_select_items_at_index_paths_ar(
        &mut self,
        collection_view: &mut CollectionView,
        index_paths: &Set<IndexPath>,
    ) -> Rar<Set<IndexPath>> { ... }
    fn collection_view_should_select_items_at_index_paths(
        &mut self,
        collection_view: &mut CollectionView,
        index_paths: &Set<IndexPath>,
    ) -> R<Set<IndexPath>> { ... }
    fn sel_collection_view_should_select_items_at_index_paths() -> &'static Sel { ... }
    fn collection_view_should_deselect_items_at_index_paths_ar(
        &mut self,
        collection_view: &mut CollectionView,
        index_paths: &Set<IndexPath>,
    ) -> Rar<Set<IndexPath>> { ... }
    fn collection_view_should_deselect_items_at_index_paths(
        &mut self,
        collection_view: &mut CollectionView,
        index_paths: &Set<IndexPath>,
    ) -> R<Set<IndexPath>> { ... }
    fn sel_collection_view_should_deselect_items_at_index_paths(    ) -> &'static Sel { ... }
    fn collection_view_did_select_items_at_index_paths(
        &mut self,
        collection_view: &mut CollectionView,
        index_paths: &Set<IndexPath>,
    ) { ... }
    fn sel_collection_view_did_select_items_at_index_paths() -> &'static Sel { ... }
    fn collection_view_did_deselect_items_at_index_paths(
        &mut self,
        collection_view: &mut CollectionView,
        index_paths: &Set<IndexPath>,
    ) { ... }
    fn sel_collection_view_did_deselect_items_at_index_paths() -> &'static Sel { ... }
    fn collection_view_will_display_item_for_represented_object_at_index_path(
        &mut self,
        collection_view: &mut CollectionView,
        item: &mut CollectionViewItem,
        index_path: &IndexPath,
    ) { ... }
    fn sel_collection_view_will_display_item_for_represented_object_at_index_path(    ) -> &'static Sel { ... }
    fn collection_view_will_display_supplementary_view_for_element_kind_at_index_path(
        &mut self,
        collection_view: &mut CollectionView,
        view: &mut View,
        element_kind: &CollectionViewElementKind,
        index_path: &IndexPath,
    ) { ... }
    fn sel_collection_view_will_display_supplementary_view_for_element_kind_at_index_path(    ) -> &'static Sel { ... }
    fn collection_view_did_end_displaying_item_for_represented_object_at_index_path(
        &mut self,
        collection_view: &mut CollectionView,
        item: &mut CollectionViewItem,
        index_path: &IndexPath,
    ) { ... }
    fn sel_collection_view_did_end_displaying_item_for_represented_object_at_index_path(    ) -> &'static Sel { ... }
    fn collection_view_did_end_displaying_supplementary_view_for_element_of_kind_at_index_path(
        &mut self,
        collection_view: &mut CollectionView,
        view: &mut View,
        element_kind: &CollectionViewElementKind,
        index_path: &IndexPath,
    ) { ... }
    fn sel_collection_view_did_end_displaying_supplementary_view_for_element_of_kind_at_index_path(    ) -> &'static Sel { ... }
    fn collection_view_transition_layout_for_old_layout_new_layout_ar(
        &mut self,
        collection_view: &mut CollectionView,
        from_layout: &CollectionViewLayout,
        to_layout: &CollectionViewLayout,
    ) -> Rar<CollectionViewTransitionLayout> { ... }
    fn collection_view_transition_layout_for_old_layout_new_layout(
        &mut self,
        collection_view: &mut CollectionView,
        from_layout: &CollectionViewLayout,
        to_layout: &CollectionViewLayout,
    ) -> R<CollectionViewTransitionLayout> { ... }
    fn sel_collection_view_transition_layout_for_old_layout_new_layout(    ) -> &'static Sel { ... }

}

Provided Methods

fn collection_view_can_drag_items_at_index_paths_with_event( &mut self, collection_view: &mut CollectionView, index_paths: &Set<IndexPath>, event: &Event, ) -> bool

fn sel_collection_view_can_drag_items_at_index_paths_with_event() -> &'static Sel

@selector(collectionView:canDragItemsAtIndexPaths:withEvent:) but dynamic use this function to check if object responds to selector

fn collection_view_can_drag_items_at_indexes_with_event( &mut self, collection_view: &mut CollectionView, indexes: &IndexSet, event: &Event, ) -> bool

fn sel_collection_view_can_drag_items_at_indexes_with_event() -> &'static Sel

@selector(collectionView:canDragItemsAtIndexes:withEvent:) but dynamic use this function to check if object responds to selector

fn collection_view_write_items_at_index_paths_to_pasteboard( &mut self, collection_view: &mut CollectionView, index_paths: &Set<IndexPath>, pasteboard: &Id, ) -> bool

fn sel_collection_view_write_items_at_index_paths_to_pasteboard() -> &'static Sel

@selector(collectionView:writeItemsAtIndexPaths:toPasteboard:) but dynamic use this function to check if object responds to selector

fn collection_view_write_items_at_indexes_to_pasteboard( &mut self, collection_view: &mut CollectionView, indexes: &IndexSet, pasteboard: &Id, ) -> bool

fn sel_collection_view_write_items_at_indexes_to_pasteboard() -> &'static Sel

@selector(collectionView:writeItemsAtIndexes:toPasteboard:) but dynamic use this function to check if object responds to selector

fn collection_view_names_of_promised_files_dropped_at_destination_for_dragged_items_at_index_paths_ar( &mut self, collection_view: &mut CollectionView, drop_url: &Url, index_paths: &Set<IndexPath>, ) -> Rar<Array<String>>

fn collection_view_names_of_promised_files_dropped_at_destination_for_dragged_items_at_index_paths( &mut self, collection_view: &mut CollectionView, drop_url: &Url, index_paths: &Set<IndexPath>, ) -> R<Array<String>>

fn sel_collection_view_names_of_promised_files_dropped_at_destination_for_dragged_items_at_index_paths() -> &'static Sel

@selector(collectionView:namesOfPromisedFilesDroppedAtDestination:forDraggedItemsAtIndexPaths:) but dynamic use this function to check if object responds to selector

fn collection_view_names_of_promised_files_dropped_at_destination_for_dragged_items_at_indexes_ar( &mut self, collection_view: &mut CollectionView, drop_url: &Url, indexes: &IndexSet, ) -> Rar<Array<String>>

fn collection_view_names_of_promised_files_dropped_at_destination_for_dragged_items_at_indexes( &mut self, collection_view: &mut CollectionView, drop_url: &Url, indexes: &IndexSet, ) -> R<Array<String>>

fn sel_collection_view_names_of_promised_files_dropped_at_destination_for_dragged_items_at_indexes() -> &'static Sel

@selector(collectionView:namesOfPromisedFilesDroppedAtDestination:forDraggedItemsAtIndexes:) but dynamic use this function to check if object responds to selector

fn collection_view_dragging_image_for_items_at_index_paths_with_event_offset_ar( &mut self, collection_view: &mut CollectionView, index_paths: &Set<IndexPath>, event: &Event, drag_image_offset: *mut Point, ) -> Rar<Image>

fn collection_view_dragging_image_for_items_at_index_paths_with_event_offset( &mut self, collection_view: &mut CollectionView, index_paths: &Set<IndexPath>, event: &Event, drag_image_offset: *mut Point, ) -> R<Image>

fn sel_collection_view_dragging_image_for_items_at_index_paths_with_event_offset() -> &'static Sel

@selector(collectionView:draggingImageForItemsAtIndexPaths:withEvent:offset:) but dynamic use this function to check if object responds to selector

fn collection_view_dragging_image_for_items_at_indexes_with_event_offset_ar( &mut self, collection_view: &mut CollectionView, indexes: &IndexSet, event: &Event, drag_image_offset: *mut Point, ) -> Rar<Image>

fn collection_view_dragging_image_for_items_at_indexes_with_event_offset( &mut self, collection_view: &mut CollectionView, indexes: &IndexSet, event: &Event, drag_image_offset: *mut Point, ) -> R<Image>

fn sel_collection_view_dragging_image_for_items_at_indexes_with_event_offset() -> &'static Sel

@selector(collectionView:draggingImageForItemsAtIndexes:withEvent:offset:) but dynamic use this function to check if object responds to selector

fn collection_view_validate_drop_proposed_index_path_drop_operation( &mut self, collection_view: &mut CollectionView, dragging_info: &Id, proposed_drop_index_path: *mut *mut IndexPath, proposed_drop_operation: *mut CollectionViewDropOp, ) -> UInteger

fn sel_collection_view_validate_drop_proposed_index_path_drop_operation() -> &'static Sel

@selector(collectionView:validateDrop:proposedIndexPath:dropOperation:) but dynamic use this function to check if object responds to selector

fn collection_view_validate_drop_proposed_index_drop_operation( &mut self, collection_view: &mut CollectionView, dragging_info: &Id, proposed_drop_index: *mut Integer, proposed_drop_operation: *mut CollectionViewDropOp, ) -> UInteger

fn sel_collection_view_validate_drop_proposed_index_drop_operation() -> &'static Sel

@selector(collectionView:validateDrop:proposedIndex:dropOperation:) but dynamic use this function to check if object responds to selector

fn collection_view_accept_drop_index_path_drop_operation( &mut self, collection_view: &mut CollectionView, dragging_info: &Id, index_path: &IndexPath, drop_operation: CollectionViewDropOp, ) -> bool

fn sel_collection_view_accept_drop_index_path_drop_operation() -> &'static Sel

@selector(collectionView:acceptDrop:indexPath:dropOperation:) but dynamic use this function to check if object responds to selector

fn collection_view_accept_drop_index_drop_operation( &mut self, collection_view: &mut CollectionView, dragging_info: &Id, index: Integer, drop_operation: CollectionViewDropOp, ) -> bool

fn sel_collection_view_accept_drop_index_drop_operation() -> &'static Sel

@selector(collectionView:acceptDrop:index:dropOperation:) but dynamic use this function to check if object responds to selector

fn collection_view_pasteboard_writer_for_item_at_index_path_ar( &mut self, collection_view: &mut CollectionView, index_path: &IndexPath, ) -> Option<Rar<Id>>

fn collection_view_pasteboard_writer_for_item_at_index_path( &mut self, collection_view: &mut CollectionView, index_path: &IndexPath, ) -> Option<R<Id>>

fn sel_collection_view_pasteboard_writer_for_item_at_index_path() -> &'static Sel

@selector(collectionView:pasteboardWriterForItemAtIndexPath:) but dynamic use this function to check if object responds to selector

fn collection_view_pasteboard_writer_for_item_at_index_ar( &mut self, collection_view: &mut CollectionView, index: usize, ) -> Option<Rar<Id>>

fn collection_view_pasteboard_writer_for_item_at_index( &mut self, collection_view: &mut CollectionView, index: usize, ) -> Option<R<Id>>

fn sel_collection_view_pasteboard_writer_for_item_at_index() -> &'static Sel

@selector(collectionView:pasteboardWriterForItemAtIndex:) but dynamic use this function to check if object responds to selector

fn collection_view_dragging_session_will_begin_at_point_for_items_at_index_paths( &mut self, collection_view: &mut CollectionView, session: &Id, screen_point: Point, index_paths: &Set<IndexPath>, )

fn sel_collection_view_dragging_session_will_begin_at_point_for_items_at_index_paths() -> &'static Sel

@selector(collectionView:draggingSession:willBeginAtPoint:forItemsAtIndexPaths:) but dynamic use this function to check if object responds to selector

fn collection_view_dragging_session_will_begin_at_point_for_items_at_indexes( &mut self, collection_view: &mut CollectionView, session: &Id, screen_point: Point, indexes: &IndexSet, )

fn sel_collection_view_dragging_session_will_begin_at_point_for_items_at_indexes() -> &'static Sel

@selector(collectionView:draggingSession:willBeginAtPoint:forItemsAtIndexes:) but dynamic use this function to check if object responds to selector

fn collection_view_dragging_session_ended_at_point_drag_operation( &mut self, collection_view: &mut CollectionView, session: &Id, screen_point: Point, operation: UInteger, )

fn sel_collection_view_dragging_session_ended_at_point_drag_operation() -> &'static Sel

@selector(collectionView:draggingSession:endedAtPoint:dragOperation:) but dynamic use this function to check if object responds to selector

fn collection_view_update_dragging_items_for_drag( &mut self, collection_view: &mut CollectionView, dragging_info: &Id, )

fn sel_collection_view_update_dragging_items_for_drag() -> &'static Sel

@selector(collectionView:updateDraggingItemsForDrag:) but dynamic use this function to check if object responds to selector

fn collection_view_should_change_items_at_index_paths_to_highlight_state_ar( &mut self, collection_view: &mut CollectionView, index_paths: &Set<IndexPath>, highlight_state: CollectionViewItemHighlightState, ) -> Rar<Set<IndexPath>>

fn collection_view_should_change_items_at_index_paths_to_highlight_state( &mut self, collection_view: &mut CollectionView, index_paths: &Set<IndexPath>, highlight_state: CollectionViewItemHighlightState, ) -> R<Set<IndexPath>>

fn sel_collection_view_should_change_items_at_index_paths_to_highlight_state() -> &'static Sel

@selector(collectionView:shouldChangeItemsAtIndexPaths:toHighlightState:) but dynamic use this function to check if object responds to selector

fn collection_view_did_change_items_at_index_paths_to_highlight_state( &mut self, collection_view: &mut CollectionView, index_paths: &Set<IndexPath>, highlight_state: CollectionViewItemHighlightState, )

fn sel_collection_view_did_change_items_at_index_paths_to_highlight_state() -> &'static Sel

@selector(collectionView:didChangeItemsAtIndexPaths:toHighlightState:) but dynamic use this function to check if object responds to selector

fn collection_view_should_select_items_at_index_paths_ar( &mut self, collection_view: &mut CollectionView, index_paths: &Set<IndexPath>, ) -> Rar<Set<IndexPath>>

fn collection_view_should_select_items_at_index_paths( &mut self, collection_view: &mut CollectionView, index_paths: &Set<IndexPath>, ) -> R<Set<IndexPath>>

fn sel_collection_view_should_select_items_at_index_paths() -> &'static Sel

@selector(collectionView:shouldSelectItemsAtIndexPaths:) but dynamic use this function to check if object responds to selector

fn collection_view_should_deselect_items_at_index_paths_ar( &mut self, collection_view: &mut CollectionView, index_paths: &Set<IndexPath>, ) -> Rar<Set<IndexPath>>

fn collection_view_should_deselect_items_at_index_paths( &mut self, collection_view: &mut CollectionView, index_paths: &Set<IndexPath>, ) -> R<Set<IndexPath>>

fn sel_collection_view_should_deselect_items_at_index_paths() -> &'static Sel

@selector(collectionView:shouldDeselectItemsAtIndexPaths:) but dynamic use this function to check if object responds to selector

fn collection_view_did_select_items_at_index_paths( &mut self, collection_view: &mut CollectionView, index_paths: &Set<IndexPath>, )

fn sel_collection_view_did_select_items_at_index_paths() -> &'static Sel

@selector(collectionView:didSelectItemsAtIndexPaths:) but dynamic use this function to check if object responds to selector

fn collection_view_did_deselect_items_at_index_paths( &mut self, collection_view: &mut CollectionView, index_paths: &Set<IndexPath>, )

fn sel_collection_view_did_deselect_items_at_index_paths() -> &'static Sel

@selector(collectionView:didDeselectItemsAtIndexPaths:) but dynamic use this function to check if object responds to selector

fn collection_view_will_display_item_for_represented_object_at_index_path( &mut self, collection_view: &mut CollectionView, item: &mut CollectionViewItem, index_path: &IndexPath, )

fn sel_collection_view_will_display_item_for_represented_object_at_index_path() -> &'static Sel

@selector(collectionView:willDisplayItem:forRepresentedObjectAtIndexPath:) but dynamic use this function to check if object responds to selector

fn collection_view_will_display_supplementary_view_for_element_kind_at_index_path( &mut self, collection_view: &mut CollectionView, view: &mut View, element_kind: &CollectionViewElementKind, index_path: &IndexPath, )

fn sel_collection_view_will_display_supplementary_view_for_element_kind_at_index_path() -> &'static Sel

@selector(collectionView:willDisplaySupplementaryView:forElementKind:atIndexPath:) but dynamic use this function to check if object responds to selector

fn collection_view_did_end_displaying_item_for_represented_object_at_index_path( &mut self, collection_view: &mut CollectionView, item: &mut CollectionViewItem, index_path: &IndexPath, )

fn sel_collection_view_did_end_displaying_item_for_represented_object_at_index_path() -> &'static Sel

@selector(collectionView:didEndDisplayingItem:forRepresentedObjectAtIndexPath:) but dynamic use this function to check if object responds to selector

fn collection_view_did_end_displaying_supplementary_view_for_element_of_kind_at_index_path( &mut self, collection_view: &mut CollectionView, view: &mut View, element_kind: &CollectionViewElementKind, index_path: &IndexPath, )

fn sel_collection_view_did_end_displaying_supplementary_view_for_element_of_kind_at_index_path() -> &'static Sel

@selector(collectionView:didEndDisplayingSupplementaryView:forElementOfKind:atIndexPath:) but dynamic use this function to check if object responds to selector

fn collection_view_transition_layout_for_old_layout_new_layout_ar( &mut self, collection_view: &mut CollectionView, from_layout: &CollectionViewLayout, to_layout: &CollectionViewLayout, ) -> Rar<CollectionViewTransitionLayout>

fn collection_view_transition_layout_for_old_layout_new_layout( &mut self, collection_view: &mut CollectionView, from_layout: &CollectionViewLayout, to_layout: &CollectionViewLayout, ) -> R<CollectionViewTransitionLayout>

fn sel_collection_view_transition_layout_for_old_layout_new_layout() -> &'static Sel

@selector(collectionView:transitionLayoutForOldLayout:newLayout:) but dynamic use this function to check if object responds to selector

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.