cidre::ns::collection_view

Trait CollectionViewElement

pub trait CollectionViewElement: Obj {
    // Provided methods
    fn prepare_for_reuse(&mut self) { ... }
    fn sel_prepare_for_reuse() -> &'static Sel { ... }
    fn apply_layout_attributes(
        &mut self,
        layout_attributes: &CollectionViewLayoutAttrs,
    ) { ... }
    fn sel_apply_layout_attributes() -> &'static Sel { ... }
    fn will_transition_from_layout(
        &mut self,
        old_layout: &CollectionViewLayout,
        new_layout: &CollectionViewLayout,
    ) { ... }
    fn sel_will_transition_from_layout() -> &'static Sel { ... }
    fn did_transition_from_layout(
        &mut self,
        old_layout: &CollectionViewLayout,
        new_layout: &CollectionViewLayout,
    ) { ... }
    fn sel_did_transition_from_layout() -> &'static Sel { ... }
    fn preferred_layout_attributes_fitting_attributes_ar(
        &mut self,
        layout_attributes: &CollectionViewLayoutAttrs,
    ) -> Rar<CollectionViewLayoutAttrs> { ... }
    fn preferred_layout_attributes_fitting_attributes(
        &mut self,
        layout_attributes: &CollectionViewLayoutAttrs,
    ) -> R<CollectionViewLayoutAttrs> { ... }
    fn sel_preferred_layout_attributes_fitting_attributes() -> &'static Sel { ... }
}

Provided Methods

fn prepare_for_reuse(&mut self)

fn sel_prepare_for_reuse() -> &'static Sel

@selector(prepareForReuse) but dynamic use this function to check if object responds to selector

fn apply_layout_attributes( &mut self, layout_attributes: &CollectionViewLayoutAttrs, )

fn sel_apply_layout_attributes() -> &'static Sel

@selector(applyLayoutAttributes:) but dynamic use this function to check if object responds to selector

fn will_transition_from_layout( &mut self, old_layout: &CollectionViewLayout, new_layout: &CollectionViewLayout, )

fn sel_will_transition_from_layout() -> &'static Sel

@selector(willTransitionFromLayout:toLayout:) but dynamic use this function to check if object responds to selector

fn did_transition_from_layout( &mut self, old_layout: &CollectionViewLayout, new_layout: &CollectionViewLayout, )

fn sel_did_transition_from_layout() -> &'static Sel

@selector(didTransitionFromLayout:toLayout:) but dynamic use this function to check if object responds to selector

fn preferred_layout_attributes_fitting_attributes_ar( &mut self, layout_attributes: &CollectionViewLayoutAttrs, ) -> Rar<CollectionViewLayoutAttrs>

fn preferred_layout_attributes_fitting_attributes( &mut self, layout_attributes: &CollectionViewLayoutAttrs, ) -> R<CollectionViewLayoutAttrs>

fn sel_preferred_layout_attributes_fitting_attributes() -> &'static Sel

@selector(preferredLayoutAttributesFittingAttributes:) but dynamic use this function to check if object responds to selector

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.