cidre::ns::color

Struct Color

pub struct Color(/* private fields */);

Implementations

impl Color

pub fn retained(&self) -> R<Self>

impl Color

pub fn cls() -> &'static Class<Self>

pub fn cls_ptr() -> *const c_void

pub fn alloc() -> A<Self>

pub fn new() -> R<Color>

shortcut to Self::alloc().init()

impl Color

pub fn with_white_alpha_ar(white: Float, alpha: Float) -> Rar<Self>

pub fn with_white_alpha(white: Float, alpha: Float) -> R<Self>

pub fn with_rgba_ar(r: Float, g: Float, b: Float, a: Float) -> Rar<Self>

pub fn with_rgba(r: Float, g: Float, b: Float, a: Float) -> R<Self>

pub fn with_hsba_ar(h: Float, s: Float, b: Float, a: Float) -> Rar<Self>

pub fn with_hsba(h: Float, s: Float, b: Float, a: Float) -> R<Self>

pub fn color_named_ar(name: &String) -> Option<Rar<Self>>

pub fn color_named(name: &String) -> Option<R<Self>>

pub fn named(name: impl AsRef<String>) -> Option<R<Self>>

pub unsafe fn white_component_throws(&self) -> Float

pub fn white_component<'ar>(&self) -> Result<Float, &'ar Exception>

pub fn alpha_component(&self) -> Float

pub unsafe fn red_component_throws(&self) -> Float

pub unsafe fn green_component_throws(&self) -> Float

pub unsafe fn blue_component_throws(&self) -> Float

pub fn red_component<'ar>(&self) -> Result<Float, &'ar Exception>

pub fn green_component<'ar>(&self) -> Result<Float, &'ar Exception>

pub fn blue_component<'ar>(&self) -> Result<Float, &'ar Exception>

pub unsafe fn hue_component_throws(&self) -> Float

pub unsafe fn saturation_component_throws(&self) -> Float

pub unsafe fn brightness_component_throws(&self) -> Float

pub fn hue_component<'ar>(&self) -> Result<Float, &'ar Exception>

pub fn saturation_component<'ar>(&self) -> Result<Float, &'ar Exception>

pub fn brightness_component<'ar>(&self) -> Result<Float, &'ar Exception>

pub unsafe fn color_space_throws_ar(&self) -> Rar<ColorSpace>

pub unsafe fn color_space_throws(&self) -> R<ColorSpace>

pub fn color_space<'ear>(&self) -> ExResult<'ear, R<ColorSpace>>

pub fn sel_with_exposure() -> &'static Sel

@selector(colorWithRed:green:blue:alpha:exposure:) but dynamic use this function to check if object responds to selector

pub unsafe fn with_exposure_ar( r: Float, g: Float, b: Float, a: Float, exposure: Float, ) -> Rar<Self>

pub unsafe fn with_exposure( r: Float, g: Float, b: Float, a: Float, exposure: Float, ) -> R<Self>

Check availability with selector Self::sel_with_exposure()

pub fn sel_with_linear_exposure() -> &'static Sel

@selector(colorWithRed:green:blue:alpha:linearExposure:) but dynamic use this function to check if object responds to selector

pub unsafe fn with_linear_exposure_ar( r: Float, g: Float, b: Float, a: Float, exposure: Float, ) -> Rar<Self>

pub unsafe fn with_linear_exposure( r: Float, g: Float, b: Float, a: Float, exposure: Float, ) -> R<Self>

Check availability with selector Self::sel_with_linear_exposure()

pub fn sel_applying_content_headroom() -> &'static Sel

@selector(colorByApplyingContentHeadroom:) but dynamic use this function to check if object responds to selector

pub unsafe fn applying_content_headroom_ar( &self, content_headroom: Float, ) -> Rar<Self>

Reinterpret the color by applying a new content_headroom without changing the color components. Changing the content_headroom redefines the color relative to a different peak white, changing its behavior under tone mapping and the result of calling sdr. The new color will have a content_headroom >= 1.0.

pub unsafe fn applying_content_headroom( &self, content_headroom: Float, ) -> R<Self>

Check availability with selector Self::sel_applying_content_headroom() Reinterpret the color by applying a new content_headroom without changing the color components. Changing the content_headroom redefines the color relative to a different peak white, changing its behavior under tone mapping and the result of calling sdr. The new color will have a content_headroom >= 1.0.

pub fn sel_linear_exposure() -> &'static Sel

@selector(linearExposure) but dynamic use this function to check if object responds to selector

pub unsafe fn linear_exposure(&self) -> Float

The linear brightness multiplier that was applied when generating this color. Colors created with an exposure by ui::Color create cg::Colors that are tagged with a contentHeadroom value. While cg::Colors created without a contentHeadroom tag will return 0 from cg::Color::content_headroom, ui::Colors generated in a similar fashion return a linear_exposure of 1.0.

pub fn sel_sdr() -> &'static Sel

@selector(standardDynamicRangeColor) but dynamic use this function to check if object responds to selector

pub unsafe fn sdr_ar(&self) -> Rar<Self>

pub unsafe fn sdr(&self) -> R<Self>

Check availability with selector Self::sel_sdr()

pub fn cg_color(&self) -> Option<&Color>

impl Color

pub fn black_ar() -> Rar<Self>

pub fn black() -> R<Self>

pub fn dark_gray_ar() -> Rar<Self>

pub fn dark_gray() -> R<Self>

pub fn light_gray_ar() -> Rar<Self>

pub fn light_gray() -> R<Self>

pub fn white_ar() -> Rar<Self>

pub fn white() -> R<Self>

pub fn gray_ar() -> Rar<Self>

pub fn gray() -> R<Self>

pub fn red_ar() -> Rar<Self>

pub fn red() -> R<Self>

pub fn green_ar() -> Rar<Self>

pub fn green() -> R<Self>

pub fn blue_ar() -> Rar<Self>

pub fn blue() -> R<Self>

pub fn cyan_ar() -> Rar<Self>

pub fn cyan() -> R<Self>

pub fn yellow_ar() -> Rar<Self>

pub fn yellow() -> R<Self>

pub fn magenta_ar() -> Rar<Self>

pub fn magenta() -> R<Self>

pub fn orange_ar() -> Rar<Self>

pub fn orange() -> R<Self>

pub fn purple_ar() -> Rar<Self>

pub fn purple() -> R<Self>

pub fn brown_ar() -> Rar<Self>

pub fn brown() -> R<Self>

pub fn clear_ar() -> Rar<Self>

pub fn clear() -> R<Self>

impl Color

Foreground Colors

pub fn label_ar() -> Rar<Self>

Foreground color for static text and related elements

pub fn label() -> R<Self>

Foreground color for static text and related elements

pub fn secondary_label_ar() -> Rar<Self>

Foreground color for secondary static text and related elements

pub fn secondary_label() -> R<Self>

Foreground color for secondary static text and related elements

pub fn tertiary_label_ar() -> Rar<Self>

Foreground color for disabled static text and related elements

pub fn tertiary_label() -> R<Self>

Foreground color for disabled static text and related elements

pub fn quaternary_label_ar() -> Rar<Self>

Foreground color for large secondary or disabled static text, separators, large glyphs/icons, etc

pub fn quaternary_label() -> R<Self>

Foreground color for large secondary or disabled static text, separators, large glyphs/icons, etc

pub fn quinary_label_ar() -> Rar<Self>

Used for large scale images or subtle decorative elements; not for general foreground content.

pub fn quinary_label() -> R<Self>

Used for large scale images or subtle decorative elements; not for general foreground content.

pub fn link_ar() -> Rar<Self>

Foreground color for standard system links

pub fn link() -> R<Self>

Foreground color for standard system links

pub fn placeholder_text_ar() -> Rar<Self>

Foreground color for placeholder text in controls or text views

pub fn placeholder_text() -> R<Self>

Foreground color for placeholder text in controls or text views

pub fn window_frame_text_ar() -> Rar<Self>

Foreground color on window chrome

pub fn window_frame_text() -> R<Self>

Foreground color on window chrome

pub fn selected_menu_item_text_ar() -> Rar<Self>

Foreground color inside selected menu items. Equivalent to +labelColor in a NSBackgroundStyleEmphasized context.

pub fn selected_menu_item_text() -> R<Self>

Foreground color inside selected menu items. Equivalent to +labelColor in a NSBackgroundStyleEmphasized context.

pub fn alternate_selected_control_text_ar() -> Rar<Self>

Foreground color inside emphasized and selected content: table views rows, collection views, etc. Equivalent to +labelColor in a ns::BackgroundStyleEmphasized context.

pub fn alternate_selected_control_text() -> R<Self>

Foreground color inside emphasized and selected content: table views rows, collection views, etc. Equivalent to +labelColor in a ns::BackgroundStyleEmphasized context.

pub fn header_text_ar() -> Rar<Self>

Foreground color for header cells in Table/OutlineView

pub fn header_text() -> R<Self>

Foreground color for header cells in Table/OutlineView

pub fn separator_ar() -> Rar<Self>

Foreground used for separators between sections of content. Not appropriate for use as split view or window chrome dividers.

pub fn separator() -> R<Self>

Foreground used for separators between sections of content. Not appropriate for use as split view or window chrome dividers.

pub fn grid_ar() -> Rar<Self>

Grids in controls

pub fn grid() -> R<Self>

Grids in controls

impl Color

Background Colors

pub fn window_bg_ar() -> Rar<Self>

Background color for windows. This should not be used for drawing, and NSVisualEffectMaterialWindowBackground should be used instead.

pub fn window_bg() -> R<Self>

Background color for windows. This should not be used for drawing, and NSVisualEffectMaterialWindowBackground should be used instead.

pub fn under_page_bg_ar() -> Rar<Self>

Background areas revealed behind documents. This should not be used for drawing, and NSVisualEffectMaterialUnderPageBackground should be used instead.

pub fn under_page_bg() -> R<Self>

Background areas revealed behind documents. This should not be used for drawing, and NSVisualEffectMaterialUnderPageBackground should be used instead.

pub fn control_bg_ar() -> Rar<Self>

Background for content areas: scroll views, table views, collection views. This should not be used for drawing, and NSVisualEffectMaterialContentBackground should be used instead.

pub fn control_bg() -> R<Self>

Background for content areas: scroll views, table views, collection views. This should not be used for drawing, and NSVisualEffectMaterialContentBackground should be used instead.

pub fn selected_content_bg_ar() -> Rar<Self>

The background color of selected and emphasized (focused) content: table views rows, collection views, etc. Alias for +alternateSelectedControlColor

pub fn selected_content_bg() -> R<Self>

The background color of selected and emphasized (focused) content: table views rows, collection views, etc. Alias for +alternateSelectedControlColor

pub fn unemphasized_selected_content_bg_ar() -> Rar<Self>

The background color of selected and unemphasized content: table views rows, collection views, etc. Alias for +secondarySelectedControlColor

pub fn unemphasized_selected_content_bg() -> R<Self>

The background color of selected and unemphasized content: table views rows, collection views, etc. Alias for +secondarySelectedControlColor

pub fn alternating_content_bg_colors_ar() -> Rar<Array<Self>>

The background colors for alternating content items: such as table view rows, collection view items. Alias for +controlAlternatingRowBackgroundColors

pub fn alternating_content_bg_colors() -> R<Array<Self>>

The background colors for alternating content items: such as table view rows, collection view items. Alias for +controlAlternatingRowBackgroundColors

impl Color

Text Colors

pub fn text_ar() -> Rar<Self>

Document text

pub fn text() -> R<Self>

Document text

pub fn text_bg_ar() -> Rar<Self>

Document text background

pub fn text_bg() -> R<Self>

Document text background

pub fn text_insertion_point_ar() -> Rar<Self>

pub fn text_insertion_point() -> R<Self>

pub fn selected_text_ar() -> Rar<Self>

Selected document text

pub fn selected_text() -> R<Self>

Selected document text

pub fn selected_text_bg_ar() -> Rar<Self>

Selected document text background

pub fn selected_text_bg() -> R<Self>

Selected document text background

pub fn unemphasized_selected_text_bg_ar() -> Rar<Self>

The background for unemphasized text selection (e.g. when the associated control/view/window is not key)

pub fn unemphasized_selected_text_bg() -> R<Self>

The background for unemphasized text selection (e.g. when the associated control/view/window is not key)

pub fn unemphasized_selected_text_ar() -> Rar<Self>

The foreground for unemphasized text selection (e.g. when the associated control/view/window is not key)

pub fn unemphasized_selected_text() -> R<Self>

The foreground for unemphasized text selection (e.g. when the associated control/view/window is not key)

impl Color

Control Colors

pub fn control_ar() -> Rar<Self>

Approximate background for the color of control faces (such as buttons).

pub fn control() -> R<Self>

Approximate background for the color of control faces (such as buttons).

pub fn control_text_ar() -> Rar<Self>

Text on controls

pub fn control_text() -> R<Self>

Text on controls

pub fn selected_control_ar() -> Rar<Self>

Control face for selected controls

pub fn selected_control() -> R<Self>

Control face for selected controls

pub fn selected_control_text_ar() -> Rar<Self>

Text on selected controls

pub fn selected_control_text() -> R<Self>

Text on selected controls

pub fn disabled_control_text_ar() -> Rar<Self>

Text on disabled controls

pub fn disabled_control_text() -> R<Self>

Text on disabled controls

pub fn keyboard_focus_indicator_ar() -> Rar<Self>

Keyboard focus ring around controls

pub fn keyboard_focus_indicator() -> R<Self>

Keyboard focus ring around controls

pub fn scrubber_textured_bg_ar() -> Rar<Self>

Patterned background color for use in NSScrubber

pub fn scrubber_textured_bg() -> R<Self>

Patterned background color for use in NSScrubber

impl Color

pub fn sys_red_ar() -> Rar<Self>

pub fn sys_red() -> R<Self>

pub fn sys_green_ar() -> Rar<Self>

pub fn sys_green() -> R<Self>

pub fn sys_blue_ar() -> Rar<Self>

pub fn sys_blue() -> R<Self>

pub fn sys_orange_ar() -> Rar<Self>

pub fn sys_orange() -> R<Self>

pub fn sys_yellow_ar() -> Rar<Self>

pub fn sys_yellow() -> R<Self>

pub fn sys_brown_ar() -> Rar<Self>

pub fn sys_brown() -> R<Self>

pub fn sys_pink_ar() -> Rar<Self>

pub fn sys_pink() -> R<Self>

pub fn sys_purple_ar() -> Rar<Self>

pub fn sys_purple() -> R<Self>

pub fn sys_gray_ar() -> Rar<Self>

pub fn sys_gray() -> R<Self>

pub fn sys_teal_ar() -> Rar<Self>

pub fn sys_teal() -> R<Self>

pub fn sys_indigo_ar() -> Rar<Self>

pub fn sys_indigo() -> R<Self>

pub fn sys_mint_ar() -> Rar<Self>

pub fn sys_mint() -> R<Self>

pub fn sys_cyan_ar() -> Rar<Self>

pub fn sys_cyan() -> R<Self>

impl Color

Fill colors for UI elements.

These are meant to be used over the background colors, since their alpha component is less than 1.

pub fn sys_fill_ar() -> Rar<Self>

systemFillColor is appropriate for filling thin shapes, such as the track of a slider.

pub fn sys_fill() -> R<Self>

systemFillColor is appropriate for filling thin shapes, such as the track of a slider.

pub fn sys_secondary_fill_ar() -> Rar<Self>

secondarySystemFillColor is appropriate for filling small-size shapes, such as the backing of a progress indicator.

pub fn sys_secondary_fill() -> R<Self>

secondarySystemFillColor is appropriate for filling small-size shapes, such as the backing of a progress indicator.

pub fn sys_tertiary_fill_ar() -> Rar<Self>

tertiarySystemFillColor is appropriate for filling medium-size shapes, such as the backing of a switch.

pub fn sys_tertiary_fill() -> R<Self>

tertiarySystemFillColor is appropriate for filling medium-size shapes, such as the backing of a switch.

pub fn sys_quaternary_fill_ar() -> Rar<Self>

quaternarySystemFillColor is appropriate for filling large areas, such as a group box or tab pane.

pub fn sys_quaternary_fill() -> R<Self>

quaternarySystemFillColor is appropriate for filling large areas, such as a group box or tab pane.

pub fn sys_quinary_fill_ar() -> Rar<Self>

quinarySystemFillColor is appropriate for filling large areas that require subtle emphasis, such as content of a form..

pub fn sys_quinary_fill() -> R<Self>

quinarySystemFillColor is appropriate for filling large areas that require subtle emphasis, such as content of a form..

pub fn control_accent_ar() -> Rar<Self>

A dynamic color that reflects the user’s current preferred accent color. This color automatically updates when the accent color preference changes. Do not make assumptions about the color space of this color, which may change across releases.

pub fn control_accent() -> R<Self>

A dynamic color that reflects the user’s current preferred accent color. This color automatically updates when the accent color preference changes. Do not make assumptions about the color space of this color, which may change across releases.

pub fn current_control_tint() -> ControlTint

The current system control tint.

pub fn highlight_ar() -> Rar<Self>

Highlight color for UI elements (this is abstract and defines the color all highlights tend toward)

pub fn highlight() -> R<Self>

Highlight color for UI elements (this is abstract and defines the color all highlights tend toward)

pub fn shadow_ar() -> Rar<Self>

Shadow color for UI elements (this is abstract and defines the color all shadows tend toward)

pub fn shadow() -> R<Self>

Shadow color for UI elements (this is abstract and defines the color all shadows tend toward)

pub fn highlight_with_level_ar(&self, val: Float) -> Option<Rar<Self>>

val = 0 => receiver, val = 1 => highlightColor

pub fn highlight_with_level(&self, val: Float) -> Option<R<Self>>

val = 0 => receiver, val = 1 => highlightColor

pub fn shadow_with_level_ar(&self, val: Float) -> Option<Rar<Self>>

val = 0 => receiver, val = 1 => shadowColor

pub fn shadow_with_level(&self, val: Float) -> Option<R<Self>>

val = 0 => receiver, val = 1 => shadowColor

pub fn color_with_sys_effect_ar(&self, sys_effect: ColorSysEffect) -> Rar<Self>

Returns a color representing the base color with a system defined effect applied to it. This color is safe to create before draw time, as the resolution of the final color only happens when being -set, retrieving its CGColor, resolving with -colorWithType:, etc. The return color type is .named.

pub fn color_with_sys_effect(&self, sys_effect: ColorSysEffect) -> R<Self>

Returns a color representing the base color with a system defined effect applied to it. This color is safe to create before draw time, as the resolution of the final color only happens when being -set, retrieving its CGColor, resolving with -colorWithType:, etc. The return color type is .named.

impl Color

pub fn set(&self)

Set the color: Sets the fill and stroke colors in the current drawing context. If the color doesn’t know about alpha, it’s set to 1.0. Should be implemented by subclassers.

pub fn set_fill(&self)

Set the fill or stroke colors individually. These should be implemented by subclassers.

pub fn set_stroke(&self)

Set the fill or stroke colors individually. These should be implemented by subclassers.

impl Color

pub fn blended_color_with_fraction_ar( &self, fraction: Float, of_color: &Self, ) -> Option<Rar<Self>>

Blend using the NSCalibratedRGB color space. Both colors are converted into the calibrated RGB color space, and they are blended by taking fraction of color and 1 - fraction of the receiver. The result is in the calibrated RGB color space. If the colors cannot be converted into the calibrated RGB color space the blending fails and nil is returned.

pub fn blended_color_with_fraction( &self, fraction: Float, of_color: &Self, ) -> Option<R<Self>>

Blend using the NSCalibratedRGB color space. Both colors are converted into the calibrated RGB color space, and they are blended by taking fraction of color and 1 - fraction of the receiver. The result is in the calibrated RGB color space. If the colors cannot be converted into the calibrated RGB color space the blending fails and nil is returned.