cidre::ns

Enum EventType

#[repr(usize)]
pub enum EventType {

    LeftMouseDown = 1,
    LeftMouseUp = 2,
    RightMouseDown = 3,
    RightMouseUp = 4,
    MouseMoved = 5,
    LeftMouseDragged = 6,
    RightMouseDragged = 7,
    MouseEntered = 8,
    MouseExited = 9,
    KeyDown = 10,
    KeyUp = 11,
    FlagsChanged = 12,
    AppKitDefined = 13,
    SysDefined = 14,
    AppDefined = 15,
    Periodic = 16,
    CursorUpdate = 17,
    ScrollWheel = 22,
    TabletPoint = 23,
    TabletProximity = 24,
    OtherMouseDown = 25,
    OtherMouseUp = 26,
    OtherMouseDragged = 27,
    Gesture = 29,
    Magnify = 30,
    Swipe = 31,
    Rotate = 18,
    BeginGesture = 19,
    EndGesture = 20,
    SmartMagnify = 32,
    QuickLook = 33,
    Pressure = 34,
    DirectTouch = 37,
    ChangeMode = 38,

}

Variants

LeftMouseDown = 1

The user pressed the left mouse button

LeftMouseUp = 2

The user released the left mouse button.

RightMouseDown = 3

The user pressed the right mouse button.

RightMouseUp = 4

The user released the right mouse button.

MouseMoved = 5

The user moved the mouse in a way that caused the cursor to move onscreen.

LeftMouseDragged = 6

The user moved the mouse while holding down the left mouse button.

RightMouseDragged = 7

The user moved the mouse while holding down the right mouse button.

MouseEntered = 8

The cursor entered a well-defined area, such as a view.

MouseExited = 9

The cursor exited a well-defined area, such as a view.

KeyDown = 10

The user pressed a key on the keyboard.

KeyUp = 11

The user released a key on the keyboard.

FlagsChanged = 12

The event flags changed.

AppKitDefined = 13

An AppKit-related event occurred.

SysDefined = 14

A system-related event occurred.

AppDefined = 15

An app-defined event occurred.

Periodic = 16

An event that provides execution time to periodic tasks.

CursorUpdate = 17

An event that updates the cursor.

ScrollWheel = 22

The scroll wheel position changed.

TabletPoint = 23

The user touched a point on a tablet.

Tablets generate tablet-point events between a mouse-down event and the first mouse drag event

TabletProximity = 24

A pointing device is near, but not touching, the associated tablet.

OtherMouseDown = 25

The user pressed a tertiary mouse button.

OtherMouseUp = 26

The user released a tertiary mouse button.

OtherMouseDragged = 27

The user moved the mouse while holding down a tertiary mouse button.

Gesture = 29

The user performed a nonspecific type of gesture.

Magnify = 30

The user performed a pinch-open or pinch-close gesture.

Swipe = 31

The user performed a swipe gesture.

Rotate = 18

The user performed a rotate gesture.

BeginGesture = 19

An event marking the beginning of a gesture.

EndGesture = 20

An event that marks the end of a gesture.

SmartMagnify = 32

The user performed a smart-zoom gesture.

QuickLook = 33

An event that initiates a Quick Look request.

Pressure = 34

An event that reports a change in pressure on a pressure-sensitive device.

DirectTouch = 37

The user touched a portion of the touch bar.

ChangeMode = 38

The user changed the mode of a connected device.

Implementations

impl EventType

pub const fn as_mask(self) -> EventMask