cidre::ns

Struct ProcessInfo

pub struct ProcessInfo(/* private fields */);

Implementations

impl ProcessInfo

pub fn retained(&self) -> R<Self>

impl ProcessInfo

pub fn cls() -> &'static Class<Self>

pub fn cls_ptr() -> *const c_void

pub fn alloc() -> A<Self>

pub fn current() -> &'static mut ProcessInfo

use cidre::ns;

let pi = ns::ProcessInfo::current();

assert_ne!(pi.thermal_state(), ns::ProcessInfoThermalState::Critical);
assert_eq!(pi.is_low_power_mode_enabled(), false);
assert!(pi.processor_count() > 1);
assert!(pi.active_processor_count() > 1);

pub fn env_ar(&self) -> Rar<Dictionary<String, String>>

pub fn env(&self) -> R<Dictionary<String, String>>

pub fn args_ar(&self) -> Rar<Array<String>>

pub fn args(&self) -> R<Array<String>>

pub fn host_name_ar(&self) -> Rar<String>

pub fn host_name(&self) -> R<String>

pub fn process_name_ar(&self) -> Rar<String>

pub fn process_name(&self) -> R<String>

pub fn process_id(&self) -> Pid

pub fn globally_unique_string_ar(&self) -> Rar<String>

pub fn globally_unique_string(&self) -> R<String>

pub fn os_version_string_ar(&self) -> Rar<String>

Human readable, localized; appropriate for displaying to user or using in bug emails and such; NOT appropriate for parsing

pub fn os_version_string(&self) -> R<String>

Human readable, localized; appropriate for displaying to user or using in bug emails and such; NOT appropriate for parsing

pub fn os_version(&self) -> OsVersion

pub fn sel_os_version() -> &'static Sel

@selector(operatingSystemVersion) but dynamic use this function to check if object responds to selector

pub fn thermal_state(&self) -> ThermalState

pub fn is_low_power_mode_enabled(&self) -> bool

pub fn sel_is_low_power_mode_enabled() -> &'static Sel

@selector(isLowPowerModeEnabled) but dynamic use this function to check if object responds to selector

pub fn processor_count(&self) -> usize

pub fn sel_processor_count() -> &'static Sel

@selector(processorCount) but dynamic use this function to check if object responds to selector

pub fn active_processor_count(&self) -> usize

pub fn sel_active_processor_count() -> &'static Sel

@selector(activeProcessorCount) but dynamic use this function to check if object responds to selector

pub fn physical_mem(&self) -> usize

pub fn sel_physical_mem() -> &'static Sel

@selector(physicalMemory) but dynamic use this function to check if object responds to selector

pub fn is_os_at_least_version(&self, ver: OsVersion) -> bool

pub fn sel_is_os_at_least_version() -> &'static Sel

@selector(isOperatingSystemAtLeastVersion:) but dynamic use this function to check if object responds to selector

pub fn system_uptime(&self) -> TimeInterval

pub fn sel_system_uptime() -> &'static Sel

@selector(systemUptime) but dynamic use this function to check if object responds to selector

impl ProcessInfo

NSProcessInfoPlatform

pub fn is_mac_catalyst_app(&self) -> bool

pub fn sel_is_mac_catalyst_app() -> &'static Sel

@selector(isMacCatalystApp) but dynamic use this function to check if object responds to selector

pub fn is_ios_app_on_mac(&self) -> bool

pub fn sel_is_ios_app_on_mac() -> &'static Sel

@selector(isiOSAppOnMac) but dynamic use this function to check if object responds to selector

pub fn sel_is_ios_app_on_vision() -> &'static Sel

@selector(isiOSAppOnVision) but dynamic use this function to check if object responds to selector

pub unsafe fn is_ios_app_on_vision(&self) -> bool