cidre::ns

Struct SetMut

pub struct SetMut<T: Obj>(/* private fields */);

Implementations

impl<T: Obj> SetMut<T>

NSPredicateSupport

pub fn filter_using_predicate(&mut self, predicate: &Predicate)

impl<T: Obj> SetMut<T>

pub fn cls() -> &'static Class<Self>

pub fn cls_ptr() -> *const c_void

pub fn alloc() -> A<Self>

pub fn new() -> R<Self>

pub fn with_capacity(capacity: usize) -> R<Self>

pub fn add_obj(&mut self, val: &T)

pub fn add(&mut self, val: impl AsRef<T>)

pub fn remove_obj(&mut self, val: &T)

pub fn remove(&mut self, val: impl AsRef<T>)

pub fn as_cf(&self) -> &SetMut

impl<T: Obj> SetMut<T>

NSExtendedMutableSet

pub fn add_objs_from_array(&mut self, val: &Array<T>)

pub fn intersect_set(&mut self, val: &Set<T>)

Removes from the receiving set each object that isn’t a member of another given set.

pub fn intersect(&mut self, val: impl AsRef<Set<T>>)

pub fn minus_set(&mut self, val: &Set<T>)

Removes each object in another given set from the receiving set, if present.

pub fn minus(&mut self, val: impl AsRef<Set<T>>)

pub fn clear(&mut self)

pub fn union_set(&mut self, val: &Set<T>)

Adds each object in another given set to the receiving set, if not present.

pub fn union(&mut self, val: impl AsRef<Set<T>>)

pub fn set_set(&mut self, val: &Set<T>)

Empties the receiving set, then adds each object contained in another given set.

pub fn set(&mut self, val: impl AsRef<Set<T>>)