cidre::ns

Trait AppDelegate

pub trait AppDelegate {

    // Provided methods
    fn app_supports_secure_restorable_state(&mut self) -> bool { ... }
    fn sel_app_supports_secure_restorable_state() -> &'static Sel { ... }
    fn app_will_finish_launching(&mut self, n: &Notification) { ... }
    fn sel_app_will_finish_launching() -> &'static Sel { ... }
    fn app_did_finish_launching(&mut self, n: &Notification) { ... }
    fn sel_app_did_finish_launching() -> &'static Sel { ... }
    fn app_will_hide(&mut self, n: &Notification) { ... }
    fn sel_app_will_hide() -> &'static Sel { ... }
    fn app_did_hide(&mut self, n: &Notification) { ... }
    fn sel_app_did_hide() -> &'static Sel { ... }
    fn app_will_unhide(&mut self, n: &Notification) { ... }
    fn sel_app_will_unhide() -> &'static Sel { ... }
    fn app_did_unhide(&mut self, n: &Notification) { ... }
    fn sel_app_did_unhide() -> &'static Sel { ... }
    fn app_will_become_active(&mut self, n: &Notification) { ... }
    fn sel_app_will_become_active() -> &'static Sel { ... }
    fn app_did_become_active(&mut self, n: &Notification) { ... }
    fn sel_app_did_become_active() -> &'static Sel { ... }
    fn app_will_resign_active(&mut self, n: &Notification) { ... }
    fn sel_app_will_resign_active() -> &'static Sel { ... }
    fn app_did_resign_active(&mut self, n: &Notification) { ... }
    fn sel_app_did_resign_active() -> &'static Sel { ... }
    fn app_will_update(&mut self, n: &Notification) { ... }
    fn sel_app_will_update() -> &'static Sel { ... }
    fn app_did_update(&mut self, n: &Notification) { ... }
    fn sel_app_did_update() -> &'static Sel { ... }
    fn app_will_terminate(&mut self, n: &Notification) { ... }
    fn sel_app_will_terminate() -> &'static Sel { ... }
    fn app_did_change_screen_params(&mut self, n: &Notification) { ... }
    fn sel_app_did_change_screen_params() -> &'static Sel { ... }
    fn app_did_change_occlusion_state(&mut self, n: &Notification) { ... }
    fn sel_app_did_change_occlusion_state() -> &'static Sel { ... }
    fn app_protected_data_will_become_unavailable(&mut self, n: &Notification) { ... }
    fn sel_app_protected_data_will_become_unavailable() -> &'static Sel { ... }
    fn app_protected_data_did_become_available(&mut self, n: &Notification) { ... }
    fn sel_app_protected_data_did_become_available() -> &'static Sel { ... }

}

Provided Methods

fn app_supports_secure_restorable_state(&mut self) -> bool

fn sel_app_supports_secure_restorable_state() -> &'static Sel

@selector(applicationSupportsSecureRestorableState) but dynamic use this function to check if object responds to selector

fn app_will_finish_launching(&mut self, n: &Notification)

fn sel_app_will_finish_launching() -> &'static Sel

@selector(applicationWillFinishLaunching:) but dynamic use this function to check if object responds to selector

fn app_did_finish_launching(&mut self, n: &Notification)

fn sel_app_did_finish_launching() -> &'static Sel

@selector(applicationDidFinishLaunching:) but dynamic use this function to check if object responds to selector

fn app_will_hide(&mut self, n: &Notification)

fn sel_app_will_hide() -> &'static Sel

@selector(applicationWillHide:) but dynamic use this function to check if object responds to selector

fn app_did_hide(&mut self, n: &Notification)

fn sel_app_did_hide() -> &'static Sel

@selector(applicationDidHide:) but dynamic use this function to check if object responds to selector

fn app_will_unhide(&mut self, n: &Notification)

fn sel_app_will_unhide() -> &'static Sel

@selector(applicationWillUnhide:) but dynamic use this function to check if object responds to selector

fn app_did_unhide(&mut self, n: &Notification)

fn sel_app_did_unhide() -> &'static Sel

@selector(applicationDidUnhide:) but dynamic use this function to check if object responds to selector

fn app_will_become_active(&mut self, n: &Notification)

fn sel_app_will_become_active() -> &'static Sel

@selector(applicationWillBecomeActive:) but dynamic use this function to check if object responds to selector

fn app_did_become_active(&mut self, n: &Notification)

fn sel_app_did_become_active() -> &'static Sel

@selector(applicationDidBecomeActive:) but dynamic use this function to check if object responds to selector

fn app_will_resign_active(&mut self, n: &Notification)

fn sel_app_will_resign_active() -> &'static Sel

@selector(applicationWillResignActive:) but dynamic use this function to check if object responds to selector

fn app_did_resign_active(&mut self, n: &Notification)

fn sel_app_did_resign_active() -> &'static Sel

@selector(applicationDidResignActive:) but dynamic use this function to check if object responds to selector

fn app_will_update(&mut self, n: &Notification)

fn sel_app_will_update() -> &'static Sel

@selector(applicationWillUpdate:) but dynamic use this function to check if object responds to selector

fn app_did_update(&mut self, n: &Notification)

fn sel_app_did_update() -> &'static Sel

@selector(applicationDidUpdate:) but dynamic use this function to check if object responds to selector

fn app_will_terminate(&mut self, n: &Notification)

fn sel_app_will_terminate() -> &'static Sel

@selector(applicationWillTerminate:) but dynamic use this function to check if object responds to selector

fn app_did_change_screen_params(&mut self, n: &Notification)

fn sel_app_did_change_screen_params() -> &'static Sel

@selector(applicationDidChangeScreenParameters:) but dynamic use this function to check if object responds to selector

fn app_did_change_occlusion_state(&mut self, n: &Notification)

fn sel_app_did_change_occlusion_state() -> &'static Sel

@selector(applicationDidChangeOcclusionState:) but dynamic use this function to check if object responds to selector

fn app_protected_data_will_become_unavailable(&mut self, n: &Notification)

fn sel_app_protected_data_will_become_unavailable() -> &'static Sel

@selector(applicationProtectedDataWillBecomeUnavailable:) but dynamic use this function to check if object responds to selector

fn app_protected_data_did_become_available(&mut self, n: &Notification)

fn sel_app_protected_data_did_become_available() -> &'static Sel

@selector(applicationProtectedDataDidBecomeAvailable:) but dynamic use this function to check if object responds to selector

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.