cidre::ns

Trait GestureRecognizerDelegate

pub trait GestureRecognizerDelegate: Obj {
    // Provided methods
    fn gr_should_attempt_to_recognize_with_event(
        &mut self,
        gr: &GestureRecognizer,
        event: &Event,
    ) -> bool { ... }
    fn sel_gr_should_attempt_to_recognize_with_event() -> &'static Sel { ... }
    fn gr_should_begin(&mut self, gr: &GestureRecognizer) -> bool { ... }
    fn sel_gr_should_begin() -> &'static Sel { ... }
    fn gr_should_recognize_simultaneously_with_gr(
        &mut self,
        gr: &GestureRecognizer,
        other_gr: &GestureRecognizer,
    ) -> bool { ... }
    fn sel_gr_should_recognize_simultaneously_with_gr() -> &'static Sel { ... }
    fn gr_should_require_failure_of_gr(
        &mut self,
        gr: &GestureRecognizer,
        other_gr: &GestureRecognizer,
    ) -> bool { ... }
    fn sel_gr_should_require_failure_of_gr() -> &'static Sel { ... }
    fn gr_should_be_required_to_fail_by_gr(
        &mut self,
        gr: &GestureRecognizer,
        other_gr: &GestureRecognizer,
    ) -> bool { ... }
    fn sel_gr_should_be_required_to_fail_by_gr() -> &'static Sel { ... }
    fn gr_should_receive_touch(
        &mut self,
        gr: &GestureRecognizer,
        touch: &Touch,
    ) -> bool { ... }
    fn sel_gr_should_receive_touch() -> &'static Sel { ... }
}

Provided Methods

fn gr_should_attempt_to_recognize_with_event( &mut self, gr: &GestureRecognizer, event: &Event, ) -> bool

fn sel_gr_should_attempt_to_recognize_with_event() -> &'static Sel

@selector(gestureRecognizer:shouldAttemptToRecognizeWithEvent:) but dynamic use this function to check if object responds to selector

fn gr_should_begin(&mut self, gr: &GestureRecognizer) -> bool

fn sel_gr_should_begin() -> &'static Sel

@selector(gestureRecognizerShouldBegin:) but dynamic use this function to check if object responds to selector

fn gr_should_recognize_simultaneously_with_gr( &mut self, gr: &GestureRecognizer, other_gr: &GestureRecognizer, ) -> bool

fn sel_gr_should_recognize_simultaneously_with_gr() -> &'static Sel

@selector(gestureRecognizer:shouldRecognizeSimultaneouslyWithGestureRecognizer:) but dynamic use this function to check if object responds to selector

fn gr_should_require_failure_of_gr( &mut self, gr: &GestureRecognizer, other_gr: &GestureRecognizer, ) -> bool

fn sel_gr_should_require_failure_of_gr() -> &'static Sel

@selector(gestureRecognizer:shouldRequireFailureOfGestureRecognizer:) but dynamic use this function to check if object responds to selector

fn gr_should_be_required_to_fail_by_gr( &mut self, gr: &GestureRecognizer, other_gr: &GestureRecognizer, ) -> bool

fn sel_gr_should_be_required_to_fail_by_gr() -> &'static Sel

@selector(gestureRecognizer:shouldBeRequiredToFailByGestureRecognizer:) but dynamic use this function to check if object responds to selector

fn gr_should_receive_touch( &mut self, gr: &GestureRecognizer, touch: &Touch, ) -> bool

fn sel_gr_should_receive_touch() -> &'static Sel

@selector(gestureRecognizer:shouldReceiveTouch:) but dynamic use this function to check if object responds to selector

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.