cidre::ns

Trait WindowDelegate

pub trait WindowDelegate: Obj {
    // Provided methods
    fn window_should_close(&mut self, sender: &Window) -> bool { ... }
    fn sel_window_should_close() -> &'static Sel { ... }
    fn window_will_resize(&mut self, sender: &Window, to_size: Size) -> Size { ... }
    fn sel_window_will_resize() -> &'static Sel { ... }
}

Provided Methods

fn window_should_close(&mut self, sender: &Window) -> bool

fn sel_window_should_close() -> &'static Sel

@selector(windowShouldClose:) but dynamic use this function to check if object responds to selector

fn window_will_resize(&mut self, sender: &Window, to_size: Size) -> Size

fn sel_window_will_resize() -> &'static Sel

@selector(windowWillResize:toSize:) but dynamic use this function to check if object responds to selector

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.