cidre

Module nw

Network

Re-exports

pub use connection::Connection;

pub use connection::State as ConnectionState;

pub use connection::send_completion as connection_send_completion;

pub use connection_group::ConnectionGroup;

pub use content_context::ContentCtx;

pub use group_descriptor::EnumerateEndpointsBlock;

pub use group_descriptor::GroupDesc;

pub use group_descriptor::MulticastGroupDesc;

pub use path_monitor::PathMonitor;

Modules

connection

connection_group

content_context

group_descriptor

path_monitor

udp_options

Structs

AdvertiseDesc

BrowseDesc

BrowseResult

BrowseResultChange

Browser

Endpoint

Error

A network error object with a domain and error code.

Iface

Listener

Obj

Params

Path

PrivacyContext

ProtocolDefinition

ProtocolMetadata

ProtocolOpts

ProtocolStack

ProxyCfg

RelayHop

ResolverCfg

TxtRecord

Enums

BrowserState

EndpointType

ErrorDomain

IfaceRadioType

IfaceType

ListenerState

States progress forward and do not move backwards.

MultipathService

PathStatus

PathUnsatisfiedReason

ServiceClass

The network service class represents the network queuing priority to use for traffic generated by a connection.

TxtRecordFindKey

Type Aliases

BrowserBrowseResultsChandedHandler

BrowserChangedHandler

BrowserResultEnumerateIface

IterIfaceBlock

IterIfaceTypeBlock

ListenerAdvertisedEndpointChangedHandler

ListenerNewConnectionGroupHandler

ListenerNewConnectionHandler

ListenerStateChangedHandler

ParamsCfgProtocolBlock

PathEnumerateIfaces

TxtRecordAccessKey

TxtRecordApplier