cidre::nw::connection::send_completion

Function idempotent

pub fn idempotent() -> &'static mut SendCompletion

A send callback override that causes the write call to be treated as idempotent. Idempotent content is allowed to be sent before the connection is ready, and may be replayed across parallel connection attempts. This content can be sent as part of fast-open protocols, which allows the data to be sent out sooner than if it were required to wait for connection establishment.

This override intentionally disallows the client from receiving callbacks for the write calls, since the content may be sent multiple times internally. For any large content, or content that need to be sensitive to sending backpressure, an explicit callback should be used.