cidre::nw

Struct Params

pub struct Params(/* private fields */);

Implementations

impl Params

pub fn retained(&self) -> R<Self>

impl Params

pub fn new() -> R<Self>

pub fn copy(&self) -> Option<R<Self>>

pub fn secure_tcp( configure_tls: &mut ParamsCfgProtocolBlock, configure_tcp: &mut ParamsCfgProtocolBlock, ) -> Option<R<Self>>

pub fn default_tcp() -> R<Self>

pub fn secure_udp( configure_dtls: &mut ParamsCfgProtocolBlock, configure_udp: &mut ParamsCfgProtocolBlock, ) -> Option<R<Self>>

pub fn udp() -> Option<R<Self>>

pub fn tcp() -> Option<R<Self>>

pub fn secure_custom_ip( custom_ip_protocol_number: u32, configure_ip: &mut ParamsCfgProtocolBlock, ) -> Option<R<Self>>

pub fn quic(configure_quic: &mut ParamsCfgProtocolBlock) -> Option<R<Self>>

pub fn app_service() -> R<Self>

pub fn default_cfg() -> &'static mut ParamsCfgProtocolBlock

pub fn disable_protocol() -> &'static mut ParamsCfgProtocolBlock

impl Params

Path Selection

pub fn set_required_iface(&mut self, iface: Option<&Iface>)

Require any connections or listeners using these parameters to use the provided network interface, or none if None is passed.

pub fn required_iface(&self) -> Option<R<Iface>>

pub fn prohibit_iface(&mut self, iface: &Iface)

pub fn clear_prohibit_ifaces(&mut self)

pub fn iterate_prohibit_ifaces_block(&self, block: &mut IterIfaceBlock)

pub fn iterate_prohibit_ifaces(&self, f: impl FnMut(&Iface) -> bool)

pub fn set_required_iface_type(&mut self, val: IfaceType)

pub fn required_iface_type(&self) -> IfaceType

pub fn prohibit_iface_type(&mut self, val: IfaceType)

pub fn clear_prohibited_iface_types(&mut self)

pub fn iterate_prohibit_iface_types_block(&self, block: &mut IterIfaceTypeBlock)

pub fn iterate_prohibit_iface_types(&self, f: impl FnMut(IfaceType) -> bool)

pub fn set_prohibit_expensive(&mut self, val: bool)

pub fn prohibit_expensive(&self) -> bool

Returns true if expensive interfaces are prohibited, or false otherwise.

pub fn prohibit_constrained(&self) -> bool

pub fn set_prohibit_constrained(&mut self, val: bool)

pub fn allow_ultra_constrained(&mut self) -> bool

Availability

Not macos_26_0, ios_26_0, tvos_26_0, watchos_26_0, visionos_26_0

pub fn set_allow_ultra_constrained(&mut self, val: bool)

Availability

Not macos_26_0, ios_26_0, tvos_26_0, watchos_26_0, visionos_26_0

pub fn reuse_local_address(&self) -> bool

pub fn set_reuse_local_address(&mut self, val: bool)

pub fn local_endpoint(&self) -> Option<Retained<Endpoint>>

pub fn set_local_endpoint(&mut self, val: Option<&Endpoint>)

pub fn include_peer_to_peer(&self) -> bool

pub fn set_include_peer_to_peer(&mut self, val: bool)

impl Params

Data Handling

pub fn fast_open_enabled(&self) -> bool

pub fn set_fast_open_enabled(&mut self, val: bool)