cidre::sc::error

Module code

Constants

ATTEMPT_TO_CONFIG_STATE

Failed due to attempting to update stream config on a stream

ATTEMPT_TO_START_STREAM_STATE

Failed due to attempting to start a stream thats already in a recording state

ATTEMPT_TO_STOP_STREAM_STATE

Failed due to attempting to stop a stream thats already in a recording state

ATTEMPT_TO_UPDATE_FILTER_STATE

Failed due to attempting to update the filter on a stream

FAILED_APPLICATION_CONNECTION_INTERRUPTED

Failed during recording appliation connection interrupted

FAILED_APPLICATION_CONNECTION_INVALID

Failed during recording application connection invalid

FAILED_NO_MATCHING_APPLICATION_CONTEXT

Failed during recording context id does not match application

FAILED_TO_START

The stream failed to start

FAILED_TO_START_AUDIO_CAPTURE

The stream failed to start audio

FAILED_TO_START_MIC_CAPTURE

The stream failed to start microphone

FAILED_TO_STOP_AUDIO_CAPTURE

The stream failed to stop audio

INTERNAL_ERR

Failed to start due to video/audio capture failure

INVALID_PARAM

Failed due to invalid parameter

MISSING_ENTITLEMENTS

The stream failed due to missing entitlements

NO_CAPTURE_SOURCE

Failed due to no display or window list to capture

NO_DISPLAY_LIST

Failed due to no display list

NO_WINDOW_LIST

Failed due to no window list

REMOVING_STREAM

Failed to remove stream

SYSTEM_STOPPED_STREAM

The stream was stopped by the system

USER_DECLINED

The user chose not to authorize capture

USER_STOPPED

The stream was stopped by the user