cidre::sc

Module stream

Structs

AnyDelegate

Cfg

ContentFilter

FrameInfo

Keys you use to retrieve metadata from a frame the system captures.

Stream

An object that represents a stream of shareable content.

Enums

CaptureDynamicRange

CaptureResolution

CfgPreset

Client can use sc::StreamCfgPreset to create sc::StreamCfg with suggested values of properties for various use cases

FrameStatus

Denotes the status of frame sample buffer.

OutputType

PresenterOverlayAlertSetting

Denotes the setting that can be set to determine when to show the presenter overlay alert for any stream

Traits

Delegate

DelegateImpl

Output

OutputImpl