cidre::sec

Module err

Constants

ACL_ADD_FAILED

An ACL add operation has failed.

ACL_CHANGE_FAILED

An ACL change operation has failed.

ACL_DELETE_FAILED

An ACL delete operation has failed.

ACL_NOT_SIMPLE

The specified access control list is not in standard (simple) form.

ACL_REPLACE_FAILED

An ACL replace operation has failed.

ADDIN_LOAD_FAILED

The add-in load operation has failed.

ADDIN_UNLOAD_FAILED

The add-in unload operation has failed.

ALGORITHM_MISMATCH

An algorithm mismatch was encountered.

ALLOCATE

Failed to allocate memory.

ALREADY_LOGGED_IN

The user is already logged in.

APPLE_ADD_APP_ACL_SUBJECT

Adding an application ACL subject failed.

APPLE_INVALID_KEY_END_DATE

The specified key has an invalid end date.

APPLE_INVALID_KEY_START_DATE

The specified key has an invalid start date.

APPLE_PUBLIC_KEY_INCOMPLETE

The public key is incomplete.

APPLE_SIGNATURE_MISMATCH

A signature mismatch has occurred.

APPLE_SSL_V2_ROLLBACK

A SSLv2 rollback error has occurred.

ATTACH_HANDLE_BUSY

The CSP handle was busy.

ATTRIBUTE_NOT_IN_CONTEXT

An attribute was not in the context.

AUTH_FAILED

The user name or passphrase you entered is not correct.

BAD_REQ

Bad parameter or invalid state for operation.

BLOCK_SIZE_MISMATCH

A block size mismatch occurred.

BUF_TOO_SMALL

There is not enough memory available to use the specified item.

CALLBACK_FAILED

A callback has failed.

CERT_CANNOT_OPERATE

The certificate cannot operate.

CERT_DUPLICATE_EXTENSION

The certificate contains multiple extensions with the same extension ID.

CERT_EXPIRED

An expired certificate was detected.

CERT_IS_CA

The verified certificate is a CA rather than an end-entity

CERT_NAME_NOT_ALLOWED

The requested name is not allowed for this certificate.

CERT_NOT_VALID_YET

The certificate is not yet valid.

CERT_POLICY_NOT_ALLOWED

The requested policy is not allowed for this certificate.

CERT_REVOKED

The certificate was revoked.

CERT_SUSPENDED

The certificate was suspended.

CERT_VALIDITY_PERIOD_TOO_LONG

The validity period in the certificate exceeds the maximum allowed.

CODE_SIGNING_BAD_CERT_CHAIN_LENGTH

Code signing encountered an incorrect certificate chain length.

CODE_SIGNING_BAD_PATH_LENGTH_CONSTRAINT

Code signing encountered an incorrect path length constraint.

CODE_SIGNING_DEVELOPMENT

Code signing indicated use of a development-only certificate.

CODE_SIGNING_NO_BASIC_CONSTRAINTS

Code signing found no basic constraints.

CODE_SIGNING_NO_EXTENDED_KEY_USAGE

Code signing found no extended key usage.

CONVERSION_ERROR

A conversion error has occurred.

CORE_FOUNDATION_UNKNOWN

CREATE_CHAIN_FAILED

One or more certificates required to validate this certificate cannot be found.

CRL_ALREADY_SIGNED

The certificate revocation list is already signed.

CRL_BAD_URI

The CRL has a bad Uniform Resource Identifier.

CRL_EXPIRED

The CRL has expired.

CRL_NOT_FOUND

The CRL was not found.

CRL_NOT_TRUSTED

The CRL is not trusted.

CRL_NOT_VALID_YET

The CRL is not yet valid.

CRL_POLICY_FAILED

The CRL policy failed.

CRL_SERVER_DOWN

The CRL server is down.

DATABASE_LOCKED

The database is locked.

DATASTORE_IS_OPEN

The data store is open.

DATA_NOT_AVAILABLE

The contents of this item cannot be retrieved.

DATA_NOT_MODIFIABLE

The contents of this item cannot be modified.

DATA_TOO_LARGE

This item contains information which is too large or in a format that cannot be displayed.

DECODE

Unable to decode the provided data.

DEVICE_ERROR

A device error was encountered.

DEVICE_FAILED

A device failure has occurred.

DEVICE_RESET

A device reset has occurred.

DEVICE_VERIFY_FAILED

A device verification failure has occurred.

DISK_FULL

Disk Full error.

DUPLICATE_CB

The specified callback function is already installed.

DUPLICATE_ITEM

The specified item already exists in the keychain.

DUPLICATE_KEYCHAIN

A keychain with the same name already exists.

EMM_LOAD_FAILED

The EMM load has failed.

EMM_UNLOAD_FAILED

The EMM unload has failed.

END_OF_DATA

An end-of-data was detected.

ERR_SEC_UNSUPPORTED_KEY_LABEL

The key label is not supported.

EVENT_NOTIFICATION_CALLBACK_NOT_FOUND

An event notification callback was not found.

EXTENDED_KEY_USAGE_NOT_CRITICAL

The extended key usage extension was not marked critical.

FIELD_SPECIFIED_MULTIPLE

Too many fields were specified.

FILE_TOO_BIG

The file is too big.

FUNCTION_FAILED

A function has failed.

FUNCTION_INTEGRITY_FAIL

A function address was not within the verified module.

HOST_NAME_MISMATCH

A host name mismatch has occurred.

IDP_FAILURE

The issuing distribution point was not valid.

INCOMPATIBLE_DATABASE_BLOB

The specified database has an incompatible blob.

INCOMPATIBLE_FIELD_FORMAT

The field format was incompatible.

INCOMPATIBLE_KEY_BLOB

The specified database has an incompatible key blob.

INCOMPATIBLE_VERSION

An incompatible version was encountered.

INCOMPLETE_CERT_REVOCATION_CHECK

An incomplete certificate revocation check occurred.

INPUT_LENGTH_ERROR

An input length error was encountered.

INSUFFICIENT_CLIENT_ID

The client ID is not correct.

INSUFFICIENT_CREDENTIALS

Insufficient credentials were detected.

INTERACTION_NOT_ALLOWED

User interaction is not allowed.

INTERACTION_REQUIRED

User interaction is required, but is currently not allowed.

INTERNAL_COMPONENT

INTERNAL_ERR

An internal error has occurred.

INVALID_ACCESS_CREDENTIALS

INVALID_ACCESS_REQUEST

The access request is not valid.

INVALID_ACL

An invalid ACL was encountered.

INVALID_ACTION

The action was not valid.

INVALID_ADDIN_FUNCTION_TABLE

An invalid add-in function table was encountered.

INVALID_ALGORITHM

An invalid algorithm was encountered.

INVALID_ALGORITHM_PARMS

An algorithm parameters attribute was not valid.

INVALID_ATTRIBUTE_ACCESS_CREDENTIALS

An access credentials attribute was not valid.

INVALID_ATTRIBUTE_BASE

A base attribute was not valid.

INVALID_ATTRIBUTE_BLOCK_SIZE

A block size attribute was not valid.

INVALID_ATTRIBUTE_DLDB_HANDLE

A database handle attribute was not valid.

INVALID_ATTRIBUTE_EFFECTIVE_BITS

An effective bits attribute was not valid.

INVALID_ATTRIBUTE_END_DATE

An end date attribute was not valid.

INVALID_ATTRIBUTE_INIT_VECTOR

An init vector attribute was not valid.

INVALID_ATTRIBUTE_ITERATION_COUNT

An iteration count attribute was not valid.

INVALID_ATTRIBUTE_KEY

A key attribute was not valid.

INVALID_ATTRIBUTE_KEY_LENGTH

A key length attribute was not valid.

INVALID_ATTRIBUTE_KEY_TYPE

A key type attribute was not valid.

INVALID_ATTRIBUTE_LABEL

A label attribute was not valid.

INVALID_ATTRIBUTE_MODE

A mode attribute was not valid.

INVALID_ATTRIBUTE_OUTPUT_SIZE

An output size attribute was not valid.

INVALID_ATTRIBUTE_PADDING

A padding attribute was not valid.

INVALID_ATTRIBUTE_PASSPHRASE

A passphrase attribute was not valid.

INVALID_ATTRIBUTE_PRIME

A prime attribute was not valid.

INVALID_ATTRIBUTE_PRIVATE_KEY_FORMAT

A private key format attribute was not valid.

INVALID_ATTRIBUTE_PUBLIC_KEY_FORMAT

A public key format attribute was not valid.

INVALID_ATTRIBUTE_RANDOM

A random number attribute was not valid.

INVALID_ATTRIBUTE_ROUNDS

The number of rounds attribute was not valid.

INVALID_ATTRIBUTE_SALT

A salt attribute was not valid.

INVALID_ATTRIBUTE_SEED

A seed attribute was not valid.

INVALID_ATTRIBUTE_START_DATE

A start date attribute was not valid.

INVALID_ATTRIBUTE_SUBPRIME

A subprime attribute was not valid.

INVALID_ATTRIBUTE_SYMMETRIC_KEY_FORMAT

A symmetric key format attribute was not valid.

INVALID_ATTRIBUTE_VERSION

A version attribute was not valid.

INVALID_ATTRIBUTE_WRAPPED_KEY_FORMAT

A wrapped key format attribute was not valid.

INVALID_AUTHORITY

The authority was not valid.

INVALID_AUTHORITY_KEY_ID

The authority key ID is not valid.

INVALID_BASE_ACLS

The base ACLs are not valid.

INVALID_BUNDLE_INFO

The bundle information was not valid.

INVALID_CB

The specified callback function is not valid.

INVALID_CERT_AUTHORITY

The certificate authority was not valid.

INVALID_CERT_GROUP

An invalid certificate group was encountered.

INVALID_CERT_REF

An invalid certificate reference was encountered.

INVALID_CONTEXT

An invalid context was encountered.

INVALID_CRL

The CRL was not valid.

INVALID_CRL_AUTHORITY

The CRL authority was not valid.

INVALID_CRL_ENCODING

The CRL encoding was not valid.

INVALID_CRL_GROUP

An invalid CRL group was detected.

INVALID_CRL_INDEX

The CRL index was not valid.

INVALID_CRL_TYPE

The CRL type was not valid.

INVALID_DATA

Invalid data was encountered.

INVALID_DATABASE_BLOB

The specified database has an invalid blob.

INVALID_DB_LIST

An invalid DB list was encountered.

INVALID_DB_LOCATION

The database location is not valid.

INVALID_DIGEST_ALGORITHM

An invalid digest algorithm was detected.

INVALID_ENCODING

The encoding was not valid.

INVALID_EXTENDED_KEY_USAGE

The extended key usage is not valid.

INVALID_FORM_TYPE

The form type was not valid.

INVALID_GUID

An invalid GUID was encountered.

INVALID_HANDLE

An invalid handle was encountered.

INVALID_HANDLE_USAGE

The CSSM handle does not match with the service type.

INVALID_ID

The ID was not valid.

INVALID_IDENTIFIER

The identifier was not valid.

INVALID_ID_LINKAGE

The ID linkage is not valid.

INVALID_INDEX

The index was not valid.

INVALID_INDEX_INFO

The index information is not valid.

INVALID_INPUT_VECTOR

The input vector is not valid.

INVALID_ITEM_REF

The specified item is no longer valid. It may have been deleted from the keychain.

INVALID_KEYCHAIN

The specified keychain is not a valid keychain file.

INVALID_KEY_ATTRIBUTE_MASK

The key attribute mask is not valid.

INVALID_KEY_BLOB

The specified database has an invalid key blob.

INVALID_KEY_FORMAT

The key format is not valid.

INVALID_KEY_HIERARCHY

An invalid key hierarchy was encountered.

INVALID_KEY_LABEL

The key label is not valid.

INVALID_KEY_REF

An invalid key was encountered.

INVALID_KEY_USAGE_FOR_POLICY

The key usage is not valid for the specified policy.

INVALID_KEY_USAGE_MASK

The key usage mask is not valid.

INVALID_LOGIN_NAME

An invalid login name was detected.

INVALID_MODIFY_MODE

The modify mode is not valid.

INVALID_NAME

An invalid name was encountered.

INVALID_NETWORK_ADDRESS

An invalid network address was encountered.

INVALID_NEW_OWNER

The new owner is not valid.

INVALID_NUMBER_OF_FIELDS

An invalid number of fields were encountered.

INVALID_OUTPUT_VECTOR

The output vector is not valid.

INVALID_OWNER_EDIT

Invalid attempt to change the owner of this item.

INVALID_PARSING_MODULE

The parsing module was not valid.

INVALID_PASSTHROUGH_ID

An invalid passthrough ID was encountered.

INVALID_PASSWORD_REF

The password reference was invalid.

INVALID_POINTER

An invalid pointer was encountered.

INVALID_POLICY_IDENTIFIERS

The policy identifiers are not valid.

INVALID_PREFS_DOMAIN

The specified preferences domain is not valid.

INVALID_PVC

An invalid PVC was encountered.

INVALID_QUERY

The specified query was not valid.

INVALID_REASON

The trust policy reason was not valid.

INVALID_RECORD

An invalid record was encountered.

INVALID_REQUESTOR

The requestor was not valid.

INVALID_REQUEST_INPUTS

The request inputs are not valid.

INVALID_RESPONSE_VECTOR

The response vector was not valid.

INVALID_ROOT

The root or anchor certificate is not valid.

INVALID_SAMPLE_VALUE

An invalid sample value was encountered.

INVALID_SCOPE

An invalid scope was encountered.

INVALID_SEARCH_REF

Unable to search the current keychain.

INVALID_SERVICE_MASK

An invalid service mask was encountered.

INVALID_SIGNATURE

An invalid signature was encountered.

INVALID_STOP_ON_POLICY

The stop-on policy was not valid.

INVALID_SUBJECT_KEY_ID

The subject key ID is not valid.

INVALID_SUBJECT_NAME

An invalid certificate subject name was encountered.

INVALID_SUB_SERVICE_ID

An invalid subservice ID was encountered.

INVALID_TIME_STRING

The time specified was not valid.

INVALID_TRUST_SETTING

The specified trust setting is invalid.

INVALID_TRUST_SETTINGS

The Trust Settings Record was corrupted.

INVALID_TUPLE

The tuple was not valid.

INVALID_TUPLE_CREDENTIALS

The tuple credentials are not valid.

INVALID_TUPLE_GROUP

The tuple group was not valid.

INVALID_VALIDITY_PERIOD

The validity period was not valid.

INVALID_VALUE

An invalid value was detected.

IN_DARK_WAKE

In dark wake, no UI possible

IO

I/O error.

ITEM_NOT_FOUND

The specified item could not be found in the keychain.

KEY_BLOB_TYPE_INCORRECT

The key blob type is incorrect.

KEY_HEADER_INCONSISTENT

The key header is inconsistent.

KEY_IS_SENSITIVE

Key material must be wrapped for export.

KEY_SIZE_NOT_ALLOWED

This item specifies a key size which is too large or too small.

KEY_USAGE_INCORRECT

The key usage is incorrect.

LIBRARY_REFERENCE_NOT_FOUND

A library reference was not found.

MDS_ERROR

A Module Directory Service error has occurred.

MEMORY_ERR

A memory error has occurred.

MISSING_ALGORITHM_PARMS

An algorithm parameters attribute was missing.

MISSING_ATTRIBUTE_ACCESS_CREDENTIALS

An access credentials attribute was missing.

MISSING_ATTRIBUTE_BASE

A base attribute was missing.

MISSING_ATTRIBUTE_BLOCK_SIZE

A block size attribute was missing.

MISSING_ATTRIBUTE_DLDBHANDLE

A database handle attribute was missing.

MISSING_ATTRIBUTE_EFFECTIVE_BITS

An effective bits attribute was missing.

MISSING_ATTRIBUTE_END_DATE

An end date attribute was missing.

MISSING_ATTRIBUTE_INIT_VECTOR

An init vector attribute was missing.

MISSING_ATTRIBUTE_ITERATION_COUNT

An iteration count attribute was missing.

MISSING_ATTRIBUTE_KEY

A key attribute was missing.

MISSING_ATTRIBUTE_KEY_LENGTH

A key length attribute was missing.

MISSING_ATTRIBUTE_KEY_TYPE

A key type attribute was missing.

MISSING_ATTRIBUTE_LABEL

A label attribute was missing.

MISSING_ATTRIBUTE_MODE

A mode attribute was missing.

MISSING_ATTRIBUTE_OUTPUT_SIZE

An output size attribute was missing.

MISSING_ATTRIBUTE_PADDING

A padding attribute was missing.

MISSING_ATTRIBUTE_PASSPHRASE

A passphrase attribute was missing.

MISSING_ATTRIBUTE_PRIME

A prime attribute was missing.

MISSING_ATTRIBUTE_PRIVATE_KEY_FORMAT

A private key format attribute was missing.

MISSING_ATTRIBUTE_PUBLIC_KEY_FORMAT

A public key format attribute was missing.

MISSING_ATTRIBUTE_RANDOM

A random number attribute was missing.

MISSING_ATTRIBUTE_ROUNDS

The number of rounds attribute was missing.

MISSING_ATTRIBUTE_SALT

A salt attribute was missing.

MISSING_ATTRIBUTE_SEED

A seed attribute was missing.

MISSING_ATTRIBUTE_START_DATE

A start date attribute was missing.

MISSING_ATTRIBUTE_SUBPRIME

A subprime attribute was missing.

MISSING_ATTRIBUTE_SYMMETRIC_KEY_FORMAT

A symmetric key format attribute was missing.

MISSING_ATTRIBUTE_VERSION

A version attribute was missing.

MISSING_ATTRIBUTE_WRAPPED_KEY_FORMAT

A wrapped key format attribute was missing.

MISSING_ENTITLEMENT

A required entitlement isn’t present.

MISSING_REQUIRED_EXTENSION

A required certificate extension is missing.

MISSING_VALUE

A missing value was detected.

MOBILE_ME_CSR_VERIFY_FAILURE

MOBILE_ME_FAILED_CONSISTENCY_CHECK

MOBILE_ME_NO_REQUEST_PENDING

MOBILE_ME_REQUEST_ALREADY_PENDING

MOBILE_ME_REQUEST_QUEUED

MOBILE_ME_REQUEST_REDIRECTED

MOBILE_ME_SERVER_ALREADY_EXISTS

MOBILE_ME_SERVER_ERROR

MOBILE_ME_SERVER_NOT_AVAILABLE

MOBILE_ME_SERVER_SERVICE_ERR

MODULE_MANAGER_INITIALIZE_FAILED

A module failed to initialize.

MODULE_MANAGER_NOT_FOUND

A module was not found.

MODULE_MANIFEST_VERIFY_FAILED

A module manifest verification failure has occurred.

MODULE_NOT_LOADED

A module was not loaded.

MULTIPLE_PRIV_KEYS

An attempt was made to import multiple private keys.

MULTIPLE_VALUES_UNSUPPORTED

Multiple values are not supported.

NETWORK_FAILURE

A network failure occurred.

NOT_AVAILABLE

No keychain is available. You may need to restart your computer.

NOT_INITIALIZED

A function was called without initializing CSSM.

NOT_LOGGED_IN

You are not logged in.

NOT_SIGNER

A certificate was not signed by its proposed parent.

NOT_TRUSTED

The certificate was not trusted.

NO_ACCESS_FOR_ITEM

The specified item has no access control.

NO_BASIC_CONSTRAINTS

No basic constraints were found.

NO_BASIC_CONSTRAINTS_CA

No basic CA constraints were found.

NO_CERT_MODULE

A required component (certificate module) could not be loaded. You may need to restart your computer.

NO_DEFAULT_AUTHORITY

No default authority was detected.

NO_DEFAULT_KEYCHAIN

A default keychain could not be found.

NO_FIELD_VALUES

No field values were detected.

NO_POLICY_MODULE

A required component (policy module) could not be loaded. You may need to restart your computer.

NO_STORAGE_MODULE

A required component (data storage module) could not be loaded. You may need to restart your computer.

NO_SUCH_ATTR

The specified attribute does not exist.

NO_SUCH_CLASS

The specified item does not appear to be a valid keychain item.

NO_SUCH_KEYCHAIN

The specified keychain could not be found.

NO_TRUST_SETTINGS

No Trust Settings were found.

OCSP_BAD_REQUEST

The OCSP request was incorrect or could not be parsed.

OCSP_BAD_RESPONSE

The OCSP response was incorrect or could not be parsed.

OCSP_NOT_TRUSTED_TO_ANCHOR

The OCSP response was not trusted to a root or anchor certificate.

OCSP_NO_SIGNER

The OCSP response had no signer.

OCSP_RESPONDER_INTERNAL_ERR

The OCSP responder encountered an internal error.

OCSP_RESPONDER_MALFORMED_REQ

The OCSP responder was given a malformed request.

OCSP_RESPONDER_SIGNATURE_REQUIRED

The OCSP responder requires a signature.

OCSP_RESPONDER_TRY_LATER

The OCSP responder is busy, try again later.

OCSP_RESPONDER_UNAUTHORIZED

The OCSP responder rejected this request as unauthorized.

OCSP_RESPONSE_NONCE_MISMATCH

The OCSP response nonce did not match the request.

OCSP_SIGNATURE_ERROR

The OCSP response had an invalid signature.

OCSP_STATUS_UNRECOGNIZED

The OCSP server did not recognize this certificate.

OCSP_UNAVAILABLE

OCSP service is unavailable.

OP_WR

File already open with write permission.

OUTPUT_LENGTH_ERROR

An output length error was encountered.

PARAM

One or more parameters passed to a function were not valid.

PASSPHRASE_REQUIRED

Passphrase is required for import/export.

PATH_LENGTH_CONSTRAINT_EXCEEDED

The path length constraint was exceeded.

PKCS12_VERIFY_FAILURE

MAC verification failed during PKCS12 import (wrong password?)

POLICY_NOT_FOUND

The specified policy cannot be found.

PRIVILEGE_NOT_GRANTED

The privilege was not granted.

PRIVILEGE_NOT_SUPPORTED

The privilege is not supported.

PUBLIC_KEY_INCONSISTENT

The public key was inconsistent.

PVC_ALREADY_CONFIGURED

The PVC is already configured.

PVC_REFERENT_NOT_FOUND

A reference to the calling module was not found in the list of authorized callers.

QUERY_SIZE_UNKNOWN

The query size is unknown.

QUOTA_EXCEEDED

The quota was exceeded.

READ_ONLY

This keychain cannot be modified.

READ_ONLY_ATTR

The specified attribute could not be modified.

RECORD_MODIFIED

The record was modified.

REJECTED_FORM

The trust policy had a rejected form.

REQUEST_DESCRIPTOR

The request descriptor was not valid.

REQUEST_LOST

The request was lost.

REQUEST_REJECTED

The request was rejected.

RESOURCE_SIGN_BAD_CERT_CHAIN_LENGTH

Resource signing has encountered an incorrect certificate chain length.

RESOURCE_SIGN_BAD_EXT_KEY_USAGE

Resource signing has encountered an error in the extended key usage.

RESTRICTED_API

Client is restricted and is not permitted to perform this operation.

SELF_CHECK_FAILED

Self-check has failed.

SERVICE_NOT_AVAILABLE

The required service is not available.

SIGNING_TIME_MISSING

A signing time was expected but was not found.

SMIME_BAD_EXTENDED_KEY_USAGE

The appropriate extended key usage for SMIME was not found.

SMIME_BAD_KEY_USAGE

The key usage is not compatible with SMIME.

SMIME_EMAIL_ADDRESSES_NOT_FOUND

An email address mismatch was encountered.

SMIME_KEY_USAGE_NOT_CRITICAL

The key usage extension is not marked as critical.

SMIME_NO_EMAIL_ADDRESS

No email address was found in the certificate.

SMIME_SUBJ_ALT_NAME_NOT_CRITICAL

The subject alternative name extension is not marked as critical.

SSL_BAD_EXTENDED_KEY_USAGE

The appropriate extended key usage for SSL was not found.

STAGED_OPERATION_IN_PROGRESS

A staged operation is in progress.

STAGED_OPERATION_NOT_STARTED

A staged operation was not started.

TAG_NOT_FOUND

The specified tag was not found.

TIMESTAMP_ADD_INFO_NOT_AVAILABLE

The additional information requested is not available.

TIMESTAMP_BAD_ALG

An unrecognized or unsupported Algorithm Identifier in timestamp.

TIMESTAMP_BAD_DATA_FORMAT

The timestamp data submitted has the wrong format.

TIMESTAMP_BAD_REQUEST

The timestamp transaction is not permitted or supported.

TIMESTAMP_INVALID

The timestamp was not valid.

TIMESTAMP_MISSING

A timestamp was expected but was not found.

TIMESTAMP_NOT_TRUSTED

The timestamp was not trusted.

TIMESTAMP_REJECTION

A timestamp transaction was rejected.

TIMESTAMP_REVOCATION_NOTIFICATION

A timestamp authority revocation notification was issued.

TIMESTAMP_REVOCATION_WARNING

A timestamp authority revocation warning was issued.

TIMESTAMP_SERVICE_NOT_AVAILABLE

The timestamp service is not available.

TIMESTAMP_SYSTEM_FAILURE

The timestamp request cannot be handled due to system failure.

TIMESTAMP_TIME_NOT_AVAILABLE

The time source for the Timestamp Authority is not available.

TIMESTAMP_UNACCEPTED_EXTENSION

The requested extension is not supported by the Timestamp Authority.

TIMESTAMP_UNACCEPTED_POLICY

The requested policy is not supported by the Timestamp Authority.

TIMESTAMP_WAITING

A timestamp transaction is waiting.

TRUST_NOT_AVAILABLE

No trust results are available.

TRUST_SETTING_DENY

The trust setting for this policy was set to Deny.

UNIMPLEMENTED

Function or operation not implemented.

UNKNOWN_CERT_EXTENSION

An unknown certificate extension was encountered.

UNKNOWN_CRITICAL_EXTENSION_FLAG

There is an unknown critical extension flag.

UNKNOWN_CRL_EXTENSION

An unknown CRL extension was encountered.

UNKNOWN_FORMAT

Unknown format in import.

UNKNOWN_QUALIFIED_CERT_STATEMENT

An unknown qualified certificate statement was encountered.

UNKNOWN_TAG

An unknown tag was encountered.

UNSUPPORTED_ADDRESS_TYPE

The address type is not supported.

UNSUPPORTED_FIELD_FORMAT

The field format is not supported.

UNSUPPORTED_FORMAT

Import/Export format unsupported.

UNSUPPORTED_INDEX_INFO

The index information is not supported.

UNSUPPORTED_KEY_ATTRIBUTE_MASK

The key attribute mask is not supported.

UNSUPPORTED_KEY_FORMAT

The key header format is not supported.

UNSUPPORTED_KEY_SIZE

The key size is not supported.

UNSUPPORTED_KEY_USAGE_MASK

The key usage mask is not supported.

UNSUPPORTED_LOCALITY

The locality is not supported.

UNSUPPORTED_NUM_ATTRIBUTES

The number of attributes is not supported.

UNSUPPORTED_NUM_INDEXES

The number of indexes is not supported.

UNSUPPORTED_NUM_RECORD_TYPES

The number of record types is not supported.

UNSUPPORTED_NUM_SELECTION_PREDS

The number of selection predicates is not supported.

UNSUPPORTED_OPERATOR

The operator is not supported.

UNSUPPORTED_QUERY_LIMITS

The query limits are not supported.

UNSUPPORTED_SERVICE

The service is not supported.

UNSUPPORTED_VECTOR_OF_BUFFERS

The vector of buffers is not supported.

USER_CANCELED

User canceled the operation.

VERIFICATION_FAILURE

A verification failure occurred.

VERIFY_ACTION_FAILED

A verify action has failed.

VERIFY_FAILED

A cryptographic verification failure has occurred.

WRONG_SEC_VERSION

This keychain was created by a different version of the system software and cannot be opened.

WR_PERM

Write permissions error.