cidre::sec::item

Module match_keys

Predefined search constants used to set values in a query dictionary. You can specify a combination of search attributes and item attributes when looking for matching items with the SecItemCopyMatching function.

Functions

case_insesitive

Specifies a dictionary key whose value is a CFBooleanRef. If this value is kCFBooleanFalse, or is not provided, then case-sensitive string matching is performed.

diacritic_case_insesitive

macOS only. Specifies a dictionary key whose value is a CFBooleanRef. If this value is kCFBooleanFalse, or is not provided, then diacritic-sensitive string matching is performed.

email_address_if_present

Specifies a dictionary key whose value is a &cf::String containing an RFC822 email address. If provided, returned certificates or identities will be limited to those that contain the address in their subject or subject alternative name.

host_or_subdomain_of_host

Specifies a dictionary key whose value is a &cf::String. If provided, returned internet passwords will be limited to those which have a server host that is equal to or a subdomain of this string. This filter only works on the Data Protection Keychain on macOS.

issuers

Specifies a dictionary key whose value is a CFArray of X.500 names (of type CFDataRef). If provided, returned certificates or identities will be limited to those whose certificate chain contains one of the issuers provided in this list.

item_list

Specifies a dictionary key whose value is a CFArray of SecKeychainItemRef items. If provided, returned items will be limited to the subset which are contained in this list.

limit

Specifies a dictionary key whose value is a CFNumberRef. If provided, this value specifies the maximum number of results to return. If not provided, results are limited to the first item found. Predefined values are provided for a single item (kSecMatchLimitOne) and all matching items (kSecMatchLimitAll).

policy

Specifies a dictionary key whose value is a SecPolicyRef. If provided, returned certificates or identities must verify with this policy.

subject_contains

Specifies a dictionary key whose value is a &cf::String. If provided, returned certificates or identities will be limited to those containing this string in the subject.

subject_ends_with

macOS only. Specifies a dictionary key whose value is a &cf::String. If provided, returned certificates or identities will be limited to those with subject names that end with this string.

subject_starts_with

macOS only. Specifies a dictionary key whose value is a &cf::String. If provided, returned certificates or identities will be limited to those with subject names that start with this string.

subject_whole_string

macOS only. Specifies a dictionary key whose value is a &cf::String. If provided, returned certificates or identities will be limited to those matching this string exactly in the subject.

trusted_only

Specifies a dictionary key whose value is a CFBooleanRef. If provided with a value of kCFBooleanTrue, only certificates which can be verified back to a trusted anchor will be returned. If this value is kCFBooleanFalse, or is not provided, then both trusted and untrusted certificates may be returned.

valid_on_date

Specifies a dictionary key whose value is of type CFDateRef. If provided, returned keys, certificates or identities will be limited to those which are valid for the given date. Pass a value of kCFNull to indicate the current date.

width_insesitive

macOS only. Specifies a dictionary key whose value is a CFBooleanRef. If this value is kCFBooleanFalse, or is not provided, then string matching is width-sensitive (e.g. ‘a’ != 0xFF41).