cidre::simd

Module packed

Type Aliases

f32x2

f32x4

i8x2

i8x4

i16x2

i16x4

i32x2

A vector of two 32-bit signed (twos-complement) integers with relaxed alignment. In C++ and Metal, this type is also available as simd::packed::int2. The alignment of this type is that of the underlying scalar element type, so you can use it to load or store from an array of that type. */

i32x4

i64x2

i64x4

u8x2

u8x4

u16x2

u16x4

u32x2

u32x4

u64x2

u64x4