cidre::ut

Struct Type

pub struct Type(/* private fields */);

Implementations

impl Type

pub fn retained(&self) -> R<Self>

impl Type

pub fn cls() -> &'static Class<Self>

pub fn cls_ptr() -> *const c_void

pub fn alloc() -> A<Self>

pub fn with_id_ar(identifier: &String) -> Option<Rar<Self>>

pub fn with_id(identifier: &String) -> Option<R<Self>>

pub fn with_file_ext_ar(file_name_extension: &String) -> Option<Rar<Self>>

pub fn with_file_ext(file_name_extension: &String) -> Option<R<Self>>

pub fn id_ar(&self) -> Rar<String>

pub fn id(&self) -> R<String>

pub fn preferred_file_ext_ar(&self) -> Option<Rar<String>>

pub fn preferred_file_ext(&self) -> Option<R<String>>

pub fn preferred_mime_type_ar(&self) -> Option<Rar<String>>

pub fn preferred_mime_type(&self) -> Option<R<String>>

pub fn localized_desc_ar(&self) -> Option<Rar<String>>

pub fn localized_desc(&self) -> Option<R<String>>

pub fn version_ar(&self) -> Option<Rar<String>>

pub fn version(&self) -> Option<R<String>>

pub fn reference_url_ar(&self) -> Option<Rar<Url>>

pub fn reference_url(&self) -> Option<R<Url>>

pub fn is_dynamic(&self) -> bool

pub fn is_declared(&self) -> bool

pub fn is_public_type(&self) -> bool

impl Type

Conformance

pub fn conforms_to_type(&self, other: &Self) -> bool

pub fn is_supertype_of_type(&self, other: &Self) -> bool

pub fn is_subtype_of_type(&self, other: &Self) -> bool

pub fn supertypes_ar(&self) -> Rar<Set<Self>>

pub fn supertypes(&self) -> R<Set<Self>>

impl Type

Abstract base types

pub fn item() -> &'static Self

A generic base type for most things (files, directories.)

UTI: public.item

pub fn content() -> &'static Self

A base type for anything containing user-viewable document content (documents, pasteboard data, and document packages.)

UTI: public.content

pub fn composite_content() -> &'static Self

A base type for content formats supporting mixed embedded content (i.e., compound documents).

UTI: public.composite-content conforms to: public.content

pub fn disk_image() -> &'static Self

A data item mountable as a volume

UTI: public.disk-image

impl Type

Concrete base types

pub fn data() -> &'static Self

A base type for any sort of simple byte stream, including files and in-memory data.

UTI: public.data conforms to: public.item

pub fn directory() -> &'static Self

A file system directory (includes packages \em and folders.)

UTI: public.directory conforms to: public.item

pub fn resolvable() -> &'static Self

Symbolic link and alias file types conform to this type.

UTI: com.apple.resolvable

pub fn symbolic_link() -> &'static Self

A symbolic link.

UTI: public.symlink conforms to: public.item, com.apple.resolvable

pub fn executable() -> &'static Self

An executable item.

UTI: public.executable conforms to: public.item

pub fn mount_point() -> &'static Self

A volume mount point (resolvable, resolves to the root directory of a volume.)

UTI: com.apple.mount-point conforms to: public.item, com.apple.resolvable

pub fn alias_file() -> &'static Self

A fully-formed alias file.

UTI: com.apple.alias-file conforms to: public.data, com.apple.resolvable

impl Type

URL types

pub fn url_bookmark_data() -> &'static Self

A URL bookmark.

UTI: com.apple.bookmark conforms to: public.data, com.apple.resolvable

pub fn url() -> &'static Self

Any URL.

UTI: public.url conforms to: public.data

pub fn file_url() -> &'static Self

A URL with the scheme \c “file:”.

UTI: public.file-url conforms to: public.url

impl Type

Text types

pub fn text() -> &'static Self

The base type for all text-encoded data, including text with markup (HTML, RTF, etc.).

UTI: public.text conforms to: public.data, public.content

pub fn plain_text() -> &'static Self

Text with no markup and an unspecified encoding.

UTI: public.plain-text conforms to: public.text

pub fn utf8_plain_text() -> &'static Self

Plain text encoded as UTF-8.

UTI: public.utf8-plain-text conforms to: public.plain-text

pub fn utf16_external_plain_text() -> &'static Self

Plain text encoded as UTF-16 with a BOM, or if a BOM is not present, using “external representation” byte order (big-endian).

UTI: public.utf16-external-plain-text conforms to: public.plain-text

pub fn utf16_plain_text() -> &'static Self

Plain text encoded as UTF-16, in native byte order, with an optional BOM.

UTI: public.utf16-plain-text conforms to: public.plain-text

pub fn delimited_text() -> &'static Self

Text containing delimited values.

UTI: public.delimited-values-text conforms to: public.text

pub fn comma_separated_text() -> &'static Self

Text containing comma-separated values (.csv).

UTI: public.comma-separated-values-text conforms to: public.delimited-values-text

pub fn tab_separated_text() -> &'static Self

Text containing tab-separated values.

UTI: public.tab-separated-values-text conforms to: public.delimited-values-text

pub fn utf8_tab_separated_text() -> &'static Self

UTF-8 encoded text containing tab-separated values.

UTI: public.utf8-tab-separated-values-text conforms to: public.tab-separated-values-text, public.utf8-plain-text

pub fn rtf() -> &'static Self

Rich Text Format data.

UTI: public.rtf conforms to: public.text

impl Type

Markup languages

pub fn html() -> &'static Self

Any version of HTML.

UTI: public.html conforms to: public.text

pub fn xml() -> &'static Self

Generic XML.

UTI: public.xml conforms to: public.text

pub fn yaml() -> &'static Self

Yet Another Markup Language.

UTI: public.yaml conforms to: public.text

impl Type

Programming languages

pub fn src_code() -> &'static Self

Abstract type for source code of any language.

UTI: public.source-code conforms to: public.plain-text

pub fn asm_src() -> &'static Self

Assembly language source (.s)

UTI: public.assembly-source conforms to: public.source-code

pub fn c_src() -> &'static Self

C source code (.c)

UTI: public.c-source conforms to: public.source-code

pub fn obj_c_src() -> &'static Self

Objective-C source code (.m)

UTI: public.objective-c-source conforms to: public.source-code

pub fn swift_src() -> &'static Self

Swift source code (.swift)

UTI: public.swift-source conforms to: public.source-code

pub fn cpp_src() -> &'static Self

C++ source code (.cp, etc.)

UTI: public.c-plus-plus-source conforms to: public.source-code

pub fn obj_cpp_src() -> &'static Self

Objective-C++ source code.

UTI: public.objective-c-plus-plus-source conforms to: public.source-code

pub fn c_header() -> &'static Self

A C header.

UTI: public.c-header conforms to: public.source-code

pub fn cpp_header() -> &'static Self

A C++ header.

UTI: public.c-plus-plus-header conforms to: public.source-code

impl Type

Scripting languages

pub fn script() -> &'static Self

A base type for any scripting language source.

UTI: public.script conforms to: public.source-code

pub fn apple_script() -> &'static Self

An AppleScript text-based script (.applescript).

UTI: com.apple.applescript.text conforms to: public.script

pub fn osa_script() -> &'static Self

An Open Scripting Architecture binary script (.scpt).

UTI: com.apple.applescript.script conforms to: public.data, public.script

pub fn osa_script_bundle() -> &'static Self

An Open Scripting Architecture script bundle (.scptd).

UTI: com.apple.applescript.script-bundle conforms to: com.apple.bundle, com.apple.package, public.script

pub fn java_script() -> &'static Self

JavaScript source code

UTI: com.netscape.javascript-source conforms to: public.source-code, public.executable

pub fn shell_script() -> &'static Self

The base type for shell scripts.

UTI: public.shell-script conforms to: public.script

pub fn perl_script() -> &'static Self

A Perl script.

UTI: public.perl-script conforms to: public.shell-script

pub fn python_script() -> &'static Self

A Python script.

UTI: public.python-script conforms to: public.shell-script

pub fn ruby_script() -> &'static Self

A Ruby script.

UTI: public.ruby-script conforms to: public.shell-script

pub fn php_script() -> &'static Self

A PHP script.

UTI: public.php-script conforms to: public.shell-script

pub fn makefile() -> &'static Self

A makefile.

UTI: public.make-source conforms to: public.script

impl Type

Serialized data types

pub fn json() -> &'static Self

JavaScript object notation (JSON) data

UTI: public.json conforms to: public.text JSON almost (but doesn’t quite) conforms to com.netscape.javascript-source.

pub fn prop_list() -> &'static Self

A base type for property lists.

UTI: com.apple.property-list conforms to: public.data

pub fn xml_prop_list() -> &'static Self

An XML property list.

UTI: com.apple.xml-property-list conforms to: public.xml, com.apple.property-list

pub fn binary_prop_list() -> &'static Self

A binary property list.

UTI: com.apple.binary-property-list conforms to: com.apple.property-list

impl Type

Composite content types

pub fn pdf() -> &'static Self

An Adobe PDF document.

UTI: com.adobe.pdf conforms to: public.data, public.composite-content

pub fn rtfd() -> &'static Self

A Rich Text Format Directory document (RTF with content embedding in its on-disk format.)

UTI: com.apple.rtfd conforms to: com.apple.package, public.composite-content

pub fn flat_rtfd() -> &'static Self

A flattened RTFD document (formatted for the pasteboard.)

UTI: com.apple.flat-rtfd conforms to: public.data, public.composite-content

pub fn web_archive() -> &'static Self

The WebKit webarchive format.

UTI: com.apple.webarchive conforms to: public.data, public.composite-content

impl Type

Image content types

pub fn image() -> &'static Self

A base type for abstract image data.

UTI: public.image conforms to: public.data, public.content

pub fn jpeg() -> &'static Self

A JPEG image.

UTI: public.jpeg conforms to: public.image

pub fn tiff() -> &'static Self

A TIFF image.

UTI: public.tiff conforms to: public.image

pub fn gif() -> &'static Self

A GIF image.

UTI: com.compuserve.gif conforms to: public.image

pub fn png() -> &'static Self

A PNG image.

UTI: public.png conforms to: public.image

pub fn icns() -> &'static Self

Apple icon data

UTI: com.apple.icns conforms to: public.image

pub fn bmp() -> &'static Self

A Windows bitmap.

UTI: com.microsoft.bmp conforms to: public.image

pub fn ico() -> &'static Self

Windows icon data

UTI: com.microsoft.ico conforms to: public.image

pub fn raw_image() -> &'static Self

A base type for raw image data (.raw).

UTI: public.camera-raw-image conforms to: public.image

pub fn svg() -> &'static Self

A Scalable Vector Graphics image.

UTI: public.svg-image conforms to: public.image

pub fn live_photo() -> &'static Self

A Live Photo.

UTI: com.apple.live-photo

pub fn heif() -> &'static Self

A High Efficiency Image File Format image.

UTI: public.heif conforms to: public.heif-standard

pub fn heic() -> &'static Self

A High Efficiency Image Coding image.

UTI: public.heic conforms to: public.heif-standard

pub fn webp() -> &'static Self

The WebP image format.

UTI: org.webmproject.webp conforms to: public.image

impl Type

3D Content

pub fn _3d_content() -> &'static Self

A base type for 3D content.

UTI: public.3d-content conforms to: public.content

pub fn usd() -> &'static Self

Universal Scene Description content.

UTI: com.pixar.universal-scene-description conforms to: public.3d-content, public.data

pub fn usdz() -> &'static Self

Universal Scene Description Package content.

UTI: com.pixar.universal-scene-description-mobile conforms to: public.3d-content, public.data

pub fn reality_file() -> &'static Self

A Reality File.

UTI: com.apple.reality conforms to: public.data

pub fn sk_scene() -> &'static Self

A SceneKit serialized scene.

UTI: com.apple.scenekit.scene conforms to: public.3d-content, public.data

pub fn ar_reference_obj() -> &'static Self

An AR reference object.

UTI: com.apple.arobject conforms to: public.data

pub fn av_content() -> &'static Self

Any audio and/or video content.

UTI: public.audiovisual-content conforms to: public.data, public.content

pub fn movie() -> &'static Self

A media format which may contain both video and audio.

This type corresponds to what users would label a “movie”.

UTI: public.movie conforms to: public.audiovisual-content

pub fn video() -> &'static Self

Pure video data with no audio data.

UTI: public.video conforms to: public.movie

pub fn audio() -> &'static Self

Pure audio data with no video data.

UTI: public.audio conforms to: public.audiovisual-content

pub fn quick_time_movie() -> &'static Self

A QuickTime movie.

UTI: com.apple.quicktime-movie conforms to: public.movie

pub fn mpeg() -> &'static Self

An MPEG-1 or MPEG-2 movie.

UTI: public.mpeg conforms to: public.movie

pub fn mpeg2video() -> &'static Self

An MPEG-2 video.

UTI: public.mpeg-2-video conforms to: public.video

pub fn mpeg2ts() -> &'static Self

The MPEG-2 Transport Stream movie format.

UTI: public.mpeg-2-transport-stream conforms to: public.movie

pub fn mp3() -> &'static Self

MP3 audio.

UTI: public.mp3 conforms to: public.audio

pub fn mpeg4movie() -> &'static Self

MPEG-4 movie

UTI: public.mpeg-4 conforms to: public.movie

pub fn mpeg4audio() -> &'static Self

An MPEG-4 audio layer file.

UTI: public.mpeg-4-audio conforms to: public.mpeg-4, public.audio

pub fn apple_protected_mpeg4audio() -> &'static Self

The Apple protected MPEG4 format (.m4p, iTunes music store format.)

UTI: com.apple.protected-mpeg-4-audio conforms to: public.audio

pub fn apple_protected_mpeg4video() -> &'static Self

An Apple protected MPEG-4 movie.

UTI: com.apple.protected-mpeg-4-video conforms to: com.apple.m4v-video

pub fn avi() -> &'static Self

The AVI movie format.

UTI: public.avi conforms to: public.movie

pub fn aiff() -> &'static Self

The AIFF audio format

UTI: public.aiff-audio conforms to: public.aifc-audio

pub fn wav() -> &'static Self

The Microsoft waveform audio format (.wav).

UTI: com.microsoft.waveform-audio conforms to: public.audio

pub fn midi() -> &'static Self

The MIDI audio format.

UTI: public.midi-audio conforms to: public.audio

pub fn playlist() -> &'static Self

The base type for playlists.

UTI: public.playlist

pub fn m3u_playlist() -> &'static Self

An M3U or M3U8 playlist

UTI: public.m3u-playlist conforms to: public.text, public.playlist

impl Type

Directory types

pub fn folder() -> &'static Self

A user-browsable directory (i.e. not a package.)

UTI: public.folder conforms to: public.directory

pub fn volume() -> &'static Self

The root folder of a volume or mount point.

UTI: public.volume conforms to: public.folder

pub fn package() -> &'static Self

A packaged directory.

UTI: com.apple.package conforms to: public.directory

pub fn bundle() -> &'static Self

A directory conforming to one of the \c CFBundle layouts.

UTI: com.apple.bundle conforms to: public.directory

pub fn plugin_bundle() -> &'static Self

The base type for bundle-based plugins.

UTI: com.apple.plugin conforms to: com.apple.bundle, com.apple.package

pub fn spotlight_importer() -> &'static Self

A Spotlight metadata importer bundle.

UTI: com.apple.metadata-importer conforms to: com.apple.plugin

pub fn quick_look_generator() -> &'static Self

A QuickLook preview generator bundle.

UTI: com.apple.quicklook-generator conforms to: com.apple.plugin

pub fn xpc_service() -> &'static Self

An XPC service bundle.

UTI: com.apple.xpc-service conforms to: com.apple.bundle, com.apple.package

pub fn framework() -> &'static Self

A macOS or iOS framework bundle.

UTI: com.apple.framework conforms to: com.apple.bundle

pub fn app() -> &'static Self

The base type for macOS and iOS applications.

UTI: com.apple.application conforms to: public.executable

pub fn app_bundle() -> &'static Self

A bundled application.

UTI: com.apple.application-bundle conforms to: com.apple.application, com.apple.bundle, com.apple.package

pub fn appex() -> &'static Self

An application extension (.appex).

UTI: com.apple.application-and-system-extension conforms to: com.apple.xpc-service

pub fn unix_executable() -> &'static Self

A UNIX executable (flat file.)

UTI: public.unix-executable conforms to: public.data, public.executable

pub fn exe() -> &'static Self

A Windows executable (.exe).

UTI: com.microsoft.windows-executable conforms to: public.data, public.executable

pub fn sys_pref_pane() -> &'static Self

A System Preferences pane.

UTI: com.apple.systempreference.prefpane conforms to: com.apple.package, com.apple.bundle

impl Type

Archival and compression types

pub fn archive() -> &'static Self

An archive of files and directories

UTI: public.archive

pub fn gzip() -> &'static Self

A GNU zip archive.

UTI: org.gnu.gnu-zip-archive conforms to: public.data, public.archive

pub fn bz2() -> &'static Self

A bzip2 archive.

UTI: public.bzip2-archive conforms to: public.data, public.archive

pub fn zip() -> &'static Self

A zip archive.

UTI: public.zip-archive conforms to: com.pkware.zip-archive

pub fn apple_archive() -> &'static Self

An Apple Archive.

UTI: com.apple.archive conforms to: public.data, public.archive

impl Type

Document types

pub fn spreadsheet() -> &'static Self

A base type for spreadsheet documents.

UTI: public.spreadsheet conforms to: public.content

pub fn presentation() -> &'static Self

A base type for presentation documents.

UTI: public.presentation conforms to: public.composite-content

pub fn database() -> &'static Self

A database store.

UTI: public.database

impl Type

Messages, contacts, and calendar types

pub fn message() -> &'static Self

A base type for messages (email, IM, etc.)

UTI: public.message

pub fn contact() -> &'static Self

Contact information, e.g. for a person, group, organization

UTI: public.contact

pub fn v_card() -> &'static Self

A vCard file.

UTI: public.vcard conforms to: public.text, public.contact

pub fn todo_item() -> &'static Self

A to-do item.

UTI: public.to-do-item

pub fn calendar_event() -> &'static Self

A calendar event.

UTI: public.calendar-event

pub fn email_message() -> &'static Self

An e-mail message.

UTI: public.email-message conforms to: public.message

impl Type

Internet locations

pub fn internet_location() -> &'static Self

A base type for Apple Internet location files.

UTI: com.apple.internet-location conforms to: public.data

pub fn internet_shortcut() -> &'static Self

Microsoft Internet shortcut files (.url).

UTI: com.apple.internet-location conforms to: public.data

impl Type

Miscellaneous types

pub fn font() -> &'static Self

A base type for fonts.

UTI: public.font

pub fn bookmark() -> &'static Self

A bookmark.

UTI: public.bookmark

pub fn pkcs12() -> &'static Self

PKCS#12 data.

UTI: com.rsa.pkcs-12 conforms to: public.data

pub fn x509cert() -> &'static Self

An X.509 certificate.

UTI: public.x509-certificate conforms to: public.data

pub fn epub() -> &'static Self

The EPUB format.

UTI: org.idpf.epub-container conforms to: public.data, public.composite-content

pub fn log() -> &'static Self

A base type for console logs.

UTI: public.log

pub fn ahap() -> &'static Self

An Apple Haptics Audio Pattern file.

UTI: com.apple.haptics.ahap