cidre

Module vn

Vision

Re-exports

pub use video_processor::Cadence as VideoProcessorCadence;

pub use video_processor::FrameRateCadence as VideoProcessorFrameRateCadence;

pub use video_processor::RequestProcessingOpts as VideoProcessorRequestProcessingOpts;

pub use video_processor::TimeIntervalCadence as VideoProcessorTimeIntervalCadence;

pub use video_processor::VideoProcessor;

Modules

video_processor

Structs

AnimalId

BarcodeObservation

BarcodeSymbology

CalcImageAestheticsScoresRequest

ClassificationObservation

ClassifyImageRequest

CoreMLFeatureValueObservation

DetectBarcodesRequest

DetectDocumentSegmentationRequest

DetectFaceCaptureQualityRequest

DetectFaceRectanglesRequest

DetectHorizonRequest

DetectedObjectObservation

FaceLandmarkRegion

FaceLandmarkRegion2d

FaceLandmarks

FaceLandmarks2d

FaceObservation

FeaturePrintObservation

GenAttentionBasedSaliencyImageRequest

GenForegroundInstanceMaskRequest

GenImageFeaturePrintRequest

GenObjectnessBasedSaliencyImageRequest

GenOpticalFlowRequest

GenPersonSegmentationRequest

HorizonObservation

HumanObservation

ImageAestheticsScoresObservation

ImageBasedRequest

ImageCropAndScaleOpt

ImageRequestHandler

InstanceMaskObservation

Observation

PixelBufObservation

RecognizeAnimalsRequest

RecognizeTextRequest

RecognizedObjectObservation

RecognizedText

RecognizedTextObservation

RectangleObservation

Request

SaliencyImageObservation

SequenceRequestHandler

StatefulRequest

TextObservation

TrackingRequest

TrajectoryObservation

Enums

Chirality

ElementType

ErrorCode

GenPersonSegmentationRequestQualityLevel

Person segmentation level options to favor speed over recognition accuracy. Accurate is the default option.

PointsClassification

RequestTextRecognitionLevel

RequestTrackingRequestLevel

Traits

RequestProgressProviding

RequestProgressProvidingImpl

Type Aliases

AspectRatio

Confidence

Degrees

RequestCh

A block that is executed at the completion of a request.

RequestProgressHandler