cidre::vn

Enum ErrorCode

#[repr(isize)]
pub enum ErrorCode {

    TuriCoreErrorCode = -1,
    Ok = 0,
    RequestCancelled = 1,
    InvalidFormat = 2,
    OperationFailed = 3,
    OutOfBoundsError = 4,
    InvalidOption = 5,
    IOError = 6,
    MissingOption = 7,
    NotImplemented = 8,
    InternalError = 9,
    OutOfMemory = 10,
    UnknownError = 11,
    InvalidOperation = 12,
    InvalidImage = 13,
    InvalidArgument = 14,
    InvalidModel = 15,
    UnsupportedRevision = 16,
    DataUnavailable = 17,
    TimeStampNotFound = 18,
    UnsupportedRequest = 19,
    Timeout = 20,

}

Variants

TuriCoreErrorCode = -1

Ok = 0

The operation finished without error.

RequestCancelled = 1

An app canceled the request.

InvalidFormat = 2

The format of the image is invalid.

OperationFailed = 3

The requested operation failed.

OutOfBoundsError = 4

An app attempted to access data that’s out-of-bounds.

InvalidOption = 5

An app specified an invalid option on a request.

IOError = 6

An I/O error for an image, image sequence, or Core ML model.

MissingOption = 7

A request is missing a required option.

NotImplemented = 8

The method isn’t implemented in the underlying model.

InternalError = 9

An internal error occurred within the framework.

OutOfMemory = 10

The system doesn’t have enough memory to complete the request.

UnknownError = 11

An unidentified error occurred.

InvalidOperation = 12

An app requested an unsupported operation.

InvalidImage = 13

The image is invalid.

This error occurs when you pass an invalid image to an operation, such as passing an image with no dimensions.

InvalidArgument = 14

An app passed an invalid parameter to a request.

InvalidModel = 15

The Core ML model is incompatible with the request.

UnsupportedRevision = 16

An app specified an unsupported request revision.

DataUnavailable = 17

The data isn’t available.

TimeStampNotFound = 18

The system can’t find a timestamp.

UnsupportedRequest = 19

An app attempted an unsupported request.

Timeout = 20