cidre::vn

Struct RecognizeTextRequest

pub struct RecognizeTextRequest(/* private fields */);

Implementations

impl RecognizeTextRequest

pub fn retained(&self) -> R<Self>

impl RecognizeTextRequest

pub fn cls() -> &'static Class<Self>

pub fn cls_ptr() -> *const c_void

pub fn alloc() -> A<Self>

pub fn new() -> R<RecognizeTextRequest>

shortcut to Self::alloc().init()

impl RecognizeTextRequest

pub const REVISION_1: usize = 1usize

only supports English

pub const REVISION_2: usize = 2usize

supports English, Chinese, Portuguese, French, Italian, German and Spanish in the accurate recognition level. The fast recognition level supports English, Portuguese, French, Italian, German and Spanish. Best practice is to use supportedRecognitionLanguagesForTextRecognitionLevel to check for supported languages. As the underlying engine has changed from VNRecognizeTextRequestRevision1, results can differ but are generally more accurate.

pub const REVISION_3: usize = 3usize

supports all languages from previous revisions plus some new languages in the accurate recognition level. It also improves recognition capbilities for rotation and handwriting as well as support for the automatic language detection property. Best practice is to use supportedRecognitionLanguagesForTextRecognitionLevel to check for supported languages. As the underlying engine has changed from previous revsions, results can differ but are generally more accurate.

pub fn results_ar(&self) -> Option<Rar<Array<RecognizedTextObservation>>>

pub fn results(&self) -> Option<R<Array<RecognizedTextObservation>>>

pub unsafe fn supported_recognition_langs_err_ar<'ear>( &self, err: *mut Option<&'ear Error>, ) -> Option<Rar<Array<String>>>

pub fn sel_supported_recognition_langs_err() -> &'static Sel

@selector(supportedRecognitionLanguagesAndReturnError:) but dynamic use this function to check if object responds to selector

pub unsafe fn supported_recognition_langs_err<'ear>( &self, err: *mut Option<&'ear Error>, ) -> Option<R<Array<String>>>

pub fn supported_recognition_langs<'ear>( &self, ) -> Result<'ear, R<Array<String>>>

Availability

macos_12_0, ios_15_0, tvos_15_0, visionos_1_0, maccatalyst_15_0

pub fn recognition_langs_ar(&self) -> Rar<Array<String>>

pub fn recognition_langs(&self) -> R<Array<String>>

pub fn set_recognition_langs(&mut self, val: &Array<String>)

pub fn custom_words_ar(&self) -> Rar<Array<String>>

pub fn custom_words(&self) -> R<Array<String>>

pub fn set_custom_words(&mut self, val: &Array<String>)

pub fn recognition_level(&self) -> RecognitionLevel

pub fn set_recognition_level(&mut self, val: RecognitionLevel)

pub fn uses_lang_correction(&self) -> bool

pub fn set_uses_lang_correction(&mut self, val: bool)

pub fn automatically_detects_lang(&self) -> bool

pub fn sel_automatically_detects_lang() -> &'static Sel

@selector(automaticallyDetectsLanguage) but dynamic use this function to check if object responds to selector

pub fn set_automatically_detects_lang(&mut self, val: bool)

pub fn sel_set_automatically_detects_lang() -> &'static Sel

@selector(setAutomaticallyDetectsLanguage:) but dynamic use this function to check if object responds to selector

pub fn min_text_height(&self) -> f32

pub fn set_min_text_height(&mut self, val: f32)