cidre::vt::compression::properties

Module keys

Recommendations for configuring VTCompressionSessions

These are recommendations for configuring VTCompressionSessions for some common scenarios.
These are starting points; you may find that the details of your application necessitate further adjustments.

Video-conferencing, live capture, and live broadcast scenarios: • kVTCompressionPropertyKey_RealTime: kCFBooleanTrue • kVTCompressionPropertyKey_ExpectedFrameRate: set to real-time frame rate if possible

Offline transcode initiated by a user, who is waiting for the results: • kVTCompressionPropertyKey_RealTime: kCFBooleanFalse • kVTCompressionPropertyKey_MaximizePowerEfficiency: kCFBooleanFalse

Offline transcode in the background (when the user is not aware): • kVTCompressionPropertyKey_RealTime: kCFBooleanFalse • kVTCompressionPropertyKey_MaximizePowerEfficiency: kCFBooleanTrue

Ultra-low-latency capture / conferencing / cloud gaming (cases where every millisecond counts): • kVTCompressionPropertyKey_RealTime: kCFBooleanTrue • kVTCompressionPropertyKey_PrioritizeEncodingSpeedOverQuality: kCFBooleanTrue

Functions

allow_frame_reordering

Enables frame reordering.

allow_open_gop

Enables Open GOP (Group Of Pictures) encoding.

allow_temporal_compression

Enables temporal compression.

alpha_channel_mode

aspect_ratio_16x9

avarage_bit_rate

The long-term desired average bit rate in bits per second.

base_layer_bit_rate_fraction

base_layer_frame_rate

base_layer_frame_rate_fraction

clean_aperture

color_primaries

constant_bit_rate

Requires that the encoder use a Constant Bit Rate algorithm.

content_light_lvl_info

data_rate_limits

Zero, one or two hard limits on data rate.

depth

enable_ltr

Enable Long Term Reference (LTR) frames during encoding

encoder_id

estimated_average_bytes_per_frame

An estimate of the expected size in bytes of a single encoded frame based on the current configuration.

expected_duration

expected_frame_rate

field_count

field_detail

gamma_lvl

h264_entropy_mode

hdr_metadata_insertion_mode

icc_profile

master_display_color_volume

max_allowed_frame_qp

Specifies the minimum allowed encoded frame QP (Quantization Parameter).

max_frame_delay_count

The maximum frame delay count is the maximum number of frames that a compressor is allowed to hold before it must output a compressed frame.

max_h264_slice_bytes

max_key_frame_interval

The maximum interval between key frames, also known as the key frame rate.

max_key_frame_interval_duration

The maximum duration from one key frame to the next in seconds.

max_real_time_frame_rate

Indicates the maximum real time rate at which frames could be submitted to the VTCompressionSession CFNumber/Optional

maximize_power_efficiecy

Hints to the video encoder that it should maximize power efficiency during encode.

more_frames_after_end

more_frames_before_start

multi_pass_storage

number_of_pending_frames

The number of pending frames in the compression session.

output_bit_depth

pixel_aspect_ratio

Describes the pixel aspect ratio for encoded frames.

pixel_buf_pool_is_shared

Indicates whether the a common pixel buffer pool is shared between the video encoder and the session client.

pixel_transfer_properties

preserve_dynamic_hdr_metadata

prioritize_encoding_speed_over_quality

Hint for the video encoder that it should maximize its speed during encode, sacrificing quality if needed

profile_lvl

progressive_scan

quality

The desired compression quality.

real_time

Hints the video encoder that compression is, or is not, being performed in real time.

source_frame_count

spatial_adaptive_qp_level

Read/write, cf::Number, Optional Control spatial adaptation of the quantization parameter (QP) based on per-frame statistics.

suggested_look_ahead_frame_count

Read/write, cf::Number, Optional

support_base_frame_qp

supported_preset_dictionaries⚠

Availability

target_quality_for_alpha

transfer_fn

variable_bit_rate⚠

Requires that the encoder use a variable bitrate (VBR) rate control algorithm and specifies the desired variable bitrate in bits per second.

vbv_buf_duration⚠

Capacity of the video buffering verifier (VBV) model in seconds.

vbv_initial_delay_percentage⚠

Initial delay of the VBV model between storing the picture in the VBV buffer model and decoding of that picture, as a percentage of VBV buffer duration.

vbv_max_bit_rate⚠

Defines the maximum bitrate that can enter the video buffering verifier (VBV) model at any time in variable bitrate (VBR) mode.

video_encoder_pixel_buf_attrs

The video encoder’s pixel buffer attributes for the compression session.

ycbcr_matrix