cidre::vt::compression::properties::keys

Function constant_bit_rate

pub fn constant_bit_rate() -> &'static String

Requires that the encoder use a Constant Bit Rate algorithm.

The property kVTCompressionPropertyKey_ExpectedFrameRate should be set along with kVTCompressionPropertyKey_ConstantBitRate to ensure effective CBR rate control.

This property is not compatible with kVTCompressionPropertyKey_DataRateLimits and kVTCompressionPropertyKey_AverageBitRate.

The encoder will pad the frame if they are smaller than they need to be based on the Constant BitRate. This property is not recommended for general streaming or export scenarios. It is intended for interoperability with stremaing CDNs which specifically require that data rates not drop even during low motion and activity scenes. This is not supported in all encoders or in all encoder operating modes. kVTPropertyNotSupportedErr will be returned when this option is not supported.