cidre::vt::compression::properties::keys

Function real_time

pub fn real_time() -> &'static String

Hints the video encoder that compression is, or is not, being performed in real time.

For offline compression, clients are advised to set this property to kCFBooleanFalse. This indicates that the timeline of the frames is not connected to wall-clock time at all. This allows the video encoder to run either slower than real-time or faster than real-time, depending upon the situation and device capabilities. For real-time compression, clients are advised to set this property to kCFBooleanTrue. This indicates that the timeline of frames is connected to real time (so, for example, the client expects to encode one second of frames each second, on average). In such a case, it is important that encoding keep up, but it is also not valuable to encode substantially faster than real time.
When this property is set to kCFBooleanTrue, the video encoder may use the value of the kVTCompressionPropertyKey_ExpectedFrameRate property to optimize energy usage, by anticipating that encoding requests will happen at this rate. By default, this property is NULL, indicating unknown.