cidre::vt::compression::properties::keys

Function spatial_adaptive_qp_level

pub fn spatial_adaptive_qp_level() -> &'static String

Read/write, cf::Number, Optional Control spatial adaptation of the quantization parameter (QP) based on per-frame statistics.

If set to qp_modulation_level::disable(), spatial QP adaptation is not applied based on per-frame statistics. If set to qp_modulation_level::default(), video encoder is allowed to apply spatial QP adaptation for each macro block (or coding unit) within a video frame. QP adaptation is based on spatial characteristics of a frame and the level of spatial QP adaptation is decided internally by the rate controller.

Availability

macos_15_0