cidre::vt::compression::session

Struct Session

pub struct Session(/* private fields */);

Implementations

impl Session

pub fn retained(&self) -> R<Self>

impl Session

pub fn new<T>( width: u32, height: u32, codec: VideoCodec, encoder_spec: Option<&Dictionary>, src_image_buf_attrs: Option<&Dictionary>, compressed_data_allocator: Option<&Allocator>, output_cb: Option<OutputCallback<T>>, output_cb_ref_con: *mut T, ) -> Result<R<Self>>

pub unsafe fn create_in( width: i32, height: i32, codec_type: VideoCodec, encoder_spec: Option<&Dictionary>, src_image_buf_attrs: Option<&Dictionary>, compressed_data_allocator: Option<&Allocator>, output_cb: Option<OutputCallback<c_void>>, output_cb_ref_con: *mut c_void, compression_session_out: *mut Option<R<Session>>, allocator: Option<&Allocator>, ) -> Result

Safety

use ::new

pub fn invalidate(&mut self)

pub fn prepare(&mut self) -> Result

pub fn encode_frame( &self, image_buf: &ImageBuf, pts: Time, duration: Time, frame_props: Option<&DictionaryOf<String, Type>>, src_frame_ref_con: *mut c_void, info_flags_out: *mut Option<NonNull<EncodeInfoFlags>>, ) -> Result

Encoded frames may or may not be output before the function returns. The client should not modify the pixel data after making this call. The session and/or encoder retains the image buffer as long as necessary.

If you don’t know duration. Set it invalid. TODO: investigate. Right now we keep buf to calculate ts delta

pub fn enc_frame( &self, image_buf: &ImageBuf, pts: Time, duration: Time, info_flags_out: *mut Option<NonNull<EncodeInfoFlags>>, ) -> Result

pub fn encode_frame_with_output_handler( &mut self, image_buf: &ImageBuf, pts: Time, duration: Time, frame_props: Option<&Dictionary>, info_flags_out: *mut Option<NonNull<EncodeInfoFlags>>, block: &'static mut OutputHandler, ) -> Result

pub fn enc_frame_with_output_handler_block( &mut self, image_buf: &ImageBuf, pts: Time, duration: Time, frame_props: Option<&Dictionary>, info_flags_out: *mut Option<NonNull<EncodeInfoFlags>>, block: &mut OutputHandler, ) -> Result

pub fn enc_frame_with_output_handler( &mut self, image_buf: &ImageBuf, pts: Time, duration: Time, frame_props: Option<&Dictionary>, info_flags_out: *mut Option<NonNull<EncodeInfoFlags>>, block: impl FnMut(Status, EncodeInfoFlags, Option<&SampleBuf>) + Send + 'static, ) -> Result

pub fn pixel_buf_pool(&self) -> Option<&PixelBufPool>

pub fn complete_frames(&self, until_pts: Time) -> Result

pub fn complete_all(&self) -> Result

pub fn is_stereo_mv_hevc_encode_supported() -> bool

Indicates whether the current system supports stereo MV-HEVC encode.

This call returning true does not guarantee that encode resources will be available at all times.

Availability

macos_14_0, ios_17_0, tvos_17_0, visionos_1_0