cidre::vt::decompression::properties

Module keys

Functions

content_has_interframe_dependencies

decoder_produces_raw_output

cf::Boolean, Read, Optional, kCFBooleanFalse by default

deinterlace_mode

field_mode

generate_per_frame_hdr_display_metadata

Generates per frame HDR Metadata and attaches it to the resulting decoded cv::PixelBufs.

max_output_pts_of_frames_being_decoded

maximize_power_efficiency

min_output_pts_of_frames_being_decoded

number_of_frames_being_decoded

only_these_frames

output_pool_requested_minimum_buffer_count

pixel_buffer_is_shared

pixel_buffer_pool

pixel_formats_with_reduced_resolution_support

pixel_transfer_props

propogate_per_frame_hdr_display_metadata

Availability

real_time

Hints the video decoder that decompression is, or is not, being performed in real time.

reduced_coefficent_decode

reduced_frame_delivery

reduced_resolution_decode

request_raw_output

cf::Boolean, Read/Write, Optional, kCFBooleanFalse by default

requested_mv_hevc_video_layer_ids

Requests multi-image decoding of specific MV-HEVC VideoLayerIDs.

supported_pixel_formats_ordered_by_performance

thread_count

using_gpu_registry_id

cf::Number indicating the gpu registryID of the decoder that was selected. NULL indicates a built-in decoder or software decode was used. read-only

using_hardware_accelerated_video_decoder

cf::Boolean, Read; assumed false by default