cidre::vt::decompression::properties

Module video_decoder_specification

Functions

enable_hardware_accelerated_video_decoder

If set to cf::Boolean::value_true(), the VideoToolbox will use a hardware accelerated video decoder if available. If set to cf::Boolean::value_false(), hardware decode will never be used.

preferred_decoder_gpu_registry_id

cf::Number Optional

require_hardware_accelerated_video_decoder

if set to cf::Boolean::value_true(), the VideoToolbox will try to allocate a hardware accelerated decoder and return an error if that isn’t possible. Setting this key automatically implies kVTVideoDecoderSpecification_EnableHardwareAcceleratedVideoDecoder – there is no need to set both and the Enable key does nothing if the Require key is set.

required_decoder_gpu_registry_id

cf::Number Optional